Multiple-sources Cox proportional hazards model with group sparsity and Local-Global transfer

Multiple-sources Cox proportional hazards model with group sparsity and Local-Global transfer

Usage

coxens(
  formula,
  data,
  group,
  lambda1 = 0,
  lambda2 = 0,
  lambda3 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 2,
  init,
  control,
  ...
)

Arguments

formula

a formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

a data frame containing the variables in the model.

group

a factor specifying the group of each sample.

lambda1

a non-negative value specifying the sparse penalty parameter. The default is 0.

lambda2

a non-negative value specifying the global biased penalty parameter. The default is 0.

lambda3

a non-negative value specifying the local biased penalty parameter. The default is 0.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3.0 for MCP.

rho

a value in (2, 10) specifying the expansion factor of the augmented Lagrangian's penalty parameter. The default is 2.0.

init

a numeric vector specifying the initial value of the coefficients. The default is a zero vector.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a coxsg object.

Examples

library(survtrans)
formula <- Surv(time, status) ~ . - group - id
group <- as.factor(sim2$group)
fit <- coxens(
  formula, sim2, group,
  lambda1 = 2e-2, lambda3 = 3e-3, penalty = "SCAD"
)
summary(fit)
#> Call:
#> coxens(formula = formula, data = sim2, group = group, lambda1 = 0.02, 
#>     lambda3 = 0.003, penalty = "SCAD")
#> 
#>   n=1000, number of events=811
#> 
#>            coef exp(coef) se(coef)       z Pr(>|z|)    
#> 1 X1    0.54067   1.71716  0.04802  11.260  < 2e-16 ***
#> 1 X2    0.55958   1.74994  0.05097  10.979  < 2e-16 ***
#> 1 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 1 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 1 X5    0.00000   1.00000       NA      NA       NA    
#> 1 X6    0.00000   1.00000       NA      NA       NA    
#> 1 X7    0.00000   1.00000       NA      NA       NA    
#> 1 X8    0.00000   1.00000       NA      NA       NA    
#> 1 X9    0.00000   1.00000       NA      NA       NA    
#> 1 X10   0.00000   1.00000       NA      NA       NA    
#> 1 X11   0.00000   1.00000       NA      NA       NA    
#> 1 X12   0.00000   1.00000       NA      NA       NA    
#> 1 X13   0.00000   1.00000       NA      NA       NA    
#> 1 X14   0.00000   1.00000       NA      NA       NA    
#> 1 X15   0.00000   1.00000       NA      NA       NA    
#> 1 X16   0.00000   1.00000       NA      NA       NA    
#> 1 X17   0.00000   1.00000       NA      NA       NA    
#> 1 X18   0.00000   1.00000       NA      NA       NA    
#> 1 X19   0.00000   1.00000       NA      NA       NA    
#> 1 X20   0.00000   1.00000       NA      NA       NA    
#> 2 X1   -0.64495   0.52469  0.05096 -12.656  < 2e-16 ***
#> 2 X2   -0.62512   0.53520  0.05283 -11.833  < 2e-16 ***
#> 2 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 2 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 2 X5    0.00000   1.00000       NA      NA       NA    
#> 2 X6    0.00000   1.00000       NA      NA       NA    
#> 2 X7    0.00000   1.00000       NA      NA       NA    
#> 2 X8    0.00000   1.00000       NA      NA       NA    
#> 2 X9    0.00000   1.00000       NA      NA       NA    
#> 2 X10   0.00000   1.00000       NA      NA       NA    
#> 2 X11   0.00000   1.00000       NA      NA       NA    
#> 2 X12   0.00000   1.00000       NA      NA       NA    
#> 2 X13   0.00000   1.00000       NA      NA       NA    
#> 2 X14   0.00000   1.00000       NA      NA       NA    
#> 2 X15   0.00000   1.00000       NA      NA       NA    
#> 2 X16   0.00000   1.00000       NA      NA       NA    
#> 2 X17   0.00000   1.00000       NA      NA       NA    
#> 2 X18   0.00000   1.00000       NA      NA       NA    
#> 2 X19   0.00000   1.00000       NA      NA       NA    
#> 2 X20   0.00000   1.00000       NA      NA       NA    
#> 3 X1    0.54067   1.71716  0.04802  11.260  < 2e-16 ***
#> 3 X2    0.55958   1.74994  0.05097  10.979  < 2e-16 ***
#> 3 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 3 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 3 X5    0.00000   1.00000       NA      NA       NA    
#> 3 X6    0.00000   1.00000       NA      NA       NA    
#> 3 X7    0.00000   1.00000       NA      NA       NA    
#> 3 X8    0.00000   1.00000       NA      NA       NA    
#> 3 X9    0.00000   1.00000       NA      NA       NA    
#> 3 X10   0.00000   1.00000       NA      NA       NA    
#> 3 X11   0.00000   1.00000       NA      NA       NA    
#> 3 X12   0.00000   1.00000       NA      NA       NA    
#> 3 X13   0.00000   1.00000       NA      NA       NA    
#> 3 X14   0.00000   1.00000       NA      NA       NA    
#> 3 X15   0.00000   1.00000       NA      NA       NA    
#> 3 X16   0.00000   1.00000       NA      NA       NA    
#> 3 X17   0.00000   1.00000       NA      NA       NA    
#> 3 X18   0.00000   1.00000       NA      NA       NA    
#> 3 X19   0.00000   1.00000       NA      NA       NA    
#> 3 X20   0.00000   1.00000       NA      NA       NA    
#> 4 X1   -0.64495   0.52469  0.05096 -12.656  < 2e-16 ***
#> 4 X2   -0.62512   0.53520  0.05283 -11.833  < 2e-16 ***
#> 4 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 4 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 4 X5    0.00000   1.00000       NA      NA       NA    
#> 4 X6    0.00000   1.00000       NA      NA       NA    
#> 4 X7    0.00000   1.00000       NA      NA       NA    
#> 4 X8    0.00000   1.00000       NA      NA       NA    
#> 4 X9    0.00000   1.00000       NA      NA       NA    
#> 4 X10   0.00000   1.00000       NA      NA       NA    
#> 4 X11   0.00000   1.00000       NA      NA       NA    
#> 4 X12   0.00000   1.00000       NA      NA       NA    
#> 4 X13   0.00000   1.00000       NA      NA       NA    
#> 4 X14   0.00000   1.00000       NA      NA       NA    
#> 4 X15   0.00000   1.00000       NA      NA       NA    
#> 4 X16   0.00000   1.00000       NA      NA       NA    
#> 4 X17   0.00000   1.00000       NA      NA       NA    
#> 4 X18   0.00000   1.00000       NA      NA       NA    
#> 4 X19   0.00000   1.00000       NA      NA       NA    
#> 4 X20   0.00000   1.00000       NA      NA       NA    
#> 5 X1    0.54067   1.71716  0.04802  11.260  < 2e-16 ***
#> 5 X2    0.55958   1.74994  0.05097  10.979  < 2e-16 ***
#> 5 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 5 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 5 X5    0.00000   1.00000       NA      NA       NA    
#> 5 X6    0.00000   1.00000       NA      NA       NA    
#> 5 X7    0.00000   1.00000       NA      NA       NA    
#> 5 X8    0.00000   1.00000       NA      NA       NA    
#> 5 X9    0.00000   1.00000       NA      NA       NA    
#> 5 X10   0.00000   1.00000       NA      NA       NA    
#> 5 X11   0.00000   1.00000       NA      NA       NA    
#> 5 X12   0.00000   1.00000       NA      NA       NA    
#> 5 X13   0.00000   1.00000       NA      NA       NA    
#> 5 X14   0.00000   1.00000       NA      NA       NA    
#> 5 X15   0.00000   1.00000       NA      NA       NA    
#> 5 X16   0.00000   1.00000       NA      NA       NA    
#> 5 X17   0.00000   1.00000       NA      NA       NA    
#> 5 X18   0.00000   1.00000       NA      NA       NA    
#> 5 X19   0.00000   1.00000       NA      NA       NA    
#> 5 X20   0.00000   1.00000       NA      NA       NA    
#> 6 X1   -0.64495   0.52469  0.05096 -12.656  < 2e-16 ***
#> 6 X2   -0.62512   0.53520  0.05283 -11.833  < 2e-16 ***
#> 6 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 6 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 6 X5    0.00000   1.00000       NA      NA       NA    
#> 6 X6    0.00000   1.00000       NA      NA       NA    
#> 6 X7    0.00000   1.00000       NA      NA       NA    
#> 6 X8    0.00000   1.00000       NA      NA       NA    
#> 6 X9    0.00000   1.00000       NA      NA       NA    
#> 6 X10   0.00000   1.00000       NA      NA       NA    
#> 6 X11   0.00000   1.00000       NA      NA       NA    
#> 6 X12   0.00000   1.00000       NA      NA       NA    
#> 6 X13   0.00000   1.00000       NA      NA       NA    
#> 6 X14   0.00000   1.00000       NA      NA       NA    
#> 6 X15   0.00000   1.00000       NA      NA       NA    
#> 6 X16   0.00000   1.00000       NA      NA       NA    
#> 6 X17   0.00000   1.00000       NA      NA       NA    
#> 6 X18   0.00000   1.00000       NA      NA       NA    
#> 6 X19   0.00000   1.00000       NA      NA       NA    
#> 6 X20   0.00000   1.00000       NA      NA       NA    
#> 7 X1    0.54067   1.71716  0.04802  11.260  < 2e-16 ***
#> 7 X2    0.31086   1.36460  0.13109   2.371   0.0177 *  
#> 7 X3    0.75915   2.13646  0.12915   5.878 4.15e-09 ***
#> 7 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 7 X5    0.00000   1.00000       NA      NA       NA    
#> 7 X6    0.00000   1.00000       NA      NA       NA    
#> 7 X7    0.00000   1.00000       NA      NA       NA    
#> 7 X8    0.00000   1.00000       NA      NA       NA    
#> 7 X9    0.00000   1.00000       NA      NA       NA    
#> 7 X10   0.00000   1.00000       NA      NA       NA    
#> 7 X11   0.00000   1.00000       NA      NA       NA    
#> 7 X12   0.00000   1.00000       NA      NA       NA    
#> 7 X13   0.00000   1.00000       NA      NA       NA    
#> 7 X14   0.00000   1.00000       NA      NA       NA    
#> 7 X15   0.00000   1.00000       NA      NA       NA    
#> 7 X16   0.00000   1.00000       NA      NA       NA    
#> 7 X17   0.00000   1.00000       NA      NA       NA    
#> 7 X18   0.00000   1.00000       NA      NA       NA    
#> 7 X19   0.00000   1.00000       NA      NA       NA    
#> 7 X20   0.00000   1.00000       NA      NA       NA    
#> 8 X1   -0.64495   0.52469  0.05096 -12.656  < 2e-16 ***
#> 8 X2   -0.62512   0.53520  0.05283 -11.833  < 2e-16 ***
#> 8 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 8 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 8 X5    0.00000   1.00000       NA      NA       NA    
#> 8 X6    0.00000   1.00000       NA      NA       NA    
#> 8 X7    0.00000   1.00000       NA      NA       NA    
#> 8 X8    0.00000   1.00000       NA      NA       NA    
#> 8 X9    0.00000   1.00000       NA      NA       NA    
#> 8 X10   0.00000   1.00000       NA      NA       NA    
#> 8 X11   0.00000   1.00000       NA      NA       NA    
#> 8 X12   0.00000   1.00000       NA      NA       NA    
#> 8 X13   0.00000   1.00000       NA      NA       NA    
#> 8 X14   0.00000   1.00000       NA      NA       NA    
#> 8 X15   0.00000   1.00000       NA      NA       NA    
#> 8 X16   0.00000   1.00000       NA      NA       NA    
#> 8 X17   0.00000   1.00000       NA      NA       NA    
#> 8 X18   0.00000   1.00000       NA      NA       NA    
#> 8 X19   0.00000   1.00000       NA      NA       NA    
#> 8 X20   0.00000   1.00000       NA      NA       NA    
#> 9 X1    0.54067   1.71716  0.04802  11.260  < 2e-16 ***
#> 9 X2    0.55958   1.74994  0.05097  10.979  < 2e-16 ***
#> 9 X3    0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 9 X4    0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 9 X5    0.00000   1.00000       NA      NA       NA    
#> 9 X6    0.00000   1.00000       NA      NA       NA    
#> 9 X7    0.00000   1.00000       NA      NA       NA    
#> 9 X8    0.00000   1.00000       NA      NA       NA    
#> 9 X9    0.00000   1.00000       NA      NA       NA    
#> 9 X10   0.00000   1.00000       NA      NA       NA    
#> 9 X11   0.00000   1.00000       NA      NA       NA    
#> 9 X12   0.00000   1.00000       NA      NA       NA    
#> 9 X13   0.00000   1.00000       NA      NA       NA    
#> 9 X14   0.00000   1.00000       NA      NA       NA    
#> 9 X15   0.00000   1.00000       NA      NA       NA    
#> 9 X16   0.00000   1.00000       NA      NA       NA    
#> 9 X17   0.00000   1.00000       NA      NA       NA    
#> 9 X18   0.00000   1.00000       NA      NA       NA    
#> 9 X19   0.00000   1.00000       NA      NA       NA    
#> 9 X20   0.00000   1.00000       NA      NA       NA    
#> 10 X1  -0.64495   0.52469  0.05096 -12.656  < 2e-16 ***
#> 10 X2  -0.62512   0.53520  0.05283 -11.833  < 2e-16 ***
#> 10 X3   0.52807   1.69565  0.03905  13.523  < 2e-16 ***
#> 10 X4   0.54321   1.72152  0.03951  13.750  < 2e-16 ***
#> 10 X5   0.00000   1.00000       NA      NA       NA    
#> 10 X6   0.00000   1.00000       NA      NA       NA    
#> 10 X7   0.00000   1.00000       NA      NA       NA    
#> 10 X8   0.00000   1.00000       NA      NA       NA    
#> 10 X9   0.00000   1.00000       NA      NA       NA    
#> 10 X10  0.00000   1.00000       NA      NA       NA    
#> 10 X11  0.00000   1.00000       NA      NA       NA    
#> 10 X12  0.00000   1.00000       NA      NA       NA    
#> 10 X13  0.00000   1.00000       NA      NA       NA    
#> 10 X14  0.00000   1.00000       NA      NA       NA    
#> 10 X15  0.00000   1.00000       NA      NA       NA    
#> 10 X16  0.00000   1.00000       NA      NA       NA    
#> 10 X17  0.00000   1.00000       NA      NA       NA    
#> 10 X18  0.00000   1.00000       NA      NA       NA    
#> 10 X19  0.00000   1.00000       NA      NA       NA    
#> 10 X20  0.00000   1.00000       NA      NA       NA    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>        exp(coef) exp(-coef) lower .95 upper .95
#> 1 X1   1.7172    0.5824     1.5629    1.8866   
#> 1 X2   1.7499    0.5714     1.5836    1.9338   
#> 1 X3   1.6957    0.5897     1.5707    1.8305   
#> 1 X4   1.7215    0.5809     1.5933    1.8601   
#> 1 X5   1.0000    1.0000         NA        NA   
#> 1 X6   1.0000    1.0000         NA        NA   
#> 1 X7   1.0000    1.0000         NA        NA   
#> 1 X8   1.0000    1.0000         NA        NA   
#> 1 X9   1.0000    1.0000         NA        NA   
#> 1 X10  1.0000    1.0000         NA        NA   
#> 1 X11  1.0000    1.0000         NA        NA   
#> 1 X12  1.0000    1.0000         NA        NA   
#> 1 X13  1.0000    1.0000         NA        NA   
#> 1 X14  1.0000    1.0000         NA        NA   
#> 1 X15  1.0000    1.0000         NA        NA   
#> 1 X16  1.0000    1.0000         NA        NA   
#> 1 X17  1.0000    1.0000         NA        NA   
#> 1 X18  1.0000    1.0000         NA        NA   
#> 1 X19  1.0000    1.0000         NA        NA   
#> 1 X20  1.0000    1.0000         NA        NA   
#> 2 X1   0.5247    1.9059     0.4748    0.5798   
#> 2 X2   0.5352    1.8685     0.4826    0.5936   
#> 2 X3   1.6957    0.5897     1.5707    1.8305   
#> 2 X4   1.7215    0.5809     1.5933    1.8601   
#> 2 X5   1.0000    1.0000         NA        NA   
#> 2 X6   1.0000    1.0000         NA        NA   
#> 2 X7   1.0000    1.0000         NA        NA   
#> 2 X8   1.0000    1.0000         NA        NA   
#> 2 X9   1.0000    1.0000         NA        NA   
#> 2 X10  1.0000    1.0000         NA        NA   
#> 2 X11  1.0000    1.0000         NA        NA   
#> 2 X12  1.0000    1.0000         NA        NA   
#> 2 X13  1.0000    1.0000         NA        NA   
#> 2 X14  1.0000    1.0000         NA        NA   
#> 2 X15  1.0000    1.0000         NA        NA   
#> 2 X16  1.0000    1.0000         NA        NA   
#> 2 X17  1.0000    1.0000         NA        NA   
#> 2 X18  1.0000    1.0000         NA        NA   
#> 2 X19  1.0000    1.0000         NA        NA   
#> 2 X20  1.0000    1.0000         NA        NA   
#> 3 X1   1.7172    0.5824     1.5629    1.8866   
#> 3 X2   1.7499    0.5714     1.5836    1.9338   
#> 3 X3   1.6957    0.5897     1.5707    1.8305   
#> 3 X4   1.7215    0.5809     1.5933    1.8601   
#> 3 X5   1.0000    1.0000         NA        NA   
#> 3 X6   1.0000    1.0000         NA        NA   
#> 3 X7   1.0000    1.0000         NA        NA   
#> 3 X8   1.0000    1.0000         NA        NA   
#> 3 X9   1.0000    1.0000         NA        NA   
#> 3 X10  1.0000    1.0000         NA        NA   
#> 3 X11  1.0000    1.0000         NA        NA   
#> 3 X12  1.0000    1.0000         NA        NA   
#> 3 X13  1.0000    1.0000         NA        NA   
#> 3 X14  1.0000    1.0000         NA        NA   
#> 3 X15  1.0000    1.0000         NA        NA   
#> 3 X16  1.0000    1.0000         NA        NA   
#> 3 X17  1.0000    1.0000         NA        NA   
#> 3 X18  1.0000    1.0000         NA        NA   
#> 3 X19  1.0000    1.0000         NA        NA   
#> 3 X20  1.0000    1.0000         NA        NA   
#> 4 X1   0.5247    1.9059     0.4748    0.5798   
#> 4 X2   0.5352    1.8685     0.4826    0.5936   
#> 4 X3   1.6957    0.5897     1.5707    1.8305   
#> 4 X4   1.7215    0.5809     1.5933    1.8601   
#> 4 X5   1.0000    1.0000         NA        NA   
#> 4 X6   1.0000    1.0000         NA        NA   
#> 4 X7   1.0000    1.0000         NA        NA   
#> 4 X8   1.0000    1.0000         NA        NA   
#> 4 X9   1.0000    1.0000         NA        NA   
#> 4 X10  1.0000    1.0000         NA        NA   
#> 4 X11  1.0000    1.0000         NA        NA   
#> 4 X12  1.0000    1.0000         NA        NA   
#> 4 X13  1.0000    1.0000         NA        NA   
#> 4 X14  1.0000    1.0000         NA        NA   
#> 4 X15  1.0000    1.0000         NA        NA   
#> 4 X16  1.0000    1.0000         NA        NA   
#> 4 X17  1.0000    1.0000         NA        NA   
#> 4 X18  1.0000    1.0000         NA        NA   
#> 4 X19  1.0000    1.0000         NA        NA   
#> 4 X20  1.0000    1.0000         NA        NA   
#> 5 X1   1.7172    0.5824     1.5629    1.8866   
#> 5 X2   1.7499    0.5714     1.5836    1.9338   
#> 5 X3   1.6957    0.5897     1.5707    1.8305   
#> 5 X4   1.7215    0.5809     1.5933    1.8601   
#> 5 X5   1.0000    1.0000         NA        NA   
#> 5 X6   1.0000    1.0000         NA        NA   
#> 5 X7   1.0000    1.0000         NA        NA   
#> 5 X8   1.0000    1.0000         NA        NA   
#> 5 X9   1.0000    1.0000         NA        NA   
#> 5 X10  1.0000    1.0000         NA        NA   
#> 5 X11  1.0000    1.0000         NA        NA   
#> 5 X12  1.0000    1.0000         NA        NA   
#> 5 X13  1.0000    1.0000         NA        NA   
#> 5 X14  1.0000    1.0000         NA        NA   
#> 5 X15  1.0000    1.0000         NA        NA   
#> 5 X16  1.0000    1.0000         NA        NA   
#> 5 X17  1.0000    1.0000         NA        NA   
#> 5 X18  1.0000    1.0000         NA        NA   
#> 5 X19  1.0000    1.0000         NA        NA   
#> 5 X20  1.0000    1.0000         NA        NA   
#> 6 X1   0.5247    1.9059     0.4748    0.5798   
#> 6 X2   0.5352    1.8685     0.4826    0.5936   
#> 6 X3   1.6957    0.5897     1.5707    1.8305   
#> 6 X4   1.7215    0.5809     1.5933    1.8601   
#> 6 X5   1.0000    1.0000         NA        NA   
#> 6 X6   1.0000    1.0000         NA        NA   
#> 6 X7   1.0000    1.0000         NA        NA   
#> 6 X8   1.0000    1.0000         NA        NA   
#> 6 X9   1.0000    1.0000         NA        NA   
#> 6 X10  1.0000    1.0000         NA        NA   
#> 6 X11  1.0000    1.0000         NA        NA   
#> 6 X12  1.0000    1.0000         NA        NA   
#> 6 X13  1.0000    1.0000         NA        NA   
#> 6 X14  1.0000    1.0000         NA        NA   
#> 6 X15  1.0000    1.0000         NA        NA   
#> 6 X16  1.0000    1.0000         NA        NA   
#> 6 X17  1.0000    1.0000         NA        NA   
#> 6 X18  1.0000    1.0000         NA        NA   
#> 6 X19  1.0000    1.0000         NA        NA   
#> 6 X20  1.0000    1.0000         NA        NA   
#> 7 X1   1.7172    0.5824     1.5629    1.8866   
#> 7 X2   1.3646    0.7328     1.0554    1.7644   
#> 7 X3   2.1365    0.4681     1.6587    2.7518   
#> 7 X4   1.7215    0.5809     1.5933    1.8601   
#> 7 X5   1.0000    1.0000         NA        NA   
#> 7 X6   1.0000    1.0000         NA        NA   
#> 7 X7   1.0000    1.0000         NA        NA   
#> 7 X8   1.0000    1.0000         NA        NA   
#> 7 X9   1.0000    1.0000         NA        NA   
#> 7 X10  1.0000    1.0000         NA        NA   
#> 7 X11  1.0000    1.0000         NA        NA   
#> 7 X12  1.0000    1.0000         NA        NA   
#> 7 X13  1.0000    1.0000         NA        NA   
#> 7 X14  1.0000    1.0000         NA        NA   
#> 7 X15  1.0000    1.0000         NA        NA   
#> 7 X16  1.0000    1.0000         NA        NA   
#> 7 X17  1.0000    1.0000         NA        NA   
#> 7 X18  1.0000    1.0000         NA        NA   
#> 7 X19  1.0000    1.0000         NA        NA   
#> 7 X20  1.0000    1.0000         NA        NA   
#> 8 X1   0.5247    1.9059     0.4748    0.5798   
#> 8 X2   0.5352    1.8685     0.4826    0.5936   
#> 8 X3   1.6957    0.5897     1.5707    1.8305   
#> 8 X4   1.7215    0.5809     1.5933    1.8601   
#> 8 X5   1.0000    1.0000         NA        NA   
#> 8 X6   1.0000    1.0000         NA        NA   
#> 8 X7   1.0000    1.0000         NA        NA   
#> 8 X8   1.0000    1.0000         NA        NA   
#> 8 X9   1.0000    1.0000         NA        NA   
#> 8 X10  1.0000    1.0000         NA        NA   
#> 8 X11  1.0000    1.0000         NA        NA   
#> 8 X12  1.0000    1.0000         NA        NA   
#> 8 X13  1.0000    1.0000         NA        NA   
#> 8 X14  1.0000    1.0000         NA        NA   
#> 8 X15  1.0000    1.0000         NA        NA   
#> 8 X16  1.0000    1.0000         NA        NA   
#> 8 X17  1.0000    1.0000         NA        NA   
#> 8 X18  1.0000    1.0000         NA        NA   
#> 8 X19  1.0000    1.0000         NA        NA   
#> 8 X20  1.0000    1.0000         NA        NA   
#> 9 X1   1.7172    0.5824     1.5629    1.8866   
#> 9 X2   1.7499    0.5714     1.5836    1.9338   
#> 9 X3   1.6957    0.5897     1.5707    1.8305   
#> 9 X4   1.7215    0.5809     1.5933    1.8601   
#> 9 X5   1.0000    1.0000         NA        NA   
#> 9 X6   1.0000    1.0000         NA        NA   
#> 9 X7   1.0000    1.0000         NA        NA   
#> 9 X8   1.0000    1.0000         NA        NA   
#> 9 X9   1.0000    1.0000         NA        NA   
#> 9 X10  1.0000    1.0000         NA        NA   
#> 9 X11  1.0000    1.0000         NA        NA   
#> 9 X12  1.0000    1.0000         NA        NA   
#> 9 X13  1.0000    1.0000         NA        NA   
#> 9 X14  1.0000    1.0000         NA        NA   
#> 9 X15  1.0000    1.0000         NA        NA   
#> 9 X16  1.0000    1.0000         NA        NA   
#> 9 X17  1.0000    1.0000         NA        NA   
#> 9 X18  1.0000    1.0000         NA        NA   
#> 9 X19  1.0000    1.0000         NA        NA   
#> 9 X20  1.0000    1.0000         NA        NA   
#> 10 X1  0.5247    1.9059     0.4748    0.5798   
#> 10 X2  0.5352    1.8685     0.4826    0.5936   
#> 10 X3  1.6957    0.5897     1.5707    1.8305   
#> 10 X4  1.7215    0.5809     1.5933    1.8601   
#> 10 X5  1.0000    1.0000         NA        NA   
#> 10 X6  1.0000    1.0000         NA        NA   
#> 10 X7  1.0000    1.0000         NA        NA   
#> 10 X8  1.0000    1.0000         NA        NA   
#> 10 X9  1.0000    1.0000         NA        NA   
#> 10 X10 1.0000    1.0000         NA        NA   
#> 10 X11 1.0000    1.0000         NA        NA   
#> 10 X12 1.0000    1.0000         NA        NA   
#> 10 X13 1.0000    1.0000         NA        NA   
#> 10 X14 1.0000    1.0000         NA        NA   
#> 10 X15 1.0000    1.0000         NA        NA   
#> 10 X16 1.0000    1.0000         NA        NA   
#> 10 X17 1.0000    1.0000         NA        NA   
#> 10 X18 1.0000    1.0000         NA        NA   
#> 10 X19 1.0000    1.0000         NA        NA   
#> 10 X20 1.0000    1.0000         NA        NA