Multiple-sources Cox proportional hazards model with group sparsity

Multiple-sources Cox proportional hazards model with group sparsity

Usage

coxgrp(
  formula,
  data,
  group,
  lambda = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 2,
  init,
  control,
  ...
)

Arguments

formula

a formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

a data frame containing the variables in the model.

group

a factor specifying the group of each sample.

lambda

a non-negative value specifying the sparse penalty parameter. The default is 0.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3.0 for MCP.

rho

a value in (2, 10) specifying the expansion factor of the augmented Lagrangian's penalty parameter. The default is 2.0.

init

a numeric vector specifying the initial value of the coefficients. The default is a zero vector.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a coxsg object.

Examples

library(survtrans)
formula <- Surv(time, status) ~ . - group - id
group <- as.factor(sim2$group)
fit <- coxgrp(formula, sim2, group, lambda = 0.02, penalty = "SCAD")
fit$coefficients
#>             1          2         3          4         5          6         7
#> X1  0.5394258 -0.7628322 0.7133059 -0.5241018 0.4686284 -0.6207682 0.5623104
#> X2  0.5845159 -0.5714762 0.7693320 -0.5994700 0.4725933 -0.7439499 0.3209021
#> X3  0.4562600  0.6392264 0.8669619  0.4704888 0.5578964  0.4915979 0.7639459
#> X4  0.3481866  0.7535935 0.6419702  0.4545768 0.5752310  0.5641411 0.5731811
#> X5  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X6  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X7  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X8  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X9  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X10 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X11 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X12 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X13 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X14 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X15 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X16 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X17 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X18 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X19 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X20 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#>              8         9         10
#> X1  -0.7321618 0.5783262 -0.6895172
#> X2  -0.6494250 0.5948016 -0.6184747
#> X3   0.5293477 0.4676483  0.4903296
#> X4   0.4680858 0.3319988  0.9018439
#> X5   0.0000000 0.0000000  0.0000000
#> X6   0.0000000 0.0000000  0.0000000
#> X7   0.0000000 0.0000000  0.0000000
#> X8   0.0000000 0.0000000  0.0000000
#> X9   0.0000000 0.0000000  0.0000000
#> X10  0.0000000 0.0000000  0.0000000
#> X11  0.0000000 0.0000000  0.0000000
#> X12  0.0000000 0.0000000  0.0000000
#> X13  0.0000000 0.0000000  0.0000000
#> X14  0.0000000 0.0000000  0.0000000
#> X15  0.0000000 0.0000000  0.0000000
#> X16  0.0000000 0.0000000  0.0000000
#> X17  0.0000000 0.0000000  0.0000000
#> X18  0.0000000 0.0000000  0.0000000
#> X19  0.0000000 0.0000000  0.0000000
#> X20  0.0000000 0.0000000  0.0000000