Subgroup Analysis for Cox proportional hazards model

Subgroup Analysis for Cox proportional hazards model

Usage

coxsg(
  formula,
  data,
  group,
  lambda1 = 0,
  lambda2 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 2,
  init,
  control,
  ...
)

Arguments

formula

a formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

a data frame containing the variables in the model.

group

a factor specifying the group of each sample.

lambda1

a non-negative value specifying the sparse penalty parameter. The default is 0.

lambda2

a non-negative value specifying the group penalty parameter. The default is 0.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3.0 for MCP.

rho

a value in (2, 10) specifying the expansion factor of the augmented Lagrangian's penalty parameter. The default is 2.0.

init

a numeric vector specifying the initial value of the coefficients. The default is a zero vector.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a coxsg object.

Examples

library(survtrans)
formula <- Surv(time, status) ~ . - group - id
group <- as.factor(sim2$group)
fit <- coxsg(
  formula, sim2, group,
  lambda1 = 0.07, lambda2 = 0.05, penalty = "SCAD"
)
fit$coefficients
#>             1          2         3          4         5          6         7
#> X1  0.5346558 -0.6600533 0.5346558 -0.6600533 0.5346558 -0.6600533 0.5346558
#> X2  0.5089769 -0.6405885 0.5089769 -0.6405885 0.5089769 -0.6405885 0.5089769
#> X3  0.5567368  0.5567368 0.5567368  0.5567368 0.5567368  0.5567368 0.5567368
#> X4  0.5539575  0.5539575 0.5539575  0.5539575 0.5539575  0.5539575 0.5539575
#> X5  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X6  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X7  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X8  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X9  0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X10 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X11 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X12 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X13 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X14 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X15 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X16 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X17 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X18 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X19 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#> X20 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000  0.0000000 0.0000000
#>              8         9         10
#> X1  -0.6600533 0.5346558 -0.6600533
#> X2  -0.6405885 0.5089769 -0.6405885
#> X3   0.5567368 0.5567368  0.5567368
#> X4   0.5539575 0.5539575  0.5539575
#> X5   0.0000000 0.0000000  0.0000000
#> X6   0.0000000 0.0000000  0.0000000
#> X7   0.0000000 0.0000000  0.0000000
#> X8   0.0000000 0.0000000  0.0000000
#> X9   0.0000000 0.0000000  0.0000000
#> X10  0.0000000 0.0000000  0.0000000
#> X11  0.0000000 0.0000000  0.0000000
#> X12  0.0000000 0.0000000  0.0000000
#> X13  0.0000000 0.0000000  0.0000000
#> X14  0.0000000 0.0000000  0.0000000
#> X15  0.0000000 0.0000000  0.0000000
#> X16  0.0000000 0.0000000  0.0000000
#> X17  0.0000000 0.0000000  0.0000000
#> X18  0.0000000 0.0000000  0.0000000
#> X19  0.0000000 0.0000000  0.0000000
#> X20  0.0000000 0.0000000  0.0000000