Cross-validation for ncvcox

Cross-validation for ncvcox

Usage

cv.ncvcox(
  formula,
  data,
  group,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

formula

A formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function.

data

A data frame containing the variables in the model.

group

A factor variable indicating the group of each observation.

penalty

A character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

A non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3.0 for MCP.

nfolds

An integer specifying the number of folds.

nlambdas

An integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

An integer specifying the random seed. Default is 0.

control

An object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Additional arguments to be passed to the fitting algorithm.

Value

An object of class cv.ncvcox.

Examples

formula <- survival::Surv(time, status) ~ . - group - id
df <- sim2[sim2$group == 2 | sim2$group == 4, ]
cvfit <- cv.ncvcox(formula, df, df$group, penalty = "SCAD")
coef(cvfit)
#>            X1            X2            X3            X4            X5 
#>  0.9019220245  1.0203864953  0.3473820420  0.4246209272  0.0000000000 
#>            X6            X7            X8            X9           X10 
#>  0.0000000000  0.0000000000  0.0000000000  0.0000000000  0.0000000000 
#>           X11           X12           X13           X14           X15 
#>  0.0000000000  0.0000000000  0.0667580637  0.0000000000  0.0000000000 
#>           X16           X17           X18           X19           X20 
#>  0.0000000000  0.0000000000 -0.0001856228  0.0000000000  0.0000000000