Cross-validation for coxtrans

Cross-validation for coxtrans

Usage

cv_coxtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a coxph object fitted to the source data.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 1.5 for MCP.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_coxtrans object.

Examples

library(survival)
library(survtrans)
fit_src <- coxph(Surv(time, status) ~ ., data = sim1_src)
cv_coxtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6475044  1.4867306 -1.0656651  1.0190857 -0.8911751  1.0482614 -1.1872427 
#>         X8         X9        X10 
#>  0.9370094 -0.9985653  1.1695095 
#> 
#> $coefficients$eta
#>  [1]  0.371656042  0.423018758 -0.064907982  0.000000000  0.142054927
#>  [6]  0.000000000 -0.151363266  0.007892429  0.000000000  0.157658156
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05716259
#> 
#> $coefficients$gamma
#> [1] 1
#> 
#> 
#> $lambdas
#>   [1] 5.716259e-02 5.208442e-02 4.745739e-02 4.324140e-02 3.939996e-02
#>   [6] 3.589977e-02 3.271054e-02 2.980462e-02 2.715686e-02 2.474432e-02
#>  [11] 2.254610e-02 2.054317e-02 1.871817e-02 1.705530e-02 1.554015e-02
#>  [16] 1.415961e-02 1.290171e-02 1.175556e-02 1.071123e-02 9.759670e-03
#>  [21] 8.892648e-03 8.102649e-03 7.382832e-03 6.726962e-03 6.129357e-03
#>  [26] 5.584842e-03 5.088700e-03 4.636634e-03 4.224728e-03 3.849415e-03
#>  [31] 3.507443e-03 3.195852e-03 2.911941e-03 2.653252e-03 2.417545e-03
#>  [36] 2.202777e-03 2.007088e-03 1.828784e-03 1.666320e-03 1.518289e-03
#>  [41] 1.383408e-03 1.260510e-03 1.148530e-03 1.046497e-03 9.535295e-04
#>  [46] 8.688205e-04 7.916369e-04 7.213101e-04 6.572309e-04 5.988443e-04
#>  [51] 5.456446e-04 4.971711e-04 4.530037e-04 4.127601e-04 3.760917e-04
#>  [56] 3.426807e-04 3.122379e-04 2.844996e-04 2.592254e-04 2.361965e-04
#>  [61] 2.152135e-04 1.960945e-04 1.786740e-04 1.628011e-04 1.483383e-04
#>  [66] 1.351603e-04 1.231531e-04 1.122125e-04 1.022438e-04 9.316078e-05
#>  [71] 8.488463e-05 7.734371e-05 7.047271e-05 6.421211e-05 5.850768e-05
#>  [76] 5.331002e-05 4.857411e-05 4.425892e-05 4.032708e-05 3.674453e-05
#>  [81] 3.348025e-05 3.050595e-05 2.779589e-05 2.532658e-05 2.307664e-05
#>  [86] 2.102657e-05 1.915863e-05 1.745663e-05 1.590583e-05 1.449280e-05
#>  [91] 1.320530e-05 1.203218e-05 1.096327e-05 9.989325e-06 9.101901e-06
#>  [96] 8.293313e-06 7.556558e-06 6.885254e-06 6.273587e-06 5.716259e-06
#> 
#> $etas
#>             [,1]      [,2]        [,3]       [,4]          [,5]        [,6]
#>   [1,] 0.3716560 0.4230188 -0.06490798 0.00000000  0.1420549267 0.000000000
#>   [2,] 0.3748949 0.4545394 -0.08728060 0.00000000  0.1441319946 0.000000000
#>   [3,] 0.3777670 0.4847601 -0.10832064 0.00000000  0.1460099979 0.000000000
#>   [4,] 0.3803154 0.5136364 -0.12808388 0.00000000  0.1477127370 0.000000000
#>   [5,] 0.3825780 0.5411377 -0.14662372 0.00000000  0.1492597001 0.000000000
#>   [6,] 0.3845885 0.5672473 -0.16399248 0.00000000  0.1506669595 0.000000000
#>   [7,] 0.3863769 0.5919597 -0.18024021 0.00000000  0.1519485486 0.000000000
#>   [8,] 0.3879698 0.6152811 -0.19541621 0.00000000  0.1531166340 0.000000000
#>   [9,] 0.3893898 0.6372306 -0.20957101 0.00000000  0.1541813843 0.000000000
#>  [10,] 0.3904512 0.6584102 -0.22299685 0.00000000  0.1550078077 0.000000000
#>  [11,] 0.3881534 0.6866925 -0.23932528 0.00000000  0.1530368477 0.009030287
#>  [12,] 0.3850644 0.7154097 -0.25565565 0.00000000  0.1502911969 0.022133607
#>  [13,] 0.3821570 0.7422605 -0.27089597 0.00000000  0.1477008436 0.034278025
#>  [14,] 0.3794264 0.7673205 -0.28509521 0.00000000  0.1452630989 0.045523944
#>  [15,] 0.3734122 0.7971759 -0.30592050 0.01559381  0.1365557480 0.064133074
#>  [16,] 0.3667669 0.8270847 -0.32762463 0.03458043  0.1265128372 0.083868120
#>  [17,] 0.3605024 0.8552024 -0.34801177 0.05237963  0.1170227140 0.102360888
#>  [18,] 0.3546127 0.8815782 -0.36712160 0.06903304  0.1080807572 0.119656323
#>  [19,] 0.3490888 0.9062693 -0.38499902 0.08458600  0.0996773470 0.135802949
#>  [20,] 0.3439191 0.9293396 -0.40169313 0.09908669  0.0917987342 0.150852051
#>  [21,] 0.3390916 0.9508543 -0.41725340 0.11258291  0.0844295559 0.164854414
#>  [22,] 0.3345917 0.9708857 -0.43173414 0.12512597  0.0775503709 0.177864185
#>  [23,] 0.3304053 0.9895046 -0.44518817 0.13676534  0.0711415486 0.189933536
#>  [24,] 0.3265163 1.0067859 -0.45767102 0.14755223  0.0651809248 0.201116265
#>  [25,] 0.3229097 1.0228014 -0.46923560 0.15753516  0.0596470372 0.211463333
#>  [26,] 0.3195693 1.0376261 -0.47993712 0.16676418  0.0545163170 0.221027078
#>  [27,] 0.3164801 1.0513295 -0.48982656 0.17528531  0.0497670039 0.229855720
#>  [28,] 0.3136261 1.0639837 -0.49895662 0.18314575  0.0453757006 0.237998462
#>  [29,] 0.3109927 1.0756558 -0.50737620 0.19038910  0.0413206712 0.245500850
#>  [30,] 0.3085653 1.0864113 -0.51513309 0.19705779  0.0375803114 0.252407082
#>  [31,] 0.3063300 1.0963131 -0.52227308 0.20319228  0.0341337082 0.258759298
#>  [32,] 0.3042733 1.1054215 -0.52883984 0.20883103  0.0309607368 0.264597523
#>  [33,] 0.3023825 1.1137934 -0.53487473 0.21401036  0.0280422215 0.269959513
#>  [34,] 0.3006457 1.1214826 -0.54041678 0.21876442  0.0253599638 0.274880788
#>  [35,] 0.2990513 1.1285400 -0.54550286 0.22312543  0.0228966689 0.279394800
#>  [36,] 0.2975883 1.1350145 -0.55016841 0.22712425  0.0206355514 0.283533567
#>  [37,] 0.2962472 1.1409493 -0.55444460 0.23078800  0.0185620160 0.287325297
#>  [38,] 0.2950180 1.1463883 -0.55836316 0.23414422  0.0166608450 0.290798504
#>  [39,] 0.2938924 1.1513687 -0.56195104 0.23721626  0.0149193761 0.293977465
#>  [40,] 0.2928616 1.1559288 -0.56523599 0.24002815  0.0133242014 0.296887030
#>  [41,] 0.2919185 1.1601009 -0.56824112 0.24259984  0.0118644241 0.299547961
#>  [42,] 0.2910556 1.1639181 -0.57099052 0.24495215  0.0105283424 0.301981757
#>  [43,] 0.2902664 1.1674087 -0.57350456 0.24710262  0.0093062591 0.304206634
#>  [44,] 0.2895452 1.1705987 -0.57580196 0.24906741  0.0081892350 0.306239352
#>  [45,] 0.2888859 1.1735149 -0.57790207 0.25086315  0.0071678125 0.308097091
#>  [46,] 0.2882834 1.1761793 -0.57982076 0.25250349  0.0062344148 0.309794015
#>  [47,] 0.2877331 1.1786130 -0.58157329 0.25400156  0.0053816697 0.311343715
#>  [48,] 0.2872306 1.1808355 -0.58317369 0.25536940  0.0046028007 0.312758667
#>  [49,] 0.2867717 1.1828648 -0.58463487 0.25661810  0.0038915658 0.314050343
#>  [50,] 0.2863528 1.1847173 -0.58596870 0.25775784  0.0032422201 0.315229287
#>  [51,] 0.2859705 1.1864080 -0.58718606 0.25879795  0.0026494966 0.316305154
#>  [52,] 0.2856216 1.1879509 -0.58829692 0.25974698  0.0021085661 0.317286791
#>  [53,] 0.2853030 1.1893595 -0.58931109 0.26061333  0.0016146038 0.318182876
#>  [54,] 0.2850127 1.1906435 -0.59023552 0.26140296  0.0011643807 0.318999632
#>  [55,] 0.2847474 1.1918161 -0.59107982 0.26212408  0.0007530563 0.319745474
#>  [56,] 0.2845058 1.1928847 -0.59184911 0.26278111  0.0003783145 0.320425046
#>  [57,] 0.2842852 1.1938599 -0.59255122 0.26338072  0.0000362253 0.321045209
#>  [58,] 0.2841761 1.1944793 -0.59299734 0.26371915  0.0000000000 0.321432660
#>  [59,] 0.2840873 1.1950122 -0.59338131 0.26400317  0.0000000000 0.321764951
#>  [60,] 0.2840064 1.1954979 -0.59373123 0.26426197  0.0000000000 0.322067764
#>  [61,] 0.2839327 1.1959403 -0.59404999 0.26449770  0.0000000000 0.322343608
#>  [62,] 0.2838217 1.1964725 -0.59443334 0.26481195 -0.0001308421 0.322680064
#>  [63,] 0.2836613 1.1971315 -0.59490766 0.26523252 -0.0004285535 0.323101340
#>  [64,] 0.2835151 1.1977323 -0.59534010 0.26561594 -0.0006999662 0.323485404
#>  [65,] 0.2833821 1.1982791 -0.59573364 0.26596489 -0.0009469838 0.323834938
#>  [66,] 0.2832606 1.1987785 -0.59609303 0.26628352 -0.0011725297 0.324154103
#>  [67,] 0.2831499 1.1992337 -0.59642064 0.26657397 -0.0013781244 0.324445037
#>  [68,] 0.2830489 1.1996486 -0.59671927 0.26683872 -0.0015655257 0.324710227
#>  [69,] 0.2829569 1.2000268 -0.59699147 0.26708004 -0.0017363378 0.324951943
#>  [70,] 0.2828731 1.2003715 -0.59723957 0.26729999 -0.0018920246 0.325172256
#>  [71,] 0.2827966 1.2006857 -0.59746570 0.26750046 -0.0020339213 0.325373056
#>  [72,] 0.2827270 1.2009721 -0.59767180 0.26768316 -0.0021632462 0.325556066
#>  [73,] 0.2826635 1.2012331 -0.59785964 0.26784968 -0.0022811104 0.325722858
#>  [74,] 0.2826056 1.2014710 -0.59803083 0.26800144 -0.0023885272 0.325874867
#>  [75,] 0.2825529 1.2016878 -0.59818685 0.26813975 -0.0024864208 0.326013399
#>  [76,] 0.2825048 1.2018853 -0.59832904 0.26826579 -0.0025756339 0.326139647
#>  [77,] 0.2824610 1.2020654 -0.59845861 0.26838065 -0.0026569349 0.326254700
#>  [78,] 0.2824211 1.2022295 -0.59857670 0.26848533 -0.0027310245 0.326359547
#>  [79,] 0.2823847 1.2023790 -0.59868431 0.26858072 -0.0027985414 0.326455093
#>  [80,] 0.2823516 1.2025153 -0.59878237 0.26866765 -0.0028600679 0.326542162
#>  [81,] 0.2823213 1.2026394 -0.59887174 0.26874686 -0.0029161350 0.326621505
#>  [82,] 0.2822938 1.2027526 -0.59895317 0.26881904 -0.0029672265 0.326693807
#>  [83,] 0.2822687 1.2028557 -0.59902738 0.26888482 -0.0030137836 0.326759693
#>  [84,] 0.2822459 1.2029497 -0.59909500 0.26894476 -0.0030562083 0.326819730
#>  [85,] 0.2822251 1.2030353 -0.59915662 0.26899938 -0.0030948671 0.326874438
#>  [86,] 0.2822061 1.2031133 -0.59921277 0.26904915 -0.0031300941 0.326924290
#>  [87,] 0.2821888 1.2031844 -0.59926393 0.26909451 -0.0031621938 0.326969716
#>  [88,] 0.2821730 1.2032492 -0.59931055 0.26913583 -0.0031914435 0.327011109
#>  [89,] 0.2821587 1.2033082 -0.59935303 0.26917349 -0.0032180962 0.327048827
#>  [90,] 0.2821456 1.2033620 -0.59939175 0.26920780 -0.0032423823 0.327083196
#>  [91,] 0.2821337 1.2034110 -0.59942702 0.26923907 -0.0032645119 0.327114513
#>  [92,] 0.2821228 1.2034557 -0.59945916 0.26926756 -0.0032846764 0.327143049
#>  [93,] 0.2821129 1.2034964 -0.59948845 0.26929352 -0.0033030502 0.327169051
#>  [94,] 0.2821039 1.2035335 -0.59951513 0.26931717 -0.0033197923 0.327192744
#>  [95,] 0.2820957 1.2035673 -0.59953945 0.26933873 -0.0033350476 0.327214332
#>  [96,] 0.2820882 1.2035981 -0.59956161 0.26935836 -0.0033489480 0.327234004
#>  [97,] 0.2820813 1.2036261 -0.59958180 0.26937626 -0.0033616139 0.327251928
#>  [98,] 0.2820751 1.2036517 -0.59960019 0.26939257 -0.0033731548 0.327268261
#>  [99,] 0.2820695 1.2036750 -0.59961695 0.26940742 -0.0033836707 0.327283142
#> [100,] 0.2820643 1.2036962 -0.59963223 0.26942096 -0.0033932526 0.327296702
#>              [,7]        [,8]         [,9]     [,10]
#>   [1,] -0.1513633 0.007892429  0.000000000 0.1576582
#>   [2,] -0.1746115 0.034046427  0.000000000 0.1874781
#>   [3,] -0.1968949 0.058756204  0.000000000 0.2153354
#>   [4,] -0.2181782 0.082064136  0.000000000 0.2413447
#>   [5,] -0.2384375 0.104012130  0.000000000 0.2656114
#>   [6,] -0.2576606 0.124643125  0.000000000 0.2882336
#>   [7,] -0.2758438 0.144000406  0.000000000 0.3093022
#>   [8,] -0.2929928 0.162128826  0.000000000 0.3289033
#>   [9,] -0.3091230 0.179076351  0.000000000 0.3471201
#>  [10,] -0.3246517 0.195126023 -0.001119391 0.3643264
#>  [11,] -0.3461188 0.214965211 -0.013361944 0.3852795
#>  [12,] -0.3683290 0.235198923 -0.026340637 0.4062711
#>  [13,] -0.3890886 0.254128974 -0.038402061 0.4258114
#>  [14,] -0.4084590 0.271808692 -0.049597831 0.4439784
#>  [15,] -0.4332529 0.293406267 -0.067402502 0.4701910
#>  [16,] -0.4584292 0.315119967 -0.086167330 0.4973960
#>  [17,] -0.4820600 0.335497669 -0.103750660 0.5229089
#>  [18,] -0.5041943 0.354583441 -0.120195324 0.5467884
#>  [19,] -0.5248869 0.372425709 -0.135547536 0.5690983
#>  [20,] -0.5441970 0.389076386 -0.149856107 0.5899064
#>  [21,] -0.5621845 0.404587579 -0.163169408 0.6092802
#>  [22,] -0.5789142 0.419015315 -0.175538936 0.6272920
#>  [23,] -0.5944492 0.432414166 -0.187014346 0.6440119
#>  [24,] -0.6088552 0.444840740 -0.197646791 0.6595123
#>  [25,] -0.6221952 0.456349151 -0.207484757 0.6738620
#>  [26,] -0.6345339 0.466995144 -0.216577989 0.6871319
#>  [27,] -0.6459316 0.476830536 -0.224972360 0.6993876
#>  [28,] -0.6564500 0.485908252 -0.232714629 0.7106959
#>  [29,] -0.6661464 0.494277620 -0.239848105 0.7211192
#>  [30,] -0.6750766 0.501986638 -0.246414804 0.7307177
#>  [31,] -0.6832941 0.509081218 -0.252454784 0.7395493
#>  [32,] -0.6908496 0.515605096 -0.258006093 0.7476689
#>  [33,] -0.6977914 0.521599649 -0.263104621 0.7551284
#>  [34,] -0.7041648 0.527103897 -0.267784135 0.7619766
#>  [35,] -0.7100125 0.532154667 -0.272076436 0.7682597
#>  [36,] -0.7153756 0.536787213 -0.276011950 0.7740218
#>  [37,] -0.7202902 0.541032801 -0.279617510 0.7793019
#>  [38,] -0.7247930 0.544922879 -0.282920204 0.7841394
#>  [39,] -0.7289152 0.548484457 -0.285943132 0.7885678
#>  [40,] -0.7326887 0.551744998 -0.288709898 0.7926216
#>  [41,] -0.7361404 0.554727678 -0.291240260 0.7963296
#>  [42,] -0.7392980 0.557456292 -0.293554641 0.7997215
#>  [43,] -0.7421849 0.559951170 -0.295670365 0.8028227
#>  [44,] -0.7448229 0.562231041 -0.297603377 0.8056563
#>  [45,] -0.7472340 0.564314948 -0.299369996 0.8082463
#>  [46,] -0.7494367 0.566218750 -0.300983695 0.8106124
#>  [47,] -0.7514485 0.567957618 -0.302457398 0.8127733
#>  [48,] -0.7532855 0.569545492 -0.303802967 0.8147465
#>  [49,] -0.7549625 0.570995194 -0.305031311 0.8165479
#>  [50,] -0.7564934 0.572318512 -0.306152453 0.8181923
#>  [51,] -0.7578905 0.573526251 -0.307175577 0.8196929
#>  [52,] -0.7591653 0.574628313 -0.308109095 0.8210622
#>  [53,] -0.7603290 0.575634359 -0.308961250 0.8223122
#>  [54,] -0.7613898 0.576551480 -0.309737977 0.8234516
#>  [55,] -0.7623586 0.577388944 -0.310447255 0.8244922
#>  [56,] -0.7632413 0.578152113 -0.311093523 0.8254403
#>  [57,] -0.7640468 0.578848563 -0.311683290 0.8263055
#>  [58,] -0.7645987 0.579353857 -0.312056658 0.8268993
#>  [59,] -0.7650804 0.579799334 -0.312377680 0.8274179
#>  [60,] -0.7655195 0.580205338 -0.312670219 0.8278904
#>  [61,] -0.7659194 0.580575225 -0.312936700 0.8283208
#>  [62,] -0.7663714 0.580974592 -0.313258115 0.8288066
#>  [63,] -0.7669009 0.581422104 -0.313656959 0.8293750
#>  [64,] -0.7673837 0.581830102 -0.314020576 0.8298931
#>  [65,] -0.7678230 0.582201455 -0.314351507 0.8303647
#>  [66,] -0.7682242 0.582540531 -0.314653682 0.8307953
#>  [67,] -0.7685899 0.582849624 -0.314929129 0.8311878
#>  [68,] -0.7689232 0.583131374 -0.315180204 0.8315456
#>  [69,] -0.7692271 0.583388190 -0.315409055 0.8318718
#>  [70,] -0.7695040 0.583622272 -0.315617642 0.8321690
#>  [71,] -0.7697565 0.583835624 -0.315807755 0.8324400
#>  [72,] -0.7699865 0.584030078 -0.315981025 0.8326869
#>  [73,] -0.7701962 0.584207303 -0.316138941 0.8329120
#>  [74,] -0.7703873 0.584368822 -0.316282861 0.8331171
#>  [75,] -0.7705614 0.584516023 -0.316414021 0.8333040
#>  [76,] -0.7707202 0.584650174 -0.316533551 0.8334744
#>  [77,] -0.7708648 0.584772429 -0.316642481 0.8336296
#>  [78,] -0.7709966 0.584883841 -0.316741750 0.8337711
#>  [79,] -0.7711167 0.584985371 -0.316832212 0.8339000
#>  [80,] -0.7712262 0.585077894 -0.316914648 0.8340175
#>  [81,] -0.7713259 0.585162207 -0.316989770 0.8341246
#>  [82,] -0.7714168 0.585239039 -0.317058225 0.8342222
#>  [83,] -0.7714997 0.585309052 -0.317120605 0.8343111
#>  [84,] -0.7715751 0.585372852 -0.317177448 0.8343921
#>  [85,] -0.7716439 0.585430989 -0.317229246 0.8344659
#>  [86,] -0.7717066 0.585483965 -0.317276445 0.8345332
#>  [87,] -0.7717637 0.585532238 -0.317319454 0.8345945
#>  [88,] -0.7718157 0.585576226 -0.317358645 0.8346504
#>  [89,] -0.7718632 0.585616308 -0.317394356 0.8347013
#>  [90,] -0.7719064 0.585652832 -0.317426896 0.8347476
#>  [91,] -0.7719457 0.585686112 -0.317456547 0.8347899
#>  [92,] -0.7719816 0.585716437 -0.317483565 0.8348284
#>  [93,] -0.7720143 0.585744069 -0.317508184 0.8348635
#>  [94,] -0.7720441 0.585769248 -0.317530616 0.8348955
#>  [95,] -0.7720712 0.585792190 -0.317551056 0.8349246
#>  [96,] -0.7720960 0.585813095 -0.317569681 0.8349512
#>  [97,] -0.7721185 0.585832143 -0.317586652 0.8349753
#>  [98,] -0.7721390 0.585849500 -0.317602115 0.8349974
#>  [99,] -0.7721578 0.585865315 -0.317616205 0.8350175
#> [100,] -0.7721748 0.585879725 -0.317629044 0.8350358
#> 
#> $cvm
#>   [1] -24.45521 -24.30858 -24.15212 -23.98787 -23.85385 -23.74159 -23.64928
#>   [8] -23.54931 -23.45717 -23.38038 -23.30033 -23.22359 -23.15447 -23.06292
#>  [15] -22.95685 -22.85370 -22.76146 -22.68161 -22.61209 -22.55056 -22.49299
#>  [22] -22.43891 -22.37869 -22.32468 -22.27764 -22.23507 -22.20271 -22.17397
#>  [29] -22.14808 -22.12457 -22.10285 -22.07702 -22.05401 -22.03394 -22.01604
#>  [36] -21.99996 -21.98549 -21.97247 -21.96074 -21.95017 -21.94063 -21.93202
#>  [43] -21.92425 -21.91722 -21.91087 -21.90530 -21.90028 -21.89574 -21.89162
#>  [50] -21.88774 -21.88416 -21.88093 -21.87799 -21.87532 -21.87290 -21.87070
#>  [57] -21.86870 -21.86689 -21.86524 -21.86374 -21.86237 -21.86113 -21.86000
#>  [64] -21.85898 -21.85804 -21.85719 -21.85642 -21.85571 -21.85507 -21.85449
#>  [71] -21.85396 -21.85347 -21.85303 -21.85263 -21.85227 -21.85194 -21.85163
#>  [78] -21.85136 -21.85111 -21.85088 -21.85067 -21.85048 -21.85031 -21.85015
#>  [85] -21.85001 -21.84988 -21.84976 -21.84965 -21.84955 -21.84946 -21.84938
#>  [92] -21.84930 -21.84923 -21.84917 -21.84911 -21.84906 -21.84902 -21.84897
#>  [99] -21.84893 -21.84890
#> 
#> $cvsd
#>   [1] 7.998071 8.035566 8.059175 8.063192 8.078950 8.100795 8.133942 8.159199
#>   [9] 8.181295 8.203778 8.212187 8.225429 8.249921 8.292182 8.339873 8.397462
#>  [17] 8.456755 8.517435 8.577917 8.636310 8.688091 8.737858 8.790639 8.840773
#>  [25] 8.889658 8.942573 8.992258 9.038661 9.081675 9.121410 9.158403 9.193795
#>  [33] 9.226881 9.257938 9.286799 9.313508 9.338189 9.360968 9.381970 9.401311
#>  [41] 9.419109 9.435469 9.450500 9.464295 9.476949 9.488670 9.499419 9.509262
#>  [49] 9.518272 9.526416 9.533838 9.540622 9.546825 9.552494 9.557672 9.562403
#>  [57] 9.566723 9.570670 9.574270 9.577557 9.580556 9.583293 9.585790 9.588068
#>  [65] 9.590145 9.592040 9.593768 9.595343 9.596780 9.598090 9.599285 9.600374
#>  [73] 9.601367 9.602272 9.603097 9.603849 9.604534 9.605159 9.605728 9.606247
#>  [81] 9.606720 9.607151 9.607544 9.607902 9.608228 9.608526 9.608796 9.609046
#>  [89] 9.609271 9.609476 9.609662 9.609833 9.609988 9.610129 9.610258 9.610375
#>  [97] 9.610482 9.610580 9.610668 9.610749
#> 
#> $formula
#> Surv(time, status) ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + 
#>     X9 + X10
#> <environment: 0x55d8f9f70180>
#> 
#> attr(,"class")
#> [1] "cv_coxtrans"