Cross-validation for ncvcox

Cross-validation for ncvcox

Usage

cv_ncvcox(
  formula,
  data,
  group,
  offset,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  parallel = FALSE,
  control,
  ...
)

Arguments

formula

a formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

a data frame containing the variables in the model.

group

a factor specifying the group of each sample.

offset

a numeric vector specifying the offset..

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3 for MCP.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

parallel

a logical value specifying whether to perform the cross-validation in parallel. Default is FALSE.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_ncvcox object.

Examples

library(survtrans)
formula <- Surv(time, status) ~ . - group - id
group <- as.factor(sim1$group)
cv_fit <- cv_ncvcox(
  formula, sim1[sim1$group == 1, ],
  penalty = "SCAD"
)
coef(cv_fit)
#>           [,1]      [,2]      [,3]      [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,] 0.5029167 0.4979949 0.5088049 0.5416722    0    0    0    0    0     0
#>      [,11] [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#> [1,]     0     0     0     0     0     0     0     0     0     0