Cross-validation for survtran

Cross-validation for survtran

Usage

cv_odetrans(
  data,
  fit_source,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

cbh_func

a function specifying the cumulative baseline hazard. The default is NULL, which means the cumulative baseline hazard is estimateds from the survode object fitted to the source data.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_odetrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6132374  1.4006671 -1.0028305  1.0047907 -0.8535958  1.0413994 -1.1271410 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0881526 
#> 
#> $coefficients$eta
#>  [1]  0.390389983  0.343665828 -0.007496167  0.000000000  0.170431294
#>  [6]  0.000000000 -0.095123157  0.000000000  0.000000000  0.085553827
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05536255
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]          [,3]         [,4]        [,5]
#>   [1,] 0.08525555 0.02937811  0.0000000000  0.000000000 0.000000000
#>   [2,] 0.11560547 0.06612220  0.0000000000  0.000000000 0.000000000
#>   [3,] 0.14426287 0.10107474  0.0000000000  0.000000000 0.000000000
#>   [4,] 0.17117741 0.13349865  0.0000000000  0.000000000 0.003269571
#>   [5,] 0.19645868 0.16138757  0.0000000000  0.000000000 0.020975899
#>   [6,] 0.22020318 0.18782313  0.0000000000  0.000000000 0.037552409
#>   [7,] 0.24239759 0.21273070  0.0000000000  0.000000000 0.052994375
#>   [8,] 0.26305121 0.23606940  0.0000000000  0.000000000 0.067314565
#>   [9,] 0.28198608 0.25734278  0.0000000000  0.000000000 0.080509644
#>  [10,] 0.29905835 0.27615901  0.0000000000  0.000000000 0.092749554
#>  [11,] 0.31480521 0.29367296  0.0000000000  0.000000000 0.104192227
#>  [12,] 0.33105989 0.30374750  0.0000000000  0.000000000 0.117920610
#>  [13,] 0.34593016 0.31299549  0.0000000000  0.000000000 0.130478459
#>  [14,] 0.35949804 0.32150603  0.0000000000  0.000000000 0.141925613
#>  [15,] 0.37186105 0.32932137  0.0000000000  0.000000000 0.152347470
#>  [16,] 0.38306664 0.33649647 -0.0001209964  0.000000000 0.161826327
#>  [17,] 0.39038998 0.34366583 -0.0074961668  0.000000000 0.170431294
#>  [18,] 0.39708024 0.35021604 -0.0141585653  0.000000000 0.178268639
#>  [19,] 0.40356097 0.35602658 -0.0203460430  0.000000000 0.185418564
#>  [20,] 0.40986722 0.36116949 -0.0261142324  0.000000000 0.191967117
#>  [21,] 0.41564432 0.36586626 -0.0313414351  0.000000000 0.197953092
#>  [22,] 0.42102446 0.37015456 -0.0359936552  0.000000000 0.203489234
#>  [23,] 0.42640257 0.37409756 -0.0397605957  0.000000000 0.208896565
#>  [24,] 0.43130031 0.37771605 -0.0431490933  0.000000000 0.213837971
#>  [25,] 0.43575903 0.38103307 -0.0462003355  0.000000000 0.218351033
#>  [26,] 0.44027165 0.38678347 -0.0492628446 -0.005499923 0.223815928
#>  [27,] 0.44461956 0.39343967 -0.0521475153 -0.013378233 0.229532878
#>  [28,] 0.44860789 0.39961922 -0.0547689418 -0.020627042 0.234788993
#>  [29,] 0.45226344 0.40534189 -0.0571509443 -0.027287814 0.239615389
#>  [30,] 0.45561163 0.41062979 -0.0593152900 -0.033401135 0.244042421
#>  [31,] 0.45867645 0.41550707 -0.0612818665 -0.039006580 0.248099355
#>  [32,] 0.46148028 0.41999754 -0.0630687848 -0.044141473 0.251814140
#>  [33,] 0.46404413 0.42412656 -0.0646925809 -0.048842046 0.255213267
#>  [34,] 0.46638756 0.42791873 -0.0661682759 -0.053142309 0.258321709
#>  [35,] 0.46852871 0.43139790 -0.0675095077 -0.057074184 0.261162886
#>  [36,] 0.47048443 0.43458710 -0.0687286670 -0.060667549 0.263758647
#>  [37,] 0.47227025 0.43750829 -0.0698370000 -0.063950238 0.266129308
#>  [38,] 0.47390055 0.44018224 -0.0708447045 -0.066948071 0.268293700
#>  [39,] 0.47538853 0.44262848 -0.0717610322 -0.069684959 0.270269236
#>  [40,] 0.47674636 0.44486531 -0.0725943737 -0.072182980 0.272071972
#>  [41,] 0.47798521 0.44690980 -0.0733523381 -0.074462486 0.273716698
#>  [42,] 0.47911535 0.44877779 -0.0740418258 -0.076542201 0.275217005
#>  [43,] 0.48014619 0.45048400 -0.0746690974 -0.078439334 0.276585375
#>  [44,] 0.48108633 0.45204199 -0.0752398292 -0.080169670 0.277833251
#>  [45,] 0.48194367 0.45346430 -0.0757591713 -0.081747681 0.278971123
#>  [46,] 0.48272544 0.45476245 -0.0762317980 -0.083186624 0.280008590
#>  [47,] 0.48343824 0.45594707 -0.0766619516 -0.084498635 0.280954436
#>  [48,] 0.48408811 0.45702792 -0.0770534844 -0.085694823 0.281816692
#>  [49,] 0.48468057 0.45801394 -0.0774098930 -0.086785335 0.282602697
#>  [50,] 0.48522066 0.45891334 -0.0777343525 -0.087779444 0.283319157
#>  [51,] 0.48571298 0.45973363 -0.0780297479 -0.088685623 0.283972196
#>  [52,] 0.48616175 0.46048169 -0.0782987006 -0.089511612 0.284567404
#>  [53,] 0.48657079 0.46116382 -0.0785435926 -0.090264476 0.285109883
#>  [54,] 0.48694362 0.46178580 -0.0787665889 -0.090950665 0.285604291
#>  [55,] 0.48728342 0.46235287 -0.0789696580 -0.091576063 0.286054873
#>  [56,] 0.48759312 0.46286986 -0.0791545895 -0.092146040 0.286465507
#>  [57,] 0.48787537 0.46334116 -0.0793230108 -0.092665494 0.286839725
#>  [58,] 0.48813260 0.46377079 -0.0794764022 -0.093138893 0.287180751
#>  [59,] 0.48836702 0.46416241 -0.0796161103 -0.093570312 0.287491524
#>  [60,] 0.48858066 0.46451938 -0.0797433600 -0.093963467 0.287774724
#>  [61,] 0.48877535 0.46484475 -0.0798592661 -0.094321746 0.288032794
#>  [62,] 0.48895277 0.46514130 -0.0799648427 -0.094648238 0.288267961
#>  [63,] 0.48911444 0.46541158 -0.0800610131 -0.094945759 0.288482255
#>  [64,] 0.48926178 0.46565792 -0.0801486173 -0.095216878 0.288677528
#>  [65,] 0.48939603 0.46588242 -0.0802284203 -0.095463934 0.288855467
#>  [66,] 0.48951838 0.46608702 -0.0803011181 -0.095689062 0.289017608
#>  [67,] 0.48962986 0.46627347 -0.0803673447 -0.095894206 0.289165354
#>  [68,] 0.48973145 0.46644340 -0.0804276771 -0.096081138 0.289299982
#>  [69,] 0.48982402 0.46659825 -0.0804826407 -0.096251475 0.289422656
#>  [70,] 0.48990837 0.46673936 -0.0805327141 -0.096406688 0.289534437
#>  [71,] 0.48998524 0.46686796 -0.0805783328 -0.096548120 0.289636291
#>  [72,] 0.49005528 0.46698514 -0.0806198938 -0.096676993 0.289729100
#>  [73,] 0.49011910 0.46709193 -0.0806577583 -0.096794423 0.289813667
#>  [74,] 0.49017725 0.46718924 -0.0806922555 -0.096901425 0.289890724
#>  [75,] 0.49023024 0.46727791 -0.0807236850 -0.096998924 0.289960937
#>  [76,] 0.49027852 0.46735871 -0.0807523200 -0.097087765 0.290024914
#>  [77,] 0.49032252 0.46743233 -0.0807784091 -0.097168715 0.290083209
#>  [78,] 0.49036261 0.46749943 -0.0808021789 -0.097242476 0.290136326
#>  [79,] 0.49039913 0.46756056 -0.0808238355 -0.097309686 0.290184726
#>  [80,] 0.49043242 0.46761627 -0.0808435671 -0.097370927 0.290228826
#>  [81,] 0.49046274 0.46766703 -0.0808615448 -0.097426728 0.290269009
#>  [82,] 0.49049038 0.46771328 -0.0808779246 -0.097477573 0.290305623
#>  [83,] 0.49051556 0.46775542 -0.0808928486 -0.097523902 0.290338985
#>  [84,] 0.49053850 0.46779383 -0.0809064463 -0.097566116 0.290369383
#>  [85,] 0.49055941 0.46782882 -0.0809188355 -0.097604580 0.290397081
#>  [86,] 0.49057846 0.46786070 -0.0809301237 -0.097639627 0.290422318
#>  [87,] 0.49059581 0.46788976 -0.0809404087 -0.097671561 0.290445314
#>  [88,] 0.49061163 0.46791623 -0.0809497798 -0.097700659 0.290466267
#>  [89,] 0.49062604 0.46794035 -0.0809583182 -0.097727172 0.290485359
#>  [90,] 0.49063917 0.46796233 -0.0809660979 -0.097751330 0.290502755
#>  [91,] 0.49065113 0.46798236 -0.0809731863 -0.097773342 0.290518605
#>  [92,] 0.49066203 0.46800061 -0.0809796448 -0.097793398 0.290533048
#>  [93,] 0.49067196 0.46801723 -0.0809855295 -0.097811673 0.290546207
#>  [94,] 0.49068101 0.46803238 -0.0809908914 -0.097828324 0.290558198
#>  [95,] 0.49068926 0.46804619 -0.0809957768 -0.097843497 0.290569123
#>  [96,] 0.49069677 0.46805877 -0.0810002281 -0.097857321 0.290579078
#>  [97,] 0.49070362 0.46807023 -0.0810042840 -0.097869918 0.290588149
#>  [98,] 0.49070986 0.46808067 -0.0810079795 -0.097881395 0.290596413
#>  [99,] 0.49071554 0.46809019 -0.0810113467 -0.097891853 0.290603944
#> [100,] 0.49072072 0.46809886 -0.0810144147 -0.097901382 0.290610805
#>                [,6]         [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,]  0.000000000 -0.004351414 0.000000000 0.000000000 0.00000000
#>  [10,]  0.000000000 -0.017451541 0.000000000 0.000000000 0.00000000
#>  [11,]  0.000000000 -0.029568468 0.000000000 0.000000000 0.00000000
#>  [12,]  0.000000000 -0.043135169 0.000000000 0.000000000 0.01836350
#>  [13,]  0.000000000 -0.055618567 0.000000000 0.000000000 0.03484694
#>  [14,]  0.000000000 -0.067055437 0.000000000 0.000000000 0.04952991
#>  [15,]  0.000000000 -0.077509587 0.000000000 0.000000000 0.06262118
#>  [16,]  0.000000000 -0.087038475 0.000000000 0.000000000 0.07431871
#>  [17,]  0.000000000 -0.095123157 0.000000000 0.000000000 0.08555383
#>  [18,]  0.000000000 -0.102462574 0.000000000 0.000000000 0.09560261
#>  [19,]  0.000000000 -0.109050597 0.000000000 0.003027081 0.10459626
#>  [20,]  0.000000000 -0.114960305 0.000000000 0.008808979 0.11265652
#>  [21,]  0.000000000 -0.120323331 0.000000000 0.014108970 0.11989404
#>  [22,]  0.000000000 -0.125439517 0.001288226 0.018965859 0.12669273
#>  [23,]  0.000000000 -0.131381380 0.008975371 0.023441822 0.13428040
#>  [24,]  0.000000000 -0.136815408 0.015956094 0.027549643 0.14110716
#>  [25,]  0.000000000 -0.141780098 0.022295470 0.031315353 0.14725491
#>  [26,] -0.005097015 -0.146158635 0.027801387 0.036225053 0.15169948
#>  [27,] -0.011864061 -0.150168447 0.032728108 0.041447201 0.15511544
#>  [28,] -0.018076110 -0.153874331 0.037228630 0.046272456 0.15815270
#>  [29,] -0.023772276 -0.157291917 0.041337610 0.050723419 0.16085752
#>  [30,] -0.028990312 -0.160437904 0.045087447 0.054822996 0.16327010
#>  [31,] -0.033766497 -0.163329339 0.048508217 0.058594050 0.16542519
#>  [32,] -0.038134920 -0.165983510 0.051627941 0.062058839 0.16735337
#>  [33,] -0.042128138 -0.168417081 0.054472325 0.065239123 0.16908078
#>  [34,] -0.045776545 -0.170646285 0.057065126 0.068155730 0.17063034
#>  [35,] -0.049108479 -0.172686682 0.059428216 0.070828488 0.17202210
#>  [36,] -0.052150302 -0.174553009 0.061581658 0.073276160 0.17327357
#>  [37,] -0.054926440 -0.176259142 0.063543845 0.075516407 0.17440010
#>  [38,] -0.057459449 -0.177818092 0.065331624 0.077565765 0.17541519
#>  [39,] -0.059770127 -0.179241975 0.066960398 0.079439666 0.17633069
#>  [40,] -0.061877610 -0.180542049 0.068444242 0.081152464 0.17715711
#>  [41,] -0.063799475 -0.181728731 0.069796009 0.082717473 0.17790368
#>  [42,] -0.065551849 -0.182811645 0.071027429 0.084147014 0.17857862
#>  [43,] -0.067149511 -0.183799652 0.072149197 0.085452467 0.17918920
#>  [44,] -0.068605989 -0.184700910 0.073171073 0.086644323 0.17974190
#>  [45,] -0.069933655 -0.185522913 0.074101951 0.087732240 0.18024250
#>  [46,] -0.071143824 -0.186272532 0.074949935 0.088725099 0.18069614
#>  [47,] -0.072246835 -0.186956065 0.075722412 0.089631057 0.18110741
#>  [48,] -0.073252134 -0.187579275 0.076426107 0.090457602 0.18148045
#>  [49,] -0.074168341 -0.188147441 0.077067152 0.091211597 0.18181892
#>  [50,] -0.075003321 -0.188665389 0.077651130 0.091899329 0.18212616
#>  [51,] -0.075764255 -0.189137529 0.078183128 0.092526558 0.18240514
#>  [52,] -0.076457693 -0.189567889 0.078667778 0.093098553 0.18265852
#>  [53,] -0.077089610 -0.189960150 0.079109297 0.093620134 0.18288874
#>  [54,] -0.077665452 -0.190317669 0.079511530 0.094095710 0.18309795
#>  [55,] -0.078190189 -0.190643512 0.079877975 0.094529310 0.18328811
#>  [56,] -0.078668350 -0.190940478 0.080211820 0.094924615 0.18346101
#>  [57,] -0.079104065 -0.191211117 0.080515968 0.095284988 0.18361823
#>  [58,] -0.079501097 -0.191457759 0.080793063 0.095613499 0.18376122
#>  [59,] -0.079862878 -0.191682527 0.081045515 0.095912953 0.18389130
#>  [60,] -0.080192536 -0.191887357 0.081275516 0.096185909 0.18400964
#>  [61,] -0.080492921 -0.192074015 0.081485064 0.096434703 0.18411731
#>  [62,] -0.080766631 -0.192244110 0.081675981 0.096661466 0.18421530
#>  [63,] -0.081016034 -0.192399110 0.081849923 0.096868143 0.18430448
#>  [64,] -0.081243288 -0.192540354 0.082008400 0.097056509 0.18438566
#>  [65,] -0.081450358 -0.192669061 0.082152790 0.097228181 0.18445955
#>  [66,] -0.081639037 -0.192786343 0.082284344 0.097384636 0.18452682
#>  [67,] -0.081810958 -0.192893213 0.082404204 0.097527220 0.18458806
#>  [68,] -0.081967609 -0.192990595 0.082513411 0.097657160 0.18464383
#>  [69,] -0.082110346 -0.193079331 0.082612912 0.097775575 0.18469461
#>  [70,] -0.082240404 -0.193160188 0.082703569 0.097883487 0.18474084
#>  [71,] -0.082358910 -0.193233865 0.082786169 0.097981825 0.18478295
#>  [72,] -0.082466890 -0.193300999 0.082861428 0.098071438 0.18482130
#>  [73,] -0.082565278 -0.193362172 0.082929999 0.098153099 0.18485622
#>  [74,] -0.082654926 -0.193417912 0.082992477 0.098227513 0.18488803
#>  [75,] -0.082736611 -0.193468702 0.083049403 0.098295323 0.18491701
#>  [76,] -0.082811040 -0.193514981 0.083101270 0.098357113 0.18494340
#>  [77,] -0.082878857 -0.193557150 0.083148528 0.098413419 0.18496743
#>  [78,] -0.082940650 -0.193595573 0.083191587 0.098464726 0.18498933
#>  [79,] -0.082996953 -0.193630584 0.083230821 0.098511478 0.18500928
#>  [80,] -0.083048256 -0.193662485 0.083266568 0.098554079 0.18502744
#>  [81,] -0.083095000 -0.193691553 0.083299139 0.098592897 0.18504400
#>  [82,] -0.083137593 -0.193718038 0.083328816 0.098628269 0.18505908
#>  [83,] -0.083176401 -0.193742171 0.083355856 0.098660499 0.18507281
#>  [84,] -0.083211762 -0.193764161 0.083380494 0.098689868 0.18508533
#>  [85,] -0.083243982 -0.193784197 0.083402942 0.098716629 0.18509673
#>  [86,] -0.083273340 -0.193802454 0.083423397 0.098741013 0.18510711
#>  [87,] -0.083300090 -0.193819088 0.083442034 0.098763231 0.18511658
#>  [88,] -0.083324463 -0.193834245 0.083459015 0.098783476 0.18512520
#>  [89,] -0.083346671 -0.193848056 0.083474487 0.098801923 0.18513305
#>  [90,] -0.083366906 -0.193860640 0.083488585 0.098818732 0.18514021
#>  [91,] -0.083385344 -0.193872106 0.083501431 0.098834048 0.18514673
#>  [92,] -0.083402144 -0.193882553 0.083513135 0.098848003 0.18515267
#>  [93,] -0.083417451 -0.193892073 0.083523799 0.098860719 0.18515808
#>  [94,] -0.083431398 -0.193900747 0.083533516 0.098872305 0.18516301
#>  [95,] -0.083444107 -0.193908650 0.083542370 0.098882863 0.18516750
#>  [96,] -0.083455686 -0.193915851 0.083550437 0.098892482 0.18517159
#>  [97,] -0.083466237 -0.193922412 0.083557787 0.098901247 0.18517532
#>  [98,] -0.083475850 -0.193928391 0.083564484 0.098909234 0.18517872
#>  [99,] -0.083484610 -0.193933838 0.083570587 0.098916511 0.18518182
#> [100,] -0.083492591 -0.193938802 0.083576147 0.098923141 0.18518464
#> 
#> $cvm
#>   [1] -3.973187 -3.850573 -3.739456 -3.649509 -3.553908 -3.450061 -3.365267
#>   [8] -3.294079 -3.234938 -3.189845 -3.139442 -3.092057 -3.048299 -3.016438
#>  [15] -2.991697 -2.974506 -2.970467 -2.975804 -2.986616 -3.004907 -3.027371
#>  [22] -3.052378 -3.076390 -3.101870 -3.126524 -3.150746 -3.173053 -3.192544
#>  [29] -3.209552 -3.225423 -3.240471 -3.254325 -3.266302 -3.275628 -3.284619
#>  [36] -3.292948 -3.299853 -3.306426 -3.312652 -3.318522 -3.324035 -3.329196
#>  [43] -3.333085 -3.336376 -3.339458 -3.342335 -3.345013 -3.347070 -3.348724
#>  [50] -3.350264 -3.351696 -3.353025 -3.354255 -3.355393 -3.356443 -3.357411
#>  [57] -3.358303 -3.359123 -3.359877 -3.360569 -3.361204 -3.361787 -3.362320
#>  [64] -3.362809 -3.363257 -3.363667 -3.364042 -3.364384 -3.364698 -3.364984
#>  [71] -3.365245 -3.365484 -3.365703 -3.365902 -3.366084 -3.366250 -3.366401
#>  [78] -3.366539 -3.366665 -3.366780 -3.366885 -3.366981 -3.367068 -3.367148
#>  [85] -3.367220 -3.367286 -3.367346 -3.367401 -3.367452 -3.367497 -3.367539
#>  [92] -3.367577 -3.367611 -3.367643 -3.367672 -3.367698 -3.367722 -3.367744
#>  [99] -3.367763 -3.367781
#> 
#> $cvsd
#>   [1] 2.623310 2.581834 2.530074 2.486564 2.454138 2.432476 2.413498 2.399881
#>   [9] 2.388833 2.380883 2.354630 2.324012 2.301098 2.285220 2.277560 2.267701
#>  [17] 2.265007 2.267486 2.270067 2.271657 2.275033 2.281167 2.286147 2.293737
#>  [25] 2.302507 2.311995 2.322008 2.331967 2.343047 2.355252 2.366876 2.378488
#>  [33] 2.389886 2.402007 2.413523 2.423873 2.431627 2.438920 2.445755 2.452142
#>  [41] 2.458094 2.463627 2.469363 2.474896 2.480000 2.484702 2.489028 2.493253
#>  [49] 2.497285 2.500985 2.504377 2.507486 2.510334 2.512942 2.515328 2.517510
#>  [57] 2.519506 2.521330 2.522997 2.524521 2.525912 2.527182 2.528342 2.529401
#>  [65] 2.530368 2.531250 2.532054 2.532788 2.533458 2.534069 2.534626 2.535134
#>  [73] 2.535597 2.536020 2.536405 2.536756 2.537076 2.537368 2.537634 2.537876
#>  [81] 2.538097 2.538299 2.538482 2.538650 2.538802 2.538941 2.539068 2.539183
#>  [89] 2.539289 2.539385 2.539472 2.539552 2.539624 2.539690 2.539750 2.539805
#>  [97] 2.539855 2.539901 2.539943 2.539980
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x5605326657c8>
#> 
#> attr(,"class")
#> [1] "cv_survtran"