Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.5967129  1.4157648 -1.0187559  1.0047907 -0.8351245  1.0413994 -1.1442181 
#>         X8         X9        X10 
#>  0.9215155 -0.9925437  1.1115023 
#> 
#> $coefficients$eta
#>  [1]  0.406914407  0.358763480 -0.023421532  0.000000000  0.188902618
#>  [6]  0.000000000 -0.112200218  0.000000000  0.006101134  0.108903514
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.04378866
#> 
#> $coefficients$formula
#> Surv(time, status) ~ .
#> <environment: 0x563fe9e1e128>
#> 
#> 
#> $lambdas
#>   [1] 3.089203e-01 2.814767e-01 2.564711e-01 2.336869e-01 2.129268e-01
#>   [6] 1.940110e-01 1.767756e-01 1.610713e-01 1.467622e-01 1.337242e-01
#>  [11] 1.218445e-01 1.110202e-01 1.011575e-01 9.217093e-02 8.398272e-02
#>  [16] 7.652192e-02 6.972393e-02 6.352984e-02 5.788603e-02 5.274359e-02
#>  [21] 4.805800e-02 4.378866e-02 3.989859e-02 3.635411e-02 3.312451e-02
#>  [26] 3.018182e-02 2.750055e-02 2.505748e-02 2.283144e-02 2.080316e-02
#>  [31] 1.895506e-02 1.727115e-02 1.573683e-02 1.433881e-02 1.306499e-02
#>  [36] 1.190433e-02 1.084678e-02 9.883186e-03 9.005191e-03 8.205195e-03
#>  [41] 7.476268e-03 6.812097e-03 6.206929e-03 5.655523e-03 5.153102e-03
#>  [46] 4.695314e-03 4.278195e-03 3.898132e-03 3.551833e-03 3.236298e-03
#>  [51] 2.948794e-03 2.686831e-03 2.448141e-03 2.230655e-03 2.032489e-03
#>  [56] 1.851929e-03 1.687408e-03 1.537504e-03 1.400916e-03 1.276463e-03
#>  [61] 1.163065e-03 1.059742e-03 9.655971e-04 8.798162e-04 8.016557e-04
#>  [66] 7.304388e-04 6.655487e-04 6.064231e-04 5.525502e-04 5.034632e-04
#>  [71] 4.587369e-04 4.179840e-04 3.808514e-04 3.470176e-04 3.161895e-04
#>  [76] 2.881001e-04 2.625061e-04 2.391858e-04 2.179372e-04 1.985762e-04
#>  [81] 1.809353e-04 1.648615e-04 1.502156e-04 1.368709e-04 1.247117e-04
#>  [86] 1.136326e-04 1.035378e-04 9.433980e-05 8.595891e-05 7.832256e-05
#>  [91] 7.136460e-05 6.502477e-05 5.924815e-05 5.398470e-05 4.918885e-05
#>  [96] 4.481905e-05 4.083745e-05 3.720956e-05 3.390397e-05 3.089203e-05
#> 
#> $etas
#>              [,1]       [,2]         [,3]         [,4]       [,5]         [,6]
#>   [1,] 0.00000000 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [2,] 0.03624828 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [3,] 0.07013592 0.01117874  0.000000000  0.000000000 0.00000000  0.000000000
#>   [4,] 0.10127198 0.04873376  0.000000000  0.000000000 0.00000000  0.000000000
#>   [5,] 0.13074780 0.08455985  0.000000000  0.000000000 0.00000000  0.000000000
#>   [6,] 0.15850358 0.11853574  0.000000000  0.000000000 0.00000000  0.000000000
#>   [7,] 0.18453640 0.14820136  0.000000000  0.000000000 0.01263273  0.000000000
#>   [8,] 0.20902026 0.17534463  0.000000000  0.000000000 0.02975226  0.000000000
#>   [9,] 0.23195752 0.20099142  0.000000000  0.000000000 0.04573723  0.000000000
#>  [10,] 0.25334681 0.22508481  0.000000000  0.000000000 0.06059224  0.000000000
#>  [11,] 0.27320850 0.24760072  0.000000000  0.000000000 0.07433839  0.000000000
#>  [12,] 0.29104838 0.26730795  0.000000000  0.000000000 0.08698454  0.000000000
#>  [13,] 0.30742198 0.28544282  0.000000000  0.000000000 0.09880956  0.000000000
#>  [14,] 0.32344556 0.29904394  0.000000000  0.000000000 0.11148536  0.000000000
#>  [15,] 0.33897429 0.30865926  0.000000000  0.000000000 0.12460585  0.000000000
#>  [16,] 0.35315395 0.31751803  0.000000000  0.000000000 0.13657439  0.000000000
#>  [17,] 0.36608212 0.32566099  0.000000000  0.000000000 0.14747694  0.000000000
#>  [18,] 0.37785550 0.33313215  0.000000000  0.000000000 0.15739762  0.000000000
#>  [19,] 0.38696352 0.34031102 -0.004056183  0.000000000 0.16640844  0.000000000
#>  [20,] 0.39395077 0.34715218 -0.011050927  0.000000000 0.17460543  0.000000000
#>  [21,] 0.40034020 0.35339429 -0.017374360  0.000000000 0.18206837  0.000000000
#>  [22,] 0.40691441 0.35876348 -0.023421532  0.000000000 0.18890262  0.000000000
#>  [23,] 0.41293999 0.36366961 -0.028901172  0.000000000 0.19515251  0.000000000
#>  [24,] 0.41845722 0.36814777 -0.033867336  0.000000000 0.20086299  0.000000000
#>  [25,] 0.42388839 0.37225049 -0.038005858  0.000000000 0.20636640  0.000000000
#>  [26,] 0.42901071 0.37602180 -0.041570305  0.000000000 0.21152607  0.000000000
#>  [27,] 0.43367512 0.37948009 -0.044778140  0.000000000 0.21623997  0.000000000
#>  [28,] 0.43805284 0.38342144 -0.047780090 -0.001490395 0.22090431 -0.001646759
#>  [29,] 0.44258302 0.39031120 -0.050799861 -0.009684805 0.22685330 -0.008693639
#>  [30,] 0.44674020 0.39671675 -0.053544294 -0.017229967 0.23232630 -0.015166613
#>  [31,] 0.45055192 0.40265572 -0.056038173 -0.024167351 0.23735472 -0.021105148
#>  [32,] 0.45404429 0.40814904 -0.058304207 -0.030537963 0.24196936 -0.026547643
#>  [33,] 0.45724195 0.41321985 -0.060363171 -0.036381812 0.24620003 -0.031531066
#>  [34,] 0.46016806 0.41789277 -0.062234047 -0.041737618 0.25007530 -0.036090746
#>  [35,] 0.46284435 0.42219197 -0.063934046 -0.046641943 0.25362244 -0.040259836
#>  [36,] 0.46529102 0.42614253 -0.065478918 -0.051129898 0.25686713 -0.044069791
#>  [37,] 0.46752690 0.42976876 -0.066882956 -0.055234425 0.25983353 -0.047549946
#>  [38,] 0.46956943 0.43309414 -0.068159139 -0.058986401 0.26254423 -0.050727617
#>  [39,] 0.47143479 0.43614112 -0.069319240 -0.062414612 0.26502029 -0.053628132
#>  [40,] 0.47313793 0.43893143 -0.070373957 -0.065546067 0.26728127 -0.056275059
#>  [41,] 0.47469252 0.44148442 -0.071332965 -0.068405123 0.26934520 -0.058689776
#>  [42,] 0.47611125 0.44381940 -0.072205069 -0.071014956 0.27122881 -0.060892331
#>  [43,] 0.47740578 0.44595400 -0.072998244 -0.073396733 0.27294747 -0.062901033
#>  [44,] 0.47858679 0.44790467 -0.073719720 -0.075569942 0.27451535 -0.064732692
#>  [45,] 0.47966408 0.44968664 -0.074376055 -0.077552503 0.27594545 -0.066402718
#>  [46,] 0.48064665 0.45131401 -0.074973199 -0.079360870 0.27724969 -0.067925225
#>  [47,] 0.48154273 0.45279981 -0.075516549 -0.081010134 0.27843902 -0.069313129
#>  [48,] 0.48235985 0.45415606 -0.076011000 -0.082514124 0.27952345 -0.070578242
#>  [49,] 0.48310491 0.45539380 -0.076460997 -0.083885502 0.28051214 -0.071731360
#>  [50,] 0.48378421 0.45652319 -0.076870573 -0.085135856 0.28141349 -0.072782346
#>  [51,] 0.48440352 0.45755356 -0.077243392 -0.086275781 0.28223516 -0.073740205
#>  [52,] 0.48496810 0.45849346 -0.077582778 -0.087314963 0.28298414 -0.074613158
#>  [53,] 0.48548276 0.45935073 -0.077891752 -0.088262253 0.28366684 -0.075408708
#>  [54,] 0.48595190 0.46013256 -0.078173060 -0.089125733 0.28428909 -0.076133699
#>  [55,] 0.48637952 0.46084551 -0.078429194 -0.089912784 0.28485623 -0.076794375
#>  [56,] 0.48676928 0.46149561 -0.078662421 -0.090630143 0.28537311 -0.077396429
#>  [57,] 0.48712453 0.46208834 -0.078874802 -0.091283960 0.28584419 -0.077945056
#>  [58,] 0.48744830 0.46262874 -0.079068209 -0.091879845 0.28627350 -0.078444989
#>  [59,] 0.48774339 0.46312139 -0.079244345 -0.092422917 0.28666475 -0.078900545
#>  [60,] 0.48801232 0.46357050 -0.079404760 -0.092917845 0.28702129 -0.079315659
#>  [61,] 0.48825741 0.46397988 -0.079550862 -0.093368888 0.28734621 -0.079693918
#>  [62,] 0.48848076 0.46435305 -0.079683934 -0.093779930 0.28764230 -0.080038592
#>  [63,] 0.48868431 0.46469318 -0.079805141 -0.094154511 0.28791212 -0.080352660
#>  [64,] 0.48886981 0.46500320 -0.079915545 -0.094495861 0.28815800 -0.080638839
#>  [65,] 0.48903884 0.46528576 -0.080016111 -0.094806925 0.28838205 -0.080899604
#>  [66,] 0.48919288 0.46554328 -0.080107719 -0.095090385 0.28858622 -0.081137210
#>  [67,] 0.48933326 0.46577799 -0.080191168 -0.095348689 0.28877226 -0.081353715
#>  [68,] 0.48946117 0.46599189 -0.080267186 -0.095584067 0.28894179 -0.081550990
#>  [69,] 0.48957773 0.46618682 -0.080336437 -0.095798551 0.28909627 -0.081730745
#>  [70,] 0.48968395 0.46636447 -0.080399524 -0.095993996 0.28923703 -0.081894533
#>  [71,] 0.48978073 0.46652637 -0.080456997 -0.096172090 0.28936529 -0.082043774
#>  [72,] 0.48986893 0.46667390 -0.080509356 -0.096334372 0.28948216 -0.082179758
#>  [73,] 0.48994929 0.46680834 -0.080557057 -0.096482246 0.28958866 -0.082303664
#>  [74,] 0.49002252 0.46693086 -0.080600514 -0.096616989 0.28968570 -0.082416564
#>  [75,] 0.49008925 0.46704250 -0.080640106 -0.096739768 0.28977412 -0.082519435
#>  [76,] 0.49015005 0.46714424 -0.080676177 -0.096851643 0.28985469 -0.082613169
#>  [77,] 0.49020546 0.46723695 -0.080709041 -0.096953584 0.28992810 -0.082698576
#>  [78,] 0.49025594 0.46732143 -0.080738982 -0.097046472 0.28999499 -0.082776396
#>  [79,] 0.49030194 0.46739840 -0.080766261 -0.097131111 0.29005594 -0.082847304
#>  [80,] 0.49034386 0.46746855 -0.080791115 -0.097208232 0.29011148 -0.082911913
#>  [81,] 0.49038205 0.46753247 -0.080813759 -0.097278504 0.29016209 -0.082970782
#>  [82,] 0.49041685 0.46759071 -0.080834390 -0.097342535 0.29020820 -0.083024422
#>  [83,] 0.49044856 0.46764378 -0.080853188 -0.097400879 0.29025021 -0.083073297
#>  [84,] 0.49047745 0.46769214 -0.080870314 -0.097454040 0.29028849 -0.083117830
#>  [85,] 0.49050378 0.46773621 -0.080885919 -0.097502480 0.29032338 -0.083158408
#>  [86,] 0.49052777 0.46777636 -0.080900136 -0.097546617 0.29035516 -0.083195380
#>  [87,] 0.49054963 0.46781294 -0.080913090 -0.097586834 0.29038412 -0.083229068
#>  [88,] 0.49056954 0.46784628 -0.080924893 -0.097623478 0.29041051 -0.083259764
#>  [89,] 0.49058769 0.46787666 -0.080935647 -0.097656867 0.29043455 -0.083287732
#>  [90,] 0.49060423 0.46790433 -0.080945445 -0.097687291 0.29045646 -0.083313217
#>  [91,] 0.49061929 0.46792955 -0.080954373 -0.097715012 0.29047642 -0.083336437
#>  [92,] 0.49063302 0.46795253 -0.080962507 -0.097740270 0.29049461 -0.083357594
#>  [93,] 0.49064553 0.46797347 -0.080969919 -0.097763285 0.29051118 -0.083376872
#>  [94,] 0.49065693 0.46799255 -0.080976672 -0.097784256 0.29052628 -0.083394437
#>  [95,] 0.49066731 0.46800994 -0.080982825 -0.097803363 0.29054004 -0.083410442
#>  [96,] 0.49067678 0.46802578 -0.080988431 -0.097820774 0.29055258 -0.083425025
#>  [97,] 0.49068540 0.46804021 -0.080993539 -0.097836637 0.29056400 -0.083438313
#>  [98,] 0.49069326 0.46805336 -0.080998193 -0.097851092 0.29057441 -0.083450420
#>  [99,] 0.49070041 0.46806535 -0.081002434 -0.097864262 0.29058390 -0.083461451
#> [100,] 0.49070694 0.46807626 -0.081006298 -0.097876263 0.29059254 -0.083471503
#>               [,7]        [,8]         [,9]       [,10]
#>   [1,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [2,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [3,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [4,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [5,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [6,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [7,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [8,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [9,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [10,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [11,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [12,] -0.01129958 0.000000000 0.000000e+00 0.000000000
#>  [13,] -0.02388400 0.000000000 0.000000e+00 0.000000000
#>  [14,] -0.03676659 0.000000000 0.000000e+00 0.009760832
#>  [15,] -0.04977201 0.000000000 0.000000e+00 0.027188121
#>  [16,] -0.06170267 0.000000000 0.000000e+00 0.042706138
#>  [17,] -0.07261938 0.000000000 0.000000e+00 0.056535485
#>  [18,] -0.08258808 0.000000000 0.000000e+00 0.068872218
#>  [19,] -0.09134697 0.000000000 0.000000e+00 0.080328257
#>  [20,] -0.09903496 0.000000000 0.000000e+00 0.090927119
#>  [21,] -0.10601053 0.000000000 7.599198e-05 0.100413113
#>  [22,] -0.11220022 0.000000000 6.101134e-03 0.108903514
#>  [23,] -0.11781892 0.000000000 1.162766e-02 0.116523092
#>  [24,] -0.12291656 0.000000000 1.669034e-02 0.123368871
#>  [25,] -0.12860035 0.005384781 2.134493e-02 0.130745463
#>  [26,] -0.13427265 0.012695353 2.562583e-02 0.137925792
#>  [27,] -0.13945767 0.019334382 2.955226e-02 0.144389426
#>  [28,] -0.14412544 0.025279068 3.357612e-02 0.149913105
#>  [29,] -0.14828616 0.030423101 3.899614e-02 0.153529208
#>  [30,] -0.15213583 0.035123373 4.400875e-02 0.156741761
#>  [31,] -0.15568949 0.039415750 4.863619e-02 0.159600468
#>  [32,] -0.15896348 0.043333752 5.290126e-02 0.162148397
#>  [33,] -0.16197475 0.046908581 5.682686e-02 0.164422867
#>  [34,] -0.16474039 0.050169136 6.043564e-02 0.166456154
#>  [35,] -0.16727756 0.053142331 6.374959e-02 0.168276657
#>  [36,] -0.16960268 0.055852815 6.678999e-02 0.169908741
#>  [37,] -0.17173163 0.058323343 6.957717e-02 0.171373762
#>  [38,] -0.17367955 0.060574819 7.213042e-02 0.172690381
#>  [39,] -0.17546074 0.062626434 7.446792e-02 0.173874963
#>  [40,] -0.17708856 0.064495718 7.660681e-02 0.174941717
#>  [41,] -0.17857570 0.066198836 7.856292e-02 0.175903538
#>  [42,] -0.17993374 0.067750439 8.035118e-02 0.176771404
#>  [43,] -0.18117349 0.069163954 8.198540e-02 0.177555136
#>  [44,] -0.18230497 0.070451635 8.347837e-02 0.178263427
#>  [45,] -0.18333739 0.071624665 8.484192e-02 0.178903985
#>  [46,] -0.18427924 0.072693241 8.608694e-02 0.179483659
#>  [47,] -0.18513832 0.073666661 8.722350e-02 0.180008545
#>  [48,] -0.18592180 0.074553398 8.826085e-02 0.180484077
#>  [49,] -0.18663625 0.075361174 8.920746e-02 0.180915109
#>  [50,] -0.18728768 0.076097026 9.007116e-02 0.181305982
#>  [51,] -0.18788159 0.076767362 9.085910e-02 0.181660586
#>  [52,] -0.18842303 0.077378021 9.157782e-02 0.181982408
#>  [53,] -0.18891660 0.077934322 9.223335e-02 0.182274579
#>  [54,] -0.18936650 0.078441109 9.283118e-02 0.182539917
#>  [55,] -0.18977658 0.078902794 9.337633e-02 0.182780954
#>  [56,] -0.19015035 0.079323395 9.387343e-02 0.182999975
#>  [57,] -0.19049101 0.079706572 9.432665e-02 0.183199039
#>  [58,] -0.19080148 0.080055659 9.473987e-02 0.183380004
#>  [59,] -0.19108443 0.080373692 9.511657e-02 0.183544549
#>  [60,] -0.19134230 0.080663436 9.545998e-02 0.183694192
#>  [61,] -0.19157730 0.080927410 9.577302e-02 0.183830304
#>  [62,] -0.19179145 0.081167907 9.605837e-02 0.183954128
#>  [63,] -0.19198661 0.081387019 9.631846e-02 0.184066790
#>  [64,] -0.19216445 0.081586647 9.655552e-02 0.184169308
#>  [65,] -0.19232651 0.081768525 9.677159e-02 0.184262607
#>  [66,] -0.19247419 0.081934234 9.696852e-02 0.184347525
#>  [67,] -0.19260876 0.082085210 9.714800e-02 0.184424822
#>  [68,] -0.19273138 0.082222766 9.731156e-02 0.184495188
#>  [69,] -0.19284312 0.082348094 9.746063e-02 0.184559250
#>  [70,] -0.19294494 0.082462282 9.759648e-02 0.184617577
#>  [71,] -0.19303772 0.082566321 9.772029e-02 0.184670686
#>  [72,] -0.19312226 0.082661112 9.783311e-02 0.184719047
#>  [73,] -0.19319929 0.082747480 9.793592e-02 0.184763086
#>  [74,] -0.19326949 0.082826171 9.802962e-02 0.184803192
#>  [75,] -0.19333345 0.082897870 9.811499e-02 0.184839718
#>  [76,] -0.19339173 0.082963197 9.819280e-02 0.184872985
#>  [77,] -0.19344483 0.083022718 9.826369e-02 0.184903284
#>  [78,] -0.19349322 0.083076951 9.832830e-02 0.184930882
#>  [79,] -0.19353731 0.083126364 9.838717e-02 0.184956020
#>  [80,] -0.19357749 0.083171387 9.844081e-02 0.184978918
#>  [81,] -0.19361410 0.083212409 9.848969e-02 0.184999776
#>  [82,] -0.19364745 0.083249786 9.853423e-02 0.185018776
#>  [83,] -0.19367784 0.083283842 9.857482e-02 0.185036084
#>  [84,] -0.19370554 0.083314872 9.861180e-02 0.185051852
#>  [85,] -0.19373077 0.083343145 9.864550e-02 0.185066216
#>  [86,] -0.19375376 0.083368906 9.867621e-02 0.185079302
#>  [87,] -0.19377471 0.083392379 9.870419e-02 0.185091224
#>  [88,] -0.19379380 0.083413765 9.872968e-02 0.185102085
#>  [89,] -0.19381119 0.083433252 9.875291e-02 0.185111979
#>  [90,] -0.19382704 0.083451008 9.877408e-02 0.185120994
#>  [91,] -0.19384148 0.083467185 9.879337e-02 0.185129207
#>  [92,] -0.19385464 0.083481926 9.881094e-02 0.185136690
#>  [93,] -0.19386663 0.083495357 9.882696e-02 0.185143507
#>  [94,] -0.19387755 0.083507595 9.884155e-02 0.185149718
#>  [95,] -0.19388751 0.083518745 9.885484e-02 0.185155377
#>  [96,] -0.19389657 0.083528905 9.886696e-02 0.185160533
#>  [97,] -0.19390484 0.083538163 9.887800e-02 0.185165231
#>  [98,] -0.19391237 0.083546597 9.888805e-02 0.185169511
#>  [99,] -0.19391923 0.083554283 9.889722e-02 0.185173410
#> [100,] -0.19392548 0.083561286 9.890557e-02 0.185176963
#> 
#> $cvm
#>   [1] -4.154367 -4.116567 -4.015191 -3.884786 -3.765069 -3.634219 -3.512147
#>   [8] -3.415559 -3.333020 -3.263064 -3.205511 -3.161050 -3.124059 -3.085829
#>  [15] -3.039278 -3.002149 -2.966934 -2.942062 -2.928523 -2.920212 -2.916900
#>  [22] -2.915002 -2.918004 -2.922682 -2.929827 -2.939748 -2.950581 -2.960736
#>  [29] -2.969807 -2.978898 -2.986446 -2.994412 -3.000971 -3.006399 -3.012022
#>  [36] -3.017707 -3.022621 -3.026099 -3.029489 -3.032588 -3.035129 -3.037620
#>  [43] -3.040035 -3.042357 -3.044574 -3.046678 -3.048665 -3.050533 -3.052284
#>  [50] -3.053919 -3.055442 -3.056858 -3.058171 -3.059386 -3.060509 -3.061546
#>  [57] -3.062501 -3.063381 -3.064190 -3.064934 -3.065616 -3.066243 -3.066817
#>  [64] -3.067343 -3.067825 -3.068267 -3.068671 -3.069040 -3.069378 -3.069686
#>  [71] -3.069968 -3.070226 -3.070461 -3.070676 -3.070873 -3.071052 -3.071215
#>  [78] -3.071364 -3.071501 -3.071625 -3.071738 -3.071841 -3.071935 -3.072021
#>  [85] -3.072100 -3.072171 -3.072236 -3.072296 -3.072350 -3.072399 -3.072444
#>  [92] -3.072485 -3.072522 -3.072557 -3.072588 -3.072616 -3.072642 -3.072665
#>  [99] -3.072687 -3.072706
#> 
#> $cvsd
#>   [1] 4.058923 4.066469 4.025321 3.987233 3.934558 3.858358 3.788383 3.728663
#>   [9] 3.676526 3.633581 3.599495 3.574432 3.556153 3.540649 3.511262 3.485910
#>  [17] 3.461814 3.443909 3.432815 3.425202 3.418012 3.408240 3.403659 3.401199
#>  [25] 3.399162 3.397382 3.397904 3.400297 3.401837 3.404507 3.406938 3.409835
#>  [33] 3.411410 3.412146 3.413259 3.414631 3.416229 3.420320 3.424411 3.428554
#>  [41] 3.432959 3.437088 3.440944 3.444534 3.447868 3.450958 3.453815 3.456452
#>  [49] 3.458884 3.461123 3.463183 3.465075 3.466813 3.468407 3.469868 3.471207
#>  [57] 3.472432 3.473554 3.474581 3.475519 3.476377 3.477162 3.477878 3.478533
#>  [65] 3.479130 3.479676 3.480174 3.480629 3.481043 3.481422 3.481767 3.482082
#>  [73] 3.482370 3.482632 3.482871 3.483088 3.483287 3.483468 3.483633 3.483784
#>  [81] 3.483921 3.484046 3.484160 3.484264 3.484359 3.484445 3.484524 3.484596
#>  [89] 3.484661 3.484721 3.484775 3.484824 3.484870 3.484911 3.484948 3.484982
#>  [97] 3.485013 3.485042 3.485067 3.485091
#> 
#> attr(,"class")
#> [1] "cv_survtran"