Simulated Group Survival Data: Homogeneous Baseline Hazard Function and Heterogeneous Effects

A dataset containing 2 groups of survival data, where the baseline hazards are homogeneous and the effects are heterogeneous. The first group, also called the source group, contains 900 individuals, while the second group, also called the target group, contains 100 individuals. The survival time of group \(i\) is generated according to the following models: $$\lambda^{(i)}(t)=\lambda_{0}(t)\exp\left(x^\top(\beta+\eta^{(i)}) \right),$$ where \(\lambda_{0}(t)\) represents the baseline hazard, \(\beta+\eta^{(i)}\) represents the effects of group \(i\), and \(x\) represents the covariates. The covariates \(x=(x_{1},\ldots,x_{10})\) are generated from a multivariate normal distribution with mean 0 and covariance matrix \(\Sigma=I_{10}\). The baseline hazard \(\lambda_{0}(t)=t^2\), while the effects $$\eta^{(i)}=\left\{\begin{array}{ll}(0,\ldots,0),&i=1,\\ (0.5,0,\ldots,0),&i=2.\end{array}\right.$$ Additionally, \(\beta=(1,\ldots,1)\). The maximum censoring time is fixed at 3, and the censoring rate is approximately 20%.

Format

A data frame with 1000 rows and 14 variables:

id

Individual identifier, 1-1000.

group

Group indicator, 1=Source, 2=Target.

time

Survival time.

status

Status indicator, 0=censored, 1=event.

X1

covariate 1.

X2

covariate 2.

X3

covariate 3.

X4

covariate 4.

X5

covariate 5.

X6

covariate 6.

X7

covariate 7.

X8

covariate 8.

X9

covariate 9.

X10

covariate 10.

Source

../articles/simulate-data.html