Multiple-sources Cox proportional hazards model with group sparsity and local-global transfer penalties

Multiple-sources Cox proportional hazards model with group sparsity and local-global transfer penalties

Usage

coxens(
  formula,
  data,
  group,
  lambda1 = 0,
  lambda2 = 0,
  lambda3 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 2,
  init,
  control,
  ...
)

Arguments

formula

a formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

a data frame containing the variables in the model.

group

a factor specifying the group of each sample.

lambda1

a non-negative value specifying the sparse penalty parameter. The default is 0.

lambda2

a non-negative value specifying the global biased penalty parameter. The default is 0.

lambda3

a non-negative value specifying the local biased penalty parameter. The default is 0.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3.0 for MCP.

rho

a value in (2, 10) specifying the expansion factor of the augmented Lagrangian's penalty parameter. The default is 2.0.

init

a numeric vector specifying the initial value of the coefficients. The default is a zero vector.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a coxens object.

Examples

library(survtrans)
formula <- Surv(time, status) ~ . - group - id
fit <- coxens(
  formula, sim2, sim2$group,
  lambda1 = 0.03, lambda2 = 0.04, lambda3 = 0.01, penalty = "SCAD"
)
summary(fit)
#> Call:
#> coxens(formula = formula, data = sim2, group = sim2$group, lambda1 = 0.03, 
#>     lambda2 = 0.04, lambda3 = 0.01, penalty = "SCAD")
#> 
#>   n=500, number of events=422
#> 
#>               coef exp(coef) se(coef)      z Pr(>|z|)    
#> X1 (1)     0.34772   1.41583  0.05130  6.778 1.22e-11 ***
#> X1 (2, 4)  0.94780   2.58004  0.08575 11.053  < 2e-16 ***
#> X1 (3, 5) -0.25268   0.77672  0.06887 -3.669 0.000244 ***
#> X2 (1)     0.36087   1.43458  0.05369  6.721 1.81e-11 ***
#> X2 (2, 4)  0.96455   2.62360  0.08449 11.416  < 2e-16 ***
#> X2 (3, 5) -0.24283   0.78441  0.08020 -3.028 0.002462 ** 
#> X3 (ALL)   0.34498   1.41196  0.05541  6.226 4.78e-10 ***
#> X4 (ALL)   0.32757   1.38760  0.05335  6.140 8.23e-10 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>           exp(coef) exp(-coef) lower .95 upper .95
#> X1 (1)    1.4158    0.7063     1.2804    1.5656   
#> X1 (2, 4) 2.5800    0.3876     2.1809    3.0522   
#> X1 (3, 5) 0.7767    1.2875     0.6786    0.8890   
#> X2 (1)    1.4346    0.6971     1.2913    1.5938   
#> X2 (2, 4) 2.6236    0.3812     2.2232    3.0961   
#> X2 (3, 5) 0.7844    1.2749     0.6703    0.9179   
#> X3 (ALL)  1.4120    0.7082     1.2667    1.5739   
#> X4 (ALL)  1.3876    0.7207     1.2498    1.5405