Transfer Learning of Multi-source for Cox proportional hazards model

Transfer Learning of Multi-source for Cox proportional hazards model

Usage

coxtrans(
  formula,
  data,
  group,
  lambda1 = 0,
  lambda2 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 1.1,
  init,
  control,
  ...
)

Arguments

formula

a formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

a data frame containing the variables in the model.

group

a factor specifying the group of each sample.

lambda1

a non-negative value specifying the sparisity penalty parameter. The default is 0.

lambda2

a non-negative value specifying the group penalty parameter. The default is 0.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3.0 for MCP.

rho

a value in (2, 10) specifying the expansion factor of the augmented Lagrangian's penalty parameter. The default is 1.1.

init

a numeric vector specifying the initial value of the coefficients. The default is a zero vector.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a coxtrans object.

Examples

library(survtrans)
formula <- Surv(time, status) ~ . - group - id
group <- as.factor(sim2$group)
fit <- coxtrans(
  formula, sim2, group,
  lambda1 = 0.1, lambda2 = 0.1, penalty = "SCAD"
)
fit$coefficients
#>             [,1]          [,2]          [,3]         [,4]         [,5]
#> X1   0.000000000  0.0000000000  0.0000000000 -0.007559374  0.000000000
#> X2  -0.011920836  0.0052767105 -0.0085238122  0.008931251 -0.009581714
#> X3   0.000000000  0.0000000000  0.0000000000  0.000000000  0.000000000
#> X4  -0.007699957  0.0077475692 -0.0031367921  0.006295944 -0.007441672
#> X5   0.000000000  0.0000000000  0.0006947136  0.000000000  0.000000000
#> X6  -0.007969559  0.0096274323 -0.0027648749  0.010325169 -0.008259357
#> X7   0.000000000 -0.0007547586 -0.0030502378  0.000000000  0.000000000
#> X8  -0.006330619  0.0099703864 -0.0151433092  0.010969689 -0.010091106
#> X9   0.000000000  0.0000000000  0.0000000000  0.000000000  0.003765106
#> X10 -0.001471112  0.0070021294 -0.0111120129  0.008168972 -0.008718395
#>            [,6]          [,7]         [,8]         [,9]         [,10]
#> X1  0.004858468  0.0000000000 0.0000000000  0.000000000  0.0000000000
#> X2  0.004328498 -0.0046791766 0.0071771516 -0.003316812  0.0132194579
#> X3  0.000000000  0.0000000000 0.0007330034 -0.001816284  0.0000000000
#> X4  0.005004995 -0.0003392516 0.0067649109 -0.009185526  0.0000000000
#> X5  0.000000000  0.0000000000 0.0000000000  0.000000000 -0.0092521690
#> X6  0.010341288 -0.0051134690 0.0041857691 -0.017654829  0.0060034635
#> X7  0.000000000  0.0000000000 0.0000000000  0.000000000  0.0004208664
#> X8  0.013561582 -0.0095313162 0.0078081213 -0.009453181  0.0080466906
#> X9  0.000000000  0.0000000000 0.0000000000  0.000000000  0.0000000000
#> X10 0.000000000 -0.0052747113 0.0054296801 -0.002677745  0.0120539158
#>           [,11] [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#> X1   0.48556938     0     0     0     0     0     0     0     0     0     0
#> X2  -0.05272298     0     0     0     0     0     0     0     0     0     0
#> X3   0.38786180     0     0     0     0     0     0     0     0     0     0
#> X4   0.07208634     0     0     0     0     0     0     0     0     0     0
#> X5   0.36666275     0     0     0     0     0     0     0     0     0     0
#> X6  -0.03162015     0     0     0     0     0     0     0     0     0     0
#> X7   0.39774641     0     0     0     0     0     0     0     0     0     0
#> X8  -0.01985960     0     0     0     0     0     0     0     0     0     0
#> X9   0.34870039     0     0     0     0     0     0     0     0     0     0
#> X10  0.07548076     0     0     0     0     0     0     0     0     0     0
#>     [,22] [,23] [,24] [,25] [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33]
#> X1      0     0     0     0     0     0     0     0     0     0     0     0
#> X2      0     0     0     0     0     0     0     0     0     0     0     0
#> X3      0     0     0     0     0     0     0     0     0     0     0     0
#> X4      0     0     0     0     0     0     0     0     0     0     0     0
#> X5      0     0     0     0     0     0     0     0     0     0     0     0
#> X6      0     0     0     0     0     0     0     0     0     0     0     0
#> X7      0     0     0     0     0     0     0     0     0     0     0     0
#> X8      0     0     0     0     0     0     0     0     0     0     0     0
#> X9      0     0     0     0     0     0     0     0     0     0     0     0
#> X10     0     0     0     0     0     0     0     0     0     0     0     0
#>     [,34] [,35] [,36] [,37] [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45]
#> X1      0     0     0     0     0     0     0     0     0     0     0     0
#> X2      0     0     0     0     0     0     0     0     0     0     0     0
#> X3      0     0     0     0     0     0     0     0     0     0     0     0
#> X4      0     0     0     0     0     0     0     0     0     0     0     0
#> X5      0     0     0     0     0     0     0     0     0     0     0     0
#> X6      0     0     0     0     0     0     0     0     0     0     0     0
#> X7      0     0     0     0     0     0     0     0     0     0     0     0
#> X8      0     0     0     0     0     0     0     0     0     0     0     0
#> X9      0     0     0     0     0     0     0     0     0     0     0     0
#> X10     0     0     0     0     0     0     0     0     0     0     0     0
#>     [,46] [,47] [,48] [,49] [,50] [,51] [,52] [,53] [,54] [,55] [,56]
#> X1      0     0     0     0     0     0     0     0     0     0     0
#> X2      0     0     0     0     0     0     0     0     0     0     0
#> X3      0     0     0     0     0     0     0     0     0     0     0
#> X4      0     0     0     0     0     0     0     0     0     0     0
#> X5      0     0     0     0     0     0     0     0     0     0     0
#> X6      0     0     0     0     0     0     0     0     0     0     0
#> X7      0     0     0     0     0     0     0     0     0     0     0
#> X8      0     0     0     0     0     0     0     0     0     0     0
#> X9      0     0     0     0     0     0     0     0     0     0     0
#> X10     0     0     0     0     0     0     0     0     0     0     0
#>             [,57]         [,58]         [,59]         [,60]         [,61]
#> X1  -5.363275e-03  0.0000000000  0.0000000000  0.0101325080  0.0031610099
#> X2   8.745740e-05 -0.0549188974 -0.0112215516 -0.0680631714 -0.0079302841
#> X3  -1.083281e-03  0.0000000000  0.0000000000  0.0000000000  0.0000000000
#> X4   5.932050e-03 -0.0500327337 -0.0155550854 -0.0398077477  0.0047500563
#> X5  -1.358627e-02  0.0000000000 -0.0006947136  0.0000000000  0.0000000000
#> X6  -5.812080e-04 -0.0472227021 -0.0035558117 -0.0453629764 -0.0079885894
#> X7  -2.081915e-03  0.0007547586  0.0048915775 -0.0009796642  0.0020533325
#> X8   8.940762e-05 -0.0416550495  0.0156855604 -0.0389417695  0.0009380032
#> X9   7.465367e-03  0.0000000000  0.0000000000  0.0000000000 -0.0074653672
#> X10  2.487213e-03 -0.0200988777  0.0252008807 -0.0338975071  0.0230631097
#>            [,62]        [,63]         [,64]        [,65]       [,66]
#> X1  -0.007930243  0.000000000  0.0000000000  0.000000000  0.00000000
#> X2  -0.048089223 -0.023002875 -0.0497584023 -0.026037629 -0.06925633
#> X3   0.000000000  0.000000000 -0.0007330034  0.001816284  0.00000000
#> X4  -0.033475564 -0.015407333 -0.0332883256  0.008657786 -0.01869188
#> X5   0.000000000 -0.001163621  0.0000000000  0.000000000  0.01544461
#> X6  -0.053406897 -0.009694668 -0.0294076093  0.017422149 -0.04597006
#> X7   0.000000000  0.000000000  0.0000000000  0.000000000 -0.00463809
#> X8  -0.050069386  0.010907934 -0.0360062101  0.016791312 -0.03571865
#> X9   0.000000000  0.000000000  0.0000000000  0.000000000  0.00000000
#> X10 -0.008176654  0.008789583 -0.0233183496  0.010593986 -0.02170313
#>             [,67]        [,68]        [,69]         [,70]         [,71]
#> X1   0.0000000000  0.010132508  0.003161010 -0.0079302432  0.0000000000
#> X2   0.0436973458 -0.013144274  0.046988613  0.0068296741  0.0319160219
#> X3   0.0000000000  0.000000000  0.000000000  0.0000000000  0.0000000000
#> X4   0.0344776483  0.010224986  0.054782790  0.0165571698  0.0346254003
#> X5  -0.0006947136  0.000000000  0.000000000  0.0000000000 -0.0011636209
#> X6   0.0436668904  0.001859726  0.039234113 -0.0061841952  0.0375280344
#> X7   0.0041368188 -0.001734423  0.001298574 -0.0007547586 -0.0007547586
#> X8   0.0573406099  0.002713280  0.042593053 -0.0084143363  0.0525629835
#> X9   0.0000000000  0.000000000 -0.007465367  0.0000000000  0.0000000000
#> X10  0.0452997583 -0.013798629  0.043161987  0.0119222235  0.0288884611
#>             [,72]         [,73]        [,74]         [,75]         [,76]
#> X1   0.0000000000  0.0000000000  0.000000000  0.0101325080  0.0031610099
#> X2   0.0051604951  0.0288812688 -0.014337437 -0.0568416197  0.0032912675
#> X3  -0.0007330034  0.0018162843  0.000000000  0.0000000000  0.0000000000
#> X4   0.0167444081  0.0586905195  0.031340856 -0.0242526623  0.0203051417
#> X5   0.0000000000  0.0000000000  0.015444608  0.0006947136  0.0006947136
#> X6   0.0178150928  0.0646448513  0.001252645 -0.0418071647 -0.0044327777
#> X7  -0.0007547586 -0.0007547586 -0.005392848 -0.0058712417 -0.0028382449
#> X8   0.0056488395  0.0584463620  0.005936396 -0.0546273299 -0.0147475572
#> X9   0.0000000000  0.0000000000  0.000000000  0.0000000000 -0.0074653672
#> X10 -0.0032194719  0.0306928637 -0.001604249 -0.0590983878 -0.0021377709
#>             [,77]         [,78]         [,79]         [,80]        [,81]
#> X1  -0.0079302432  0.0000000000  0.0000000000  0.0000000000  0.000000000
#> X2  -0.0368676716 -0.0117813238 -0.0385368507 -0.0148160769 -0.058034783
#> X3   0.0000000000  0.0000000000 -0.0007330034  0.0018162843  0.000000000
#> X4  -0.0179204785  0.0001477520 -0.0177332402  0.0242128712 -0.003136792
#> X5   0.0006947136 -0.0004689073  0.0006947136  0.0006947136  0.016139321
#> X6  -0.0498510856 -0.0061388560 -0.0258517977  0.0209779609 -0.042414245
#> X7  -0.0048915775 -0.0048915775 -0.0048915775 -0.0048915775 -0.009529667
#> X8  -0.0657549463 -0.0047776264 -0.0516917705  0.0011057520 -0.051404214
#> X9   0.0000000000  0.0000000000  0.0000000000  0.0000000000  0.000000000
#> X10 -0.0333775348 -0.0164112972 -0.0485192302 -0.0146068946 -0.046904008
#>            [,82]         [,83]         [,84]         [,85]         [,86]
#> X1  -0.006971498 -0.0180627512 -0.0101325080 -0.0101325080 -0.0101325080
#> X2   0.060132887  0.0199739481  0.0450602959  0.0183047690  0.0420255428
#> X3   0.000000000  0.0000000000  0.0000000000 -0.0007330034  0.0018162843
#> X4   0.044557804  0.0063321837  0.0244004142  0.0065194221  0.0484655335
#> X5   0.000000000  0.0000000000 -0.0011636209  0.0000000000  0.0000000000
#> X6   0.037374387 -0.0080439209  0.0356683087  0.0159553670  0.0627851256
#> X7   0.003032997  0.0009796642  0.0009796642  0.0009796642  0.0009796642
#> X8   0.039879773 -0.0111276164  0.0498497035  0.0029355595  0.0557330820
#> X9  -0.007465367  0.0000000000  0.0000000000  0.0000000000  0.0000000000
#> X10  0.056960617  0.0257208530  0.0426870906  0.0105791575  0.0444914932
#>             [,87]        [,88]        [,89]         [,90]        [,91]
#> X1  -0.0101325080 -0.011091253 -0.003161010 -0.0031610099 -0.003161010
#> X2  -0.0011931631 -0.040158939 -0.015072591 -0.0418281182 -0.018107344
#> X3   0.0000000000  0.000000000  0.000000000 -0.0007330034  0.001816284
#> X4   0.0211158702 -0.038225620 -0.020157390 -0.0380383819  0.003907729
#> X5   0.0154446076  0.000000000 -0.001163621  0.0000000000  0.000000000
#> X6  -0.0006070807 -0.045418308 -0.001706078 -0.0214190200  0.025410739
#> X7  -0.0036584255 -0.002053333 -0.002053333 -0.0020533325 -0.002053333
#> X8   0.0032231164 -0.051007389  0.009969931 -0.0369442133  0.015853309
#> X9   0.0000000000  0.007465367  0.007465367  0.0074653672  0.007465367
#> X10  0.0121943800 -0.031239764 -0.014273526 -0.0463814593 -0.012469124
#>            [,92]        [,93]         [,94]       [,95]        [,96]
#> X1  -0.003161010  0.007930243  0.0079302432 0.007930243  0.007930243
#> X2  -0.061326050  0.025086348 -0.0016691791 0.022051595 -0.021167111
#> X3   0.000000000  0.000000000 -0.0007330034 0.001816284  0.000000000
#> X4  -0.023441934  0.018068231  0.0001872383 0.042133350  0.014783686
#> X5   0.015444608 -0.001163621  0.0000000000 0.000000000  0.015444608
#> X6  -0.037981468  0.043712230  0.0239992879 0.070829047  0.007436840
#> X7  -0.006691422  0.000000000  0.0000000000 0.000000000 -0.004638090
#> X8  -0.036656656  0.060977320  0.0140631758 0.066860698  0.014350733
#> X9   0.007465367  0.000000000  0.0000000000 0.000000000  0.000000000
#> X10 -0.044766237  0.016966238 -0.0151416955 0.018770640 -0.013526473
#>             [,97]        [,98]        [,99]      [,100]        [,101]
#> X1   0.0000000000  0.000000000  0.000000000 0.000000000  0.0000000000
#> X2  -0.0267555268 -0.003034753 -0.046253459 0.023720774 -0.0194979322
#> X3  -0.0007330034  0.001816284  0.000000000 0.002549288  0.0007330034
#> X4  -0.0178809922  0.024065119 -0.003284544 0.041946111  0.0145964481
#> X5   0.0011636209  0.001163621  0.016608228 0.000000000  0.0154446076
#> X6  -0.0197129417  0.027116817 -0.036275389 0.046829759 -0.0165624477
#> X7   0.0000000000  0.000000000 -0.004638090 0.000000000 -0.0046380897
#> X8  -0.0469141441  0.005883378 -0.046626587 0.052797523  0.0002875569
#> X9   0.0000000000  0.000000000  0.000000000 0.000000000  0.0000000000
#> X10 -0.0321079330  0.001804403 -0.030492711 0.033912336  0.0016152224
#>           [,102]
#> X1   0.000000000
#> X2  -0.043218706
#> X3  -0.001816284
#> X4  -0.027349663
#> X5   0.015444608
#> X6  -0.063392206
#> X7  -0.004638090
#> X8  -0.052509966
#> X9   0.000000000
#> X10 -0.032297113