Transfer Learning for Cox Model with Global and Local Shrinkage

Transfer Learning for Cox Model with Global and Local Shrinkage

Usage

coxtrans(
  formula,
  data,
  group,
  lambda1 = 0,
  lambda2 = 0,
  lambda3 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 2,
  tau = 10,
  init,
  control,
  ...
)

Arguments

formula

A formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function.

data

A data frame containing the variables in the model.

group

A factor variable indicating the group of each observation.

lambda1

A non-negative value specifying the sparse penalty parameter. The default is 0.

lambda2

A non-negative value specifying the global bias penalty parameter. The default is 0.

lambda3

A non-negative value specifying the local bias penalty parameter. The default is 0.

penalty

A character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

A non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 3.0 for MCP.

rho

A value larger than 1 specifying the increase/decrease factor for the augmented Lagrangian's penalty parameter. The default is 2.0.

tau

A value larger than 1 specifying the tolerance for the trade-off between the primal and dual residuals. The default is 10.0.

init

A numeric vector of initial values for the coefficients. The default is a zero vector.

control

An object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Additional arguments to be passed to the fitting algorithm.

Value

An object of class coxtrans.

Examples

formula <- survival::Surv(time, status) ~ . - group - id
fit <- coxtrans(
  formula, sim2, sim2$group,
  lambda1 = 0.03, lambda2 = 0.01, lambda3 = 0.01, penalty = "SCAD"
)
summary(fit)
#> Call:
#> coxtrans(formula = formula, data = sim2, group = sim2$group, 
#>     lambda1 = 0.03, lambda2 = 0.01, lambda3 = 0.01, penalty = "SCAD")
#> 
#>   n=500, number of events=422
#> 
#>               coef exp(coef) se(coef)      z Pr(>|z|)    
#> X1 (1)     0.34792   1.41612  0.05130  6.781 1.19e-11 ***
#> X1 (2, 4)  0.94905   2.58324  0.08576 11.066  < 2e-16 ***
#> X1 (3, 5) -0.25320   0.77631  0.06888 -3.676 0.000237 ***
#> X2 (1)     0.36097   1.43472  0.05370  6.722 1.79e-11 ***
#> X2 (2, 4)  0.96453   2.62354  0.08449 11.415  < 2e-16 ***
#> X2 (3, 5) -0.24258   0.78460  0.08020 -3.025 0.002488 ** 
#> X3 (ALL)   0.34501   1.41200  0.05541  6.227 4.76e-10 ***
#> X4 (ALL)   0.32763   1.38767  0.05335  6.141 8.22e-10 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>           exp(coef) exp(-coef) lower .95 upper .95
#> X1 (1)    1.4161    0.7062     1.2806    1.5659   
#> X1 (2, 4) 2.5832    0.3871     2.1836    3.0561   
#> X1 (3, 5) 0.7763    1.2881     0.6783    0.8885   
#> X2 (1)    1.4347    0.6970     1.2914    1.5940   
#> X2 (2, 4) 2.6235    0.3812     2.2231    3.0961   
#> X2 (3, 5) 0.7846    1.2745     0.6705    0.9181   
#> X3 (ALL)  1.4120    0.7082     1.2667    1.5740   
#> X4 (ALL)  1.3877    0.7206     1.2499    1.5406