Cross-validation for coxtrans

Cross-validation for coxtrans

Usage

cv_coxtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a coxph object fitted to the source data.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 1.5 for MCP.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_coxtrans object.

Examples

library(survival)
library(survtrans)
fit_src <- coxph(Surv(time, status) ~ ., data = sim1_src)
cv_coxtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$beta
#>        X1        X2        X3        X4        X5        X6        X7        X8 
#> -0.737096  2.267407 -1.600389  1.288506 -1.036623  1.375558 -1.808054  1.514996 
#>        X9       X10 
#> -1.316194  1.846887 
#> 
#> $coefficients$eta
#>  [1]  0.282064471  1.203695660 -0.599631828  0.269420624 -0.003393021
#>  [6]  0.327296367 -0.772174363  0.585879383 -0.317628730  0.835035303
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 5.716259e-06
#> 
#> $coefficients$gamma
#> [1] 1
#> 
#> 
#> $lambdas
#>   [1] 5.716259e-02 5.208442e-02 4.745739e-02 4.324140e-02 3.939996e-02
#>   [6] 3.589977e-02 3.271054e-02 2.980462e-02 2.715686e-02 2.474432e-02
#>  [11] 2.254610e-02 2.054317e-02 1.871817e-02 1.705530e-02 1.554015e-02
#>  [16] 1.415961e-02 1.290171e-02 1.175556e-02 1.071123e-02 9.759670e-03
#>  [21] 8.892648e-03 8.102649e-03 7.382832e-03 6.726962e-03 6.129357e-03
#>  [26] 5.584842e-03 5.088700e-03 4.636634e-03 4.224728e-03 3.849415e-03
#>  [31] 3.507443e-03 3.195852e-03 2.911941e-03 2.653252e-03 2.417545e-03
#>  [36] 2.202777e-03 2.007088e-03 1.828784e-03 1.666320e-03 1.518289e-03
#>  [41] 1.383408e-03 1.260510e-03 1.148530e-03 1.046497e-03 9.535295e-04
#>  [46] 8.688205e-04 7.916369e-04 7.213101e-04 6.572309e-04 5.988443e-04
#>  [51] 5.456446e-04 4.971711e-04 4.530037e-04 4.127601e-04 3.760917e-04
#>  [56] 3.426807e-04 3.122379e-04 2.844996e-04 2.592254e-04 2.361965e-04
#>  [61] 2.152135e-04 1.960945e-04 1.786740e-04 1.628011e-04 1.483383e-04
#>  [66] 1.351603e-04 1.231531e-04 1.122125e-04 1.022438e-04 9.316078e-05
#>  [71] 8.488463e-05 7.734371e-05 7.047271e-05 6.421211e-05 5.850768e-05
#>  [76] 5.331002e-05 4.857411e-05 4.425892e-05 4.032708e-05 3.674453e-05
#>  [81] 3.348025e-05 3.050595e-05 2.779589e-05 2.532658e-05 2.307664e-05
#>  [86] 2.102657e-05 1.915863e-05 1.745663e-05 1.590583e-05 1.449280e-05
#>  [91] 1.320530e-05 1.203218e-05 1.096327e-05 9.989325e-06 9.101901e-06
#>  [96] 8.293313e-06 7.556558e-06 6.885254e-06 6.273587e-06 5.716259e-06
#> 
#> $etas
#>             [,1]      [,2]        [,3]       [,4]          [,5]        [,6]
#>   [1,] 0.3716559 0.4230191 -0.06490820 0.00000000  1.420548e-01 0.000000000
#>   [2,] 0.3748949 0.4545396 -0.08728071 0.00000000  1.441320e-01 0.000000000
#>   [3,] 0.3777672 0.4847597 -0.10832035 0.00000000  1.460101e-01 0.000000000
#>   [4,] 0.3803155 0.5136360 -0.12808360 0.00000000  1.477128e-01 0.000000000
#>   [5,] 0.3825780 0.5411378 -0.14662381 0.00000000  1.492597e-01 0.000000000
#>   [6,] 0.3845885 0.5672473 -0.16399249 0.00000000  1.506670e-01 0.000000000
#>   [7,] 0.3863770 0.5919595 -0.18024006 0.00000000  1.519486e-01 0.000000000
#>   [8,] 0.3879697 0.6152814 -0.19541642 0.00000000  1.531166e-01 0.000000000
#>   [9,] 0.3893898 0.6372303 -0.20957082 0.00000000  1.541814e-01 0.000000000
#>  [10,] 0.3904513 0.6584100 -0.22299677 0.00000000  1.550078e-01 0.000000000
#>  [11,] 0.3881533 0.6866929 -0.23932556 0.00000000  1.530367e-01 0.009030493
#>  [12,] 0.3850643 0.7154102 -0.25565595 0.00000000  1.502911e-01 0.022133826
#>  [13,] 0.3821569 0.7422610 -0.27089624 0.00000000  1.477007e-01 0.034278227
#>  [14,] 0.3794264 0.7673204 -0.28509515 0.00000000  1.452631e-01 0.045523901
#>  [15,] 0.3734124 0.7971755 -0.30592016 0.01559351  1.365560e-01 0.064132775
#>  [16,] 0.3667666 0.8270854 -0.32762513 0.03458086  1.265125e-01 0.083868548
#>  [17,] 0.3605024 0.8552025 -0.34801182 0.05237967  1.170227e-01 0.102360931
#>  [18,] 0.3546127 0.8815781 -0.36712158 0.06903303  1.080808e-01 0.119656306
#>  [19,] 0.3490886 0.9062697 -0.38499935 0.08458628  9.967715e-02 0.135803231
#>  [20,] 0.3439194 0.9293387 -0.40169248 0.09908613  9.179912e-02 0.150851500
#>  [21,] 0.3390916 0.9508543 -0.41725341 0.11258292  8.442955e-02 0.164854423
#>  [22,] 0.3345917 0.9708857 -0.43173415 0.12512599  7.755036e-02 0.177864200
#>  [23,] 0.3304051 0.9895051 -0.44518856 0.13676567  7.114132e-02 0.189933872
#>  [24,] 0.3265162 1.0067860 -0.45767104 0.14755225  6.518091e-02 0.201116282
#>  [25,] 0.3229096 1.0228018 -0.46923588 0.15753540  5.964687e-02 0.211463569
#>  [26,] 0.3195695 1.0376256 -0.47993675 0.16676386  5.451654e-02 0.221026761
#>  [27,] 0.3164802 1.0513291 -0.48982630 0.17528509  4.976716e-02 0.229855498
#>  [28,] 0.3136262 1.0639835 -0.49895650 0.18314564  4.537577e-02 0.237998358
#>  [29,] 0.3109929 1.0756550 -0.50737562 0.19038862  4.132101e-02 0.245500364
#>  [30,] 0.3085655 1.0864108 -0.51513272 0.19705748  3.758053e-02 0.252406772
#>  [31,] 0.3063301 1.0963127 -0.52227276 0.20319201  3.413390e-02 0.258759028
#>  [32,] 0.3042735 1.1054210 -0.52883943 0.20883068  3.096098e-02 0.264597178
#>  [33,] 0.3023828 1.1137927 -0.53487422 0.21400992  2.804252e-02 0.269959082
#>  [34,] 0.3006456 1.1214828 -0.54041693 0.21876455  2.535987e-02 0.274880916
#>  [35,] 0.2990512 1.1285402 -0.54550306 0.22312560  2.289655e-02 0.279394968
#>  [36,] 0.2975885 1.1350140 -0.55016800 0.22712390  2.063579e-02 0.283533220
#>  [37,] 0.2962471 1.1409498 -0.55444493 0.23078828  1.856182e-02 0.287325575
#>  [38,] 0.2950181 1.1463879 -0.55836284 0.23414395  1.666103e-02 0.290798231
#>  [39,] 0.2938922 1.1513692 -0.56195147 0.23721663  1.491913e-02 0.293977828
#>  [40,] 0.2928617 1.1559285 -0.56523576 0.24002795  1.332433e-02 0.296886838
#>  [41,] 0.2919183 1.1601016 -0.56824163 0.24260028  1.186413e-02 0.299548391
#>  [42,] 0.2910556 1.1639179 -0.57099038 0.24495203  1.052843e-02 0.301981636
#>  [43,] 0.2902664 1.1674085 -0.57350444 0.24710252  9.306328e-03 0.304206535
#>  [44,] 0.2895450 1.1705996 -0.57580258 0.24906793  8.188875e-03 0.306239872
#>  [45,] 0.2888859 1.1735149 -0.57790202 0.25086311  7.167840e-03 0.308097052
#>  [46,] 0.2882834 1.1761793 -0.57982075 0.25250348  6.234420e-03 0.309794007
#>  [47,] 0.2877331 1.1786130 -0.58157331 0.25400158  5.381655e-03 0.311343736
#>  [48,] 0.2872305 1.1808356 -0.58317375 0.25536945  4.602767e-03 0.312758716
#>  [49,] 0.2867717 1.1828649 -0.58463496 0.25661818  3.891514e-03 0.314050419
#>  [50,] 0.2863528 1.1847174 -0.58596881 0.25775793  3.242156e-03 0.315229379
#>  [51,] 0.2859704 1.1864082 -0.58718619 0.25879806  2.649423e-03 0.316305259
#>  [52,] 0.2856215 1.1879511 -0.58829706 0.25974709  2.108486e-03 0.317286906
#>  [53,] 0.2853029 1.1893597 -0.58931125 0.26061346  1.614512e-03 0.318183009
#>  [54,] 0.2850126 1.1906437 -0.59023569 0.26140310  1.164287e-03 0.318999768
#>  [55,] 0.2847476 1.1918154 -0.59107930 0.26212364  7.533575e-04 0.319745039
#>  [56,] 0.2845057 1.1928849 -0.59184930 0.26278127  3.782037e-04 0.320425206
#>  [57,] 0.2842851 1.1938602 -0.59255141 0.26338088  3.611207e-05 0.321045373
#>  [58,] 0.2841760 1.1944795 -0.59299754 0.26371928  0.000000e+00 0.321432813
#>  [59,] 0.2840872 1.1950125 -0.59338151 0.26400331  0.000000e+00 0.321765112
#>  [60,] 0.2840063 1.1954981 -0.59373144 0.26426211  0.000000e+00 0.322067929
#>  [61,] 0.2839326 1.1959406 -0.59405019 0.26449784  0.000000e+00 0.322343767
#>  [62,] 0.2838219 1.1964719 -0.59443287 0.26481155 -1.305674e-04 0.322679667
#>  [63,] 0.2836615 1.1971309 -0.59490719 0.26523212 -4.282808e-04 0.323100945
#>  [64,] 0.2835153 1.1977317 -0.59533964 0.26561555 -6.996954e-04 0.323485013
#>  [65,] 0.2833820 1.1982795 -0.59573388 0.26596509 -9.471223e-04 0.323835138
#>  [66,] 0.2832605 1.1987788 -0.59609327 0.26628373 -1.172672e-03 0.324154308
#>  [67,] 0.2831498 1.1992340 -0.59642089 0.26657418 -1.378269e-03 0.324445246
#>  [68,] 0.2830488 1.1996489 -0.59671952 0.26683894 -1.565673e-03 0.324710440
#>  [69,] 0.2829568 1.2000271 -0.59699172 0.26708026 -1.736488e-03 0.324952160
#>  [70,] 0.2828729 1.2003719 -0.59723983 0.26730021 -1.892177e-03 0.325172477
#>  [71,] 0.2827965 1.2006861 -0.59746597 0.26750068 -2.034076e-03 0.325373280
#>  [72,] 0.2827268 1.2009725 -0.59767207 0.26768339 -2.163403e-03 0.325556293
#>  [73,] 0.2826633 1.2012335 -0.59785991 0.26784991 -2.281269e-03 0.325723087
#>  [74,] 0.2826055 1.2014713 -0.59803111 0.26800167 -2.388687e-03 0.325875098
#>  [75,] 0.2825527 1.2016881 -0.59818713 0.26813998 -2.486582e-03 0.326013633
#>  [76,] 0.2825047 1.2018857 -0.59832932 0.26826603 -2.575797e-03 0.326139883
#>  [77,] 0.2824609 1.2020658 -0.59845890 0.26838089 -2.657099e-03 0.326254937
#>  [78,] 0.2824210 1.2022298 -0.59857698 0.26848557 -2.731190e-03 0.326359786
#>  [79,] 0.2823846 1.2023794 -0.59868460 0.26858096 -2.798708e-03 0.326455334
#>  [80,] 0.2823514 1.2025156 -0.59878266 0.26866789 -2.860235e-03 0.326542404
#>  [81,] 0.2823212 1.2026398 -0.59887203 0.26874711 -2.916303e-03 0.326621748
#>  [82,] 0.2822937 1.2027530 -0.59895346 0.26881929 -2.967396e-03 0.326694052
#>  [83,] 0.2822686 1.2028561 -0.59902767 0.26888507 -3.013953e-03 0.326759938
#>  [84,] 0.2822458 1.2029501 -0.59909529 0.26894501 -3.056379e-03 0.326819977
#>  [85,] 0.2822249 1.2030357 -0.59915691 0.26899963 -3.095038e-03 0.326874686
#>  [86,] 0.2822059 1.2031137 -0.59921306 0.26904940 -3.130266e-03 0.326924538
#>  [87,] 0.2821887 1.2031848 -0.59926423 0.26909476 -3.162366e-03 0.326969965
#>  [88,] 0.2821729 1.2032496 -0.59931085 0.26913608 -3.191616e-03 0.327011359
#>  [89,] 0.2821588 1.2033077 -0.59935263 0.26917315 -3.217863e-03 0.327048490
#>  [90,] 0.2821458 1.2033615 -0.59939134 0.26920746 -3.242149e-03 0.327082859
#>  [91,] 0.2821338 1.2034105 -0.59942662 0.26923873 -3.264279e-03 0.327114176
#>  [92,] 0.2821230 1.2034552 -0.59945876 0.26926722 -3.284444e-03 0.327142712
#>  [93,] 0.2821131 1.2034959 -0.59948805 0.26929318 -3.302818e-03 0.327168715
#>  [94,] 0.2821040 1.2035329 -0.59951473 0.26931683 -3.319560e-03 0.327192408
#>  [95,] 0.2820958 1.2035667 -0.59953905 0.26933839 -3.334815e-03 0.327213997
#>  [96,] 0.2820883 1.2035975 -0.59956121 0.26935803 -3.348716e-03 0.327233668
#>  [97,] 0.2820815 1.2036256 -0.59958140 0.26937592 -3.361382e-03 0.327251593
#>  [98,] 0.2820753 1.2036511 -0.59959979 0.26939223 -3.372923e-03 0.327267925
#>  [99,] 0.2820696 1.2036744 -0.59961655 0.26940709 -3.383439e-03 0.327282807
#> [100,] 0.2820645 1.2036957 -0.59963183 0.26942062 -3.393021e-03 0.327296367
#>              [,7]        [,8]         [,9]     [,10]
#>   [1,] -0.1513635 0.007892585  0.000000000 0.1576584
#>   [2,] -0.1746116 0.034046514  0.000000000 0.1874782
#>   [3,] -0.1968946 0.058755987  0.000000000 0.2153351
#>   [4,] -0.2181779 0.082063916  0.000000000 0.2413444
#>   [5,] -0.2384376 0.104012200  0.000000000 0.2656115
#>   [6,] -0.2576606 0.124643135  0.000000000 0.2882336
#>   [7,] -0.2758437 0.144000285  0.000000000 0.3093021
#>   [8,] -0.2929931 0.162129001  0.000000000 0.3289035
#>   [9,] -0.3091228 0.179076197  0.000000000 0.3471199
#>  [10,] -0.3246516 0.195125952 -0.001119347 0.3643263
#>  [11,] -0.3461192 0.214965484 -0.013362138 0.3852798
#>  [12,] -0.3683294 0.235199217 -0.026340845 0.4062714
#>  [13,] -0.3890890 0.254129249 -0.038402252 0.4258117
#>  [14,] -0.4084589 0.271808632 -0.049597790 0.4439784
#>  [15,] -0.4332525 0.293405977 -0.067402226 0.4701906
#>  [16,] -0.4584298 0.315120385 -0.086167727 0.4973966
#>  [17,] -0.4820601 0.335497711 -0.103750700 0.5229090
#>  [18,] -0.5041943 0.354583425 -0.120195309 0.5467884
#>  [19,] -0.5248872 0.372425986 -0.135547797 0.5690987
#>  [20,] -0.5441963 0.389075841 -0.149855595 0.5899056
#>  [21,] -0.5621845 0.404587589 -0.163169417 0.6092802
#>  [22,] -0.5789142 0.419015330 -0.175538951 0.6272921
#>  [23,] -0.5944496 0.432414500 -0.187014659 0.6440124
#>  [24,] -0.6088553 0.444840757 -0.197646807 0.6595123
#>  [25,] -0.6221955 0.456349386 -0.207484977 0.6738623
#>  [26,] -0.6345335 0.466994827 -0.216577693 0.6871315
#>  [27,] -0.6459313 0.476830314 -0.224972154 0.6993873
#>  [28,] -0.6564499 0.485908148 -0.232714532 0.7106958
#>  [29,] -0.6661458 0.494277131 -0.239847651 0.7211185
#>  [30,] -0.6750762 0.501986325 -0.246414514 0.7307173
#>  [31,] -0.6832937 0.509080946 -0.252454532 0.7395490
#>  [32,] -0.6908492 0.515604748 -0.258005770 0.7476685
#>  [33,] -0.6977909 0.521599212 -0.263104217 0.7551278
#>  [34,] -0.7041650 0.527104027 -0.267784255 0.7619768
#>  [35,] -0.7100127 0.532154838 -0.272076593 0.7682600
#>  [36,] -0.7153751 0.536786861 -0.276011625 0.7740213
#>  [37,] -0.7202906 0.541033083 -0.279617769 0.7793023
#>  [38,] -0.7247926 0.544922601 -0.282919949 0.7841390
#>  [39,] -0.7289156 0.548484826 -0.285943471 0.7885683
#>  [40,] -0.7326885 0.551744804 -0.288709719 0.7926214
#>  [41,] -0.7361410 0.554728115 -0.291240663 0.7963302
#>  [42,] -0.7392979 0.557456168 -0.293554527 0.7997214
#>  [43,] -0.7421848 0.559951069 -0.295670272 0.8028225
#>  [44,] -0.7448236 0.562231571 -0.297603864 0.8056571
#>  [45,] -0.7472340 0.564314908 -0.299369959 0.8082463
#>  [46,] -0.7494367 0.566218742 -0.300983688 0.8106123
#>  [47,] -0.7514485 0.567957640 -0.302457418 0.8127733
#>  [48,] -0.7532855 0.569545542 -0.303803013 0.8147466
#>  [49,] -0.7549626 0.570995271 -0.305031381 0.8165480
#>  [50,] -0.7564935 0.572318606 -0.306152539 0.8181924
#>  [51,] -0.7578906 0.573526359 -0.307175676 0.8196931
#>  [52,] -0.7591654 0.574628431 -0.308109203 0.8210624
#>  [53,] -0.7603292 0.575634494 -0.308961374 0.8223124
#>  [54,] -0.7613900 0.576551619 -0.309738104 0.8234518
#>  [55,] -0.7623580 0.577388500 -0.310446847 0.8244916
#>  [56,] -0.7632415 0.578152277 -0.311093673 0.8254405
#>  [57,] -0.7640471 0.578848731 -0.311683443 0.8263057
#>  [58,] -0.7645989 0.579354055 -0.312056804 0.8268996
#>  [59,] -0.7650807 0.579799542 -0.312377834 0.8274181
#>  [60,] -0.7655197 0.580205552 -0.312670376 0.8278907
#>  [61,] -0.7659196 0.580575432 -0.312936852 0.8283211
#>  [62,] -0.7663709 0.580974186 -0.313257743 0.8288061
#>  [63,] -0.7669004 0.581421701 -0.313656590 0.8293744
#>  [64,] -0.7673831 0.581829702 -0.314020209 0.8298926
#>  [65,] -0.7678232 0.582201659 -0.314351695 0.8303650
#>  [66,] -0.7682244 0.582540741 -0.314653874 0.8307956
#>  [67,] -0.7685901 0.582849838 -0.314929325 0.8311881
#>  [68,] -0.7689235 0.583131592 -0.315180404 0.8315459
#>  [69,] -0.7692274 0.583388412 -0.315409258 0.8318721
#>  [70,] -0.7695043 0.583622497 -0.315617849 0.8321693
#>  [71,] -0.7697567 0.583835853 -0.315807964 0.8324403
#>  [72,] -0.7699868 0.584030309 -0.315981237 0.8326872
#>  [73,] -0.7701965 0.584207537 -0.316139156 0.8329123
#>  [74,] -0.7703876 0.584369058 -0.316283077 0.8331174
#>  [75,] -0.7705617 0.584516262 -0.316414240 0.8333043
#>  [76,] -0.7707205 0.584650415 -0.316533772 0.8334747
#>  [77,] -0.7708651 0.584772672 -0.316642704 0.8336300
#>  [78,] -0.7709969 0.584884086 -0.316741974 0.8337714
#>  [79,] -0.7711170 0.584985617 -0.316832437 0.8339004
#>  [80,] -0.7712265 0.585078141 -0.316914875 0.8340179
#>  [81,] -0.7713262 0.585162456 -0.316989998 0.8341249
#>  [82,] -0.7714171 0.585239289 -0.317058454 0.8342225
#>  [83,] -0.7715000 0.585309303 -0.317120835 0.8343114
#>  [84,] -0.7715755 0.585373104 -0.317177679 0.8343924
#>  [85,] -0.7716442 0.585431242 -0.317229477 0.8344663
#>  [86,] -0.7717069 0.585484219 -0.317276677 0.8345335
#>  [87,] -0.7717640 0.585532493 -0.317319687 0.8345948
#>  [88,] -0.7718161 0.585576481 -0.317358879 0.8346507
#>  [89,] -0.7718627 0.585615964 -0.317394040 0.8347008
#>  [90,] -0.7719059 0.585652487 -0.317426581 0.8347472
#>  [91,] -0.7719453 0.585685768 -0.317456232 0.8347894
#>  [92,] -0.7719812 0.585716093 -0.317483250 0.8348279
#>  [93,] -0.7720139 0.585743726 -0.317507869 0.8348630
#>  [94,] -0.7720437 0.585768904 -0.317530302 0.8348950
#>  [95,] -0.7720708 0.585791847 -0.317550742 0.8349241
#>  [96,] -0.7720955 0.585812752 -0.317569367 0.8349507
#>  [97,] -0.7721181 0.585831801 -0.317586338 0.8349749
#>  [98,] -0.7721386 0.585849157 -0.317601801 0.8349969
#>  [99,] -0.7721573 0.585864972 -0.317615891 0.8350170
#> [100,] -0.7721744 0.585879383 -0.317628730 0.8350353
#> 
#> $cvm
#>   [1] -25.11953 -25.00033 -24.90074 -24.83543 -24.78397 -24.75404 -24.74077
#>   [8] -24.75679 -24.77996 -24.79552 -24.81899 -24.87129 -24.92400 -24.91773
#>  [15] -24.87814 -24.82795 -24.76997 -24.72294 -24.70245 -24.68601 -24.68422
#>  [22] -24.69056 -24.69297 -24.69430 -24.69914 -24.70682 -24.71676 -24.73213
#>  [29] -24.76023 -24.78850 -24.81660 -24.84424 -24.86496 -24.87266 -24.88068
#>  [36] -24.88885 -24.89704 -24.92195 -24.94815 -24.97275 -24.99578 -25.01728
#>  [43] -25.03732 -25.05596 -25.07326 -25.08929 -25.10412 -25.11783 -25.13049
#>  [50] -25.14215 -25.15289 -25.16278 -25.17186 -25.18021 -25.18787 -25.19524
#>  [57] -25.20210 -25.20839 -25.21414 -25.21941 -25.22379 -25.22778 -25.23144
#>  [64] -25.23478 -25.23783 -25.24062 -25.24316 -25.24549 -25.24761 -25.24955
#>  [71] -25.25132 -25.25294 -25.25441 -25.25576 -25.25698 -25.25810 -25.25912
#>  [78] -25.26005 -25.26090 -25.26167 -25.26238 -25.26302 -25.26361 -25.26414
#>  [85] -25.26463 -25.26507 -25.26548 -25.26585 -25.26618 -25.26649 -25.26677
#>  [92] -25.26702 -25.26726 -25.26747 -25.26766 -25.26784 -25.26800 -25.26814
#>  [99] -25.26827 -25.26840
#> 
#> $cvsd
#>   [1] 13.10817 13.22026 13.33726 13.51694 13.74540 13.96887 14.18774 14.42664
#>   [9] 14.65613 14.85967 15.06804 15.32246 15.59109 15.83976 16.03834 16.23970
#>  [17] 16.46061 16.68461 16.93527 17.18033 17.41948 17.64915 17.85582 18.04487
#>  [25] 18.22422 18.39401 18.55439 18.71101 18.87550 19.03025 19.17553 19.31164
#>  [33] 19.42962 19.52106 19.60635 19.68577 19.75962 19.85283 19.94321 20.02687
#>  [41] 20.10421 20.17564 20.24153 20.30225 20.35815 20.40958 20.45685 20.50028
#>  [49] 20.54014 20.57670 20.61023 20.64095 20.66909 20.69484 20.71842 20.73943
#>  [57] 20.75847 20.77589 20.79180 20.80634 20.82034 20.83313 20.84482 20.85548
#>  [65] 20.86522 20.87410 20.88221 20.88961 20.89635 20.90251 20.90812 20.91324
#>  [73] 20.91791 20.92217 20.92605 20.92959 20.93282 20.93576 20.93844 20.94089
#>  [81] 20.94311 20.94514 20.94699 20.94868 20.95022 20.95162 20.95290 20.95406
#>  [89] 20.95512 20.95609 20.95697 20.95777 20.95850 20.95917 20.95978 20.96033
#>  [97] 20.96083 20.96129 20.96171 20.96209
#> 
#> $formula
#> Surv(time, status) ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + 
#>     X9 + X10
#> <environment: 0x55f10cb5c490>
#> 
#> attr(,"class")
#> [1] "cv_coxtrans"