Cross-validation for coxtrans

Cross-validation for coxtrans

Usage

cv_coxtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a coxph object fitted to the source data.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 1.5 for MCP.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_coxtrans object.

Examples

library(survival)
library(survtrans)
fit_src <- coxph(Surv(time, status) ~ ., data = sim1_src)
cv_coxtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6475045  1.4867309 -1.0656654  1.0190857 -0.8911752  1.0482614 -1.1872429 
#>         X8         X9        X10 
#>  0.9370096 -0.9985653  1.1695097 
#> 
#> $coefficients$eta
#>  [1]  0.371655921  0.423019109 -0.064908236  0.000000000  0.142054833
#>  [6]  0.000000000 -0.151363496  0.007892613  0.000000000  0.157658399
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05716259
#> 
#> $coefficients$gamma
#> [1] 1
#> 
#> 
#> $lambdas
#>   [1] 5.716259e-02 5.208442e-02 4.745739e-02 4.324140e-02 3.939996e-02
#>   [6] 3.589977e-02 3.271054e-02 2.980462e-02 2.715686e-02 2.474432e-02
#>  [11] 2.254610e-02 2.054317e-02 1.871817e-02 1.705530e-02 1.554015e-02
#>  [16] 1.415961e-02 1.290171e-02 1.175556e-02 1.071123e-02 9.759670e-03
#>  [21] 8.892648e-03 8.102649e-03 7.382832e-03 6.726962e-03 6.129357e-03
#>  [26] 5.584842e-03 5.088700e-03 4.636634e-03 4.224728e-03 3.849415e-03
#>  [31] 3.507443e-03 3.195852e-03 2.911941e-03 2.653252e-03 2.417545e-03
#>  [36] 2.202777e-03 2.007088e-03 1.828784e-03 1.666320e-03 1.518289e-03
#>  [41] 1.383408e-03 1.260510e-03 1.148530e-03 1.046497e-03 9.535295e-04
#>  [46] 8.688205e-04 7.916369e-04 7.213101e-04 6.572309e-04 5.988443e-04
#>  [51] 5.456446e-04 4.971711e-04 4.530037e-04 4.127601e-04 3.760917e-04
#>  [56] 3.426807e-04 3.122379e-04 2.844996e-04 2.592254e-04 2.361965e-04
#>  [61] 2.152135e-04 1.960945e-04 1.786740e-04 1.628011e-04 1.483383e-04
#>  [66] 1.351603e-04 1.231531e-04 1.122125e-04 1.022438e-04 9.316078e-05
#>  [71] 8.488463e-05 7.734371e-05 7.047271e-05 6.421211e-05 5.850768e-05
#>  [76] 5.331002e-05 4.857411e-05 4.425892e-05 4.032708e-05 3.674453e-05
#>  [81] 3.348025e-05 3.050595e-05 2.779589e-05 2.532658e-05 2.307664e-05
#>  [86] 2.102657e-05 1.915863e-05 1.745663e-05 1.590583e-05 1.449280e-05
#>  [91] 1.320530e-05 1.203218e-05 1.096327e-05 9.989325e-06 9.101901e-06
#>  [96] 8.293313e-06 7.556558e-06 6.885254e-06 6.273587e-06 5.716259e-06
#> 
#> $etas
#>             [,1]      [,2]        [,3]       [,4]          [,5]       [,6]
#>   [1,] 0.3716559 0.4230191 -0.06490824 0.00000000  1.420548e-01 0.00000000
#>   [2,] 0.3748950 0.4545393 -0.08728054 0.00000000  1.441320e-01 0.00000000
#>   [3,] 0.3777671 0.4847598 -0.10832043 0.00000000  1.460101e-01 0.00000000
#>   [4,] 0.3803155 0.5136359 -0.12808351 0.00000000  1.477129e-01 0.00000000
#>   [5,] 0.3825780 0.5411377 -0.14662375 0.00000000  1.492597e-01 0.00000000
#>   [6,] 0.3845885 0.5672473 -0.16399252 0.00000000  1.506669e-01 0.00000000
#>   [7,] 0.3863769 0.5919599 -0.18024032 0.00000000  1.519485e-01 0.00000000
#>   [8,] 0.3879697 0.6152813 -0.19541635 0.00000000  1.531166e-01 0.00000000
#>   [9,] 0.3893899 0.6372303 -0.20957080 0.00000000  1.541815e-01 0.00000000
#>  [10,] 0.3904511 0.6584108 -0.22299722 0.00000000  1.550077e-01 0.00000000
#>  [11,] 0.3881532 0.6866931 -0.23932565 0.00000000  1.530367e-01 0.00903056
#>  [12,] 0.3850644 0.7154096 -0.25565560 0.00000000  1.502912e-01 0.02213357
#>  [13,] 0.3821570 0.7422604 -0.27089591 0.00000000  1.477009e-01 0.03427799
#>  [14,] 0.3794263 0.7673209 -0.28509544 0.00000000  1.452630e-01 0.04552411
#>  [15,] 0.3734123 0.7971759 -0.30592044 0.01559376  1.365558e-01 0.06413302
#>  [16,] 0.3667668 0.8270849 -0.32762479 0.03458057  1.265127e-01 0.08386826
#>  [17,] 0.3605025 0.8552022 -0.34801158 0.05237947  1.170228e-01 0.10236073
#>  [18,] 0.3546128 0.8815780 -0.36712145 0.06903291  1.080808e-01 0.11965619
#>  [19,] 0.3490887 0.9062695 -0.38499921 0.08458616  9.967724e-02 0.13580311
#>  [20,] 0.3439192 0.9293393 -0.40169295 0.09908653  9.179884e-02 0.15085190
#>  [21,] 0.3390915 0.9508546 -0.41725361 0.11258309  8.442943e-02 0.16485460
#>  [22,] 0.3345918 0.9708856 -0.43173408 0.12512592  7.755040e-02 0.17786414
#>  [23,] 0.3304052 0.9895047 -0.44518827 0.13676542  7.114149e-02 0.18993362
#>  [24,] 0.3265164 1.0067854 -0.45767059 0.14755186  6.518118e-02 0.20111590
#>  [25,] 0.3229095 1.0228020 -0.46923603 0.15753553  5.964678e-02 0.21146370
#>  [26,] 0.3195695 1.0376256 -0.47993679 0.16676389  5.451651e-02 0.22102679
#>  [27,] 0.3164800 1.0513298 -0.48982681 0.17528552  4.976686e-02 0.22985593
#>  [28,] 0.3136261 1.0639838 -0.49895673 0.18314584  4.537564e-02 0.23799856
#>  [29,] 0.3109927 1.0756558 -0.50737623 0.19038913  4.132065e-02 0.24550088
#>  [30,] 0.3085653 1.0864113 -0.51513314 0.19705784  3.758028e-02 0.25240713
#>  [31,] 0.3063302 1.0963124 -0.52227255 0.20319184  3.413401e-02 0.25875886
#>  [32,] 0.3042735 1.1054209 -0.52883941 0.20883067  3.096099e-02 0.26459716
#>  [33,] 0.3023827 1.1137928 -0.53487426 0.21400995  2.804250e-02 0.26995911
#>  [34,] 0.3006456 1.1214828 -0.54041693 0.21876455  2.535988e-02 0.27488091
#>  [35,] 0.2990512 1.1285402 -0.54550301 0.22312556  2.289658e-02 0.27939493
#>  [36,] 0.2975885 1.1350139 -0.55016792 0.22712383  2.063584e-02 0.28353315
#>  [37,] 0.2962471 1.1409496 -0.55444483 0.23078820  1.856188e-02 0.28732549
#>  [38,] 0.2950182 1.1463877 -0.55836270 0.23414383  1.666111e-02 0.29079811
#>  [39,] 0.2938923 1.1513690 -0.56195129 0.23721648  1.491923e-02 0.29397768
#>  [40,] 0.2928618 1.1559283 -0.56523555 0.24002777  1.332446e-02 0.29688666
#>  [41,] 0.2919184 1.1601012 -0.56824139 0.24260008  1.186427e-02 0.29954819
#>  [42,] 0.2910554 1.1639185 -0.57099081 0.24495239  1.052818e-02 0.30198200
#>  [43,] 0.2902666 1.1674082 -0.57350415 0.24710228  9.306498e-03 0.30420629
#>  [44,] 0.2895451 1.1705991 -0.57580225 0.24906766  8.189065e-03 0.30623960
#>  [45,] 0.2888857 1.1735153 -0.57790238 0.25086341  7.167635e-03 0.30809735
#>  [46,] 0.2882833 1.1761797 -0.57982108 0.25250376  6.234226e-03 0.30979429
#>  [47,] 0.2877330 1.1786134 -0.58157362 0.25400184  5.381476e-03 0.31134400
#>  [48,] 0.2872307 1.1808351 -0.58317334 0.25536910  4.603004e-03 0.31275837
#>  [49,] 0.2867718 1.1828644 -0.58463454 0.25661782  3.891759e-03 0.31405006
#>  [50,] 0.2863529 1.1847168 -0.58596838 0.25775757  3.242408e-03 0.31522902
#>  [51,] 0.2859703 1.1864085 -0.58718644 0.25879827  2.649279e-03 0.31630547
#>  [52,] 0.2856214 1.1879514 -0.58829729 0.25974729  2.108352e-03 0.31728710
#>  [53,] 0.2853031 1.1893590 -0.58931077 0.26061305  1.614794e-03 0.31818260
#>  [54,] 0.2850125 1.1906439 -0.59023588 0.26140326  1.164173e-03 0.31899993
#>  [55,] 0.2847476 1.1918157 -0.59107949 0.26212380  7.532471e-04 0.31974520
#>  [56,] 0.2845056 1.1928852 -0.59184949 0.26278143  3.780939e-04 0.32042536
#>  [57,] 0.2842854 1.1938595 -0.59255089 0.26338044  3.641641e-05 0.32104493
#>  [58,] 0.2841759 1.1944798 -0.59299771 0.26371940  0.000000e+00 0.32143295
#>  [59,] 0.2840871 1.1950127 -0.59338168 0.26400342  0.000000e+00 0.32176524
#>  [60,] 0.2840062 1.1954983 -0.59373160 0.26426222  0.000000e+00 0.32206806
#>  [61,] 0.2839326 1.1959407 -0.59405034 0.26449794  0.000000e+00 0.32234388
#>  [62,] 0.2838218 1.1964721 -0.59443301 0.26481167 -1.306512e-04 0.32267979
#>  [63,] 0.2836614 1.1971311 -0.59490733 0.26523224 -4.283618e-04 0.32310106
#>  [64,] 0.2835152 1.1977319 -0.59533977 0.26561566 -6.997737e-04 0.32348513
#>  [65,] 0.2833820 1.1982796 -0.59573401 0.26596520 -9.471982e-04 0.32383525
#>  [66,] 0.2832605 1.1987790 -0.59609340 0.26628383 -1.172745e-03 0.32415441
#>  [67,] 0.2831497 1.1992342 -0.59642101 0.26657428 -1.378341e-03 0.32444535
#>  [68,] 0.2830488 1.1996491 -0.59671964 0.26683904 -1.565743e-03 0.32471054
#>  [69,] 0.2829568 1.2000273 -0.59699184 0.26708035 -1.736556e-03 0.32495226
#>  [70,] 0.2828729 1.2003720 -0.59723994 0.26730031 -1.892243e-03 0.32517257
#>  [71,] 0.2827965 1.2006862 -0.59746608 0.26750078 -2.034141e-03 0.32537337
#>  [72,] 0.2827268 1.2009726 -0.59767218 0.26768348 -2.163466e-03 0.32555638
#>  [73,] 0.2826633 1.2012336 -0.59786002 0.26785000 -2.281331e-03 0.32572318
#>  [74,] 0.2826054 1.2014715 -0.59803121 0.26800176 -2.388748e-03 0.32587519
#>  [75,] 0.2825527 1.2016883 -0.59818723 0.26814007 -2.486642e-03 0.32601372
#>  [76,] 0.2825046 1.2018858 -0.59832942 0.26826611 -2.575856e-03 0.32613997
#>  [77,] 0.2824608 1.2020659 -0.59845900 0.26838098 -2.657157e-03 0.32625502
#>  [78,] 0.2824212 1.2022290 -0.59857638 0.26848506 -2.730841e-03 0.32635928
#>  [79,] 0.2823848 1.2023786 -0.59868399 0.26858045 -2.798358e-03 0.32645483
#>  [80,] 0.2823517 1.2025148 -0.59878206 0.26866738 -2.859884e-03 0.32654190
#>  [81,] 0.2823215 1.2026390 -0.59887142 0.26874659 -2.915951e-03 0.32662124
#>  [82,] 0.2822940 1.2027522 -0.59895285 0.26881878 -2.967042e-03 0.32669354
#>  [83,] 0.2822689 1.2028553 -0.59902706 0.26888455 -3.013599e-03 0.32675943
#>  [84,] 0.2822460 1.2029492 -0.59909468 0.26894449 -3.056024e-03 0.32681946
#>  [85,] 0.2822252 1.2030349 -0.59915630 0.26899911 -3.094682e-03 0.32687417
#>  [86,] 0.2822062 1.2031129 -0.59921245 0.26904888 -3.129909e-03 0.32692402
#>  [87,] 0.2821889 1.2031840 -0.59926361 0.26909424 -3.162009e-03 0.32696945
#>  [88,] 0.2821732 1.2032488 -0.59931023 0.26913556 -3.191259e-03 0.32701084
#>  [89,] 0.2821588 1.2033078 -0.59935272 0.26917322 -3.217911e-03 0.32704856
#>  [90,] 0.2821457 1.2033616 -0.59939143 0.26920753 -3.242197e-03 0.32708293
#>  [91,] 0.2821338 1.2034106 -0.59942670 0.26923880 -3.264327e-03 0.32711425
#>  [92,] 0.2821229 1.2034553 -0.59945884 0.26926729 -3.284491e-03 0.32714278
#>  [93,] 0.2821130 1.2034960 -0.59948813 0.26929325 -3.302865e-03 0.32716878
#>  [94,] 0.2821040 1.2035331 -0.59951481 0.26931690 -3.319607e-03 0.32719248
#>  [95,] 0.2820958 1.2035668 -0.59953913 0.26933845 -3.334862e-03 0.32721406
#>  [96,] 0.2820883 1.2035976 -0.59956129 0.26935809 -3.348763e-03 0.32723374
#>  [97,] 0.2820815 1.2036257 -0.59958148 0.26937599 -3.361428e-03 0.32725166
#>  [98,] 0.2820753 1.2036512 -0.59959987 0.26939230 -3.372969e-03 0.32726799
#>  [99,] 0.2820696 1.2036745 -0.59961663 0.26940715 -3.383485e-03 0.32728287
#> [100,] 0.2820644 1.2036958 -0.59963191 0.26942069 -3.393067e-03 0.32729643
#>              [,7]        [,8]        [,9]     [,10]
#>   [1,] -0.1513635 0.007892613  0.00000000 0.1576584
#>   [2,] -0.1746115 0.034046390  0.00000000 0.1874780
#>   [3,] -0.1968947 0.058756046  0.00000000 0.2153352
#>   [4,] -0.2181778 0.082063848  0.00000000 0.2413444
#>   [5,] -0.2384375 0.104012154  0.00000000 0.2656114
#>   [6,] -0.2576606 0.124643154  0.00000000 0.2882336
#>   [7,] -0.2758439 0.144000498  0.00000000 0.3093024
#>   [8,] -0.2929930 0.162128938  0.00000000 0.3289034
#>   [9,] -0.3091228 0.179076175  0.00000000 0.3471198
#>  [10,] -0.3246521 0.195126332 -0.00111958 0.3643268
#>  [11,] -0.3461193 0.214965572 -0.01336220 0.3852799
#>  [12,] -0.3683290 0.235198877 -0.02634061 0.4062710
#>  [13,] -0.3890886 0.254128920 -0.03840202 0.4258113
#>  [14,] -0.4084593 0.271808925 -0.04959799 0.4439787
#>  [15,] -0.4332529 0.293406217 -0.06740245 0.4701909
#>  [16,] -0.4584294 0.315120107 -0.08616746 0.4973962
#>  [17,] -0.4820598 0.335497514 -0.10375051 0.5229087
#>  [18,] -0.5041942 0.354583313 -0.12019520 0.5467883
#>  [19,] -0.5248871 0.372425867 -0.13554768 0.5690985
#>  [20,] -0.5441968 0.389076232 -0.14985596 0.5899062
#>  [21,] -0.5621847 0.404587761 -0.16316958 0.6092804
#>  [22,] -0.5789141 0.419015267 -0.17553889 0.6272920
#>  [23,] -0.5944493 0.432414253 -0.18701443 0.6440121
#>  [24,] -0.6088548 0.444840378 -0.19764645 0.6595118
#>  [25,] -0.6221956 0.456349518 -0.20748510 0.6738625
#>  [26,] -0.6345336 0.466994860 -0.21657772 0.6871315
#>  [27,] -0.6459319 0.476830746 -0.22497256 0.6993878
#>  [28,] -0.6564501 0.485908347 -0.23271472 0.7106960
#>  [29,] -0.6661464 0.494277645 -0.23984813 0.7211192
#>  [30,] -0.6750766 0.501986683 -0.24641485 0.7307178
#>  [31,] -0.6832935 0.509080773 -0.25245437 0.7395487
#>  [32,] -0.6908492 0.515604730 -0.25800575 0.7476685
#>  [33,] -0.6977909 0.521599243 -0.26310425 0.7551279
#>  [34,] -0.7041650 0.527104023 -0.26778425 0.7619768
#>  [35,] -0.7100127 0.532154797 -0.27207656 0.7682599
#>  [36,] -0.7153750 0.536786795 -0.27601156 0.7740213
#>  [37,] -0.7202904 0.541032998 -0.27961769 0.7793022
#>  [38,] -0.7247925 0.544922483 -0.28291984 0.7841388
#>  [39,] -0.7289155 0.548484677 -0.28594333 0.7885681
#>  [40,] -0.7326883 0.551744623 -0.28870955 0.7926211
#>  [41,] -0.7361407 0.554727911 -0.29124047 0.7963299
#>  [42,] -0.7392983 0.557456534 -0.29355486 0.7997219
#>  [43,] -0.7421845 0.559950818 -0.29567004 0.8028222
#>  [44,] -0.7448232 0.562231291 -0.29760361 0.8056567
#>  [45,] -0.7472344 0.564315209 -0.29937024 0.8082467
#>  [46,] -0.7494370 0.566219028 -0.30098395 0.8106127
#>  [47,] -0.7514488 0.567957904 -0.30245766 0.8127737
#>  [48,] -0.7532851 0.569545192 -0.30380269 0.8147461
#>  [49,] -0.7549622 0.570994909 -0.30503105 0.8165476
#>  [50,] -0.7564930 0.572318236 -0.30615220 0.8181919
#>  [51,] -0.7578909 0.573526571 -0.30717587 0.8196933
#>  [52,] -0.7591657 0.574628629 -0.30810939 0.8210626
#>  [53,] -0.7603287 0.575634079 -0.30896099 0.8223118
#>  [54,] -0.7613902 0.576551787 -0.30973826 0.8234520
#>  [55,] -0.7623582 0.577388662 -0.31044700 0.8244918
#>  [56,] -0.7632417 0.578152439 -0.31109382 0.8254407
#>  [57,] -0.7640465 0.578848281 -0.31168303 0.8263051
#>  [58,] -0.7645992 0.579354228 -0.31205693 0.8268998
#>  [59,] -0.7650809 0.579799713 -0.31237796 0.8274183
#>  [60,] -0.7655199 0.580205716 -0.31267050 0.8278909
#>  [61,] -0.7659198 0.580575585 -0.31293696 0.8283213
#>  [62,] -0.7663710 0.580974310 -0.31325786 0.8288063
#>  [63,] -0.7669005 0.581421820 -0.31365670 0.8293746
#>  [64,] -0.7673833 0.581829818 -0.31402032 0.8298927
#>  [65,] -0.7678234 0.582201771 -0.31435180 0.8303651
#>  [66,] -0.7682246 0.582540849 -0.31465397 0.8307957
#>  [67,] -0.7685903 0.582849944 -0.31492942 0.8311883
#>  [68,] -0.7689236 0.583131695 -0.31518050 0.8315461
#>  [69,] -0.7692275 0.583388513 -0.31540935 0.8318722
#>  [70,] -0.7695044 0.583622595 -0.31561794 0.8321695
#>  [71,] -0.7697569 0.583835948 -0.31580805 0.8324404
#>  [72,] -0.7699869 0.584030403 -0.31598132 0.8326874
#>  [73,] -0.7701966 0.584207629 -0.31613924 0.8329124
#>  [74,] -0.7703877 0.584369148 -0.31628316 0.8331175
#>  [75,] -0.7705619 0.584516351 -0.31641432 0.8333045
#>  [76,] -0.7707206 0.584650502 -0.31653385 0.8334748
#>  [77,] -0.7708652 0.584772757 -0.31664278 0.8336301
#>  [78,] -0.7709963 0.584883570 -0.31674150 0.8337707
#>  [79,] -0.7711164 0.584985099 -0.31683196 0.8338997
#>  [80,] -0.7712258 0.585077622 -0.31691440 0.8340172
#>  [81,] -0.7713256 0.585161935 -0.31698952 0.8341242
#>  [82,] -0.7714165 0.585238766 -0.31705798 0.8342218
#>  [83,] -0.7714993 0.585308780 -0.31712036 0.8343107
#>  [84,] -0.7715748 0.585372579 -0.31717720 0.8343917
#>  [85,] -0.7716436 0.585430716 -0.31722900 0.8344656
#>  [86,] -0.7717062 0.585483692 -0.31727619 0.8345328
#>  [87,] -0.7717634 0.585531965 -0.31731920 0.8345941
#>  [88,] -0.7718154 0.585575953 -0.31735839 0.8346500
#>  [89,] -0.7718628 0.585616035 -0.31739411 0.8347009
#>  [90,] -0.7719060 0.585652558 -0.31742665 0.8347473
#>  [91,] -0.7719454 0.585685838 -0.31745630 0.8347895
#>  [92,] -0.7719813 0.585716163 -0.31748331 0.8348280
#>  [93,] -0.7720140 0.585743796 -0.31750793 0.8348631
#>  [94,] -0.7720438 0.585768974 -0.31753037 0.8348951
#>  [95,] -0.7720709 0.585791916 -0.31755081 0.8349242
#>  [96,] -0.7720956 0.585812821 -0.31756943 0.8349508
#>  [97,] -0.7721182 0.585831869 -0.31758640 0.8349750
#>  [98,] -0.7721387 0.585849226 -0.31760186 0.8349970
#>  [99,] -0.7721574 0.585865041 -0.31761595 0.8350171
#> [100,] -0.7721745 0.585879451 -0.31762879 0.8350354
#> 
#> $cvm
#>   [1] -24.50293 -24.40858 -24.32172 -24.21608 -24.12171 -24.04137 -23.94783
#>   [8] -23.86277 -23.77982 -23.68989 -23.58496 -23.47462 -23.37094 -23.28320
#>  [15] -23.19413 -23.10388 -23.02014 -22.94277 -22.85570 -22.78372 -22.72239
#>  [22] -22.66838 -22.62086 -22.57807 -22.53329 -22.49297 -22.46345 -22.43980
#>  [29] -22.42031 -22.40504 -22.38601 -22.36613 -22.34747 -22.33083 -22.31598
#>  [36] -22.30272 -22.29085 -22.28024 -22.27072 -22.26219 -22.25454 -22.24766
#>  [43] -22.24148 -22.23591 -22.23090 -22.22638 -22.22231 -22.21863 -22.21530
#>  [50] -22.21230 -22.20958 -22.20712 -22.20490 -22.20288 -22.20105 -22.19940
#>  [57] -22.19789 -22.19653 -22.19529 -22.19417 -22.19315 -22.19222 -22.19138
#>  [64] -22.19061 -22.18992 -22.18928 -22.18871 -22.18818 -22.18771 -22.18727
#>  [71] -22.18688 -22.18652 -22.18619 -22.18590 -22.18562 -22.18538 -22.18515
#>  [78] -22.18495 -22.18476 -22.18459 -22.18444 -22.18430 -22.18417 -22.18405
#>  [85] -22.18395 -22.18385 -22.18376 -22.18368 -22.18361 -22.18354 -22.18348
#>  [92] -22.18343 -22.18338 -22.18333 -22.18329 -22.18325 -22.18322 -22.18318
#>  [99] -22.18316 -22.18313
#> 
#> $cvsd
#>   [1]  9.312091  9.328710  9.356803  9.388886  9.423339  9.459129  9.450638
#>   [8]  9.447563  9.434115  9.416308  9.423653  9.428722  9.443608  9.468159
#>  [15]  9.502781  9.543495  9.586676  9.624637  9.675316  9.723872  9.770370
#>  [22]  9.816702  9.862431  9.908198  9.956960 10.003354 10.043768 10.080209
#>  [29] 10.114305 10.148053 10.183641 10.219458 10.254166 10.286577 10.316775
#>  [36] 10.344856 10.370923 10.395079 10.417433 10.438088 10.457153 10.474728
#>  [43] 10.490914 10.505807 10.519497 10.532072 10.543615 10.554204 10.563911
#>  [50] 10.572807 10.580951 10.588409 10.595232 10.601472 10.607180 10.612396
#>  [57] 10.617162 10.621518 10.625495 10.629128 10.632444 10.635472 10.638236
#>  [64] 10.640757 10.643057 10.645156 10.647070 10.648817 10.650410 10.651862
#>  [71] 10.653186 10.654395 10.655496 10.656500 10.657415 10.658250 10.659011
#>  [78] 10.659704 10.660336 10.660912 10.661437 10.661915 10.662351 10.662749
#>  [85] 10.663111 10.663441 10.663742 10.664016 10.664266 10.664494 10.664701
#>  [92] 10.664890 10.665063 10.665220 10.665363 10.665493 10.665612 10.665720
#>  [99] 10.665819 10.665909
#> 
#> $formula
#> Surv(time, status) ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + 
#>     X9 + X10
#> <environment: 0x55587dc31fa0>
#> 
#> attr(,"class")
#> [1] "cv_coxtrans"