Cross-validation for coxtrans

Cross-validation for coxtrans

Usage

cv_coxtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a coxph object fitted to the source data.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 1.5 for MCP.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_coxtrans object.

Examples

library(survival)
library(survtrans)
fit_src <- coxph(Surv(time, status) ~ ., data = sim1_src)
cv_coxtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6475045  1.4867308 -1.0656653  1.0190857 -0.8911751  1.0482614 -1.1872429 
#>         X8         X9        X10 
#>  0.9370096 -0.9985653  1.1695097 
#> 
#> $coefficients$eta
#>  [1]  0.371655950  0.423019026 -0.064908176  0.000000000  0.142054855
#>  [6]  0.000000000 -0.151363441  0.007892569  0.000000000  0.157658342
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05716259
#> 
#> $coefficients$gamma
#> [1] 1
#> 
#> 
#> $lambdas
#>   [1] 5.716259e-02 5.208442e-02 4.745739e-02 4.324140e-02 3.939996e-02
#>   [6] 3.589977e-02 3.271054e-02 2.980462e-02 2.715686e-02 2.474432e-02
#>  [11] 2.254610e-02 2.054317e-02 1.871817e-02 1.705530e-02 1.554015e-02
#>  [16] 1.415961e-02 1.290171e-02 1.175556e-02 1.071123e-02 9.759670e-03
#>  [21] 8.892648e-03 8.102649e-03 7.382832e-03 6.726962e-03 6.129357e-03
#>  [26] 5.584842e-03 5.088700e-03 4.636634e-03 4.224728e-03 3.849415e-03
#>  [31] 3.507443e-03 3.195852e-03 2.911941e-03 2.653252e-03 2.417545e-03
#>  [36] 2.202777e-03 2.007088e-03 1.828784e-03 1.666320e-03 1.518289e-03
#>  [41] 1.383408e-03 1.260510e-03 1.148530e-03 1.046497e-03 9.535295e-04
#>  [46] 8.688205e-04 7.916369e-04 7.213101e-04 6.572309e-04 5.988443e-04
#>  [51] 5.456446e-04 4.971711e-04 4.530037e-04 4.127601e-04 3.760917e-04
#>  [56] 3.426807e-04 3.122379e-04 2.844996e-04 2.592254e-04 2.361965e-04
#>  [61] 2.152135e-04 1.960945e-04 1.786740e-04 1.628011e-04 1.483383e-04
#>  [66] 1.351603e-04 1.231531e-04 1.122125e-04 1.022438e-04 9.316078e-05
#>  [71] 8.488463e-05 7.734371e-05 7.047271e-05 6.421211e-05 5.850768e-05
#>  [76] 5.331002e-05 4.857411e-05 4.425892e-05 4.032708e-05 3.674453e-05
#>  [81] 3.348025e-05 3.050595e-05 2.779589e-05 2.532658e-05 2.307664e-05
#>  [86] 2.102657e-05 1.915863e-05 1.745663e-05 1.590583e-05 1.449280e-05
#>  [91] 1.320530e-05 1.203218e-05 1.096327e-05 9.989325e-06 9.101901e-06
#>  [96] 8.293313e-06 7.556558e-06 6.885254e-06 6.273587e-06 5.716259e-06
#> 
#> $etas
#>             [,1]      [,2]        [,3]       [,4]          [,5]        [,6]
#>   [1,] 0.3716559 0.4230190 -0.06490818 0.00000000  1.420549e-01 0.000000000
#>   [2,] 0.3748948 0.4545400 -0.08728100 0.00000000  1.441318e-01 0.000000000
#>   [3,] 0.3777672 0.4847596 -0.10832030 0.00000000  1.460101e-01 0.000000000
#>   [4,] 0.3803154 0.5136361 -0.12808370 0.00000000  1.477128e-01 0.000000000
#>   [5,] 0.3825780 0.5411378 -0.14662382 0.00000000  1.492597e-01 0.000000000
#>   [6,] 0.3845885 0.5672472 -0.16399240 0.00000000  1.506670e-01 0.000000000
#>   [7,] 0.3863769 0.5919599 -0.18024031 0.00000000  1.519485e-01 0.000000000
#>   [8,] 0.3879697 0.6152813 -0.19541635 0.00000000  1.531166e-01 0.000000000
#>   [9,] 0.3893897 0.6372308 -0.20957116 0.00000000  1.541813e-01 0.000000000
#>  [10,] 0.3904512 0.6584102 -0.22299685 0.00000000  1.550078e-01 0.000000000
#>  [11,] 0.3881533 0.6866929 -0.23932554 0.00000000  1.530367e-01 0.009030479
#>  [12,] 0.3850643 0.7154100 -0.25565583 0.00000000  1.502911e-01 0.022133736
#>  [13,] 0.3821569 0.7422607 -0.27089610 0.00000000  1.477008e-01 0.034278123
#>  [14,] 0.3794263 0.7673208 -0.28509538 0.00000000  1.452630e-01 0.045524069
#>  [15,] 0.3734122 0.7971760 -0.30592053 0.01559383  1.365557e-01 0.064133094
#>  [16,] 0.3667668 0.8270849 -0.32762478 0.03458056  1.265127e-01 0.083868248
#>  [17,] 0.3605025 0.8552022 -0.34801163 0.05237951  1.170228e-01 0.102360775
#>  [18,] 0.3546128 0.8815781 -0.36712155 0.06903300  1.080808e-01 0.119656281
#>  [19,] 0.3490887 0.9062694 -0.38499917 0.08458613  9.967726e-02 0.135803077
#>  [20,] 0.3439192 0.9293392 -0.40169284 0.09908644  9.179891e-02 0.150851804
#>  [21,] 0.3390916 0.9508544 -0.41725350 0.11258299  8.442950e-02 0.164854498
#>  [22,] 0.3345918 0.9708854 -0.43173396 0.12512582  7.755048e-02 0.177864032
#>  [23,] 0.3304053 0.9895044 -0.44518805 0.13676523  7.114162e-02 0.189933432
#>  [24,] 0.3265163 1.0067858 -0.45767094 0.14755216  6.518097e-02 0.201116195
#>  [25,] 0.3229097 1.0228015 -0.46923564 0.15753520  5.964701e-02 0.211463371
#>  [26,] 0.3195693 1.0376259 -0.47993704 0.16676411  5.451636e-02 0.221027010
#>  [27,] 0.3164801 1.0513294 -0.48982647 0.17528524  4.976706e-02 0.229855645
#>  [28,] 0.3136261 1.0639837 -0.49895660 0.18314573  4.537571e-02 0.237998444
#>  [29,] 0.3109929 1.0756550 -0.50737561 0.19038860  4.132102e-02 0.245500349
#>  [30,] 0.3085655 1.0864108 -0.51513272 0.19705748  3.758053e-02 0.252406767
#>  [31,] 0.3063300 1.0963129 -0.52227293 0.20319215  3.413379e-02 0.258759174
#>  [32,] 0.3042733 1.1054215 -0.52883979 0.20883099  3.096077e-02 0.264597481
#>  [33,] 0.3023825 1.1137933 -0.53487469 0.21401032  2.804224e-02 0.269959481
#>  [34,] 0.3006457 1.1214824 -0.54041666 0.21876432  2.536004e-02 0.274880685
#>  [35,] 0.2990514 1.1285397 -0.54550268 0.22312528  2.289677e-02 0.279394648
#>  [36,] 0.2975884 1.1350142 -0.55016820 0.22712407  2.063567e-02 0.283533390
#>  [37,] 0.2962473 1.1409491 -0.55444439 0.23078783  1.856213e-02 0.287325128
#>  [38,] 0.2950181 1.1463880 -0.55836298 0.23414406  1.666095e-02 0.290798346
#>  [39,] 0.2938925 1.1513684 -0.56195086 0.23721611  1.491948e-02 0.293977316
#>  [40,] 0.2928616 1.1559286 -0.56523582 0.24002801  1.332430e-02 0.296886890
#>  [41,] 0.2919183 1.1601016 -0.56824165 0.24260030  1.186411e-02 0.299548415
#>  [42,] 0.2910556 1.1639179 -0.57099038 0.24495203  1.052843e-02 0.301981634
#>  [43,] 0.2902665 1.1674085 -0.57350441 0.24710250  9.306345e-03 0.304206510
#>  [44,] 0.2895450 1.1705995 -0.57580252 0.24906789  8.188908e-03 0.306239824
#>  [45,] 0.2888859 1.1735148 -0.57790195 0.25086304  7.167885e-03 0.308096987
#>  [46,] 0.2882835 1.1761791 -0.57982065 0.25250340  6.234476e-03 0.309793927
#>  [47,] 0.2877332 1.1786129 -0.58157319 0.25400148  5.381725e-03 0.311343635
#>  [48,] 0.2872306 1.1808354 -0.58317361 0.25536933  4.602850e-03 0.312758596
#>  [49,] 0.2867717 1.1828647 -0.58463480 0.25661804  3.891606e-03 0.314050285
#>  [50,] 0.2863528 1.1847172 -0.58596864 0.25775779  3.242256e-03 0.315229235
#>  [51,] 0.2859705 1.1864079 -0.58718600 0.25879790  2.649532e-03 0.316305102
#>  [52,] 0.2856216 1.1879508 -0.58829686 0.25974693  2.108600e-03 0.317286742
#>  [53,] 0.2853030 1.1893594 -0.58931103 0.26061327  1.614640e-03 0.318182824
#>  [54,] 0.2850124 1.1906443 -0.59023616 0.26140350  1.164011e-03 0.319000166
#>  [55,] 0.2847474 1.1918161 -0.59107976 0.26212403  7.530914e-04 0.319745424
#>  [56,] 0.2845058 1.1928846 -0.59184905 0.26278106  3.783471e-04 0.320424999
#>  [57,] 0.2842853 1.1938598 -0.59255116 0.26338067  3.626001e-05 0.321045159
#>  [58,] 0.2841761 1.1944792 -0.59299729 0.26371912  0.000000e+00 0.321432618
#>  [59,] 0.2840873 1.1950121 -0.59338125 0.26400313  0.000000e+00 0.321764909
#>  [60,] 0.2840064 1.1954978 -0.59373117 0.26426193  0.000000e+00 0.322067722
#>  [61,] 0.2839327 1.1959402 -0.59404994 0.26449767  0.000000e+00 0.322343569
#>  [62,] 0.2838217 1.1964724 -0.59443328 0.26481191 -1.308099e-04 0.322680018
#>  [63,] 0.2836613 1.1971314 -0.59490761 0.26523247 -4.285214e-04 0.323101293
#>  [64,] 0.2835151 1.1977323 -0.59534005 0.26561589 -6.999342e-04 0.323485358
#>  [65,] 0.2833819 1.1982800 -0.59573429 0.26596543 -9.473595e-04 0.323835481
#>  [66,] 0.2832607 1.1987784 -0.59609298 0.26628348 -1.172500e-03 0.324154060
#>  [67,] 0.2831499 1.1992336 -0.59642059 0.26657393 -1.378095e-03 0.324444994
#>  [68,] 0.2830490 1.1996485 -0.59671921 0.26683868 -1.565496e-03 0.324710185
#>  [69,] 0.2829570 1.2000267 -0.59699141 0.26707999 -1.736309e-03 0.324951901
#>  [70,] 0.2828731 1.2003715 -0.59723952 0.26729994 -1.891995e-03 0.325172214
#>  [71,] 0.2827966 1.2006857 -0.59746565 0.26750041 -2.033892e-03 0.325373014
#>  [72,] 0.2827270 1.2009720 -0.59767175 0.26768312 -2.163217e-03 0.325556024
#>  [73,] 0.2826635 1.2012330 -0.59785959 0.26784964 -2.281081e-03 0.325722816
#>  [74,] 0.2826056 1.2014709 -0.59803078 0.26800140 -2.388498e-03 0.325874825
#>  [75,] 0.2825529 1.2016877 -0.59818680 0.26813970 -2.486392e-03 0.326013357
#>  [76,] 0.2825048 1.2018853 -0.59832899 0.26826575 -2.575605e-03 0.326139606
#>  [77,] 0.2824610 1.2020653 -0.59845857 0.26838061 -2.656906e-03 0.326254658
#>  [78,] 0.2824211 1.2022294 -0.59857665 0.26848529 -2.730996e-03 0.326359505
#>  [79,] 0.2823847 1.2023789 -0.59868426 0.26858068 -2.798513e-03 0.326455051
#>  [80,] 0.2823516 1.2025152 -0.59878233 0.26866760 -2.860039e-03 0.326542120
#>  [81,] 0.2823214 1.2026394 -0.59887169 0.26874682 -2.916106e-03 0.326621464
#>  [82,] 0.2822938 1.2027525 -0.59895312 0.26881900 -2.967198e-03 0.326693766
#>  [83,] 0.2822688 1.2028556 -0.59902733 0.26888478 -3.013755e-03 0.326759651
#>  [84,] 0.2822459 1.2029496 -0.59909495 0.26894472 -3.056180e-03 0.326819689
#>  [85,] 0.2822251 1.2030352 -0.59915657 0.26899934 -3.094839e-03 0.326874397
#>  [86,] 0.2822061 1.2031132 -0.59921272 0.26904911 -3.130066e-03 0.326924249
#>  [87,] 0.2821888 1.2031843 -0.59926388 0.26909446 -3.162165e-03 0.326969675
#>  [88,] 0.2821730 1.2032491 -0.59931050 0.26913579 -3.191415e-03 0.327011068
#>  [89,] 0.2821587 1.2033082 -0.59935299 0.26917345 -3.218068e-03 0.327048786
#>  [90,] 0.2821456 1.2033620 -0.59939170 0.26920776 -3.242354e-03 0.327083155
#>  [91,] 0.2821337 1.2034110 -0.59942697 0.26923903 -3.264483e-03 0.327114472
#>  [92,] 0.2821228 1.2034556 -0.59945911 0.26926752 -3.284648e-03 0.327143008
#>  [93,] 0.2821129 1.2034963 -0.59948840 0.26929348 -3.303022e-03 0.327169010
#>  [94,] 0.2821039 1.2035334 -0.59951508 0.26931713 -3.319764e-03 0.327192703
#>  [95,] 0.2820957 1.2035672 -0.59953940 0.26933868 -3.335019e-03 0.327214291
#>  [96,] 0.2820882 1.2035980 -0.59956156 0.26935832 -3.348920e-03 0.327233963
#>  [97,] 0.2820814 1.2036260 -0.59958175 0.26937622 -3.361585e-03 0.327251887
#>  [98,] 0.2820751 1.2036516 -0.59960014 0.26939252 -3.373126e-03 0.327268219
#>  [99,] 0.2820695 1.2036749 -0.59961690 0.26940738 -3.383642e-03 0.327283101
#> [100,] 0.2820643 1.2036961 -0.59963218 0.26942092 -3.393224e-03 0.327296661
#>              [,7]        [,8]         [,9]     [,10]
#>   [1,] -0.1513634 0.007892569  0.000000000 0.1576583
#>   [2,] -0.1746119 0.034046725  0.000000000 0.1874784
#>   [3,] -0.1968946 0.058755949  0.000000000 0.2153350
#>   [4,] -0.2181780 0.082063993  0.000000000 0.2413445
#>   [5,] -0.2384376 0.104012209  0.000000000 0.2656115
#>   [6,] -0.2576605 0.124643063  0.000000000 0.2882335
#>   [7,] -0.2758439 0.144000487  0.000000000 0.3093023
#>   [8,] -0.2929930 0.162128942  0.000000000 0.3289034
#>   [9,] -0.3091231 0.179076476  0.000000000 0.3471202
#>  [10,] -0.3246517 0.195126024 -0.001119391 0.3643264
#>  [11,] -0.3461191 0.214965465 -0.013362125 0.3852798
#>  [12,] -0.3683292 0.235199097 -0.026340760 0.4062713
#>  [13,] -0.3890888 0.254129107 -0.038402154 0.4258115
#>  [14,] -0.4084592 0.271808865 -0.049597950 0.4439786
#>  [15,] -0.4332530 0.293406287 -0.067402521 0.4701910
#>  [16,] -0.4584294 0.315120092 -0.086167448 0.4973962
#>  [17,] -0.4820599 0.335497558 -0.103750555 0.5229088
#>  [18,] -0.5041943 0.354583400 -0.120195286 0.5467884
#>  [19,] -0.5248870 0.372425834 -0.135547654 0.5690985
#>  [20,] -0.5441967 0.389076142 -0.149855877 0.5899061
#>  [21,] -0.5621846 0.404587663 -0.163169486 0.6092803
#>  [22,] -0.5789140 0.419015163 -0.175538794 0.6272918
#>  [23,] -0.5944490 0.432414063 -0.187014250 0.6440118
#>  [24,] -0.6088552 0.444840670 -0.197646726 0.6595122
#>  [25,] -0.6221952 0.456349188 -0.207484792 0.6738621
#>  [26,] -0.6345338 0.466995076 -0.216577926 0.6871318
#>  [27,] -0.6459315 0.476830461 -0.224972291 0.6993875
#>  [28,] -0.6564500 0.485908234 -0.232714612 0.7106959
#>  [29,] -0.6661457 0.494277115 -0.239847637 0.7211185
#>  [30,] -0.6750762 0.501986320 -0.246414510 0.7307173
#>  [31,] -0.6832939 0.509081093 -0.252454668 0.7395492
#>  [32,] -0.6908496 0.515605054 -0.258006053 0.7476689
#>  [33,] -0.6977914 0.521599617 -0.263104591 0.7551284
#>  [34,] -0.7041647 0.527103793 -0.267784039 0.7619765
#>  [35,] -0.7100123 0.532154513 -0.272076294 0.7682595
#>  [36,] -0.7153754 0.536787033 -0.276011784 0.7740216
#>  [37,] -0.7202900 0.541032630 -0.279617352 0.7793017
#>  [38,] -0.7247928 0.544922718 -0.282920056 0.7841391
#>  [39,] -0.7289150 0.548484306 -0.285942993 0.7885676
#>  [40,] -0.7326886 0.551744856 -0.288709767 0.7926214
#>  [41,] -0.7361410 0.554728140 -0.291240685 0.7963302
#>  [42,] -0.7392979 0.557456167 -0.293554525 0.7997214
#>  [43,] -0.7421848 0.559951044 -0.295670249 0.8028225
#>  [44,] -0.7448235 0.562231522 -0.297603819 0.8056570
#>  [45,] -0.7472339 0.564314842 -0.299369898 0.8082462
#>  [46,] -0.7494366 0.566218660 -0.300983613 0.8106122
#>  [47,] -0.7514483 0.567957537 -0.302457323 0.8127732
#>  [48,] -0.7532854 0.569545420 -0.303802901 0.8147464
#>  [49,] -0.7549625 0.570995135 -0.305031256 0.8165479
#>  [50,] -0.7564933 0.572318459 -0.306152405 0.8181922
#>  [51,] -0.7578904 0.573526198 -0.307175529 0.8196928
#>  [52,] -0.7591652 0.574628263 -0.308109049 0.8210621
#>  [53,] -0.7603290 0.575634306 -0.308961201 0.8223121
#>  [54,] -0.7613905 0.576552026 -0.309738477 0.8234524
#>  [55,] -0.7623585 0.577388892 -0.310447207 0.8244921
#>  [56,] -0.7632412 0.578152065 -0.311093479 0.8254402
#>  [57,] -0.7640468 0.578848512 -0.311683243 0.8263055
#>  [58,] -0.7645986 0.579353802 -0.312056618 0.8268993
#>  [59,] -0.7650804 0.579799278 -0.312377640 0.8274178
#>  [60,] -0.7655194 0.580205283 -0.312670179 0.8278903
#>  [61,] -0.7659193 0.580575174 -0.312936662 0.8283208
#>  [62,] -0.7663713 0.580974545 -0.313258071 0.8288066
#>  [63,] -0.7669008 0.581422056 -0.313656916 0.8293749
#>  [64,] -0.7673836 0.581830055 -0.314020532 0.8298931
#>  [65,] -0.7678237 0.582202010 -0.314352016 0.8303654
#>  [66,] -0.7682241 0.582540488 -0.314653642 0.8307952
#>  [67,] -0.7685898 0.582849581 -0.314929089 0.8311878
#>  [68,] -0.7689232 0.583131331 -0.315180164 0.8315456
#>  [69,] -0.7692270 0.583388147 -0.315409015 0.8318717
#>  [70,] -0.7695040 0.583622229 -0.315617603 0.8321690
#>  [71,] -0.7697564 0.583835581 -0.315807716 0.8324399
#>  [72,] -0.7699865 0.584030035 -0.315980986 0.8326869
#>  [73,] -0.7701961 0.584207260 -0.316138902 0.8329119
#>  [74,] -0.7703872 0.584368779 -0.316282821 0.8331170
#>  [75,] -0.7705614 0.584515981 -0.316413982 0.8333040
#>  [76,] -0.7707201 0.584650131 -0.316533512 0.8334743
#>  [77,] -0.7708647 0.584772386 -0.316642442 0.8336296
#>  [78,] -0.7709965 0.584883799 -0.316741711 0.8337711
#>  [79,] -0.7711167 0.584985329 -0.316832173 0.8339000
#>  [80,] -0.7712261 0.585077851 -0.316914609 0.8340175
#>  [81,] -0.7713259 0.585162165 -0.316989731 0.8341245
#>  [82,] -0.7714168 0.585238997 -0.317058186 0.8342221
#>  [83,] -0.7714996 0.585309010 -0.317120566 0.8343110
#>  [84,] -0.7715751 0.585372810 -0.317177409 0.8343920
#>  [85,] -0.7716439 0.585430947 -0.317229207 0.8344659
#>  [86,] -0.7717065 0.585483923 -0.317276406 0.8345331
#>  [87,] -0.7717636 0.585532196 -0.317319416 0.8345944
#>  [88,] -0.7718157 0.585576184 -0.317358606 0.8346503
#>  [89,] -0.7718631 0.585616266 -0.317394317 0.8347012
#>  [90,] -0.7719063 0.585652789 -0.317426858 0.8347476
#>  [91,] -0.7719457 0.585686070 -0.317456509 0.8347898
#>  [92,] -0.7719816 0.585716395 -0.317483527 0.8348284
#>  [93,] -0.7720143 0.585744027 -0.317508145 0.8348634
#>  [94,] -0.7720440 0.585769206 -0.317530578 0.8348954
#>  [95,] -0.7720712 0.585792148 -0.317551018 0.8349245
#>  [96,] -0.7720959 0.585813053 -0.317569643 0.8349511
#>  [97,] -0.7721185 0.585832101 -0.317586613 0.8349753
#>  [98,] -0.7721390 0.585849458 -0.317602077 0.8349973
#>  [99,] -0.7721577 0.585865273 -0.317616167 0.8350174
#> [100,] -0.7721747 0.585879683 -0.317629005 0.8350357
#> 
#> $cvm
#>   [1] -24.31803 -24.17360 -24.01424 -23.85732 -23.72185 -23.60705 -23.50693
#>   [8] -23.42670 -23.36706 -23.31907 -23.24994 -23.15667 -23.05453 -22.94711
#>  [15] -22.84113 -22.75371 -22.67820 -22.60222 -22.53465 -22.47072 -22.41130
#>  [22] -22.35944 -22.30999 -22.26634 -22.22552 -22.19279 -22.16776 -22.14810
#>  [29] -22.13131 -22.11349 -22.09688 -22.08335 -22.07230 -22.06279 -22.05460
#>  [36] -22.04707 -22.03949 -22.03288 -22.02711 -22.02205 -22.01762 -22.01374
#>  [43] -22.01032 -22.00732 -22.00467 -22.00233 -22.00027 -21.99844 -21.99682
#>  [50] -21.99537 -21.99409 -21.99295 -21.99193 -21.99102 -21.99021 -21.98948
#>  [57] -21.98882 -21.98824 -21.98771 -21.98724 -21.98681 -21.98642 -21.98608
#>  [64] -21.98576 -21.98548 -21.98522 -21.98499 -21.98478 -21.98459 -21.98442
#>  [71] -21.98426 -21.98412 -21.98400 -21.98388 -21.98378 -21.98368 -21.98359
#>  [78] -21.98351 -21.98344 -21.98337 -21.98332 -21.98326 -21.98321 -21.98317
#>  [85] -21.98313 -21.98309 -21.98306 -21.98303 -21.98300 -21.98297 -21.98295
#>  [92] -21.98293 -21.98291 -21.98289 -21.98287 -21.98286 -21.98285 -21.98283
#>  [99] -21.98282 -21.98281
#> 
#> $cvsd
#>   [1] 12.54623 12.55678 12.59010 12.63016 12.67617 12.72883 12.78780 12.85089
#>   [9] 12.92686 13.01214 13.10167 13.21001 13.29816 13.40072 13.50063 13.59698
#>  [17] 13.70386 13.82982 13.95358 14.07712 14.20034 14.31858 14.43115 14.53826
#>  [25] 14.63609 14.72383 14.80677 14.88550 14.95979 15.03308 15.10289 15.16737
#>  [33] 15.22695 15.28241 15.33392 15.38209 15.42756 15.46960 15.50842 15.54423
#>  [41] 15.57722 15.60758 15.63551 15.66119 15.68476 15.70639 15.72623 15.74441
#>  [49] 15.76107 15.77633 15.79029 15.80307 15.81475 15.82544 15.83520 15.84413
#>  [57] 15.85228 15.85972 15.86652 15.87273 15.87840 15.88357 15.88829 15.89259
#>  [65] 15.89652 15.90011 15.90337 15.90636 15.90907 15.91155 15.91381 15.91587
#>  [73] 15.91775 15.91946 15.92103 15.92245 15.92375 15.92493 15.92601 15.92699
#>  [81] 15.92789 15.92870 15.92945 15.93012 15.93074 15.93130 15.93182 15.93229
#>  [89] 15.93271 15.93310 15.93345 15.93378 15.93407 15.93434 15.93458 15.93480
#>  [97] 15.93501 15.93519 15.93536 15.93551
#> 
#> $formula
#> Surv(time, status) ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + 
#>     X9 + X10
#> <environment: 0x558b941cbf88>
#> 
#> attr(,"class")
#> [1] "cv_coxtrans"