Cross-validation for coxtrans

Cross-validation for coxtrans

Usage

cv_coxtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a coxph object fitted to the source data.

penalty

a character string specifying the penalty function. The default is "lasso". Other options are "MCP" and "SCAD".

gamma

a non-negative value specifying the penalty parameter. The default is 3.7 for SCAD and 1.5 for MCP.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_coxtrans object.

Examples

library(survival)
library(survtrans)
fit_src <- coxph(Surv(time, status) ~ ., data = sim1_src)
cv_coxtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6475046  1.4867311 -1.0656655  1.0190857 -0.8911752  1.0482614 -1.1872430 
#>         X8         X9        X10 
#>  0.9370097 -0.9985653  1.1695098 
#> 
#> $coefficients$eta
#>  [1]  0.371655865  0.423019270 -0.064908353  0.000000000  0.142054790
#>  [6]  0.000000000 -0.151363602  0.007892698  0.000000000  0.157658511
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05716259
#> 
#> $coefficients$gamma
#> [1] 1
#> 
#> 
#> $lambdas
#>   [1] 5.716259e-02 5.208442e-02 4.745739e-02 4.324140e-02 3.939996e-02
#>   [6] 3.589977e-02 3.271054e-02 2.980462e-02 2.715686e-02 2.474432e-02
#>  [11] 2.254610e-02 2.054317e-02 1.871817e-02 1.705530e-02 1.554015e-02
#>  [16] 1.415961e-02 1.290171e-02 1.175556e-02 1.071123e-02 9.759670e-03
#>  [21] 8.892648e-03 8.102649e-03 7.382832e-03 6.726962e-03 6.129357e-03
#>  [26] 5.584842e-03 5.088700e-03 4.636634e-03 4.224728e-03 3.849415e-03
#>  [31] 3.507443e-03 3.195852e-03 2.911941e-03 2.653252e-03 2.417545e-03
#>  [36] 2.202777e-03 2.007088e-03 1.828784e-03 1.666320e-03 1.518289e-03
#>  [41] 1.383408e-03 1.260510e-03 1.148530e-03 1.046497e-03 9.535295e-04
#>  [46] 8.688205e-04 7.916369e-04 7.213101e-04 6.572309e-04 5.988443e-04
#>  [51] 5.456446e-04 4.971711e-04 4.530037e-04 4.127601e-04 3.760917e-04
#>  [56] 3.426807e-04 3.122379e-04 2.844996e-04 2.592254e-04 2.361965e-04
#>  [61] 2.152135e-04 1.960945e-04 1.786740e-04 1.628011e-04 1.483383e-04
#>  [66] 1.351603e-04 1.231531e-04 1.122125e-04 1.022438e-04 9.316078e-05
#>  [71] 8.488463e-05 7.734371e-05 7.047271e-05 6.421211e-05 5.850768e-05
#>  [76] 5.331002e-05 4.857411e-05 4.425892e-05 4.032708e-05 3.674453e-05
#>  [81] 3.348025e-05 3.050595e-05 2.779589e-05 2.532658e-05 2.307664e-05
#>  [86] 2.102657e-05 1.915863e-05 1.745663e-05 1.590583e-05 1.449280e-05
#>  [91] 1.320530e-05 1.203218e-05 1.096327e-05 9.989325e-06 9.101901e-06
#>  [96] 8.293313e-06 7.556558e-06 6.885254e-06 6.273587e-06 5.716259e-06
#> 
#> $etas
#>             [,1]      [,2]        [,3]       [,4]          [,5]        [,6]
#>   [1,] 0.3716559 0.4230193 -0.06490835 0.00000000  1.420548e-01 0.000000000
#>   [2,] 0.3748949 0.4545397 -0.08728078 0.00000000  1.441319e-01 0.000000000
#>   [3,] 0.3777671 0.4847599 -0.10832054 0.00000000  1.460100e-01 0.000000000
#>   [4,] 0.3803153 0.5136365 -0.12808396 0.00000000  1.477127e-01 0.000000000
#>   [5,] 0.3825780 0.5411378 -0.14662380 0.00000000  1.492597e-01 0.000000000
#>   [6,] 0.3845885 0.5672474 -0.16399256 0.00000000  1.506669e-01 0.000000000
#>   [7,] 0.3863769 0.5919598 -0.18024026 0.00000000  1.519485e-01 0.000000000
#>   [8,] 0.3879698 0.6152811 -0.19541624 0.00000000  1.531166e-01 0.000000000
#>   [9,] 0.3893897 0.6372309 -0.20957118 0.00000000  1.541813e-01 0.000000000
#>  [10,] 0.3904511 0.6584106 -0.22299709 0.00000000  1.550077e-01 0.000000000
#>  [11,] 0.3881533 0.6866928 -0.23932546 0.00000000  1.530368e-01 0.009030421
#>  [12,] 0.3850645 0.7154095 -0.25565553 0.00000000  1.502912e-01 0.022133521
#>  [13,] 0.3821569 0.7422608 -0.27089612 0.00000000  1.477008e-01 0.034278134
#>  [14,] 0.3794263 0.7673206 -0.28509530 0.00000000  1.452631e-01 0.045524009
#>  [15,] 0.3734124 0.7971756 -0.30592028 0.01559362  1.365559e-01 0.064132879
#>  [16,] 0.3667666 0.8270855 -0.32762520 0.03458093  1.265125e-01 0.083868614
#>  [17,] 0.3605022 0.8552030 -0.34801224 0.05238004  1.170224e-01 0.102361298
#>  [18,] 0.3546128 0.8815779 -0.36712140 0.06903287  1.080809e-01 0.119656148
#>  [19,] 0.3490885 0.9062700 -0.38499957 0.08458647  9.967702e-02 0.135803417
#>  [20,] 0.3439191 0.9293395 -0.40169306 0.09908663  9.179877e-02 0.150851994
#>  [21,] 0.3390914 0.9508547 -0.41725376 0.11258322  8.442934e-02 0.164854723
#>  [22,] 0.3345920 0.9708851 -0.43173367 0.12512557  7.755065e-02 0.177863790
#>  [23,] 0.3304053 0.9895044 -0.44518799 0.13676519  7.114165e-02 0.189933386
#>  [24,] 0.3265162 1.0067860 -0.45767111 0.14755231  6.518087e-02 0.201116342
#>  [25,] 0.3229096 1.0228017 -0.46923582 0.15753534  5.964691e-02 0.211463517
#>  [26,] 0.3195696 1.0376253 -0.47993654 0.16676368  5.451666e-02 0.221026584
#>  [27,] 0.3164801 1.0513295 -0.48982661 0.17528535  4.976698e-02 0.229855759
#>  [28,] 0.3136261 1.0639837 -0.49895663 0.18314576  4.537569e-02 0.237998474
#>  [29,] 0.3109929 1.0756550 -0.50737562 0.19038861  4.132101e-02 0.245500357
#>  [30,] 0.3085655 1.0864106 -0.51513256 0.19705735  3.758062e-02 0.252406638
#>  [31,] 0.3063302 1.0963124 -0.52227253 0.20319181  3.413403e-02 0.258758836
#>  [32,] 0.3042733 1.1054216 -0.52883989 0.20883107  3.096071e-02 0.264597562
#>  [33,] 0.3023826 1.1137933 -0.53487465 0.21401029  2.804227e-02 0.269959445
#>  [34,] 0.3006457 1.1214824 -0.54041662 0.21876429  2.536006e-02 0.274880653
#>  [35,] 0.2990513 1.1285398 -0.54550273 0.22312532  2.289675e-02 0.279394689
#>  [36,] 0.2975883 1.1350144 -0.55016835 0.22712420  2.063558e-02 0.283533520
#>  [37,] 0.2962472 1.1409493 -0.55444458 0.23078799  1.856202e-02 0.287325287
#>  [38,] 0.2950180 1.1463883 -0.55836319 0.23414424  1.666083e-02 0.290798525
#>  [39,] 0.2938923 1.1513688 -0.56195111 0.23721633  1.491933e-02 0.293977528
#>  [40,] 0.2928615 1.1559290 -0.56523611 0.24002825  1.332413e-02 0.296887130
#>  [41,] 0.2919185 1.1601011 -0.56824128 0.24259998  1.186433e-02 0.299548096
#>  [42,] 0.2910555 1.1639183 -0.57099072 0.24495232  1.052823e-02 0.301981923
#>  [43,] 0.2902666 1.1674081 -0.57350409 0.24710223  9.306530e-03 0.304206242
#>  [44,] 0.2895451 1.1705991 -0.57580222 0.24906763  8.189084e-03 0.306239569
#>  [45,] 0.2888857 1.1735153 -0.57790237 0.25086340  7.167641e-03 0.308097338
#>  [46,] 0.2882833 1.1761797 -0.57982109 0.25250377  6.234220e-03 0.309794297
#>  [47,] 0.2877333 1.1786126 -0.58157295 0.25400127  5.381864e-03 0.311343434
#>  [48,] 0.2872307 1.1808351 -0.58317338 0.25536914  4.602978e-03 0.312758410
#>  [49,] 0.2867718 1.1828645 -0.58463460 0.25661787  3.891724e-03 0.314050115
#>  [50,] 0.2863529 1.1847169 -0.58596846 0.25775763  3.242363e-03 0.315229079
#>  [51,] 0.2859706 1.1864077 -0.58718583 0.25879775  2.649632e-03 0.316304958
#>  [52,] 0.2856214 1.1879515 -0.58829739 0.25974738  2.108290e-03 0.317287190
#>  [53,] 0.2853031 1.1893592 -0.58931088 0.26061314  1.614728e-03 0.318182697
#>  [54,] 0.2850125 1.1906441 -0.59023601 0.26140337  1.164099e-03 0.319000039
#>  [55,] 0.2847475 1.1918159 -0.59107962 0.26212392  7.531695e-04 0.319745311
#>  [56,] 0.2845059 1.1928845 -0.59184893 0.26278095  3.784186e-04 0.320424895
#>  [57,] 0.2842853 1.1938597 -0.59255103 0.26338056  3.633247e-05 0.321045054
#>  [58,] 0.2841761 1.1944791 -0.59299718 0.26371905  0.000000e+00 0.321432535
#>  [59,] 0.2840873 1.1950120 -0.59338115 0.26400307  0.000000e+00 0.321764830
#>  [60,] 0.2840064 1.1954977 -0.59373108 0.26426187  0.000000e+00 0.322067647
#>  [61,] 0.2839325 1.1959410 -0.59405052 0.26449806  0.000000e+00 0.322344023
#>  [62,] 0.2838217 1.1964723 -0.59443319 0.26481183 -1.307557e-04 0.322679939
#>  [63,] 0.2836614 1.1971313 -0.59490752 0.26523240 -4.284695e-04 0.323101218
#>  [64,] 0.2835152 1.1977321 -0.59533996 0.26561582 -6.998844e-04 0.323485286
#>  [65,] 0.2833819 1.1982799 -0.59573420 0.26596536 -9.473116e-04 0.323835412
#>  [66,] 0.2832604 1.1987792 -0.59609360 0.26628400 -1.172861e-03 0.324154582
#>  [67,] 0.2831496 1.1992344 -0.59642121 0.26657446 -1.378459e-03 0.324445521
#>  [68,] 0.2830490 1.1996484 -0.59671915 0.26683862 -1.565457e-03 0.324710127
#>  [69,] 0.2829570 1.2000266 -0.59699135 0.26707994 -1.736270e-03 0.324951845
#>  [70,] 0.2828731 1.2003714 -0.59723945 0.26729989 -1.891958e-03 0.325172160
#>  [71,] 0.2827967 1.2006856 -0.59746559 0.26750036 -2.033856e-03 0.325372962
#>  [72,] 0.2827270 1.2009720 -0.59767169 0.26768307 -2.163182e-03 0.325555973
#>  [73,] 0.2826635 1.2012330 -0.59785953 0.26784959 -2.281047e-03 0.325722767
#>  [74,] 0.2826056 1.2014708 -0.59803073 0.26800135 -2.388465e-03 0.325874776
#>  [75,] 0.2825529 1.2016876 -0.59818675 0.26813966 -2.486359e-03 0.326013310
#>  [76,] 0.2825048 1.2018852 -0.59832893 0.26826570 -2.575573e-03 0.326139559
#>  [77,] 0.2824610 1.2020652 -0.59845851 0.26838057 -2.656875e-03 0.326254612
#>  [78,] 0.2824211 1.2022293 -0.59857660 0.26848524 -2.730965e-03 0.326359461
#>  [79,] 0.2823847 1.2023789 -0.59868421 0.26858063 -2.798482e-03 0.326455007
#>  [80,] 0.2823516 1.2025151 -0.59878227 0.26866756 -2.860009e-03 0.326542077
#>  [81,] 0.2823214 1.2026393 -0.59887164 0.26874678 -2.916077e-03 0.326621421
#>  [82,] 0.2822939 1.2027525 -0.59895307 0.26881896 -2.967169e-03 0.326693724
#>  [83,] 0.2822688 1.2028556 -0.59902728 0.26888474 -3.013726e-03 0.326759610
#>  [84,] 0.2822459 1.2029495 -0.59909490 0.26894468 -3.056151e-03 0.326819648
#>  [85,] 0.2822251 1.2030352 -0.59915652 0.26899930 -3.094811e-03 0.326874357
#>  [86,] 0.2822061 1.2031132 -0.59921267 0.26904907 -3.130038e-03 0.326924209
#>  [87,] 0.2821888 1.2031843 -0.59926383 0.26909442 -3.162138e-03 0.326969635
#>  [88,] 0.2821731 1.2032491 -0.59931046 0.26913575 -3.191388e-03 0.327011029
#>  [89,] 0.2821587 1.2033081 -0.59935294 0.26917341 -3.218041e-03 0.327048747
#>  [90,] 0.2821456 1.2033619 -0.59939165 0.26920772 -3.242327e-03 0.327083116
#>  [91,] 0.2821337 1.2034109 -0.59942692 0.26923899 -3.264457e-03 0.327114433
#>  [92,] 0.2821228 1.2034556 -0.59945907 0.26926748 -3.284621e-03 0.327142969
#>  [93,] 0.2821129 1.2034963 -0.59948835 0.26929344 -3.302995e-03 0.327168972
#>  [94,] 0.2821039 1.2035334 -0.59951504 0.26931709 -3.319738e-03 0.327192665
#>  [95,] 0.2820957 1.2035671 -0.59953936 0.26933865 -3.334993e-03 0.327214254
#>  [96,] 0.2820882 1.2035979 -0.59956151 0.26935829 -3.348894e-03 0.327233925
#>  [97,] 0.2820814 1.2036260 -0.59958170 0.26937618 -3.361560e-03 0.327251850
#>  [98,] 0.2820752 1.2036515 -0.59960010 0.26939249 -3.373101e-03 0.327268182
#>  [99,] 0.2820695 1.2036748 -0.59961686 0.26940734 -3.383617e-03 0.327283064
#> [100,] 0.2820643 1.2036961 -0.59963213 0.26942088 -3.393199e-03 0.327296624
#>              [,7]        [,8]        [,9]     [,10]
#>   [1,] -0.1513636 0.007892698  0.00000000 0.1576585
#>   [2,] -0.1746117 0.034046561  0.00000000 0.1874782
#>   [3,] -0.1968948 0.058756129  0.00000000 0.2153353
#>   [4,] -0.2181782 0.082064195  0.00000000 0.2413448
#>   [5,] -0.2384376 0.104012194  0.00000000 0.2656115
#>   [6,] -0.2576606 0.124643188  0.00000000 0.2882336
#>   [7,] -0.2758439 0.144000448  0.00000000 0.3093023
#>   [8,] -0.2929929 0.162128856  0.00000000 0.3289033
#>   [9,] -0.3091231 0.179076489  0.00000000 0.3471202
#>  [10,] -0.3246520 0.195126218 -0.00111951 0.3643266
#>  [11,] -0.3461190 0.214965389 -0.01336207 0.3852797
#>  [12,] -0.3683289 0.235198807 -0.02634056 0.4062710
#>  [13,] -0.3890888 0.254129122 -0.03840216 0.4258116
#>  [14,] -0.4084591 0.271808781 -0.04959789 0.4439785
#>  [15,] -0.4332527 0.293406077 -0.06740232 0.4701907
#>  [16,] -0.4584298 0.315120450 -0.08616779 0.4973967
#>  [17,] -0.4820606 0.335498071 -0.10375104 0.5229095
#>  [18,] -0.5041941 0.354583269 -0.12019516 0.5467882
#>  [19,] -0.5248875 0.372426170 -0.13554797 0.5690989
#>  [20,] -0.5441970 0.389076330 -0.14985605 0.5899063
#>  [21,] -0.5621849 0.404587886 -0.16316970 0.6092806
#>  [22,] -0.5789137 0.419014922 -0.17553857 0.6272915
#>  [23,] -0.5944490 0.432414017 -0.18701421 0.6440117
#>  [24,] -0.6088553 0.444840817 -0.19764686 0.6595124
#>  [25,] -0.6221954 0.456349334 -0.20748493 0.6738623
#>  [26,] -0.6345333 0.466994649 -0.21657753 0.6871313
#>  [27,] -0.6459317 0.476830576 -0.22497240 0.6993876
#>  [28,] -0.6564500 0.485908264 -0.23271464 0.7106959
#>  [29,] -0.6661457 0.494277123 -0.23984764 0.7211185
#>  [30,] -0.6750760 0.501986190 -0.24641439 0.7307171
#>  [31,] -0.6832935 0.509080752 -0.25245435 0.7395487
#>  [32,] -0.6908497 0.515605136 -0.25800613 0.7476690
#>  [33,] -0.6977914 0.521599580 -0.26310456 0.7551283
#>  [34,] -0.7041646 0.527103761 -0.26778401 0.7619765
#>  [35,] -0.7100124 0.532154555 -0.27207633 0.7682596
#>  [36,] -0.7153755 0.536787165 -0.27601191 0.7740218
#>  [37,] -0.7202902 0.541032791 -0.27961750 0.7793019
#>  [38,] -0.7247930 0.544922900 -0.28292022 0.7841394
#>  [39,] -0.7289153 0.548484521 -0.28594319 0.7885679
#>  [40,] -0.7326889 0.551745100 -0.28870999 0.7926218
#>  [41,] -0.7361406 0.554727816 -0.29124039 0.7963298
#>  [42,] -0.7392982 0.557456460 -0.29355480 0.7997218
#>  [43,] -0.7421844 0.559950771 -0.29567000 0.8028221
#>  [44,] -0.7448232 0.562231263 -0.29760358 0.8056566
#>  [45,] -0.7472343 0.564315200 -0.29937023 0.8082467
#>  [46,] -0.7494371 0.566219037 -0.30098396 0.8106127
#>  [47,] -0.7514481 0.567957332 -0.30245714 0.8127729
#>  [48,] -0.7532851 0.569545230 -0.30380273 0.8147462
#>  [49,] -0.7549623 0.570994961 -0.30503110 0.8165476
#>  [50,] -0.7564931 0.572318301 -0.30615226 0.8181920
#>  [51,] -0.7578902 0.573526051 -0.30717539 0.8196926
#>  [52,] -0.7591658 0.574628720 -0.30810947 0.8210628
#>  [53,] -0.7603288 0.575634176 -0.30896108 0.8223120
#>  [54,] -0.7613904 0.576551896 -0.30973836 0.8234522
#>  [55,] -0.7623584 0.577388777 -0.31044710 0.8244919
#>  [56,] -0.7632411 0.578151960 -0.31109338 0.8254401
#>  [57,] -0.7640466 0.578848405 -0.31168314 0.8263053
#>  [58,] -0.7645985 0.579353695 -0.31205654 0.8268991
#>  [59,] -0.7650802 0.579799177 -0.31237757 0.8274177
#>  [60,] -0.7655193 0.580205186 -0.31267011 0.8278902
#>  [61,] -0.7659200 0.580575765 -0.31293710 0.8283215
#>  [62,] -0.7663712 0.580974465 -0.31325800 0.8288065
#>  [63,] -0.7669007 0.581421979 -0.31365685 0.8293748
#>  [64,] -0.7673835 0.581829981 -0.31402046 0.8298930
#>  [65,] -0.7678236 0.582201939 -0.31435195 0.8303653
#>  [66,] -0.7682248 0.582541021 -0.31465413 0.8307960
#>  [67,] -0.7685905 0.582850119 -0.31492958 0.8311885
#>  [68,] -0.7689231 0.583131272 -0.31518011 0.8315455
#>  [69,] -0.7692270 0.583388090 -0.31540896 0.8318716
#>  [70,] -0.7695039 0.583622173 -0.31561755 0.8321689
#>  [71,] -0.7697563 0.583835528 -0.31580767 0.8324398
#>  [72,] -0.7699864 0.584029983 -0.31598094 0.8326868
#>  [73,] -0.7701961 0.584207209 -0.31613886 0.8329119
#>  [74,] -0.7703872 0.584368730 -0.31628278 0.8331170
#>  [75,] -0.7705613 0.584515932 -0.31641394 0.8333039
#>  [76,] -0.7707200 0.584650084 -0.31653347 0.8334743
#>  [77,] -0.7708647 0.584772340 -0.31664240 0.8336295
#>  [78,] -0.7709965 0.584883753 -0.31674167 0.8337710
#>  [79,] -0.7711166 0.584985284 -0.31683213 0.8338999
#>  [80,] -0.7712261 0.585077807 -0.31691457 0.8340174
#>  [81,] -0.7713258 0.585162121 -0.31698969 0.8341245
#>  [82,] -0.7714167 0.585238954 -0.31705815 0.8342221
#>  [83,] -0.7714996 0.585308968 -0.31712053 0.8343110
#>  [84,] -0.7715750 0.585372768 -0.31717737 0.8343920
#>  [85,] -0.7716438 0.585430905 -0.31722917 0.8344658
#>  [86,] -0.7717065 0.585483882 -0.31727637 0.8345331
#>  [87,] -0.7717636 0.585532156 -0.31731938 0.8345944
#>  [88,] -0.7718156 0.585576144 -0.31735857 0.8346502
#>  [89,] -0.7718631 0.585616226 -0.31739428 0.8347011
#>  [90,] -0.7719063 0.585652750 -0.31742682 0.8347475
#>  [91,] -0.7719456 0.585686031 -0.31745647 0.8347898
#>  [92,] -0.7719815 0.585716356 -0.31748349 0.8348283
#>  [93,] -0.7720142 0.585743988 -0.31750811 0.8348634
#>  [94,] -0.7720440 0.585769167 -0.31753054 0.8348954
#>  [95,] -0.7720711 0.585792110 -0.31755098 0.8349245
#>  [96,] -0.7720959 0.585813015 -0.31756961 0.8349510
#>  [97,] -0.7721184 0.585832063 -0.31758658 0.8349752
#>  [98,] -0.7721389 0.585849420 -0.31760204 0.8349973
#>  [99,] -0.7721576 0.585865235 -0.31761613 0.8350174
#> [100,] -0.7721747 0.585879646 -0.31762897 0.8350357
#> 
#> $cvm
#>   [1] -24.30861 -24.14167 -23.96299 -23.82151 -23.70165 -23.59936 -23.51203
#>   [8] -23.44491 -23.38456 -23.32331 -23.26366 -23.18710 -23.11431 -23.04337
#>  [15] -22.97471 -22.91212 -22.86158 -22.80877 -22.76298 -22.72242 -22.68505
#>  [22] -22.65423 -22.62928 -22.60935 -22.59406 -22.58214 -22.57434 -22.56925
#>  [29] -22.56490 -22.55918 -22.55497 -22.55137 -22.54909 -22.54907 -22.55031
#>  [36] -22.55099 -22.55212 -22.55322 -22.55279 -22.55271 -22.55289 -22.55328
#>  [43] -22.55382 -22.55446 -22.55518 -22.55594 -22.55674 -22.55753 -22.55833
#>  [50] -22.55910 -22.55986 -22.56059 -22.56128 -22.56194 -22.56256 -22.56315
#>  [57] -22.56370 -22.56421 -22.56469 -22.56514 -22.56555 -22.56594 -22.56629
#>  [64] -22.56662 -22.56692 -22.56720 -22.56746 -22.56769 -22.56791 -22.56811
#>  [71] -22.56829 -22.56845 -22.56861 -22.56875 -22.56887 -22.56899 -22.56910
#>  [78] -22.56920 -22.56929 -22.56937 -22.56944 -22.56951 -22.56957 -22.56963
#>  [85] -22.56968 -22.56973 -22.56977 -22.56981 -22.56985 -22.56988 -22.56991
#>  [92] -22.56994 -22.56996 -22.56999 -22.57001 -22.57003 -22.57004 -22.57006
#>  [99] -22.57007 -22.57009
#> 
#> $cvsd
#>   [1] 12.97306 13.01010 13.03308 13.07622 13.12620 13.17995 13.23528 13.29887
#>   [9] 13.35645 13.40532 13.44125 13.47799 13.52042 13.58222 13.65794 13.74063
#>  [17] 13.82289 13.90938 13.99418 14.08059 14.16710 14.25027 14.33007 14.40633
#>  [25] 14.47851 14.54697 14.60913 14.66688 14.72222 14.77638 14.82748 14.87954
#>  [33] 14.92822 14.97481 15.01867 15.05818 15.09479 15.12829 15.15744 15.18437
#>  [41] 15.20922 15.23212 15.25321 15.27261 15.29045 15.30683 15.32187 15.33567
#>  [49] 15.34832 15.35991 15.37053 15.38025 15.38915 15.39728 15.40473 15.41153
#>  [57] 15.41775 15.42343 15.42862 15.43336 15.43768 15.44163 15.44524 15.44853
#>  [65] 15.45153 15.45427 15.45677 15.45905 15.46113 15.46303 15.46476 15.46633
#>  [73] 15.46777 15.46908 15.47028 15.47136 15.47236 15.47326 15.47409 15.47484
#>  [81] 15.47553 15.47615 15.47672 15.47724 15.47771 15.47814 15.47854 15.47889
#>  [89] 15.47922 15.47952 15.47979 15.48004 15.48026 15.48047 15.48065 15.48082
#>  [97] 15.48098 15.48112 15.48125 15.48137
#> 
#> $formula
#> Surv(time, status) ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + 
#>     X9 + X10
#> <environment: 0x562e9590a678>
#> 
#> attr(,"class")
#> [1] "cv_coxtrans"