Cross-validation for survtran

Cross-validation for survtran

Usage

cv_odetrans(
  data,
  fit_source,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

cbh_func

a function specifying the cumulative baseline hazard. The default is NULL, which means the cumulative baseline hazard is estimateds from the survode object fitted to the source data.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_odetrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.5937602  1.4181708 -1.0214486  1.0047907 -0.8320600  1.0413994 -1.1469782 
#>         X8         X9        X10 
#>  0.9215155 -0.9898359  1.1152553 
#> 
#> $coefficients$eta
#>  [1]  0.409867118  0.361169525 -0.026114251  0.000000000  0.191967110
#>  [6]  0.000000000 -0.114960303  0.000000000  0.008808969  0.112656509
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.04187974
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]          [,3]         [,4]        [,5]
#>   [1,] 0.08525555 0.02937811  0.0000000000  0.000000000 0.000000000
#>   [2,] 0.11560547 0.06612220  0.0000000000  0.000000000 0.000000000
#>   [3,] 0.14426287 0.10107474  0.0000000000  0.000000000 0.000000000
#>   [4,] 0.17117741 0.13349865  0.0000000000  0.000000000 0.003269571
#>   [5,] 0.19645868 0.16138756  0.0000000000  0.000000000 0.020975900
#>   [6,] 0.22020318 0.18782312  0.0000000000  0.000000000 0.037552409
#>   [7,] 0.24239759 0.21273070  0.0000000000  0.000000000 0.052994376
#>   [8,] 0.26305121 0.23606940  0.0000000000  0.000000000 0.067314565
#>   [9,] 0.28198608 0.25734278  0.0000000000  0.000000000 0.080509645
#>  [10,] 0.29905835 0.27615901  0.0000000000  0.000000000 0.092749554
#>  [11,] 0.31480521 0.29367296  0.0000000000  0.000000000 0.104192227
#>  [12,] 0.33105989 0.30374751  0.0000000000  0.000000000 0.117920609
#>  [13,] 0.34593016 0.31299552  0.0000000000  0.000000000 0.130478449
#>  [14,] 0.35949805 0.32150600  0.0000000000  0.000000000 0.141925626
#>  [15,] 0.37186101 0.32932164  0.0000000000  0.000000000 0.152347371
#>  [16,] 0.38306665 0.33649650 -0.0001209869  0.000000000 0.161826316
#>  [17,] 0.39038998 0.34366589 -0.0074961600  0.000000000 0.170431270
#>  [18,] 0.39708022 0.35021612 -0.0141585624  0.000000000 0.178268610
#>  [19,] 0.40356096 0.35602664 -0.0203460395  0.000000000 0.185418543
#>  [20,] 0.40986712 0.36116952 -0.0261142514  0.000000000 0.191967110
#>  [21,] 0.41564430 0.36586636 -0.0313414290  0.000000000 0.197953055
#>  [22,] 0.42102437 0.37015469 -0.0359936988  0.000000000 0.203489176
#>  [23,] 0.42640259 0.37409754 -0.0397606008  0.000000000 0.208896573
#>  [24,] 0.43130020 0.37771619 -0.0431491599  0.000000000 0.213837906
#>  [25,] 0.43575921 0.38103287 -0.0462001951  0.000000000 0.218351133
#>  [26,] 0.44027164 0.38678347 -0.0492628453 -0.005499919 0.223815928
#>  [27,] 0.44461955 0.39343965 -0.0521475189 -0.013378219 0.229532874
#>  [28,] 0.44860788 0.39961919 -0.0547689441 -0.020627022 0.234788990
#>  [29,] 0.45226343 0.40534186 -0.0571509475 -0.027287791 0.239615385
#>  [30,] 0.45561162 0.41062978 -0.0593152960 -0.033401115 0.244042414
#>  [31,] 0.45867638 0.41550672 -0.0612818729 -0.039006356 0.248099335
#>  [32,] 0.46148020 0.41999766 -0.0630688315 -0.044141469 0.251814090
#>  [33,] 0.46404404 0.42412675 -0.0646926374 -0.048842065 0.255213209
#>  [34,] 0.46638746 0.42791891 -0.0661683343 -0.053142326 0.258321649
#>  [35,] 0.46852861 0.43139813 -0.0675095730 -0.057074219 0.261162821
#>  [36,] 0.47048443 0.43458782 -0.0687287263 -0.060667909 0.263758626
#>  [37,] 0.47227026 0.43750902 -0.0698370597 -0.063950599 0.266129288
#>  [38,] 0.47390055 0.44018296 -0.0708447644 -0.066948432 0.268293682
#>  [39,] 0.47538853 0.44262920 -0.0717610923 -0.069685318 0.270269218
#>  [40,] 0.47674636 0.44486602 -0.0725944339 -0.072183338 0.272071956
#>  [41,] 0.47798522 0.44691051 -0.0733523982 -0.074462842 0.273716682
#>  [42,] 0.47911536 0.44877849 -0.0740418860 -0.076542554 0.275216991
#>  [43,] 0.48014619 0.45048468 -0.0746691559 -0.078439678 0.276585362
#>  [44,] 0.48108634 0.45204265 -0.0752398854 -0.080170003 0.277833240
#>  [45,] 0.48194368 0.45346493 -0.0757592255 -0.081748002 0.278971113
#>  [46,] 0.48272545 0.45476306 -0.0762318503 -0.083186934 0.280008582
#>  [47,] 0.48343825 0.45594767 -0.0766620040 -0.084498946 0.280954428
#>  [48,] 0.48408812 0.45702853 -0.0770535371 -0.085695139 0.281816684
#>  [49,] 0.48468058 0.45801456 -0.0774099451 -0.086785653 0.282602690
#>  [50,] 0.48522067 0.45891395 -0.0777344040 -0.087779762 0.283319151
#>  [51,] 0.48571300 0.45973424 -0.0780297990 -0.088685943 0.283972191
#>  [52,] 0.48616177 0.46048230 -0.0782987512 -0.089511933 0.284567400
#>  [53,] 0.48657081 0.46116444 -0.0785436428 -0.090264798 0.285109880
#>  [54,] 0.48694364 0.46178641 -0.0787666387 -0.090950988 0.285604288
#>  [55,] 0.48728344 0.46235349 -0.0789697071 -0.091576389 0.286054872
#>  [56,] 0.48759314 0.46287048 -0.0791546379 -0.092146368 0.286465506
#>  [57,] 0.48787539 0.46334178 -0.0793230586 -0.092665824 0.286839725
#>  [58,] 0.48813262 0.46377141 -0.0794764495 -0.093139225 0.287180752
#>  [59,] 0.48836705 0.46416304 -0.0796161571 -0.093570646 0.287491526
#>  [60,] 0.48858069 0.46452001 -0.0797434064 -0.093963802 0.287774727
#>  [61,] 0.48877538 0.46484537 -0.0798593121 -0.094322082 0.288032797
#>  [62,] 0.48895280 0.46514193 -0.0799648884 -0.094648576 0.288267965
#>  [63,] 0.48911448 0.46541221 -0.0800610584 -0.094946098 0.288482260
#>  [64,] 0.48926181 0.46565855 -0.0801486623 -0.095217218 0.288677533
#>  [65,] 0.48939607 0.46588305 -0.0802284650 -0.095464275 0.288855472
#>  [66,] 0.48951841 0.46608765 -0.0803011626 -0.095689404 0.289017614
#>  [67,] 0.48962990 0.46627411 -0.0803673890 -0.095894549 0.289165360
#>  [68,] 0.48973149 0.46644403 -0.0804277211 -0.096081482 0.289299988
#>  [69,] 0.48982406 0.46659888 -0.0804826846 -0.096251819 0.289422663
#>  [70,] 0.48990841 0.46674000 -0.0805327578 -0.096407033 0.289534444
#>  [71,] 0.48998527 0.46686859 -0.0805783764 -0.096548465 0.289636298
#>  [72,] 0.49005531 0.46698578 -0.0806199372 -0.096677339 0.289729108
#>  [73,] 0.49011913 0.46709256 -0.0806578015 -0.096794769 0.289813675
#>  [74,] 0.49017729 0.46718987 -0.0806922986 -0.096901771 0.289890732
#>  [75,] 0.49023028 0.46727854 -0.0807237281 -0.096999271 0.289960945
#>  [76,] 0.49027856 0.46735934 -0.0807523630 -0.097088112 0.290024923
#>  [77,] 0.49032255 0.46743297 -0.0807784520 -0.097169063 0.290083218
#>  [78,] 0.49036264 0.46750006 -0.0808022216 -0.097242824 0.290136335
#>  [79,] 0.49039917 0.46756120 -0.0808238782 -0.097310034 0.290184734
#>  [80,] 0.49043246 0.46761690 -0.0808436097 -0.097371275 0.290228835
#>  [81,] 0.49046278 0.46766766 -0.0808615874 -0.097427077 0.290269018
#>  [82,] 0.49049042 0.46771392 -0.0808779671 -0.097477922 0.290305632
#>  [83,] 0.49051560 0.46775606 -0.0808928911 -0.097524251 0.290338994
#>  [84,] 0.49053854 0.46779447 -0.0809064887 -0.097566465 0.290369392
#>  [85,] 0.49055945 0.46782946 -0.0809188778 -0.097604929 0.290397090
#>  [86,] 0.49057849 0.46786134 -0.0809301660 -0.097639976 0.290422328
#>  [87,] 0.49059585 0.46789040 -0.0809404510 -0.097671911 0.290445324
#>  [88,] 0.49061167 0.46791687 -0.0809498221 -0.097701009 0.290466277
#>  [89,] 0.49062608 0.46794099 -0.0809583604 -0.097727522 0.290485369
#>  [90,] 0.49063921 0.46796297 -0.0809661401 -0.097751679 0.290502764
#>  [91,] 0.49065117 0.46798300 -0.0809732284 -0.097773691 0.290518615
#>  [92,] 0.49066207 0.46800125 -0.0809796870 -0.097793748 0.290533057
#>  [93,] 0.49067200 0.46801787 -0.0809855716 -0.097812023 0.290546217
#>  [94,] 0.49068105 0.46803302 -0.0809909334 -0.097828674 0.290558208
#>  [95,] 0.49068930 0.46804683 -0.0809958188 -0.097843847 0.290569133
#>  [96,] 0.49069681 0.46805941 -0.0810002702 -0.097857671 0.290579088
#>  [97,] 0.49070366 0.46807087 -0.0810043260 -0.097870268 0.290588158
#>  [98,] 0.49070990 0.46808131 -0.0810080215 -0.097881745 0.290596423
#>  [99,] 0.49071558 0.46809083 -0.0810113887 -0.097892203 0.290603954
#> [100,] 0.49072076 0.46809950 -0.0810144567 -0.097901732 0.290610815
#>                [,6]         [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,]  0.000000000 -0.004351415 0.000000000 0.000000000 0.00000000
#>  [10,]  0.000000000 -0.017451542 0.000000000 0.000000000 0.00000000
#>  [11,]  0.000000000 -0.029568468 0.000000000 0.000000000 0.00000000
#>  [12,]  0.000000000 -0.043135167 0.000000000 0.000000000 0.01836350
#>  [13,]  0.000000000 -0.055618559 0.000000000 0.000000000 0.03484693
#>  [14,]  0.000000000 -0.067055447 0.000000000 0.000000000 0.04952993
#>  [15,]  0.000000000 -0.077509508 0.000000000 0.000000000 0.06262109
#>  [16,]  0.000000000 -0.087038465 0.000000000 0.000000000 0.07431869
#>  [17,]  0.000000000 -0.095123138 0.000000000 0.000000000 0.08555380
#>  [18,]  0.000000000 -0.102462550 0.000000000 0.000000000 0.09560258
#>  [19,]  0.000000000 -0.109050580 0.000000000 0.003027081 0.10459624
#>  [20,]  0.000000000 -0.114960303 0.000000000 0.008808969 0.11265651
#>  [21,]  0.000000000 -0.120323300 0.000000000 0.014108969 0.11989400
#>  [22,]  0.000000000 -0.125439442 0.001288170 0.018965855 0.12669267
#>  [23,]  0.000000000 -0.131381391 0.008975383 0.023441824 0.13428041
#>  [24,]  0.000000000 -0.136815326 0.015956038 0.027549637 0.14110709
#>  [25,]  0.000000000 -0.141780213 0.022295538 0.031315362 0.14725500
#>  [26,] -0.005097013 -0.146158636 0.027801388 0.036225052 0.15169948
#>  [27,] -0.011864055 -0.150168447 0.032728108 0.041447197 0.15511545
#>  [28,] -0.018076101 -0.153874335 0.037228633 0.046272451 0.15815271
#>  [29,] -0.023772265 -0.157291920 0.041337612 0.050723413 0.16085753
#>  [30,] -0.028990305 -0.160437902 0.045087446 0.054822991 0.16327011
#>  [31,] -0.033766380 -0.163329400 0.048508256 0.058593991 0.16542532
#>  [32,] -0.038134944 -0.165983442 0.051627897 0.062058837 0.16735332
#>  [33,] -0.042128180 -0.168416992 0.054472267 0.065239128 0.16908070
#>  [34,] -0.045776586 -0.170646195 0.057065068 0.068155735 0.17063026
#>  [35,] -0.049108533 -0.172686577 0.059428148 0.070828497 0.17202201
#>  [36,] -0.052150524 -0.174552834 0.061581548 0.073276259 0.17327331
#>  [37,] -0.054926662 -0.176258969 0.063543737 0.075516507 0.17439984
#>  [38,] -0.057459671 -0.177817919 0.065331517 0.077565865 0.17541492
#>  [39,] -0.059770348 -0.179241805 0.066960293 0.079439766 0.17633043
#>  [40,] -0.061877830 -0.180541879 0.068444138 0.081152565 0.17715685
#>  [41,] -0.063799695 -0.181728564 0.069795907 0.082717574 0.17790342
#>  [42,] -0.065552067 -0.182811479 0.071027328 0.084147114 0.17857836
#>  [43,] -0.067149724 -0.183799491 0.072149100 0.085452566 0.17918895
#>  [44,] -0.068606194 -0.184700757 0.073170981 0.086644419 0.17974167
#>  [45,] -0.069933852 -0.185522766 0.074101863 0.087732333 0.18024227
#>  [46,] -0.071144015 -0.186272391 0.074949851 0.088725189 0.18069592
#>  [47,] -0.072247026 -0.186955924 0.075722328 0.089631147 0.18110720
#>  [48,] -0.073252328 -0.187579133 0.076426023 0.090457694 0.18148022
#>  [49,] -0.074168535 -0.188147299 0.077067068 0.091211690 0.18181870
#>  [50,] -0.075003516 -0.188665247 0.077651047 0.091899423 0.18212594
#>  [51,] -0.075764451 -0.189137387 0.078183045 0.092526652 0.18240491
#>  [52,] -0.076457889 -0.189567748 0.078667695 0.093098647 0.18265830
#>  [53,] -0.077089806 -0.189960009 0.079109215 0.093620229 0.18288852
#>  [54,] -0.077665649 -0.190317529 0.079511448 0.094095806 0.18309773
#>  [55,] -0.078190387 -0.190643372 0.079877893 0.094529406 0.18328789
#>  [56,] -0.078668549 -0.190940338 0.080211739 0.094924712 0.18346079
#>  [57,] -0.079104264 -0.191210978 0.080515887 0.095285086 0.18361801
#>  [58,] -0.079501297 -0.191457621 0.080792983 0.095613598 0.18376100
#>  [59,] -0.079863079 -0.191682389 0.081045434 0.095913052 0.18389107
#>  [60,] -0.080192737 -0.191887219 0.081275435 0.096186009 0.18400941
#>  [61,] -0.080493123 -0.192073877 0.081484984 0.096434803 0.18411709
#>  [62,] -0.080766833 -0.192243972 0.081675901 0.096661566 0.18421508
#>  [63,] -0.081016237 -0.192398973 0.081849843 0.096868244 0.18430426
#>  [64,] -0.081243490 -0.192540217 0.082008321 0.097056610 0.18438543
#>  [65,] -0.081450561 -0.192668924 0.082152711 0.097228283 0.18445932
#>  [66,] -0.081639241 -0.192786206 0.082284265 0.097384738 0.18452659
#>  [67,] -0.081811162 -0.192893077 0.082404125 0.097527322 0.18458784
#>  [68,] -0.081967813 -0.192990459 0.082513332 0.097657262 0.18464360
#>  [69,] -0.082110550 -0.193079195 0.082612833 0.097775678 0.18469438
#>  [70,] -0.082240609 -0.193160052 0.082703490 0.097883590 0.18474062
#>  [71,] -0.082359115 -0.193233729 0.082786090 0.097981928 0.18478272
#>  [72,] -0.082467095 -0.193300863 0.082861349 0.098071541 0.18482107
#>  [73,] -0.082565483 -0.193362036 0.082929921 0.098153202 0.18485600
#>  [74,] -0.082655131 -0.193417776 0.082992398 0.098227617 0.18488781
#>  [75,] -0.082736816 -0.193468566 0.083049324 0.098295426 0.18491678
#>  [76,] -0.082811245 -0.193514845 0.083101192 0.098357217 0.18494317
#>  [77,] -0.082879062 -0.193557014 0.083148450 0.098413522 0.18496721
#>  [78,] -0.082940855 -0.193595437 0.083191509 0.098464829 0.18498910
#>  [79,] -0.082997159 -0.193630448 0.083230742 0.098511581 0.18500905
#>  [80,] -0.083048461 -0.193662349 0.083266490 0.098554183 0.18502722
#>  [81,] -0.083095206 -0.193691417 0.083299061 0.098593001 0.18504377
#>  [82,] -0.083137799 -0.193717903 0.083328738 0.098628373 0.18505885
#>  [83,] -0.083176607 -0.193742036 0.083355778 0.098660604 0.18507258
#>  [84,] -0.083211969 -0.193764025 0.083380416 0.098689972 0.18508510
#>  [85,] -0.083244189 -0.193784062 0.083402864 0.098716733 0.18509650
#>  [86,] -0.083273546 -0.193802318 0.083423319 0.098741117 0.18510689
#>  [87,] -0.083300296 -0.193818953 0.083441956 0.098763335 0.18511635
#>  [88,] -0.083324669 -0.193834110 0.083458937 0.098783581 0.18512497
#>  [89,] -0.083346877 -0.193847921 0.083474409 0.098802028 0.18513282
#>  [90,] -0.083367113 -0.193860504 0.083488507 0.098818836 0.18513998
#>  [91,] -0.083385550 -0.193871970 0.083501353 0.098834152 0.18514650
#>  [92,] -0.083402350 -0.193882418 0.083513057 0.098848108 0.18515244
#>  [93,] -0.083417657 -0.193891937 0.083523721 0.098860823 0.18515785
#>  [94,] -0.083431605 -0.193900611 0.083533438 0.098872410 0.18516278
#>  [95,] -0.083444313 -0.193908514 0.083542292 0.098882967 0.18516727
#>  [96,] -0.083455893 -0.193915716 0.083550359 0.098892587 0.18517137
#>  [97,] -0.083466443 -0.193922277 0.083557709 0.098901352 0.18517510
#>  [98,] -0.083476057 -0.193928256 0.083564407 0.098909338 0.18517849
#>  [99,] -0.083484816 -0.193933703 0.083570509 0.098916615 0.18518159
#> [100,] -0.083492798 -0.193938667 0.083576069 0.098923246 0.18518441
#> 
#> $cvm
#>   [1] -3.975746 -3.867617 -3.759561 -3.655831 -3.532870 -3.418594 -3.322376
#>   [8] -3.243515 -3.173135 -3.112331 -3.061351 -3.014275 -2.973069 -2.937584
#>  [15] -2.909217 -2.882938 -2.862527 -2.849101 -2.843470 -2.842837 -2.849734
#>  [22] -2.862608 -2.878762 -2.895044 -2.911361 -2.927897 -2.944385 -2.959518
#>  [29] -2.972874 -2.984397 -2.995178 -3.004088 -3.012730 -3.021089 -3.027592
#>  [36] -3.033074 -3.038344 -3.043375 -3.048149 -3.052658 -3.056899 -3.060873
#>  [43] -3.064585 -3.068043 -3.071257 -3.074238 -3.076582 -3.078877 -3.081059
#>  [50] -3.083071 -3.084925 -3.086631 -3.088199 -3.089586 -3.090738 -3.091796
#>  [57] -3.092766 -3.093656 -3.094471 -3.095218 -3.095902 -3.096527 -3.097099
#>  [64] -3.097622 -3.098100 -3.098537 -3.098936 -3.099300 -3.099633 -3.099936
#>  [71] -3.100214 -3.100467 -3.100698 -3.100908 -3.101100 -3.101276 -3.101436
#>  [78] -3.101581 -3.101714 -3.101836 -3.101946 -3.102047 -3.102139 -3.102223
#>  [85] -3.102299 -3.102369 -3.102432 -3.102490 -3.102543 -3.102591 -3.102635
#>  [92] -3.102674 -3.102711 -3.102744 -3.102774 -3.102802 -3.102827 -3.102850
#>  [99] -3.102871 -3.102890
#> 
#> $cvsd
#>   [1] 3.031767 3.012409 2.972252 2.931335 2.891401 2.854972 2.827880 2.805452
#>   [9] 2.776903 2.743312 2.718036 2.688311 2.659730 2.634179 2.610092 2.584378
#>  [17] 2.560953 2.537843 2.514859 2.491748 2.468577 2.451877 2.438955 2.427524
#>  [25] 2.416971 2.407746 2.399619 2.391891 2.384967 2.380225 2.375684 2.374735
#>  [33] 2.374109 2.373655 2.373303 2.372692 2.372152 2.371675 2.371256 2.370887
#>  [41] 2.370564 2.370280 2.370032 2.369813 2.369622 2.369454 2.369242 2.368448
#>  [49] 2.367656 2.366934 2.366276 2.365677 2.365130 2.364554 2.363856 2.363221
#>  [57] 2.362641 2.362114 2.361633 2.361196 2.360797 2.360434 2.360103 2.359801
#>  [65] 2.359527 2.359277 2.359049 2.358841 2.358652 2.358480 2.358323 2.358180
#>  [73] 2.358050 2.357931 2.357823 2.357725 2.357635 2.357553 2.357479 2.357411
#>  [81] 2.357349 2.357293 2.357241 2.357194 2.357152 2.357113 2.357078 2.357045
#>  [89] 2.357016 2.356989 2.356965 2.356943 2.356922 2.356904 2.356887 2.356872
#>  [97] 2.356858 2.356845 2.356834 2.356823
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x55fceba1a2b8>
#> 
#> attr(,"class")
#> [1] "cv_survtran"