Cross-validation for survtran

Cross-validation for survtran

Usage

cv_odetrans(
  data,
  fit_source,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

cbh_func

a function specifying the cumulative baseline hazard. The default is NULL, which means the cumulative baseline hazard is estimated from the survode object fitted to the source data.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_odetrans object.

Examples

formula <- Surv(time, status) ~ . - group - id
fit_src <- survode(formula, data = sim1[sim1$group == 1, ], df = 10)
cv_fit <- cv_odetrans(sim1[sim1$group == 2, ], fit_src, lambda = 0.1)