Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

cbh_func

a function specifying the cumulative baseline hazard. The default is NULL, which means the cumulative baseline hazard is estimateds from the survode object fitted to the source data.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6065471  1.4072174 -1.0094929  1.0047907 -0.8457585  1.0413994 -1.1344804 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0982014 
#> 
#> $coefficients$eta
#>  [1]  0.39708024  0.35021608 -0.01415856  0.00000000  0.17826862  0.00000000
#>  [7] -0.10246256  0.00000000  0.00000000  0.09560259
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05044429
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]          [,3]         [,4]        [,5]
#>   [1,] 0.08525555 0.02937811  0.0000000000  0.000000000 0.000000000
#>   [2,] 0.11560547 0.06612220  0.0000000000  0.000000000 0.000000000
#>   [3,] 0.14426287 0.10107474  0.0000000000  0.000000000 0.000000000
#>   [4,] 0.17117741 0.13349865  0.0000000000  0.000000000 0.003269571
#>   [5,] 0.19645868 0.16138757  0.0000000000  0.000000000 0.020975899
#>   [6,] 0.22020318 0.18782313  0.0000000000  0.000000000 0.037552408
#>   [7,] 0.24239759 0.21273070  0.0000000000  0.000000000 0.052994375
#>   [8,] 0.26305121 0.23606940  0.0000000000  0.000000000 0.067314565
#>   [9,] 0.28198608 0.25734278  0.0000000000  0.000000000 0.080509644
#>  [10,] 0.29905835 0.27615901  0.0000000000  0.000000000 0.092749554
#>  [11,] 0.31480521 0.29367296  0.0000000000  0.000000000 0.104192227
#>  [12,] 0.33105988 0.30374760  0.0000000000  0.000000000 0.117920575
#>  [13,] 0.34593014 0.31299567  0.0000000000  0.000000000 0.130478392
#>  [14,] 0.35949801 0.32150622  0.0000000000  0.000000000 0.141925544
#>  [15,] 0.37186099 0.32932172  0.0000000000  0.000000000 0.152347342
#>  [16,] 0.38306673 0.33649631 -0.0001209789  0.000000000 0.161826386
#>  [17,] 0.39038998 0.34366591 -0.0074961589  0.000000000 0.170431262
#>  [18,] 0.39708024 0.35021608 -0.0141585597  0.000000000 0.178268622
#>  [19,] 0.40356097 0.35602660 -0.0203460405  0.000000000 0.185418557
#>  [20,] 0.40986722 0.36116951 -0.0261142301  0.000000000 0.191967110
#>  [21,] 0.41564432 0.36586631 -0.0313414304  0.000000000 0.197953072
#>  [22,] 0.42102439 0.37015466 -0.0359936884  0.000000000 0.203489190
#>  [23,] 0.42640250 0.37409765 -0.0397606264  0.000000000 0.208896524
#>  [24,] 0.43130031 0.37771603 -0.0431491078  0.000000000 0.213837977
#>  [25,] 0.43575924 0.38103283 -0.0462001802  0.000000000 0.218351152
#>  [26,] 0.44027165 0.38678346 -0.0492628434 -0.005499917 0.223815930
#>  [27,] 0.44461956 0.39343959 -0.0521475089 -0.013378203 0.229532884
#>  [28,] 0.44860790 0.39961912 -0.0547689311 -0.020626998 0.234789003
#>  [29,] 0.45226345 0.40534177 -0.0571509317 -0.027287761 0.239615401
#>  [30,] 0.45561167 0.41062992 -0.0593152815 -0.033401228 0.244042435
#>  [31,] 0.45867645 0.41550685 -0.0612818536 -0.039006467 0.248099362
#>  [32,] 0.46148047 0.41999785 -0.0630687254 -0.044141759 0.251814220
#>  [33,] 0.46404433 0.42412684 -0.0646925154 -0.048842327 0.255213355
#>  [34,] 0.46638777 0.42791891 -0.0661681970 -0.053142556 0.258321811
#>  [35,] 0.46852894 0.43139797 -0.0675094134 -0.057074384 0.261163001
#>  [36,] 0.47048439 0.43458715 -0.0687286885 -0.060667546 0.263758624
#>  [37,] 0.47227021 0.43750837 -0.0698370237 -0.063950244 0.266129283
#>  [38,] 0.47390051 0.44018233 -0.0708447300 -0.066948085 0.268293675
#>  [39,] 0.47538848 0.44262859 -0.0717610595 -0.069684981 0.270269209
#>  [40,] 0.47674631 0.44486544 -0.0725944027 -0.072183010 0.272071945
#>  [41,] 0.47798517 0.44690994 -0.0733523685 -0.074462523 0.273716670
#>  [42,] 0.47911531 0.44877795 -0.0740418576 -0.076542245 0.275216977
#>  [43,] 0.48014614 0.45048416 -0.0746691288 -0.078439378 0.276585346
#>  [44,] 0.48108629 0.45204214 -0.0752398596 -0.080169712 0.277833224
#>  [45,] 0.48194363 0.45346444 -0.0757592008 -0.081747719 0.278971097
#>  [46,] 0.48272540 0.45476259 -0.0762318267 -0.083186660 0.280008564
#>  [47,] 0.48343820 0.45594720 -0.0766619795 -0.084498669 0.280954411
#>  [48,] 0.48408807 0.45702803 -0.0770535106 -0.085694850 0.281816667
#>  [49,] 0.48468053 0.45801403 -0.0774099169 -0.086785351 0.282602673
#>  [50,] 0.48522062 0.45891341 -0.0777343743 -0.087779450 0.283319135
#>  [51,] 0.48571295 0.45973368 -0.0780297678 -0.088685621 0.283972174
#>  [52,] 0.48616171 0.46048172 -0.0782987188 -0.089511602 0.284567383
#>  [53,] 0.48657076 0.46116384 -0.0785436091 -0.090264458 0.285109863
#>  [54,] 0.48694358 0.46178580 -0.0787666041 -0.090950640 0.285604271
#>  [55,] 0.48728339 0.46235286 -0.0789696718 -0.091576032 0.286054854
#>  [56,] 0.48759308 0.46286984 -0.0791546021 -0.092146003 0.286465488
#>  [57,] 0.48787533 0.46334112 -0.0793230223 -0.092665451 0.286839707
#>  [58,] 0.48813256 0.46377074 -0.0794764127 -0.093138846 0.287180733
#>  [59,] 0.48836699 0.46416236 -0.0796161199 -0.093570260 0.287491507
#>  [60,] 0.48858063 0.46451932 -0.0797433688 -0.093963411 0.287774707
#>  [61,] 0.48877531 0.46484468 -0.0798592741 -0.094321686 0.288032777
#>  [62,] 0.48895273 0.46514122 -0.0799648500 -0.094648174 0.288267944
#>  [63,] 0.48911441 0.46541150 -0.0800610197 -0.094945693 0.288482239
#>  [64,] 0.48926174 0.46565783 -0.0801486234 -0.095216808 0.288677512
#>  [65,] 0.48939600 0.46588232 -0.0802284259 -0.095463862 0.288855451
#>  [66,] 0.48951835 0.46608692 -0.0803011232 -0.095688987 0.289017592
#>  [67,] 0.48962983 0.46627337 -0.0803673494 -0.095894129 0.289165338
#>  [68,] 0.48973142 0.46644329 -0.0804276814 -0.096081059 0.289299966
#>  [69,] 0.48982399 0.46659814 -0.0804826446 -0.096251394 0.289422640
#>  [70,] 0.48990834 0.46673925 -0.0805327177 -0.096406605 0.289534421
#>  [71,] 0.48998521 0.46686784 -0.0805783361 -0.096548036 0.289636276
#>  [72,] 0.49005524 0.46698502 -0.0806198968 -0.096676908 0.289729085
#>  [73,] 0.49011906 0.46709180 -0.0806577611 -0.096794336 0.289813652
#>  [74,] 0.49017722 0.46718911 -0.0806922580 -0.096901337 0.289890709
#>  [75,] 0.49023021 0.46727778 -0.0807236874 -0.096998835 0.289960922
#>  [76,] 0.49027849 0.46735858 -0.0807523222 -0.097087675 0.290024899
#>  [77,] 0.49032249 0.46743220 -0.0807784111 -0.097168624 0.290083194
#>  [78,] 0.49036257 0.46749929 -0.0808021807 -0.097242385 0.290136312
#>  [79,] 0.49039910 0.46756042 -0.0808238372 -0.097309594 0.290184711
#>  [80,] 0.49043239 0.46761613 -0.0808435686 -0.097370834 0.290228811
#>  [81,] 0.49046271 0.46766689 -0.0808615462 -0.097426635 0.290268994
#>  [82,] 0.49049035 0.46771314 -0.0808779259 -0.097477479 0.290305608
#>  [83,] 0.49051553 0.46775528 -0.0808928499 -0.097523807 0.290338970
#>  [84,] 0.49053847 0.46779369 -0.0809064474 -0.097566020 0.290369368
#>  [85,] 0.49055938 0.46782868 -0.0809188365 -0.097604484 0.290397066
#>  [86,] 0.49057842 0.46786056 -0.0809301246 -0.097639531 0.290422304
#>  [87,] 0.49059578 0.46788961 -0.0809404096 -0.097671465 0.290445300
#>  [88,] 0.49061160 0.46791609 -0.0809497807 -0.097700562 0.290466253
#>  [89,] 0.49062601 0.46794021 -0.0809583190 -0.097727075 0.290485345
#>  [90,] 0.49063914 0.46796218 -0.0809660986 -0.097751232 0.290502740
#>  [91,] 0.49065110 0.46798221 -0.0809731870 -0.097773244 0.290518591
#>  [92,] 0.49066200 0.46800046 -0.0809796455 -0.097793300 0.290533033
#>  [93,] 0.49067193 0.46801709 -0.0809855301 -0.097811575 0.290546193
#>  [94,] 0.49068098 0.46803224 -0.0809908919 -0.097828226 0.290558183
#>  [95,] 0.49068923 0.46804604 -0.0809957773 -0.097843398 0.290569109
#>  [96,] 0.49069674 0.46805862 -0.0810002286 -0.097857223 0.290579064
#>  [97,] 0.49070359 0.46807008 -0.0810042845 -0.097869819 0.290588134
#>  [98,] 0.49070983 0.46808052 -0.0810079799 -0.097881296 0.290596399
#>  [99,] 0.49071551 0.46809004 -0.0810113471 -0.097891754 0.290603929
#> [100,] 0.49072069 0.46809871 -0.0810144151 -0.097901283 0.290610791
#>                [,6]         [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,]  0.000000000 -0.004351414 0.000000000 0.000000000 0.00000000
#>  [10,]  0.000000000 -0.017451541 0.000000000 0.000000000 0.00000000
#>  [11,]  0.000000000 -0.029568468 0.000000000 0.000000000 0.00000000
#>  [12,]  0.000000000 -0.043135141 0.000000000 0.000000000 0.01836346
#>  [13,]  0.000000000 -0.055618515 0.000000000 0.000000000 0.03484688
#>  [14,]  0.000000000 -0.067055382 0.000000000 0.000000000 0.04952984
#>  [15,]  0.000000000 -0.077509485 0.000000000 0.000000000 0.06262106
#>  [16,]  0.000000000 -0.087038519 0.000000000 0.000000000 0.07431877
#>  [17,]  0.000000000 -0.095123132 0.000000000 0.000000000 0.08555380
#>  [18,]  0.000000000 -0.102462560 0.000000000 0.000000000 0.09560259
#>  [19,]  0.000000000 -0.109050591 0.000000000 0.003027081 0.10459625
#>  [20,]  0.000000000 -0.114960299 0.000000000 0.008808980 0.11265652
#>  [21,]  0.000000000 -0.120323314 0.000000000 0.014108970 0.11989401
#>  [22,]  0.000000000 -0.125439460 0.001288183 0.018965856 0.12669268
#>  [23,]  0.000000000 -0.131381327 0.008975331 0.023441818 0.13428035
#>  [24,]  0.000000000 -0.136815421 0.015956107 0.027549643 0.14110717
#>  [25,]  0.000000000 -0.141780239 0.022295557 0.031315364 0.14725502
#>  [26,] -0.005097011 -0.146158640 0.027801391 0.036225051 0.15169949
#>  [27,] -0.011864041 -0.150168468 0.032728122 0.041447194 0.15511547
#>  [28,] -0.018076080 -0.153874363 0.037228652 0.046272445 0.15815274
#>  [29,] -0.023772240 -0.157291953 0.041337635 0.050723406 0.16085757
#>  [30,] -0.028990358 -0.160437887 0.045087436 0.054823020 0.16327006
#>  [31,] -0.033766429 -0.163329394 0.048508252 0.058594021 0.16542528
#>  [32,] -0.038135044 -0.165983506 0.051627941 0.062058917 0.16735327
#>  [33,] -0.042128255 -0.168417090 0.054472333 0.065239200 0.16908068
#>  [34,] -0.045776635 -0.170646327 0.057065155 0.068155799 0.17063028
#>  [35,] -0.049108536 -0.172686761 0.059428268 0.070828544 0.17202209
#>  [36,] -0.052150313 -0.174552977 0.061581637 0.073276159 0.17327355
#>  [37,] -0.054926456 -0.176259105 0.063543821 0.075516409 0.17440007
#>  [38,] -0.057459471 -0.177818049 0.065331597 0.077565769 0.17541515
#>  [39,] -0.059770154 -0.179241928 0.066960368 0.079439672 0.17633065
#>  [40,] -0.061877641 -0.180541998 0.068444210 0.081152472 0.17715706
#>  [41,] -0.063799512 -0.181728676 0.069795975 0.082717484 0.17790362
#>  [42,] -0.065551890 -0.182811586 0.071027392 0.084147026 0.17857856
#>  [43,] -0.067149552 -0.183799594 0.072149161 0.085452480 0.17918914
#>  [44,] -0.068606027 -0.184700854 0.073171038 0.086644335 0.17974185
#>  [45,] -0.069933691 -0.185522859 0.074101917 0.087732251 0.18024245
#>  [46,] -0.071143859 -0.186272480 0.074949903 0.088725109 0.18069609
#>  [47,] -0.072246868 -0.186956014 0.075722381 0.089631067 0.18110736
#>  [48,] -0.073252163 -0.187579229 0.076426079 0.090457610 0.18148040
#>  [49,] -0.074168362 -0.188147401 0.077067127 0.091211602 0.18181889
#>  [50,] -0.075003337 -0.188665354 0.077651109 0.091899332 0.18212613
#>  [51,] -0.075764265 -0.189137499 0.078183110 0.092526558 0.18240511
#>  [52,] -0.076457698 -0.189567864 0.078667762 0.093098550 0.18265851
#>  [53,] -0.077089609 -0.189960128 0.079109284 0.093620129 0.18288873
#>  [54,] -0.077665447 -0.190317651 0.079511519 0.094095703 0.18309794
#>  [55,] -0.078190180 -0.190643498 0.079877966 0.094529301 0.18328812
#>  [56,] -0.078668338 -0.190940467 0.080211813 0.094924604 0.18346101
#>  [57,] -0.079104049 -0.191211110 0.080515963 0.095284976 0.18361824
#>  [58,] -0.079501078 -0.191457754 0.080793060 0.095613485 0.18376124
#>  [59,] -0.079862856 -0.191682524 0.081045513 0.095912938 0.18389131
#>  [60,] -0.080192511 -0.191887356 0.081275515 0.096185893 0.18400966
#>  [61,] -0.080492894 -0.192074016 0.081485064 0.096434685 0.18411734
#>  [62,] -0.080766602 -0.192244113 0.081675982 0.096661447 0.18421533
#>  [63,] -0.081016003 -0.192399115 0.081849925 0.096868124 0.18430451
#>  [64,] -0.081243254 -0.192540361 0.082008404 0.097056489 0.18438569
#>  [65,] -0.081450323 -0.192669069 0.082152794 0.097228160 0.18445958
#>  [66,] -0.081639000 -0.192786352 0.082284349 0.097384614 0.18452685
#>  [67,] -0.081810920 -0.192893224 0.082404210 0.097527197 0.18458810
#>  [68,] -0.081967570 -0.192990607 0.082513417 0.097657137 0.18464387
#>  [69,] -0.082110305 -0.193079343 0.082612918 0.097775552 0.18469464
#>  [70,] -0.082240363 -0.193160201 0.082703576 0.097883463 0.18474088
#>  [71,] -0.082358868 -0.193233879 0.082786176 0.097981800 0.18478299
#>  [72,] -0.082466846 -0.193301014 0.082861436 0.098071413 0.18482134
#>  [73,] -0.082565233 -0.193362187 0.082930008 0.098153074 0.18485627
#>  [74,] -0.082654881 -0.193417928 0.082992486 0.098227487 0.18488808
#>  [75,] -0.082736565 -0.193468718 0.083049412 0.098295296 0.18491705
#>  [76,] -0.082810993 -0.193514998 0.083101279 0.098357086 0.18494344
#>  [77,] -0.082878810 -0.193557167 0.083148538 0.098413392 0.18496748
#>  [78,] -0.082940602 -0.193595591 0.083191597 0.098464698 0.18498938
#>  [79,] -0.082996906 -0.193630602 0.083230831 0.098511450 0.18500932
#>  [80,] -0.083048207 -0.193662504 0.083266578 0.098554051 0.18502749
#>  [81,] -0.083094952 -0.193691572 0.083299149 0.098592869 0.18504405
#>  [82,] -0.083137544 -0.193718057 0.083328826 0.098628241 0.18505912
#>  [83,] -0.083176352 -0.193742191 0.083355867 0.098660471 0.18507286
#>  [84,] -0.083211713 -0.193764181 0.083380505 0.098689840 0.18508538
#>  [85,] -0.083243932 -0.193784217 0.083402954 0.098716600 0.18509678
#>  [86,] -0.083273290 -0.193802474 0.083423408 0.098740984 0.18510716
#>  [87,] -0.083300039 -0.193819109 0.083442045 0.098763202 0.18511663
#>  [88,] -0.083324412 -0.193834266 0.083459026 0.098783448 0.18512525
#>  [89,] -0.083346620 -0.193848077 0.083474499 0.098801895 0.18513310
#>  [90,] -0.083366855 -0.193860661 0.083488597 0.098818703 0.18514026
#>  [91,] -0.083385293 -0.193872127 0.083501442 0.098834019 0.18514678
#>  [92,] -0.083402092 -0.193882574 0.083513146 0.098847974 0.18515272
#>  [93,] -0.083417400 -0.193892094 0.083523811 0.098860690 0.18515813
#>  [94,] -0.083431347 -0.193900768 0.083533528 0.098872276 0.18516306
#>  [95,] -0.083444055 -0.193908671 0.083542381 0.098882833 0.18516755
#>  [96,] -0.083455635 -0.193915872 0.083550448 0.098892453 0.18517165
#>  [97,] -0.083466185 -0.193922434 0.083557799 0.098901218 0.18517538
#>  [98,] -0.083475799 -0.193928412 0.083564496 0.098909204 0.18517877
#>  [99,] -0.083484558 -0.193933860 0.083570599 0.098916481 0.18518187
#> [100,] -0.083492539 -0.193938824 0.083576159 0.098923112 0.18518469
#> 
#> $cvm
#>   [1] -3.993332 -3.894951 -3.802526 -3.677765 -3.550553 -3.445412 -3.352596
#>   [8] -3.268037 -3.198323 -3.143405 -3.095355 -3.056812 -3.020265 -2.982587
#>  [15] -2.950562 -2.926088 -2.909003 -2.903660 -2.906856 -2.920165 -2.938443
#>  [22] -2.961584 -2.985909 -3.010305 -3.031671 -3.049812 -3.066248 -3.080234
#>  [29] -3.090530 -3.098124 -3.102989 -3.108000 -3.113166 -3.118487 -3.124224
#>  [36] -3.129778 -3.134051 -3.138191 -3.142168 -3.145964 -3.149565 -3.152964
#>  [43] -3.156160 -3.159155 -3.161951 -3.164557 -3.166978 -3.169223 -3.171302
#>  [50] -3.173224 -3.174997 -3.176632 -3.178137 -3.179522 -3.180794 -3.181963
#>  [57] -3.183035 -3.184018 -3.184919 -3.185744 -3.186500 -3.187191 -3.187824
#>  [64] -3.188402 -3.188930 -3.189413 -3.189855 -3.190258 -3.190626 -3.190962
#>  [71] -3.191269 -3.191549 -3.191804 -3.192037 -3.192250 -3.192444 -3.192621
#>  [78] -3.192783 -3.192930 -3.193064 -3.193186 -3.193298 -3.193400 -3.193492
#>  [85] -3.193577 -3.193654 -3.193724 -3.193788 -3.193847 -3.193900 -3.193948
#>  [92] -3.193992 -3.194033 -3.194069 -3.194103 -3.194133 -3.194161 -3.194186
#>  [99] -3.194209 -3.194230
#> 
#> $cvsd
#>   [1] 1.894587 1.834778 1.773851 1.724251 1.685235 1.666254 1.656698 1.659705
#>   [9] 1.668257 1.684130 1.694208 1.703346 1.710917 1.725220 1.742333 1.758367
#>  [17] 1.773121 1.786328 1.804734 1.835762 1.865731 1.895419 1.923980 1.951867
#>  [25] 1.976770 1.996932 2.016370 2.035843 2.055566 2.074736 2.093426 2.110672
#>  [33] 2.126581 2.141299 2.155102 2.167804 2.178964 2.189268 2.198767 2.207517
#>  [41] 2.215567 2.222967 2.229764 2.236002 2.241724 2.246968 2.251772 2.256171
#>  [49] 2.260197 2.263880 2.267248 2.270327 2.273141 2.275712 2.278061 2.280206
#>  [57] 2.282164 2.283952 2.285584 2.287073 2.288431 2.289671 2.290802 2.291833
#>  [65] 2.292774 2.293632 2.294414 2.295127 2.295778 2.296371 2.296911 2.297404
#>  [73] 2.297854 2.298263 2.298636 2.298977 2.299287 2.299569 2.299827 2.300062
#>  [81] 2.300276 2.300471 2.300648 2.300810 2.300958 2.301092 2.301215 2.301326
#>  [89] 2.301428 2.301521 2.301605 2.301682 2.301753 2.301817 2.301875 2.301928
#>  [97] 2.301976 2.302020 2.302061 2.302097
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x55f11b7690f0>
#> 
#> attr(,"class")
#> [1] "cv_survtran"