Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

cbh_func

a function specifying the cumulative baseline hazard. The default is NULL, which means the cumulative baseline hazard is estimateds from the survode object fitted to the source data.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6205606  1.3934976 -0.9954553  1.0047907 -0.8622007  1.0413994 -1.1190564 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0769175 
#> 
#> $coefficients$eta
#>  [1]  0.3830667260  0.3364963324 -0.0001209766  0.0000000000  0.1618263763
#>  [6]  0.0000000000 -0.0870385118  0.0000000000  0.0000000000  0.0743187586
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.06076032
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]          [,3]         [,4]        [,5]
#>   [1,] 0.08525555 0.02937811  0.0000000000  0.000000000 0.000000000
#>   [2,] 0.11560547 0.06612220  0.0000000000  0.000000000 0.000000000
#>   [3,] 0.14426287 0.10107474  0.0000000000  0.000000000 0.000000000
#>   [4,] 0.17117741 0.13349865  0.0000000000  0.000000000 0.003269571
#>   [5,] 0.19645868 0.16138757  0.0000000000  0.000000000 0.020975899
#>   [6,] 0.22020318 0.18782313  0.0000000000  0.000000000 0.037552408
#>   [7,] 0.24239759 0.21273070  0.0000000000  0.000000000 0.052994375
#>   [8,] 0.26305121 0.23606940  0.0000000000  0.000000000 0.067314565
#>   [9,] 0.28198608 0.25734278  0.0000000000  0.000000000 0.080509645
#>  [10,] 0.29905835 0.27615901  0.0000000000  0.000000000 0.092749554
#>  [11,] 0.31480521 0.29367296  0.0000000000  0.000000000 0.104192227
#>  [12,] 0.33105988 0.30374756  0.0000000000  0.000000000 0.117920588
#>  [13,] 0.34593013 0.31299569  0.0000000000  0.000000000 0.130478385
#>  [14,] 0.35949803 0.32150615  0.0000000000  0.000000000 0.141925568
#>  [15,] 0.37186101 0.32932163  0.0000000000  0.000000000 0.152347374
#>  [16,] 0.38306673 0.33649633 -0.0001209766  0.000000000 0.161826376
#>  [17,] 0.39039005 0.34366575 -0.0074961483  0.000000000 0.170431322
#>  [18,] 0.39708030 0.35021599 -0.0141585485  0.000000000 0.178268654
#>  [19,] 0.40356098 0.35602671 -0.0203460267  0.000000000 0.185418518
#>  [20,] 0.40986723 0.36116961 -0.0261142168  0.000000000 0.191967074
#>  [21,] 0.41564433 0.36586641 -0.0313414176  0.000000000 0.197953038
#>  [22,] 0.42102432 0.37015474 -0.0359937173  0.000000000 0.203489154
#>  [23,] 0.42640240 0.37409773 -0.0397606832  0.000000000 0.208896491
#>  [24,] 0.43129996 0.37771644 -0.0431492635  0.000000000 0.213837795
#>  [25,] 0.43575917 0.38103291 -0.0462002101  0.000000000 0.218351114
#>  [26,] 0.44027160 0.38678371 -0.0492628810 -0.005500001 0.223815888
#>  [27,] 0.44461956 0.39343950 -0.0521475023 -0.013378161 0.229532891
#>  [28,] 0.44860790 0.39961901 -0.0547689233 -0.020626950 0.234789011
#>  [29,] 0.45226343 0.40534186 -0.0571509430 -0.027287796 0.239615391
#>  [30,] 0.45561166 0.41062979 -0.0593152744 -0.033401180 0.244042454
#>  [31,] 0.45867647 0.41550630 -0.0612817990 -0.039006247 0.248099422
#>  [32,] 0.46148024 0.41999695 -0.0630687602 -0.044141160 0.251814146
#>  [33,] 0.46404408 0.42412597 -0.0646925592 -0.048841725 0.255213267
#>  [34,] 0.46638746 0.42791842 -0.0661682981 -0.053142084 0.258321668
#>  [35,] 0.46852861 0.43139763 -0.0675095365 -0.057073971 0.261162838
#>  [36,] 0.47048432 0.43458686 -0.0687287018 -0.060667348 0.263758593
#>  [37,] 0.47227014 0.43750807 -0.0698370374 -0.063950038 0.266129250
#>  [38,] 0.47390043 0.44018203 -0.0708447442 -0.066947874 0.268293641
#>  [39,] 0.47538840 0.44262828 -0.0717610740 -0.069684764 0.270269173
#>  [40,] 0.47674623 0.44486513 -0.0725944184 -0.072182792 0.272071907
#>  [41,] 0.47798508 0.44690964 -0.0733523855 -0.074462305 0.273716630
#>  [42,] 0.47911522 0.44877765 -0.0740418758 -0.076542027 0.275216936
#>  [43,] 0.48014605 0.45048386 -0.0746691481 -0.078439160 0.276585304
#>  [44,] 0.48108620 0.45204185 -0.0752398798 -0.080169493 0.277833180
#>  [45,] 0.48194354 0.45346415 -0.0757592226 -0.081747503 0.278971051
#>  [46,] 0.48272531 0.45476231 -0.0762318500 -0.083186448 0.280008518
#>  [47,] 0.48343811 0.45594693 -0.0766620043 -0.084498461 0.280954363
#>  [48,] 0.48408798 0.45702777 -0.0770535367 -0.085694645 0.281816618
#>  [49,] 0.48468043 0.45801378 -0.0774099442 -0.086785149 0.282602623
#>  [50,] 0.48522052 0.45891317 -0.0777344027 -0.087779251 0.283319083
#>  [51,] 0.48571285 0.45973344 -0.0780297973 -0.088685424 0.283972122
#>  [52,] 0.48616161 0.46048149 -0.0782987491 -0.089511407 0.284567330
#>  [53,] 0.48657065 0.46116362 -0.0785436403 -0.090264267 0.285109810
#>  [54,] 0.48694348 0.46178558 -0.0787666360 -0.090950450 0.285604217
#>  [55,] 0.48728328 0.46235265 -0.0789697045 -0.091575844 0.286054800
#>  [56,] 0.48759298 0.46286963 -0.0791546353 -0.092145817 0.286465433
#>  [57,] 0.48787523 0.46334092 -0.0793230561 -0.092665267 0.286839651
#>  [58,] 0.48813246 0.46377054 -0.0794764471 -0.093138663 0.287180678
#>  [59,] 0.48836688 0.46416216 -0.0796161547 -0.093570079 0.287491451
#>  [60,] 0.48858052 0.46451912 -0.0797434040 -0.093963231 0.287774650
#>  [61,] 0.48877521 0.46484448 -0.0798593097 -0.094321507 0.288032720
#>  [62,] 0.48895263 0.46514103 -0.0799648860 -0.094647996 0.288267887
#>  [63,] 0.48911430 0.46541131 -0.0800610561 -0.094945516 0.288482182
#>  [64,] 0.48926164 0.46565764 -0.0801486601 -0.095216632 0.288677455
#>  [65,] 0.48939590 0.46588214 -0.0802284628 -0.095463687 0.288855393
#>  [66,] 0.48951824 0.46608674 -0.0803011604 -0.095688813 0.289017535
#>  [67,] 0.48962972 0.46627319 -0.0803673868 -0.095893955 0.289165281
#>  [68,] 0.48973131 0.46644311 -0.0804277190 -0.096080886 0.289299908
#>  [69,] 0.48982388 0.46659796 -0.0804826824 -0.096251222 0.289422582
#>  [70,] 0.48990823 0.46673907 -0.0805327556 -0.096406434 0.289534363
#>  [71,] 0.48998510 0.46686766 -0.0805783742 -0.096547864 0.289636217
#>  [72,] 0.49005514 0.46698485 -0.0806199351 -0.096676737 0.289729027
#>  [73,] 0.49011896 0.46709163 -0.0806577995 -0.096794166 0.289813594
#>  [74,] 0.49017711 0.46718894 -0.0806922965 -0.096901166 0.289890651
#>  [75,] 0.49023010 0.46727761 -0.0807237260 -0.096998665 0.289960864
#>  [76,] 0.49027838 0.46735841 -0.0807523609 -0.097087505 0.290024841
#>  [77,] 0.49032238 0.46743203 -0.0807784499 -0.097168455 0.290083136
#>  [78,] 0.49036246 0.46749912 -0.0808022196 -0.097242216 0.290136253
#>  [79,] 0.49039899 0.46756026 -0.0808238762 -0.097309425 0.290184652
#>  [80,] 0.49043228 0.46761596 -0.0808436077 -0.097370665 0.290228752
#>  [81,] 0.49046260 0.46766672 -0.0808615853 -0.097426466 0.290268935
#>  [82,] 0.49049024 0.46771297 -0.0808779651 -0.097477311 0.290305549
#>  [83,] 0.49051542 0.46775512 -0.0808928891 -0.097523639 0.290338911
#>  [84,] 0.49053836 0.46779352 -0.0809064867 -0.097565852 0.290369309
#>  [85,] 0.49055927 0.46782851 -0.0809188758 -0.097604316 0.290397007
#>  [86,] 0.49057831 0.46786040 -0.0809301640 -0.097639363 0.290422245
#>  [87,] 0.49059567 0.46788945 -0.0809404490 -0.097671297 0.290445240
#>  [88,] 0.49061149 0.46791592 -0.0809498201 -0.097700395 0.290466194
#>  [89,] 0.49062590 0.46794004 -0.0809583584 -0.097726907 0.290485285
#>  [90,] 0.49063903 0.46796202 -0.0809661381 -0.097751065 0.290502681
#>  [91,] 0.49065099 0.46798205 -0.0809732264 -0.097773077 0.290518532
#>  [92,] 0.49066189 0.46800030 -0.0809796850 -0.097793133 0.290532974
#>  [93,] 0.49067182 0.46801692 -0.0809855696 -0.097811408 0.290546134
#>  [94,] 0.49068087 0.46803207 -0.0809909315 -0.097828059 0.290558124
#>  [95,] 0.49068912 0.46804588 -0.0809958169 -0.097843231 0.290569050
#>  [96,] 0.49069663 0.46805846 -0.0810002682 -0.097857056 0.290579004
#>  [97,] 0.49070348 0.46806992 -0.0810043240 -0.097869652 0.290588075
#>  [98,] 0.49070972 0.46808036 -0.0810080195 -0.097881129 0.290596340
#>  [99,] 0.49071540 0.46808987 -0.0810113867 -0.097891587 0.290603870
#> [100,] 0.49072058 0.46809854 -0.0810144547 -0.097901116 0.290610732
#>                [,6]         [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,]  0.000000000 -0.004351415 0.000000000 0.000000000 0.00000000
#>  [10,]  0.000000000 -0.017451541 0.000000000 0.000000000 0.00000000
#>  [11,]  0.000000000 -0.029568468 0.000000000 0.000000000 0.00000000
#>  [12,]  0.000000000 -0.043135151 0.000000000 0.000000000 0.01836347
#>  [13,]  0.000000000 -0.055618510 0.000000000 0.000000000 0.03484687
#>  [14,]  0.000000000 -0.067055401 0.000000000 0.000000000 0.04952987
#>  [15,]  0.000000000 -0.077509510 0.000000000 0.000000000 0.06262109
#>  [16,]  0.000000000 -0.087038512 0.000000000 0.000000000 0.07431876
#>  [17,]  0.000000000 -0.095123178 0.000000000 0.000000000 0.08555386
#>  [18,]  0.000000000 -0.102462585 0.000000000 0.000000000 0.09560263
#>  [19,]  0.000000000 -0.109050558 0.000000000 0.003027083 0.10459621
#>  [20,]  0.000000000 -0.114960268 0.000000000 0.008808982 0.11265648
#>  [21,]  0.000000000 -0.120323285 0.000000000 0.014108973 0.11989398
#>  [22,]  0.000000000 -0.125439416 0.001288149 0.018965852 0.12669264
#>  [23,]  0.000000000 -0.131381292 0.008975309 0.023441812 0.13428032
#>  [24,]  0.000000000 -0.136815188 0.015955935 0.027549622 0.14110697
#>  [25,]  0.000000000 -0.141780189 0.022295520 0.031315359 0.14725498
#>  [26,] -0.005097079 -0.146158551 0.027801328 0.036225070 0.15169939
#>  [27,] -0.011864014 -0.150168493 0.032728140 0.041447183 0.15511550
#>  [28,] -0.018076049 -0.153874392 0.037228672 0.046272433 0.15815278
#>  [29,] -0.023772264 -0.157291927 0.041337617 0.050723414 0.16085753
#>  [30,] -0.028990319 -0.160437936 0.045087469 0.054823005 0.16327011
#>  [31,] -0.033766270 -0.163329568 0.048508367 0.058593960 0.16542549
#>  [32,] -0.038134737 -0.165983643 0.051628028 0.062058755 0.16735359
#>  [33,] -0.042127953 -0.168417210 0.054472408 0.065239036 0.16908099
#>  [34,] -0.045776437 -0.170646317 0.057065147 0.068155670 0.17063045
#>  [35,] -0.049108382 -0.172686700 0.059428227 0.070828430 0.17202219
#>  [36,] -0.052150215 -0.174553014 0.061581660 0.073276105 0.17327365
#>  [37,] -0.054926355 -0.176259143 0.063543844 0.075516352 0.17440018
#>  [38,] -0.057459366 -0.177818087 0.065331620 0.077565710 0.17541526
#>  [39,] -0.059770047 -0.179241966 0.066960392 0.079439612 0.17633076
#>  [40,] -0.061877535 -0.180542034 0.068444232 0.081152411 0.17715717
#>  [41,] -0.063799406 -0.181728710 0.069795995 0.082717422 0.17790373
#>  [42,] -0.065551784 -0.182811618 0.071027411 0.084146965 0.17857866
#>  [43,] -0.067149447 -0.183799623 0.072149179 0.085452418 0.17918925
#>  [44,] -0.068605923 -0.184700882 0.073171055 0.086644273 0.17974195
#>  [45,] -0.069933589 -0.185522884 0.074101932 0.087732190 0.18024255
#>  [46,] -0.071143760 -0.186272501 0.074949915 0.088725049 0.18069618
#>  [47,] -0.072246772 -0.186956032 0.075722391 0.089631008 0.18110746
#>  [48,] -0.073252069 -0.187579244 0.076426087 0.090457551 0.18148049
#>  [49,] -0.074168271 -0.188147413 0.077067134 0.091211544 0.18181897
#>  [50,] -0.075003247 -0.188665364 0.077651115 0.091899275 0.18212622
#>  [51,] -0.075764178 -0.189137507 0.078183114 0.092526501 0.18240520
#>  [52,] -0.076457612 -0.189567870 0.078667765 0.093098494 0.18265859
#>  [53,] -0.077089525 -0.189960132 0.079109286 0.093620074 0.18288881
#>  [54,] -0.077665365 -0.190317654 0.079511520 0.094095649 0.18309802
#>  [55,] -0.078190099 -0.190643499 0.079877966 0.094529247 0.18328819
#>  [56,] -0.078668258 -0.190940466 0.080211812 0.094924551 0.18346109
#>  [57,] -0.079103970 -0.191211108 0.080515961 0.095284923 0.18361831
#>  [58,] -0.079501000 -0.191457751 0.080793057 0.095613433 0.18376131
#>  [59,] -0.079862780 -0.191682520 0.081045509 0.095912886 0.18389138
#>  [60,] -0.080192436 -0.191887351 0.081275511 0.096185841 0.18400972
#>  [61,] -0.080492819 -0.192074010 0.081485060 0.096434634 0.18411740
#>  [62,] -0.080766528 -0.192244106 0.081675977 0.096661396 0.18421539
#>  [63,] -0.081015930 -0.192399108 0.081849920 0.096868073 0.18430458
#>  [64,] -0.081243182 -0.192540353 0.082008398 0.097056438 0.18438575
#>  [65,] -0.081450251 -0.192669060 0.082152788 0.097228110 0.18445965
#>  [66,] -0.081638929 -0.192786343 0.082284342 0.097384564 0.18452692
#>  [67,] -0.081810849 -0.192893214 0.082404203 0.097527147 0.18458816
#>  [68,] -0.081967499 -0.192990596 0.082513410 0.097657087 0.18464393
#>  [69,] -0.082110235 -0.193079333 0.082612911 0.097775502 0.18469470
#>  [70,] -0.082240293 -0.193160190 0.082703568 0.097883413 0.18474094
#>  [71,] -0.082358798 -0.193233867 0.082786169 0.097981751 0.18478305
#>  [72,] -0.082466777 -0.193301002 0.082861428 0.098071364 0.18482140
#>  [73,] -0.082565165 -0.193362175 0.082930000 0.098153025 0.18485633
#>  [74,] -0.082654812 -0.193417916 0.082992478 0.098227438 0.18488814
#>  [75,] -0.082736497 -0.193468706 0.083049403 0.098295247 0.18491711
#>  [76,] -0.082810925 -0.193514985 0.083101271 0.098357038 0.18494350
#>  [77,] -0.082878742 -0.193557154 0.083148529 0.098413343 0.18496754
#>  [78,] -0.082940535 -0.193595578 0.083191589 0.098464650 0.18498943
#>  [79,] -0.082996838 -0.193630589 0.083230822 0.098511402 0.18500938
#>  [80,] -0.083048140 -0.193662490 0.083266569 0.098554002 0.18502755
#>  [81,] -0.083094884 -0.193691558 0.083299140 0.098592821 0.18504410
#>  [82,] -0.083137477 -0.193718044 0.083328817 0.098628192 0.18505918
#>  [83,] -0.083176285 -0.193742177 0.083355858 0.098660423 0.18507292
#>  [84,] -0.083211646 -0.193764167 0.083380496 0.098689792 0.18508543
#>  [85,] -0.083243866 -0.193784203 0.083402944 0.098716552 0.18509683
#>  [86,] -0.083273223 -0.193802459 0.083423399 0.098740936 0.18510722
#>  [87,] -0.083299973 -0.193819094 0.083442036 0.098763154 0.18511668
#>  [88,] -0.083324346 -0.193834251 0.083459017 0.098783399 0.18512530
#>  [89,] -0.083346554 -0.193848062 0.083474490 0.098801846 0.18513316
#>  [90,] -0.083366789 -0.193860646 0.083488587 0.098818655 0.18514031
#>  [91,] -0.083385226 -0.193872112 0.083501433 0.098833971 0.18514683
#>  [92,] -0.083402026 -0.193882560 0.083513137 0.098847926 0.18515277
#>  [93,] -0.083417333 -0.193892079 0.083523801 0.098860642 0.18515819
#>  [94,] -0.083431281 -0.193900753 0.083533518 0.098872228 0.18516312
#>  [95,] -0.083443989 -0.193908656 0.083542372 0.098882785 0.18516761
#>  [96,] -0.083455568 -0.193915858 0.083550439 0.098892405 0.18517170
#>  [97,] -0.083466119 -0.193922419 0.083557789 0.098901170 0.18517543
#>  [98,] -0.083475732 -0.193928398 0.083564487 0.098909156 0.18517883
#>  [99,] -0.083484492 -0.193933845 0.083570589 0.098916433 0.18518193
#> [100,] -0.083492473 -0.193938809 0.083576149 0.098923064 0.18518475
#> 
#> $cvm
#>   [1] -3.998878 -3.911110 -3.817449 -3.734273 -3.671634 -3.607018 -3.508456
#>   [8] -3.425097 -3.353999 -3.272095 -3.206994 -3.154859 -3.114412 -3.078199
#>  [15] -3.055049 -3.048757 -3.051801 -3.052299 -3.058034 -3.071865 -3.089866
#>  [22] -3.109337 -3.129565 -3.149673 -3.167834 -3.185862 -3.201621 -3.216307
#>  [29] -3.226624 -3.234955 -3.242617 -3.250369 -3.258259 -3.266110 -3.272252
#>  [36] -3.277175 -3.281592 -3.286485 -3.291566 -3.295459 -3.298914 -3.302192
#>  [43] -3.305286 -3.308193 -3.310116 -3.311621 -3.313044 -3.314382 -3.315637
#>  [50] -3.316810 -3.317902 -3.318917 -3.319858 -3.320729 -3.321534 -3.322277
#>  [57] -3.322961 -3.323592 -3.324171 -3.324703 -3.325192 -3.325640 -3.326051
#>  [64] -3.326427 -3.326772 -3.327088 -3.327377 -3.327641 -3.327882 -3.328103
#>  [71] -3.328304 -3.328488 -3.328656 -3.328810 -3.328950 -3.329078 -3.329195
#>  [78] -3.329301 -3.329399 -3.329487 -3.329568 -3.329642 -3.329709 -3.329770
#>  [85] -3.329826 -3.329877 -3.329924 -3.329966 -3.330005 -3.330040 -3.330072
#>  [92] -3.330102 -3.330128 -3.330153 -3.330175 -3.330195 -3.330213 -3.330230
#>  [99] -3.330245 -3.330259
#> 
#> $cvsd
#>   [1] 3.805667 3.744811 3.676833 3.612918 3.542860 3.474846 3.371383 3.283278
#>   [9] 3.205588 3.129886 3.074452 3.031800 2.992634 2.948941 2.903302 2.859445
#>  [17] 2.815374 2.773610 2.736013 2.697561 2.661770 2.629052 2.598931 2.570916
#>  [25] 2.546997 2.525221 2.505841 2.488415 2.472673 2.457690 2.443916 2.431319
#>  [33] 2.419864 2.409445 2.402307 2.397323 2.392972 2.389106 2.385575 2.381548
#>  [41] 2.377680 2.374181 2.371014 2.368149 2.365905 2.364016 2.362311 2.360773
#>  [49] 2.359384 2.358128 2.356994 2.355967 2.355039 2.354198 2.353437 2.352747
#>  [57] 2.352122 2.351555 2.351041 2.350575 2.350151 2.349767 2.349418 2.349100
#>  [65] 2.348812 2.348550 2.348312 2.348095 2.347899 2.347719 2.347557 2.347408
#>  [73] 2.347273 2.347151 2.347039 2.346937 2.346845 2.346760 2.346684 2.346614
#>  [81] 2.346550 2.346492 2.346440 2.346392 2.346348 2.346308 2.346272 2.346239
#>  [89] 2.346209 2.346181 2.346156 2.346133 2.346113 2.346094 2.346077 2.346061
#>  [97] 2.346047 2.346034 2.346022 2.346011
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x55a585b91420>
#> 
#> attr(,"class")
#> [1] "cv_survtran"