Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6576972  1.3699967 -0.9953343  1.0047907 -0.8935486  1.0413994 -1.0876365 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0374458 
#> 
#> $coefficients$eta
#>  [1]  0.34593018  0.31299539  0.00000000  0.00000000  0.13047849  0.00000000
#>  [7] -0.05561859  0.00000000  0.00000000  0.03484698
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.08032157
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]          [,3]         [,4]        [,5]
#>   [1,] 0.08525555 0.02937811  0.0000000000  0.000000000 0.000000000
#>   [2,] 0.11560547 0.06612220  0.0000000000  0.000000000 0.000000000
#>   [3,] 0.14426287 0.10107474  0.0000000000  0.000000000 0.000000000
#>   [4,] 0.17117741 0.13349865  0.0000000000  0.000000000 0.003269571
#>   [5,] 0.19645868 0.16138757  0.0000000000  0.000000000 0.020975899
#>   [6,] 0.22020318 0.18782314  0.0000000000  0.000000000 0.037552408
#>   [7,] 0.24239758 0.21273069  0.0000000000  0.000000000 0.052994376
#>   [8,] 0.26305120 0.23606940  0.0000000000  0.000000000 0.067314566
#>   [9,] 0.28198608 0.25734278  0.0000000000  0.000000000 0.080509645
#>  [10,] 0.29905835 0.27615901  0.0000000000  0.000000000 0.092749554
#>  [11,] 0.31480521 0.29367296  0.0000000000  0.000000000 0.104192227
#>  [12,] 0.33105989 0.30374754  0.0000000000  0.000000000 0.117920597
#>  [13,] 0.34593018 0.31299539  0.0000000000  0.000000000 0.130478494
#>  [14,] 0.35949803 0.32150614  0.0000000000  0.000000000 0.141925574
#>  [15,] 0.37186101 0.32932161  0.0000000000  0.000000000 0.152347381
#>  [16,] 0.38306673 0.33649631 -0.0001209767  0.000000000 0.161826386
#>  [17,] 0.39039003 0.34366587 -0.0074961420  0.000000000 0.170431277
#>  [18,] 0.39708030 0.35021602 -0.0141585422  0.000000000 0.178268642
#>  [19,] 0.40356103 0.35602653 -0.0203460330  0.000000000 0.185418582
#>  [20,] 0.40986728 0.36116944 -0.0261142242  0.000000000 0.191967133
#>  [21,] 0.41564437 0.36586624 -0.0313414265  0.000000000 0.197953097
#>  [22,] 0.42102437 0.37015467 -0.0359936933  0.000000000 0.203489186
#>  [23,] 0.42640255 0.37409749 -0.0397606091  0.000000000 0.208896601
#>  [24,] 0.43130013 0.37771619 -0.0431491808  0.000000000 0.213837913
#>  [25,] 0.43575918 0.38103289 -0.0462002050  0.000000000 0.218351123
#>  [26,] 0.44027164 0.38678348 -0.0492628468 -0.005499923 0.223815926
#>  [27,] 0.44461955 0.39343967 -0.0521475196 -0.013378227 0.229532873
#>  [28,] 0.44860789 0.39961917 -0.0547689391 -0.020627018 0.234788995
#>  [29,] 0.45226344 0.40534180 -0.0571509358 -0.027287770 0.239615396
#>  [30,] 0.45561167 0.41062984 -0.0593152738 -0.033401190 0.244042440
#>  [31,] 0.45867665 0.41550650 -0.0612817320 -0.039006470 0.248099498
#>  [32,] 0.46148029 0.41999696 -0.0630687343 -0.044141203 0.251814175
#>  [33,] 0.46404414 0.42412600 -0.0646925352 -0.048841778 0.255213295
#>  [34,] 0.46638755 0.42791816 -0.0661682320 -0.053142035 0.258321733
#>  [35,] 0.46852862 0.43139770 -0.0675095367 -0.057074015 0.261162841
#>  [36,] 0.47048432 0.43458696 -0.0687287049 -0.060667404 0.263758595
#>  [37,] 0.47227015 0.43750820 -0.0698370432 -0.063950107 0.266129250
#>  [38,] 0.47390043 0.44018218 -0.0708447525 -0.066947953 0.268293639
#>  [39,] 0.47538841 0.44262845 -0.0717610847 -0.069684853 0.270269171
#>  [40,] 0.47674624 0.44486532 -0.0725944303 -0.072182887 0.272071905
#>  [41,] 0.47798509 0.44690984 -0.0733523991 -0.074462407 0.273716627
#>  [42,] 0.47911523 0.44877787 -0.0740418912 -0.076542137 0.275216932
#>  [43,] 0.48014606 0.45048409 -0.0746691651 -0.078439278 0.276585299
#>  [44,] 0.48108620 0.45204210 -0.0752398984 -0.080169618 0.277833175
#>  [45,] 0.48194354 0.45346441 -0.0757592419 -0.081747633 0.278971046
#>  [46,] 0.48272531 0.45476257 -0.0762318698 -0.083186579 0.280008512
#>  [47,] 0.48343811 0.45594720 -0.0766620245 -0.084498595 0.280954358
#>  [48,] 0.48408798 0.45702804 -0.0770535573 -0.085694781 0.281816613
#>  [49,] 0.48468044 0.45801406 -0.0774099651 -0.086785287 0.282602618
#>  [50,] 0.48522052 0.45891344 -0.0777344240 -0.087779391 0.283319078
#>  [51,] 0.48571285 0.45973372 -0.0780298188 -0.088685565 0.283972117
#>  [52,] 0.48616161 0.46048178 -0.0782987709 -0.089511550 0.284567326
#>  [53,] 0.48657066 0.46116391 -0.0785436624 -0.090264411 0.285109805
#>  [54,] 0.48694348 0.46178587 -0.0787666583 -0.090950596 0.285604212
#>  [55,] 0.48728329 0.46235294 -0.0789697270 -0.091575991 0.286054795
#>  [56,] 0.48759298 0.46286992 -0.0791546580 -0.092145965 0.286465429
#>  [57,] 0.48787523 0.46334121 -0.0793230790 -0.092665416 0.286839647
#>  [58,] 0.48813246 0.46377084 -0.0794764701 -0.093138813 0.287180673
#>  [59,] 0.48836689 0.46416246 -0.0796161779 -0.093570230 0.287491446
#>  [60,] 0.48858052 0.46451942 -0.0797434273 -0.093963382 0.287774646
#>  [61,] 0.48877521 0.46484479 -0.0798593331 -0.094321659 0.288032716
#>  [62,] 0.48895263 0.46514133 -0.0799649096 -0.094648149 0.288267883
#>  [63,] 0.48911431 0.46541161 -0.0800610797 -0.094945669 0.288482177
#>  [64,] 0.48926164 0.46565795 -0.0801486838 -0.095216786 0.288677450
#>  [65,] 0.48939590 0.46588244 -0.0802284866 -0.095463841 0.288855389
#>  [66,] 0.48951824 0.46608704 -0.0803011843 -0.095688968 0.289017530
#>  [67,] 0.48962973 0.46627350 -0.0803674107 -0.095894110 0.289165276
#>  [68,] 0.48973131 0.46644342 -0.0804277430 -0.096081042 0.289299904
#>  [69,] 0.48982388 0.46659827 -0.0804827065 -0.096251377 0.289422578
#>  [70,] 0.48990824 0.46673938 -0.0805327797 -0.096406590 0.289534359
#>  [71,] 0.48998510 0.46686797 -0.0805783984 -0.096548021 0.289636213
#>  [72,] 0.49005514 0.46698516 -0.0806199592 -0.096676893 0.289729022
#>  [73,] 0.49011896 0.46709194 -0.0806578237 -0.096794322 0.289813590
#>  [74,] 0.49017711 0.46718925 -0.0806923208 -0.096901323 0.289890646
#>  [75,] 0.49023010 0.46727792 -0.0807237503 -0.096998822 0.289960859
#>  [76,] 0.49027838 0.46735872 -0.0807523852 -0.097087662 0.290024836
#>  [77,] 0.49032238 0.46743234 -0.0807784743 -0.097168613 0.290083131
#>  [78,] 0.49036247 0.46749943 -0.0808022439 -0.097242373 0.290136248
#>  [79,] 0.49039900 0.46756057 -0.0808239006 -0.097309583 0.290184648
#>  [80,] 0.49043228 0.46761627 -0.0808436321 -0.097370823 0.290228748
#>  [81,] 0.49046261 0.46766703 -0.0808616098 -0.097426624 0.290268931
#>  [82,] 0.49049024 0.46771329 -0.0808779895 -0.097477469 0.290305545
#>  [83,] 0.49051542 0.46775543 -0.0808929135 -0.097523797 0.290338907
#>  [84,] 0.49053836 0.46779383 -0.0809065111 -0.097566011 0.290369305
#>  [85,] 0.49055927 0.46782882 -0.0809189003 -0.097604475 0.290397003
#>  [86,] 0.49057832 0.46786071 -0.0809301885 -0.097639522 0.290422241
#>  [87,] 0.49059567 0.46788976 -0.0809404735 -0.097671456 0.290445236
#>  [88,] 0.49061149 0.46791623 -0.0809498446 -0.097700553 0.290466189
#>  [89,] 0.49062590 0.46794036 -0.0809583830 -0.097727066 0.290485281
#>  [90,] 0.49063903 0.46796233 -0.0809661626 -0.097751224 0.290502677
#>  [91,] 0.49065099 0.46798236 -0.0809732510 -0.097773235 0.290518527
#>  [92,] 0.49066189 0.46800061 -0.0809797095 -0.097793292 0.290532970
#>  [93,] 0.49067183 0.46801724 -0.0809855942 -0.097811567 0.290546129
#>  [94,] 0.49068088 0.46803239 -0.0809909560 -0.097828218 0.290558120
#>  [95,] 0.49068912 0.46804619 -0.0809958414 -0.097843390 0.290569045
#>  [96,] 0.49069664 0.46805877 -0.0810002928 -0.097857215 0.290579000
#>  [97,] 0.49070348 0.46807023 -0.0810043486 -0.097869811 0.290588071
#>  [98,] 0.49070972 0.46808067 -0.0810080441 -0.097881288 0.290596335
#>  [99,] 0.49071540 0.46809019 -0.0810114113 -0.097891746 0.290603866
#> [100,] 0.49072058 0.46809886 -0.0810144793 -0.097901275 0.290610727
#>                [,6]         [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,]  0.000000000 -0.004351415 0.000000000 0.000000000 0.00000000
#>  [10,]  0.000000000 -0.017451541 0.000000000 0.000000000 0.00000000
#>  [11,]  0.000000000 -0.029568468 0.000000000 0.000000000 0.00000000
#>  [12,]  0.000000000 -0.043135158 0.000000000 0.000000000 0.01836348
#>  [13,]  0.000000000 -0.055618595 0.000000000 0.000000000 0.03484698
#>  [14,]  0.000000000 -0.067055406 0.000000000 0.000000000 0.04952987
#>  [15,]  0.000000000 -0.077509516 0.000000000 0.000000000 0.06262109
#>  [16,]  0.000000000 -0.087038519 0.000000000 0.000000000 0.07431877
#>  [17,]  0.000000000 -0.095123143 0.000000000 0.000000000 0.08555381
#>  [18,]  0.000000000 -0.102462575 0.000000000 0.000000000 0.09560262
#>  [19,]  0.000000000 -0.109050610 0.000000000 0.003027086 0.10459628
#>  [20,]  0.000000000 -0.114960317 0.000000000 0.008808985 0.11265655
#>  [21,]  0.000000000 -0.120323333 0.000000000 0.014108975 0.11989404
#>  [22,]  0.000000000 -0.125439457 0.001288179 0.018965854 0.12669268
#>  [23,]  0.000000000 -0.131381441 0.008975417 0.023441818 0.13428044
#>  [24,]  0.000000000 -0.136815348 0.015956050 0.027549630 0.14110710
#>  [25,]  0.000000000 -0.141780203 0.022295529 0.031315359 0.14725498
#>  [26,] -0.005097016 -0.146158632 0.027801385 0.036225053 0.15169948
#>  [27,] -0.011864060 -0.150168442 0.032728105 0.041447199 0.15511544
#>  [28,] -0.018076096 -0.153874343 0.037228639 0.046272450 0.15815272
#>  [29,] -0.023772248 -0.157291943 0.041337628 0.050723408 0.16085756
#>  [30,] -0.028990332 -0.160437910 0.045087451 0.054823009 0.16327009
#>  [31,] -0.033766356 -0.163329585 0.048508379 0.058594020 0.16542542
#>  [32,] -0.038134746 -0.165983666 0.051628043 0.062058767 0.16735359
#>  [33,] -0.042127968 -0.168417227 0.054472419 0.065239051 0.16908099
#>  [34,] -0.045776373 -0.170646429 0.057065219 0.068155656 0.17063055
#>  [35,] -0.049108405 -0.172686687 0.059428218 0.070828442 0.17202217
#>  [36,] -0.052150246 -0.174552993 0.061581647 0.073276121 0.17327361
#>  [37,] -0.054926395 -0.176259115 0.063543827 0.075516371 0.17440013
#>  [38,] -0.057459413 -0.177818053 0.065331599 0.077565732 0.17541520
#>  [39,] -0.059770100 -0.179241927 0.066960367 0.079439637 0.17633070
#>  [40,] -0.061877592 -0.180541991 0.068444205 0.081152438 0.17715710
#>  [41,] -0.063799467 -0.181728663 0.069795966 0.082717451 0.17790366
#>  [42,] -0.065551851 -0.182811566 0.071027379 0.084146996 0.17857859
#>  [43,] -0.067149519 -0.183799568 0.072149145 0.085452452 0.17918916
#>  [44,] -0.068605999 -0.184700823 0.073171019 0.086644309 0.17974186
#>  [45,] -0.069933668 -0.185522823 0.074101895 0.087732227 0.18024245
#>  [46,] -0.071143840 -0.186272439 0.074949878 0.088725087 0.18069609
#>  [47,] -0.072246854 -0.186955969 0.075722353 0.089631046 0.18110736
#>  [48,] -0.073252152 -0.187579180 0.076426049 0.090457591 0.18148039
#>  [49,] -0.074168355 -0.188147348 0.077067095 0.091211584 0.18181888
#>  [50,] -0.075003333 -0.188665298 0.077651075 0.091899315 0.18212612
#>  [51,] -0.075764264 -0.189137440 0.078183074 0.092526543 0.18240510
#>  [52,] -0.076457700 -0.189567803 0.078667725 0.093098536 0.18265849
#>  [53,] -0.077089613 -0.189960065 0.079109246 0.093620116 0.18288870
#>  [54,] -0.077665454 -0.190317585 0.079511479 0.094095692 0.18309792
#>  [55,] -0.078190189 -0.190643430 0.079877925 0.094529291 0.18328808
#>  [56,] -0.078668348 -0.190940397 0.080211771 0.094924595 0.18346098
#>  [57,] -0.079104061 -0.191211038 0.080515920 0.095284967 0.18361820
#>  [58,] -0.079501092 -0.191457681 0.080793016 0.095613477 0.18376120
#>  [59,] -0.079862872 -0.191682450 0.081045468 0.095912930 0.18389127
#>  [60,] -0.080192528 -0.191887281 0.081275469 0.096185886 0.18400962
#>  [61,] -0.080492912 -0.192073939 0.081485018 0.096434679 0.18411729
#>  [62,] -0.080766621 -0.192244035 0.081675935 0.096661441 0.18421528
#>  [63,] -0.081016023 -0.192399037 0.081849878 0.096868118 0.18430447
#>  [64,] -0.081243276 -0.192540281 0.082008356 0.097056483 0.18438564
#>  [65,] -0.081450345 -0.192668989 0.082152746 0.097228155 0.18445954
#>  [66,] -0.081639023 -0.192786271 0.082284300 0.097384610 0.18452681
#>  [67,] -0.081810944 -0.192893142 0.082404161 0.097527193 0.18458805
#>  [68,] -0.081967594 -0.192990524 0.082513368 0.097657133 0.18464382
#>  [69,] -0.082110330 -0.193079260 0.082612868 0.097775548 0.18469459
#>  [70,] -0.082240388 -0.193160118 0.082703526 0.097883460 0.18474083
#>  [71,] -0.082358894 -0.193233795 0.082786126 0.097981798 0.18478294
#>  [72,] -0.082466873 -0.193300930 0.082861386 0.098071410 0.18482129
#>  [73,] -0.082565260 -0.193362103 0.082929957 0.098153071 0.18485621
#>  [74,] -0.082654908 -0.193417843 0.082992435 0.098227485 0.18488802
#>  [75,] -0.082736593 -0.193468633 0.083049361 0.098295294 0.18491700
#>  [76,] -0.082811021 -0.193514912 0.083101228 0.098357084 0.18494339
#>  [77,] -0.082878838 -0.193557081 0.083148487 0.098413390 0.18496743
#>  [78,] -0.082940631 -0.193595505 0.083191546 0.098464697 0.18498932
#>  [79,] -0.082996934 -0.193630516 0.083230779 0.098511448 0.18500927
#>  [80,] -0.083048236 -0.193662417 0.083266526 0.098554049 0.18502744
#>  [81,] -0.083094981 -0.193691485 0.083299097 0.098592868 0.18504399
#>  [82,] -0.083137573 -0.193717970 0.083328774 0.098628239 0.18505907
#>  [83,] -0.083176382 -0.193742104 0.083355815 0.098660470 0.18507281
#>  [84,] -0.083211743 -0.193764093 0.083380453 0.098689839 0.18508532
#>  [85,] -0.083243962 -0.193784130 0.083402901 0.098716599 0.18509672
#>  [86,] -0.083273320 -0.193802386 0.083423356 0.098740983 0.18510711
#>  [87,] -0.083300069 -0.193819021 0.083441993 0.098763201 0.18511657
#>  [88,] -0.083324442 -0.193834178 0.083458974 0.098783446 0.18512519
#>  [89,] -0.083346650 -0.193847989 0.083474446 0.098801894 0.18513305
#>  [90,] -0.083366886 -0.193860573 0.083488544 0.098818702 0.18514020
#>  [91,] -0.083385323 -0.193872039 0.083501390 0.098834018 0.18514672
#>  [92,] -0.083402123 -0.193882486 0.083513094 0.098847973 0.18515266
#>  [93,] -0.083417430 -0.193892006 0.083523758 0.098860689 0.18515807
#>  [94,] -0.083431377 -0.193900679 0.083533475 0.098872275 0.18516300
#>  [95,] -0.083444086 -0.193908583 0.083542329 0.098882833 0.18516750
#>  [96,] -0.083455665 -0.193915784 0.083550396 0.098892452 0.18517159
#>  [97,] -0.083466216 -0.193922345 0.083557746 0.098901217 0.18517532
#>  [98,] -0.083475829 -0.193928324 0.083564444 0.098909204 0.18517872
#>  [99,] -0.083484589 -0.193933772 0.083570546 0.098916481 0.18518181
#> [100,] -0.083492570 -0.193938735 0.083576106 0.098923111 0.18518463
#> 
#> $cvm
#>   [1] -4.006600 -3.891598 -3.783997 -3.687380 -3.584368 -3.501974 -3.433691
#>   [8] -3.372122 -3.323343 -3.285117 -3.254858 -3.235091 -3.232500 -3.246034
#>  [15] -3.263981 -3.288842 -3.319415 -3.344767 -3.370410 -3.397251 -3.425547
#>  [22] -3.456533 -3.488408 -3.518098 -3.546343 -3.572260 -3.596276 -3.618543
#>  [29] -3.637531 -3.654492 -3.667308 -3.677867 -3.683865 -3.688344 -3.690808
#>  [36] -3.693341 -3.695886 -3.698403 -3.700860 -3.703234 -3.704825 -3.706126
#>  [43] -3.707388 -3.708601 -3.709759 -3.710858 -3.711895 -3.712869 -3.713782
#>  [50] -3.714635 -3.715428 -3.716166 -3.716849 -3.717482 -3.718066 -3.718606
#>  [57] -3.719103 -3.719560 -3.719981 -3.720367 -3.720722 -3.721048 -3.721346
#>  [64] -3.721619 -3.721870 -3.722099 -3.722309 -3.722500 -3.722676 -3.722836
#>  [71] -3.722982 -3.723116 -3.723238 -3.723350 -3.723452 -3.723544 -3.723629
#>  [78] -3.723707 -3.723777 -3.723842 -3.723900 -3.723954 -3.724003 -3.724047
#>  [85] -3.724088 -3.724125 -3.724159 -3.724190 -3.724218 -3.724243 -3.724267
#>  [92] -3.724288 -3.724307 -3.724325 -3.724341 -3.724356 -3.724369 -3.724381
#>  [99] -3.724393 -3.724403
#> 
#> $cvsd
#>   [1] 3.920051 3.923339 3.922857 3.923044 3.917475 3.920818 3.919911 3.903572
#>   [9] 3.893580 3.888930 3.889385 3.883621 3.877217 3.861804 3.843251 3.826059
#>  [17] 3.810606 3.786994 3.760718 3.735032 3.714528 3.701204 3.691456 3.686182
#>  [25] 3.682888 3.680617 3.680542 3.680420 3.679198 3.677771 3.673692 3.669899
#>  [33] 3.668609 3.669240 3.671431 3.673774 3.676196 3.678641 3.681068 3.683444
#>  [41] 3.687158 3.691150 3.694882 3.698361 3.701597 3.704598 3.707377 3.709947
#>  [49] 3.712319 3.714506 3.716519 3.718371 3.720073 3.721636 3.723070 3.724385
#>  [57] 3.725589 3.726693 3.727703 3.728628 3.729473 3.730246 3.730953 3.731598
#>  [65] 3.732188 3.732727 3.733219 3.733668 3.734078 3.734452 3.734794 3.735105
#>  [73] 3.735389 3.735649 3.735885 3.736101 3.736297 3.736477 3.736640 3.736789
#>  [81] 3.736925 3.737049 3.737162 3.737265 3.737359 3.737444 3.737522 3.737593
#>  [89] 3.737658 3.737717 3.737771 3.737820 3.737864 3.737905 3.737942 3.737976
#>  [97] 3.738007 3.738035 3.738061 3.738084
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x555889b26d90>
#> 
#> attr(,"class")
#> [1] "cv_survtran"