Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6166638  1.3973123 -0.9993905  1.0047907 -0.8576187  1.0413994 -1.1233648 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0829270 
#> 
#> $coefficients$eta
#>  [1]  0.386963517  0.340311025 -0.004056184  0.000000000  0.166408437
#>  [6]  0.000000000 -0.091346969  0.000000000  0.000000000  0.080328254
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05788603
#> 
#> 
#> $lambdas
#>   [1] 3.089203e-01 2.814767e-01 2.564711e-01 2.336869e-01 2.129268e-01
#>   [6] 1.940110e-01 1.767756e-01 1.610713e-01 1.467622e-01 1.337242e-01
#>  [11] 1.218445e-01 1.110202e-01 1.011575e-01 9.217093e-02 8.398272e-02
#>  [16] 7.652192e-02 6.972393e-02 6.352984e-02 5.788603e-02 5.274359e-02
#>  [21] 4.805800e-02 4.378866e-02 3.989859e-02 3.635411e-02 3.312451e-02
#>  [26] 3.018182e-02 2.750055e-02 2.505748e-02 2.283144e-02 2.080316e-02
#>  [31] 1.895506e-02 1.727115e-02 1.573683e-02 1.433881e-02 1.306499e-02
#>  [36] 1.190433e-02 1.084678e-02 9.883186e-03 9.005191e-03 8.205195e-03
#>  [41] 7.476268e-03 6.812097e-03 6.206929e-03 5.655523e-03 5.153102e-03
#>  [46] 4.695314e-03 4.278195e-03 3.898132e-03 3.551833e-03 3.236298e-03
#>  [51] 2.948794e-03 2.686831e-03 2.448141e-03 2.230655e-03 2.032489e-03
#>  [56] 1.851929e-03 1.687408e-03 1.537504e-03 1.400916e-03 1.276463e-03
#>  [61] 1.163065e-03 1.059742e-03 9.655971e-04 8.798162e-04 8.016557e-04
#>  [66] 7.304388e-04 6.655487e-04 6.064231e-04 5.525502e-04 5.034632e-04
#>  [71] 4.587369e-04 4.179840e-04 3.808514e-04 3.470176e-04 3.161895e-04
#>  [76] 2.881001e-04 2.625061e-04 2.391858e-04 2.179372e-04 1.985762e-04
#>  [81] 1.809353e-04 1.648615e-04 1.502156e-04 1.368709e-04 1.247117e-04
#>  [86] 1.136326e-04 1.035378e-04 9.433980e-05 8.595891e-05 7.832256e-05
#>  [91] 7.136460e-05 6.502477e-05 5.924815e-05 5.398470e-05 4.918885e-05
#>  [96] 4.481905e-05 4.083745e-05 3.720956e-05 3.390397e-05 3.089203e-05
#> 
#> $etas
#>              [,1]       [,2]         [,3]         [,4]       [,5]         [,6]
#>   [1,] 0.00000000 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [2,] 0.03624828 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [3,] 0.07013592 0.01117874  0.000000000  0.000000000 0.00000000  0.000000000
#>   [4,] 0.10127198 0.04873376  0.000000000  0.000000000 0.00000000  0.000000000
#>   [5,] 0.13074780 0.08455985  0.000000000  0.000000000 0.00000000  0.000000000
#>   [6,] 0.15850358 0.11853574  0.000000000  0.000000000 0.00000000  0.000000000
#>   [7,] 0.18453640 0.14820134  0.000000000  0.000000000 0.01263273  0.000000000
#>   [8,] 0.20902026 0.17534463  0.000000000  0.000000000 0.02975226  0.000000000
#>   [9,] 0.23195752 0.20099142  0.000000000  0.000000000 0.04573723  0.000000000
#>  [10,] 0.25334681 0.22508481  0.000000000  0.000000000 0.06059224  0.000000000
#>  [11,] 0.27320850 0.24760071  0.000000000  0.000000000 0.07433839  0.000000000
#>  [12,] 0.29104838 0.26730795  0.000000000  0.000000000 0.08698454  0.000000000
#>  [13,] 0.30742198 0.28544282  0.000000000  0.000000000 0.09880956  0.000000000
#>  [14,] 0.32344555 0.29904399  0.000000000  0.000000000 0.11148534  0.000000000
#>  [15,] 0.33897429 0.30865927  0.000000000  0.000000000 0.12460584  0.000000000
#>  [16,] 0.35315394 0.31751814  0.000000000  0.000000000 0.13657435  0.000000000
#>  [17,] 0.36608210 0.32566115  0.000000000  0.000000000 0.14747688  0.000000000
#>  [18,] 0.37785551 0.33313205  0.000000000  0.000000000 0.15739765  0.000000000
#>  [19,] 0.38696352 0.34031102 -0.004056184  0.000000000 0.16640844  0.000000000
#>  [20,] 0.39395075 0.34715215 -0.011050942  0.000000000 0.17460544  0.000000000
#>  [21,] 0.40034018 0.35339422 -0.017374373  0.000000000 0.18206840  0.000000000
#>  [22,] 0.40691442 0.35876341 -0.023421529  0.000000000 0.18890264  0.000000000
#>  [23,] 0.41294000 0.36366954 -0.028901179  0.000000000 0.19515254  0.000000000
#>  [24,] 0.41845724 0.36814767 -0.033867342  0.000000000 0.20086302  0.000000000
#>  [25,] 0.42388848 0.37225042 -0.038005811  0.000000000 0.20636643  0.000000000
#>  [26,] 0.42901097 0.37602152 -0.041570181  0.000000000 0.21152620  0.000000000
#>  [27,] 0.43367506 0.37948010 -0.044778211  0.000000000 0.21623996  0.000000000
#>  [28,] 0.43805289 0.38342118 -0.047780051 -0.001490315 0.22090435 -0.001646692
#>  [29,] 0.44258306 0.39031099 -0.050799829 -0.009684736 0.22685334 -0.008693582
#>  [30,] 0.44674021 0.39671685 -0.053544299 -0.017230016 0.23232629 -0.015166643
#>  [31,] 0.45055193 0.40265577 -0.056038170 -0.024167387 0.23735473 -0.021105166
#>  [32,] 0.45404431 0.40814906 -0.058304198 -0.030537985 0.24196937 -0.026547650
#>  [33,] 0.45724201 0.41322008 -0.060363153 -0.036381975 0.24620006 -0.031531142
#>  [34,] 0.46016837 0.41789274 -0.062233897 -0.041737859 0.25007549 -0.036090782
#>  [35,] 0.46284470 0.42219137 -0.063933831 -0.046641950 0.25362268 -0.040259709
#>  [36,] 0.46529114 0.42614205 -0.065478823 -0.051129760 0.25686722 -0.044069663
#>  [37,] 0.46752702 0.42976821 -0.066882853 -0.055234257 0.25983362 -0.047549798
#>  [38,] 0.46956956 0.43309353 -0.068159027 -0.058986203 0.26254433 -0.050727448
#>  [39,] 0.47143482 0.43614076 -0.069319195 -0.062414458 0.26502033 -0.053628026
#>  [40,] 0.47313791 0.43893067 -0.070373904 -0.065545677 0.26728128 -0.056274824
#>  [41,] 0.47469249 0.44148368 -0.071332917 -0.068404733 0.26934521 -0.058689544
#>  [42,] 0.47611121 0.44381867 -0.072205026 -0.071014565 0.27122881 -0.060892101
#>  [43,] 0.47740573 0.44595330 -0.072998205 -0.073396347 0.27294747 -0.062900809
#>  [44,] 0.47858673 0.44790399 -0.073719686 -0.075569562 0.27451534 -0.064732474
#>  [45,] 0.47966402 0.44968598 -0.074376026 -0.077552130 0.27594543 -0.066402506
#>  [46,] 0.48064658 0.45131337 -0.074973174 -0.079360502 0.27724967 -0.067925018
#>  [47,] 0.48154266 0.45279920 -0.075516527 -0.081009772 0.27843899 -0.069312927
#>  [48,] 0.48235977 0.45415546 -0.076010982 -0.082513767 0.27952342 -0.070578045
#>  [49,] 0.48310483 0.45539322 -0.076460982 -0.083885151 0.28051211 -0.071731168
#>  [50,] 0.48378413 0.45652263 -0.076870562 -0.085135510 0.28141346 -0.072782158
#>  [51,] 0.48440343 0.45755301 -0.077243383 -0.086275440 0.28223512 -0.073740021
#>  [52,] 0.48496801 0.45849292 -0.077582771 -0.087314627 0.28298410 -0.074612979
#>  [53,] 0.48548267 0.45935021 -0.077891748 -0.088261921 0.28366680 -0.075408532
#>  [54,] 0.48595181 0.46013205 -0.078173058 -0.089125404 0.28428905 -0.076133526
#>  [55,] 0.48637942 0.46084501 -0.078429194 -0.089912459 0.28485618 -0.076794205
#>  [56,] 0.48676919 0.46149512 -0.078662423 -0.090629822 0.28537307 -0.077396262
#>  [57,] 0.48712443 0.46208786 -0.078874805 -0.091283642 0.28584414 -0.077944891
#>  [58,] 0.48744820 0.46262827 -0.079068213 -0.091879530 0.28627345 -0.078444827
#>  [59,] 0.48774340 0.46312064 -0.079244267 -0.092422553 0.28666477 -0.078900311
#>  [60,] 0.48801233 0.46356975 -0.079404683 -0.092917485 0.28702132 -0.079315428
#>  [61,] 0.48825742 0.46397915 -0.079550786 -0.093368532 0.28734623 -0.079693689
#>  [62,] 0.48848077 0.46435232 -0.079683859 -0.093779576 0.28764232 -0.080038365
#>  [63,] 0.48868432 0.46469246 -0.079805067 -0.094154161 0.28791214 -0.080352436
#>  [64,] 0.48886982 0.46500249 -0.079915472 -0.094495514 0.28815802 -0.080638616
#>  [65,] 0.48903885 0.46528505 -0.080016039 -0.094806580 0.28838207 -0.080899383
#>  [66,] 0.48919289 0.46554258 -0.080107647 -0.095090042 0.28858624 -0.081136991
#>  [67,] 0.48933326 0.46577729 -0.080191096 -0.095348348 0.28877228 -0.081353497
#>  [68,] 0.48946118 0.46599120 -0.080267115 -0.095583728 0.28894181 -0.081550774
#>  [69,] 0.48957774 0.46618614 -0.080336367 -0.095798215 0.28909628 -0.081730530
#>  [70,] 0.48968395 0.46636379 -0.080399454 -0.095993661 0.28923705 -0.081894319
#>  [71,] 0.48978074 0.46652569 -0.080456927 -0.096171757 0.28936531 -0.082043561
#>  [72,] 0.48986894 0.46667322 -0.080509287 -0.096334040 0.28948218 -0.082179547
#>  [73,] 0.48994930 0.46680767 -0.080556988 -0.096481915 0.28958868 -0.082303453
#>  [74,] 0.49002253 0.46693019 -0.080600446 -0.096616659 0.28968571 -0.082416354
#>  [75,] 0.49008926 0.46704184 -0.080640038 -0.096739439 0.28977414 -0.082519226
#>  [76,] 0.49015006 0.46714357 -0.080676109 -0.096851316 0.28985470 -0.082612960
#>  [77,] 0.49020546 0.46723628 -0.080708973 -0.096953258 0.28992812 -0.082698368
#>  [78,] 0.49025595 0.46732076 -0.080738914 -0.097046146 0.28999501 -0.082776189
#>  [79,] 0.49030195 0.46739774 -0.080766193 -0.097130785 0.29005596 -0.082847097
#>  [80,] 0.49034386 0.46746789 -0.080791047 -0.097207908 0.29011150 -0.082911706
#>  [81,] 0.49038205 0.46753181 -0.080813692 -0.097278180 0.29016210 -0.082970576
#>  [82,] 0.49041685 0.46759005 -0.080834323 -0.097342212 0.29020821 -0.083024216
#>  [83,] 0.49044856 0.46764313 -0.080853121 -0.097400556 0.29025023 -0.083073092
#>  [84,] 0.49047746 0.46769149 -0.080870248 -0.097453718 0.29028851 -0.083117625
#>  [85,] 0.49050379 0.46773555 -0.080885852 -0.097502158 0.29032339 -0.083158203
#>  [86,] 0.49052777 0.46777570 -0.080900070 -0.097546295 0.29035518 -0.083195176
#>  [87,] 0.49054963 0.46781229 -0.080913024 -0.097586512 0.29038414 -0.083228864
#>  [88,] 0.49056955 0.46784563 -0.080924827 -0.097623157 0.29041052 -0.083259560
#>  [89,] 0.49058770 0.46787601 -0.080935581 -0.097656547 0.29043457 -0.083287528
#>  [90,] 0.49060423 0.46790368 -0.080945379 -0.097686971 0.29045648 -0.083313013
#>  [91,] 0.49061930 0.46792891 -0.080954307 -0.097714692 0.29047644 -0.083336233
#>  [92,] 0.49063303 0.46795189 -0.080962441 -0.097739951 0.29049463 -0.083357391
#>  [93,] 0.49064554 0.46797283 -0.080969853 -0.097762966 0.29051120 -0.083376669
#>  [94,] 0.49065693 0.46799191 -0.080976606 -0.097783936 0.29052630 -0.083394234
#>  [95,] 0.49066732 0.46800929 -0.080982759 -0.097803044 0.29054006 -0.083410239
#>  [96,] 0.49067678 0.46802513 -0.080988365 -0.097820455 0.29055260 -0.083424822
#>  [97,] 0.49068540 0.46803956 -0.080993473 -0.097836318 0.29056402 -0.083438110
#>  [98,] 0.49069326 0.46805272 -0.080998127 -0.097850773 0.29057443 -0.083450217
#>  [99,] 0.49070042 0.46806470 -0.081002368 -0.097863943 0.29058391 -0.083461249
#> [100,] 0.49070694 0.46807562 -0.081006232 -0.097875944 0.29059255 -0.083471300
#>               [,7]        [,8]         [,9]       [,10]
#>   [1,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [2,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [3,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [4,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [5,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [6,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [7,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [8,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [9,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [10,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [11,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [12,] -0.01129958 0.000000000 0.000000e+00 0.000000000
#>  [13,] -0.02388400 0.000000000 0.000000e+00 0.000000000
#>  [14,] -0.03676658 0.000000000 0.000000e+00 0.009760814
#>  [15,] -0.04977201 0.000000000 0.000000e+00 0.027188119
#>  [16,] -0.06170263 0.000000000 0.000000e+00 0.042706099
#>  [17,] -0.07261933 0.000000000 0.000000e+00 0.056535427
#>  [18,] -0.08258811 0.000000000 0.000000e+00 0.068872253
#>  [19,] -0.09134697 0.000000000 0.000000e+00 0.080328254
#>  [20,] -0.09903497 0.000000000 0.000000e+00 0.090927129
#>  [21,] -0.10601056 0.000000000 7.598927e-05 0.100413141
#>  [22,] -0.11220024 0.000000000 6.101134e-03 0.108903539
#>  [23,] -0.11781895 0.000000000 1.162766e-02 0.116523120
#>  [24,] -0.12291659 0.000000000 1.669035e-02 0.123368911
#>  [25,] -0.12860038 0.005384804 2.134493e-02 0.130745494
#>  [26,] -0.13427281 0.012695468 2.562584e-02 0.137925926
#>  [27,] -0.13945767 0.019334394 2.955225e-02 0.144389429
#>  [28,] -0.14412553 0.025279131 3.357610e-02 0.149913207
#>  [29,] -0.14828624 0.030423153 3.899612e-02 0.153529293
#>  [30,] -0.15213580 0.035123356 4.400877e-02 0.156741723
#>  [31,] -0.15568948 0.039415744 4.863620e-02 0.159600447
#>  [32,] -0.15896348 0.043333756 5.290127e-02 0.162148393
#>  [33,] -0.16197472 0.046908563 5.682690e-02 0.164422786
#>  [34,] -0.16474056 0.050169247 6.043571e-02 0.166456190
#>  [35,] -0.16727789 0.053142549 6.374959e-02 0.168276904
#>  [36,] -0.16960286 0.055852930 6.678995e-02 0.169908925
#>  [37,] -0.17173183 0.058323469 6.957713e-02 0.171373970
#>  [38,] -0.17367977 0.060574957 7.213036e-02 0.172690615
#>  [39,] -0.17546085 0.062626502 7.446788e-02 0.173875099
#>  [40,] -0.17708874 0.064495827 7.660670e-02 0.174941992
#>  [41,] -0.17857586 0.066198939 7.856281e-02 0.175903807
#>  [42,] -0.17993389 0.067750536 8.035107e-02 0.176771667
#>  [43,] -0.18117364 0.069164044 8.198529e-02 0.177555390
#>  [44,] -0.18230511 0.070451720 8.347827e-02 0.178263672
#>  [45,] -0.18333752 0.071624744 8.484181e-02 0.178904221
#>  [46,] -0.18427936 0.072693315 8.608684e-02 0.179483887
#>  [47,] -0.18513843 0.073666729 8.722340e-02 0.180008765
#>  [48,] -0.18592191 0.074553462 8.826074e-02 0.180484289
#>  [49,] -0.18663635 0.075361234 8.920736e-02 0.180915315
#>  [50,] -0.18728777 0.076097082 9.007106e-02 0.181306182
#>  [51,] -0.18788168 0.076767414 9.085900e-02 0.181660781
#>  [52,] -0.18842312 0.077378071 9.157772e-02 0.181982597
#>  [53,] -0.18891668 0.077934369 9.223325e-02 0.182274764
#>  [54,] -0.18936658 0.078441153 9.283108e-02 0.182540097
#>  [55,] -0.18977665 0.078902835 9.337624e-02 0.182781130
#>  [56,] -0.19015042 0.079323434 9.387333e-02 0.183000148
#>  [57,] -0.19049107 0.079706610 9.432656e-02 0.183199208
#>  [58,] -0.19080154 0.080055695 9.473978e-02 0.183380170
#>  [59,] -0.19108462 0.080373803 9.511647e-02 0.183544821
#>  [60,] -0.19134249 0.080663546 9.545988e-02 0.183694459
#>  [61,] -0.19157748 0.080927518 9.577292e-02 0.183830568
#>  [62,] -0.19179163 0.081168014 9.605826e-02 0.183954390
#>  [63,] -0.19198679 0.081387124 9.631836e-02 0.184067049
#>  [64,] -0.19216463 0.081586751 9.655542e-02 0.184169564
#>  [65,] -0.19232668 0.081768628 9.677149e-02 0.184262861
#>  [66,] -0.19247436 0.081934335 9.696842e-02 0.184347777
#>  [67,] -0.19260893 0.082085311 9.714789e-02 0.184425072
#>  [68,] -0.19273155 0.082222866 9.731146e-02 0.184495436
#>  [69,] -0.19284329 0.082348193 9.746053e-02 0.184559497
#>  [70,] -0.19294511 0.082462380 9.759638e-02 0.184617823
#>  [71,] -0.19303789 0.082566418 9.772019e-02 0.184670930
#>  [72,] -0.19312243 0.082661210 9.783301e-02 0.184719290
#>  [73,] -0.19319946 0.082747576 9.793582e-02 0.184763328
#>  [74,] -0.19326965 0.082826268 9.802952e-02 0.184803433
#>  [75,] -0.19333361 0.082897966 9.811490e-02 0.184839958
#>  [76,] -0.19339189 0.082963292 9.819270e-02 0.184873224
#>  [77,] -0.19344500 0.083022813 9.826359e-02 0.184903522
#>  [78,] -0.19349339 0.083077045 9.832820e-02 0.184931119
#>  [79,] -0.19353748 0.083126458 9.838707e-02 0.184956256
#>  [80,] -0.19357765 0.083171481 9.844071e-02 0.184979154
#>  [81,] -0.19361426 0.083212503 9.848959e-02 0.185000011
#>  [82,] -0.19364761 0.083249880 9.853414e-02 0.185019011
#>  [83,] -0.19367801 0.083283935 9.857472e-02 0.185036319
#>  [84,] -0.19370570 0.083314965 9.861171e-02 0.185052086
#>  [85,] -0.19373093 0.083343238 9.864541e-02 0.185066450
#>  [86,] -0.19375392 0.083368999 9.867611e-02 0.185079536
#>  [87,] -0.19377487 0.083392471 9.870409e-02 0.185091457
#>  [88,] -0.19379396 0.083413858 9.872959e-02 0.185102318
#>  [89,] -0.19381135 0.083433345 9.875282e-02 0.185112212
#>  [90,] -0.19382720 0.083451100 9.877399e-02 0.185121227
#>  [91,] -0.19384164 0.083467278 9.879327e-02 0.185129439
#>  [92,] -0.19385480 0.083482018 9.881085e-02 0.185136922
#>  [93,] -0.19386679 0.083495449 9.882686e-02 0.185143739
#>  [94,] -0.19387771 0.083507687 9.884145e-02 0.185149950
#>  [95,] -0.19388766 0.083518837 9.885475e-02 0.185155609
#>  [96,] -0.19389673 0.083528997 9.886686e-02 0.185160765
#>  [97,] -0.19390500 0.083538254 9.887790e-02 0.185165462
#>  [98,] -0.19391253 0.083546689 9.888796e-02 0.185169742
#>  [99,] -0.19391939 0.083554375 9.889712e-02 0.185173641
#> [100,] -0.19392564 0.083561377 9.890547e-02 0.185177194
#> 
#> $cvm
#>   [1] -4.190334 -4.178084 -4.129881 -4.045388 -3.953466 -3.833948 -3.707463
#>   [8] -3.584668 -3.476963 -3.388387 -3.326500 -3.278578 -3.233284 -3.179984
#>  [15] -3.135079 -3.093548 -3.049406 -3.028854 -3.020483 -3.020512 -3.029580
#>  [22] -3.049236 -3.073112 -3.098564 -3.124167 -3.149905 -3.175122 -3.199374
#>  [29] -3.222430 -3.246757 -3.273195 -3.298217 -3.321407 -3.342652 -3.362772
#>  [36] -3.379331 -3.393996 -3.406455 -3.417978 -3.428437 -3.436671 -3.444199
#>  [43] -3.449711 -3.454713 -3.459407 -3.463796 -3.467890 -3.471697 -3.475232
#>  [50] -3.478505 -3.481533 -3.484329 -3.486908 -3.489283 -3.491468 -3.493477
#>  [57] -3.495322 -3.497015 -3.498568 -3.499991 -3.501295 -3.502489 -3.503582
#>  [64] -3.504581 -3.505495 -3.506331 -3.507094 -3.507792 -3.508429 -3.509011
#>  [71] -3.509543 -3.510028 -3.510470 -3.510874 -3.511243 -3.511579 -3.511886
#>  [78] -3.512166 -3.512421 -3.512654 -3.512866 -3.513060 -3.513236 -3.513397
#>  [85] -3.513544 -3.513677 -3.513799 -3.513910 -3.514012 -3.514104 -3.514188
#>  [92] -3.514265 -3.514334 -3.514398 -3.514456 -3.514509 -3.514557 -3.514601
#>  [99] -3.514641 -3.514678
#> 
#> $cvsd
#>   [1] 3.486862 3.469744 3.432242 3.395391 3.367223 3.298621 3.225892 3.154963
#>   [9] 3.097740 3.049386 3.010566 2.981488 2.953597 2.911512 2.880787 2.868417
#>  [17] 2.840548 2.812824 2.789785 2.770954 2.756685 2.744644 2.734116 2.724079
#>  [25] 2.715552 2.707967 2.700879 2.695161 2.690523 2.684334 2.680573 2.678353
#>  [33] 2.675400 2.673955 2.673815 2.676120 2.679354 2.683394 2.687683 2.692048
#>  [41] 2.695631 2.699523 2.703570 2.707462 2.711243 2.714888 2.718377 2.721698
#>  [49] 2.724843 2.727808 2.730594 2.733203 2.735639 2.737908 2.740016 2.741972
#>  [57] 2.743782 2.745456 2.747000 2.748425 2.749736 2.750943 2.752052 2.753071
#>  [65] 2.754005 2.754863 2.755648 2.756368 2.757027 2.757630 2.758182 2.758687
#>  [73] 2.759148 2.759569 2.759955 2.760306 2.760628 2.760921 2.761189 2.761433
#>  [81] 2.761656 2.761860 2.762046 2.762215 2.762369 2.762510 2.762639 2.762756
#>  [89] 2.762863 2.762960 2.763049 2.763130 2.763204 2.763271 2.763333 2.763389
#>  [97] 2.763440 2.763486 2.763528 2.763567
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x562e9b60fed8>
#> 
#> attr(,"class")
#> [1] "cv_survtran"