Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

cbh_func

a function specifying the cumulative baseline hazard. The default is NULL, which means the cumulative baseline hazard is estimateds from the survode object fitted to the source data.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6205606  1.3934976 -0.9954553  1.0047907 -0.8622007  1.0413994 -1.1190564 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0769175 
#> 
#> $coefficients$eta
#>  [1]  0.383066730  0.336496307 -0.000120978  0.000000000  0.161826385
#>  [6]  0.000000000 -0.087038519  0.000000000  0.000000000  0.074318768
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.06076032
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]         [,3]         [,4]        [,5]         [,6]
#>   [1,] 0.08525555 0.02937811  0.000000000  0.000000000 0.000000000  0.000000000
#>   [2,] 0.11560547 0.06612220  0.000000000  0.000000000 0.000000000  0.000000000
#>   [3,] 0.14426287 0.10107474  0.000000000  0.000000000 0.000000000  0.000000000
#>   [4,] 0.17117741 0.13349865  0.000000000  0.000000000 0.003269571  0.000000000
#>   [5,] 0.19645868 0.16138756  0.000000000  0.000000000 0.020975900  0.000000000
#>   [6,] 0.22020318 0.18782312  0.000000000  0.000000000 0.037552409  0.000000000
#>   [7,] 0.24239759 0.21273070  0.000000000  0.000000000 0.052994375  0.000000000
#>   [8,] 0.26305121 0.23606940  0.000000000  0.000000000 0.067314565  0.000000000
#>   [9,] 0.28198608 0.25734278  0.000000000  0.000000000 0.080509645  0.000000000
#>  [10,] 0.29905835 0.27615901  0.000000000  0.000000000 0.092749554  0.000000000
#>  [11,] 0.31480521 0.29367296  0.000000000  0.000000000 0.104192227  0.000000000
#>  [12,] 0.33105989 0.30374752  0.000000000  0.000000000 0.117920603  0.000000000
#>  [13,] 0.34593018 0.31299541  0.000000000  0.000000000 0.130478490  0.000000000
#>  [14,] 0.35949801 0.32150627  0.000000000  0.000000000 0.141925526  0.000000000
#>  [15,] 0.37186101 0.32932159  0.000000000  0.000000000 0.152347388  0.000000000
#>  [16,] 0.38306673 0.33649631 -0.000120978  0.000000000 0.161826385  0.000000000
#>  [17,] 0.39039001 0.34366588 -0.007496150  0.000000000 0.170431273  0.000000000
#>  [18,] 0.39708027 0.35021608 -0.014158549  0.000000000 0.178268621  0.000000000
#>  [19,] 0.40356100 0.35602662 -0.020346032  0.000000000 0.185418549  0.000000000
#>  [20,] 0.40986724 0.36116956 -0.026114220  0.000000000 0.191967092  0.000000000
#>  [21,] 0.41564433 0.36586641 -0.031341417  0.000000000 0.197953038  0.000000000
#>  [22,] 0.42102435 0.37015466 -0.035993716  0.000000000 0.203489192  0.000000000
#>  [23,] 0.42640225 0.37409791 -0.039760748  0.000000000 0.208896412  0.000000000
#>  [24,] 0.43130012 0.37771629 -0.043149177  0.000000000 0.213837865  0.000000000
#>  [25,] 0.43575932 0.38103276 -0.046200147  0.000000000 0.218351183  0.000000000
#>  [26,] 0.44027164 0.38678349 -0.049262850 -0.005499923 0.223815923 -0.005097018
#>  [27,] 0.44461958 0.39343938 -0.052147487 -0.013378114 0.229532908 -0.011863979
#>  [28,] 0.44860785 0.39961930 -0.054768963 -0.020627054 0.234788970 -0.018076129
#>  [29,] 0.45226340 0.40534198 -0.057150966 -0.027287827 0.239615366 -0.023772296
#>  [30,] 0.45561160 0.41062987 -0.059315311 -0.033401148 0.244042399 -0.028990331
#>  [31,] 0.45867637 0.41550678 -0.061281884 -0.039006380 0.248099325 -0.033766399
#>  [32,] 0.46148018 0.41999738 -0.063068819 -0.044141318 0.251814093 -0.038134857
#>  [33,] 0.46404402 0.42412686 -0.064692654 -0.048842106 0.255213195 -0.042128210
#>  [34,] 0.46638744 0.42791904 -0.066168350 -0.053142379 0.258321638 -0.045776623
#>  [35,] 0.46852860 0.43139823 -0.067509583 -0.057074266 0.261162815 -0.049108564
#>  [36,] 0.47048432 0.43458747 -0.068728746 -0.060667648 0.263758575 -0.052150398
#>  [37,] 0.47227015 0.43750868 -0.069837079 -0.063950343 0.266129236 -0.054926539
#>  [38,] 0.47390058 0.44018304 -0.070844753 -0.066948501 0.268293703 -0.057459699
#>  [39,] 0.47538857 0.44262921 -0.071761070 -0.069685365 0.270269248 -0.059770358
#>  [40,] 0.47674642 0.44486598 -0.072594401 -0.072183364 0.272071994 -0.061877823
#>  [41,] 0.47798528 0.44691040 -0.073352356 -0.074462847 0.273716727 -0.063799671
#>  [42,] 0.47911543 0.44877834 -0.074041835 -0.076542540 0.275217042 -0.065552029
#>  [43,] 0.48014627 0.45048448 -0.074669097 -0.078439646 0.276585418 -0.067149672
#>  [44,] 0.48108642 0.45204240 -0.075239820 -0.080169954 0.277833302 -0.068606129
#>  [45,] 0.48194377 0.45346463 -0.075759153 -0.081747937 0.278971180 -0.069933776
#>  [46,] 0.48272555 0.45476273 -0.076231772 -0.083186855 0.280008652 -0.071143927
#>  [47,] 0.48343835 0.45594729 -0.076661919 -0.084498843 0.280954502 -0.072246922
#>  [48,] 0.48408823 0.45702807 -0.077053444 -0.085695003 0.281816762 -0.073252202
#>  [49,] 0.48468069 0.45801403 -0.077409845 -0.086785486 0.282602771 -0.074168389
#>  [50,] 0.48522057 0.45891359 -0.077734411 -0.087779506 0.283319105 -0.075003386
#>  [51,] 0.48571290 0.45973386 -0.078029804 -0.088685674 0.283972145 -0.075764313
#>  [52,] 0.48616167 0.46048190 -0.078298755 -0.089511653 0.284567354 -0.076457744
#>  [53,] 0.48657071 0.46116401 -0.078543644 -0.090264507 0.285109835 -0.077089653
#>  [54,] 0.48694354 0.46178596 -0.078766639 -0.090950687 0.285604243 -0.077665490
#>  [55,] 0.48728334 0.46235302 -0.078969706 -0.091576077 0.286054826 -0.078190222
#>  [56,] 0.48759304 0.46286999 -0.079154636 -0.092146046 0.286465461 -0.078668378
#>  [57,] 0.48787529 0.46334128 -0.079323055 -0.092665493 0.286839679 -0.079104088
#>  [58,] 0.48813252 0.46377089 -0.079476446 -0.093138886 0.287180706 -0.079501116
#>  [59,] 0.48836694 0.46416250 -0.079616152 -0.093570299 0.287491480 -0.079862893
#>  [60,] 0.48858058 0.46451946 -0.079743401 -0.093963448 0.287774680 -0.080192548
#>  [61,] 0.48877527 0.46484482 -0.079859306 -0.094321722 0.288032750 -0.080492930
#>  [62,] 0.48895269 0.46514136 -0.079964882 -0.094648209 0.288267918 -0.080766637
#>  [63,] 0.48911437 0.46541163 -0.080061051 -0.094945727 0.288482213 -0.081016037
#>  [64,] 0.48926170 0.46565796 -0.080148654 -0.095216841 0.288677486 -0.081243288
#>  [65,] 0.48939596 0.46588245 -0.080228457 -0.095463894 0.288855425 -0.081450356
#>  [66,] 0.48951830 0.46608704 -0.080301154 -0.095689019 0.289017566 -0.081639033
#>  [67,] 0.48962979 0.46627350 -0.080367380 -0.095894160 0.289165313 -0.081810952
#>  [68,] 0.48973138 0.46644341 -0.080427712 -0.096081090 0.289299941 -0.081967601
#>  [69,] 0.48982395 0.46659826 -0.080482675 -0.096251424 0.289422615 -0.082110337
#>  [70,] 0.48990830 0.46673937 -0.080532748 -0.096406635 0.289534396 -0.082240394
#>  [71,] 0.48998516 0.46686796 -0.080578366 -0.096548064 0.289636250 -0.082358898
#>  [72,] 0.49005520 0.46698514 -0.080619927 -0.096676936 0.289729060 -0.082466877
#>  [73,] 0.49011902 0.46709192 -0.080657791 -0.096794364 0.289813627 -0.082565263
#>  [74,] 0.49017717 0.46718923 -0.080692288 -0.096901364 0.289890684 -0.082654911
#>  [75,] 0.49023016 0.46727790 -0.080723717 -0.096998862 0.289960897 -0.082736595
#>  [76,] 0.49027845 0.46735869 -0.080752352 -0.097087701 0.290024874 -0.082811023
#>  [77,] 0.49032244 0.46743232 -0.080778440 -0.097168651 0.290083169 -0.082878839
#>  [78,] 0.49036253 0.46749941 -0.080802210 -0.097242411 0.290136286 -0.082940631
#>  [79,] 0.49039906 0.46756054 -0.080823866 -0.097309620 0.290184686 -0.082996934
#>  [80,] 0.49043234 0.46761624 -0.080843598 -0.097370860 0.290228786 -0.083048236
#>  [81,] 0.49046267 0.46766700 -0.080861575 -0.097426660 0.290268969 -0.083094980
#>  [82,] 0.49049030 0.46771325 -0.080877955 -0.097477504 0.290305583 -0.083137572
#>  [83,] 0.49051548 0.46775540 -0.080892879 -0.097523832 0.290338945 -0.083176380
#>  [84,] 0.49053843 0.46779380 -0.080906476 -0.097566045 0.290369343 -0.083211741
#>  [85,] 0.49055933 0.46782879 -0.080918866 -0.097604509 0.290397041 -0.083243961
#>  [86,] 0.49057838 0.46786067 -0.080930154 -0.097639556 0.290422279 -0.083273318
#>  [87,] 0.49059574 0.46788973 -0.080940439 -0.097671490 0.290445275 -0.083300067
#>  [88,] 0.49061155 0.46791620 -0.080949810 -0.097700587 0.290466228 -0.083324440
#>  [89,] 0.49062596 0.46794032 -0.080958348 -0.097727099 0.290485320 -0.083346648
#>  [90,] 0.49063909 0.46796230 -0.080966127 -0.097751257 0.290502716 -0.083366883
#>  [91,] 0.49065106 0.46798232 -0.080973216 -0.097773268 0.290518566 -0.083385320
#>  [92,] 0.49066196 0.46800057 -0.080979674 -0.097793324 0.290533009 -0.083402120
#>  [93,] 0.49067189 0.46801720 -0.080985559 -0.097811599 0.290546168 -0.083417427
#>  [94,] 0.49068094 0.46803235 -0.080990921 -0.097828250 0.290558159 -0.083431374
#>  [95,] 0.49068919 0.46804615 -0.080995806 -0.097843422 0.290569084 -0.083444083
#>  [96,] 0.49069670 0.46805873 -0.081000257 -0.097857247 0.290579039 -0.083455662
#>  [97,] 0.49070355 0.46807019 -0.081004313 -0.097869843 0.290588109 -0.083466213
#>  [98,] 0.49070978 0.46808063 -0.081008009 -0.097881320 0.290596374 -0.083475826
#>  [99,] 0.49071547 0.46809015 -0.081011376 -0.097891778 0.290603905 -0.083484585
#> [100,] 0.49072065 0.46809882 -0.081014444 -0.097901307 0.290610766 -0.083492567
#>                [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,] -0.004351415 0.000000000 0.000000000 0.00000000
#>  [10,] -0.017451542 0.000000000 0.000000000 0.00000000
#>  [11,] -0.029568468 0.000000000 0.000000000 0.00000000
#>  [12,] -0.043135163 0.000000000 0.000000000 0.01836349
#>  [13,] -0.055618591 0.000000000 0.000000000 0.03484697
#>  [14,] -0.067055368 0.000000000 0.000000000 0.04952983
#>  [15,] -0.077509522 0.000000000 0.000000000 0.06262110
#>  [16,] -0.087038519 0.000000000 0.000000000 0.07431877
#>  [17,] -0.095123139 0.000000000 0.000000000 0.08555381
#>  [18,] -0.102462559 0.000000000 0.000000000 0.09560259
#>  [19,] -0.109050584 0.000000000 0.003027084 0.10459625
#>  [20,] -0.114960283 0.000000000 0.008808983 0.11265650
#>  [21,] -0.120323284 0.000000000 0.014108973 0.11989398
#>  [22,] -0.125439473 0.001288195 0.018965852 0.12669269
#>  [23,] -0.131381192 0.008975232 0.023441803 0.13428024
#>  [24,] -0.136815270 0.015955991 0.027549632 0.14110704
#>  [25,] -0.141780275 0.022295584 0.031315369 0.14725505
#>  [26,] -0.146158628 0.027801382 0.036225052 0.15169948
#>  [27,] -0.150168533 0.032728168 0.041447172 0.15511555
#>  [28,] -0.153874296 0.037228606 0.046272458 0.15815267
#>  [29,] -0.157291879 0.041337584 0.050723422 0.16085748
#>  [30,] -0.160437868 0.045087424 0.054822999 0.16327007
#>  [31,] -0.163329376 0.048508241 0.058593997 0.16542529
#>  [32,] -0.165983505 0.051627938 0.062058798 0.16735342
#>  [33,] -0.168416957 0.054472244 0.065239138 0.16908066
#>  [34,] -0.170646157 0.057065044 0.068155749 0.17063021
#>  [35,] -0.172686549 0.059428130 0.070828510 0.17202197
#>  [36,] -0.174552867 0.061581567 0.073276187 0.17327343
#>  [37,] -0.176259000 0.063543754 0.075516436 0.17439996
#>  [38,] -0.177817922 0.065331519 0.077565883 0.17541490
#>  [39,] -0.179241829 0.066960308 0.079439778 0.17633043
#>  [40,] -0.180541924 0.068444166 0.081152570 0.17715687
#>  [41,] -0.181728627 0.069795946 0.082717573 0.17790346
#>  [42,] -0.182811560 0.071027377 0.084147108 0.17857843
#>  [43,] -0.183799588 0.072149159 0.085452554 0.17918903
#>  [44,] -0.184700867 0.073171048 0.086644402 0.17974176
#>  [45,] -0.185522889 0.074101937 0.087732311 0.18024238
#>  [46,] -0.186272526 0.074949933 0.088725163 0.18069604
#>  [47,] -0.186956075 0.075722419 0.089631114 0.18110734
#>  [48,] -0.187579303 0.076426126 0.090457651 0.18148039
#>  [49,] -0.188147487 0.077067181 0.091211638 0.18181890
#>  [50,] -0.188665288 0.077651069 0.091899348 0.18212606
#>  [51,] -0.189137435 0.078183071 0.092526573 0.18240505
#>  [52,] -0.189567801 0.078667724 0.093098565 0.18265845
#>  [53,] -0.189960067 0.079109248 0.093620143 0.18288867
#>  [54,] -0.190317592 0.079511483 0.094095717 0.18309788
#>  [55,] -0.190643440 0.079877931 0.094529315 0.18328806
#>  [56,] -0.190940410 0.080211779 0.094924617 0.18346096
#>  [57,] -0.191211053 0.080515929 0.095284988 0.18361818
#>  [58,] -0.191457699 0.080793027 0.095613497 0.18376118
#>  [59,] -0.191682470 0.081045480 0.095912950 0.18389126
#>  [60,] -0.191887303 0.081275483 0.096185904 0.18400960
#>  [61,] -0.192073963 0.081485033 0.096434696 0.18411729
#>  [62,] -0.192244061 0.081675951 0.096661458 0.18421528
#>  [63,] -0.192399064 0.081849894 0.096868134 0.18430446
#>  [64,] -0.192540310 0.082008373 0.097056498 0.18438564
#>  [65,] -0.192669019 0.082152764 0.097228170 0.18445954
#>  [66,] -0.192786302 0.082284319 0.097384624 0.18452681
#>  [67,] -0.192893174 0.082404180 0.097527207 0.18458805
#>  [68,] -0.192990558 0.082513388 0.097657146 0.18464382
#>  [69,] -0.193079295 0.082612889 0.097775561 0.18469460
#>  [70,] -0.193160153 0.082703547 0.097883472 0.18474084
#>  [71,] -0.193233831 0.082786148 0.097981809 0.18478295
#>  [72,] -0.193300966 0.082861407 0.098071422 0.18482130
#>  [73,] -0.193362140 0.082929979 0.098153082 0.18485622
#>  [74,] -0.193417881 0.082992457 0.098227495 0.18488803
#>  [75,] -0.193468671 0.083049383 0.098295304 0.18491701
#>  [76,] -0.193514951 0.083101251 0.098357095 0.18494340
#>  [77,] -0.193557120 0.083148510 0.098413400 0.18496744
#>  [78,] -0.193595544 0.083191569 0.098464706 0.18498933
#>  [79,] -0.193630556 0.083230803 0.098511458 0.18500928
#>  [80,] -0.193662457 0.083266550 0.098554059 0.18502745
#>  [81,] -0.193691525 0.083299122 0.098592877 0.18504400
#>  [82,] -0.193718011 0.083328799 0.098628248 0.18505908
#>  [83,] -0.193742145 0.083355839 0.098660479 0.18507282
#>  [84,] -0.193764135 0.083380477 0.098689847 0.18508533
#>  [85,] -0.193784171 0.083402926 0.098716608 0.18509673
#>  [86,] -0.193802428 0.083423381 0.098740992 0.18510712
#>  [87,] -0.193819063 0.083442018 0.098763210 0.18511658
#>  [88,] -0.193834220 0.083458999 0.098783455 0.18512521
#>  [89,] -0.193848031 0.083474472 0.098801902 0.18513306
#>  [90,] -0.193860615 0.083488570 0.098818710 0.18514022
#>  [91,] -0.193872081 0.083501415 0.098834026 0.18514674
#>  [92,] -0.193882529 0.083513119 0.098847981 0.18515268
#>  [93,] -0.193892048 0.083523784 0.098860697 0.18515809
#>  [94,] -0.193900722 0.083533501 0.098872284 0.18516302
#>  [95,] -0.193908626 0.083542354 0.098882841 0.18516751
#>  [96,] -0.193915827 0.083550421 0.098892460 0.18517161
#>  [97,] -0.193922389 0.083557772 0.098901225 0.18517534
#>  [98,] -0.193928367 0.083564469 0.098909211 0.18517873
#>  [99,] -0.193933815 0.083570572 0.098916488 0.18518183
#> [100,] -0.193938778 0.083576132 0.098923119 0.18518465
#> 
#> $cvm
#>   [1] -3.965866 -3.831605 -3.719759 -3.626059 -3.529807 -3.430873 -3.347054
#>   [8] -3.277667 -3.228877 -3.182558 -3.146578 -3.117998 -3.093948 -3.072167
#>  [15] -3.054902 -3.051728 -3.056650 -3.067763 -3.082526 -3.099901 -3.117715
#>  [22] -3.137688 -3.160386 -3.183297 -3.203597 -3.226202 -3.252386 -3.277850
#>  [29] -3.301852 -3.324976 -3.347093 -3.366000 -3.380644 -3.392943 -3.404899
#>  [36] -3.415914 -3.426338 -3.436159 -3.445375 -3.453996 -3.462036 -3.469517
#>  [43] -3.476461 -3.482896 -3.488847 -3.494343 -3.499412 -3.504081 -3.508378
#>  [50] -3.512328 -3.515957 -3.519287 -3.522342 -3.525142 -3.527707 -3.530056
#>  [57] -3.532206 -3.534172 -3.535971 -3.537615 -3.539117 -3.540491 -3.541745
#>  [64] -3.542890 -3.543936 -3.544890 -3.545762 -3.546557 -3.547283 -3.547945
#>  [71] -3.548548 -3.549099 -3.549602 -3.550060 -3.550477 -3.550858 -3.551206
#>  [78] -3.551522 -3.551811 -3.552074 -3.552314 -3.552533 -3.552732 -3.552914
#>  [85] -3.553080 -3.553230 -3.553368 -3.553493 -3.553608 -3.553712 -3.553807
#>  [92] -3.553893 -3.553972 -3.554044 -3.554109 -3.554169 -3.554223 -3.554273
#>  [99] -3.554318 -3.554359
#> 
#> $cvsd
#>   [1] 1.913293 1.855278 1.813828 1.782930 1.755094 1.736276 1.725301 1.720365
#>   [9] 1.715618 1.722385 1.736921 1.759492 1.779146 1.799721 1.813456 1.815849
#>  [17] 1.816915 1.812045 1.803266 1.795421 1.788080 1.782072 1.777734 1.774242
#>  [25] 1.772306 1.773599 1.777281 1.781756 1.786781 1.792357 1.798311 1.803654
#>  [33] 1.809015 1.815236 1.821598 1.827654 1.833689 1.839623 1.845395 1.850961
#>  [41] 1.856290 1.861360 1.866159 1.870683 1.874929 1.878904 1.882612 1.886063
#>  [49] 1.889269 1.892240 1.894989 1.897529 1.899873 1.902032 1.904020 1.905847
#>  [57] 1.907527 1.909068 1.910483 1.911780 1.912968 1.914056 1.915052 1.915963
#>  [65] 1.916797 1.917559 1.918256 1.918892 1.919474 1.920005 1.920490 1.920932
#>  [73] 1.921336 1.921705 1.922042 1.922349 1.922629 1.922884 1.923117 1.923330
#>  [81] 1.923524 1.923700 1.923862 1.924009 1.924143 1.924265 1.924376 1.924477
#>  [89] 1.924570 1.924654 1.924731 1.924801 1.924865 1.924923 1.924976 1.925024
#>  [97] 1.925068 1.925109 1.925145 1.925179
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x559796719170>
#> 
#> attr(,"class")
#> [1] "cv_survtran"