Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6166638  1.3973123 -0.9993905  1.0047907 -0.8576187  1.0413994 -1.1233648 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0829270 
#> 
#> $coefficients$eta
#>  [1]  0.386963515  0.340311054 -0.004056181  0.000000000  0.166408426
#>  [6]  0.000000000 -0.091346960  0.000000000  0.000000000  0.080328243
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05788603
#> 
#> $coefficients$formula
#> Surv(time, status) ~ .
#> <environment: 0x5591f475f1f8>
#> 
#> 
#> $lambdas
#>   [1] 3.089203e-01 2.814767e-01 2.564711e-01 2.336869e-01 2.129268e-01
#>   [6] 1.940110e-01 1.767756e-01 1.610713e-01 1.467622e-01 1.337242e-01
#>  [11] 1.218445e-01 1.110202e-01 1.011575e-01 9.217093e-02 8.398272e-02
#>  [16] 7.652192e-02 6.972393e-02 6.352984e-02 5.788603e-02 5.274359e-02
#>  [21] 4.805800e-02 4.378866e-02 3.989859e-02 3.635411e-02 3.312451e-02
#>  [26] 3.018182e-02 2.750055e-02 2.505748e-02 2.283144e-02 2.080316e-02
#>  [31] 1.895506e-02 1.727115e-02 1.573683e-02 1.433881e-02 1.306499e-02
#>  [36] 1.190433e-02 1.084678e-02 9.883186e-03 9.005191e-03 8.205195e-03
#>  [41] 7.476268e-03 6.812097e-03 6.206929e-03 5.655523e-03 5.153102e-03
#>  [46] 4.695314e-03 4.278195e-03 3.898132e-03 3.551833e-03 3.236298e-03
#>  [51] 2.948794e-03 2.686831e-03 2.448141e-03 2.230655e-03 2.032489e-03
#>  [56] 1.851929e-03 1.687408e-03 1.537504e-03 1.400916e-03 1.276463e-03
#>  [61] 1.163065e-03 1.059742e-03 9.655971e-04 8.798162e-04 8.016557e-04
#>  [66] 7.304388e-04 6.655487e-04 6.064231e-04 5.525502e-04 5.034632e-04
#>  [71] 4.587369e-04 4.179840e-04 3.808514e-04 3.470176e-04 3.161895e-04
#>  [76] 2.881001e-04 2.625061e-04 2.391858e-04 2.179372e-04 1.985762e-04
#>  [81] 1.809353e-04 1.648615e-04 1.502156e-04 1.368709e-04 1.247117e-04
#>  [86] 1.136326e-04 1.035378e-04 9.433980e-05 8.595891e-05 7.832256e-05
#>  [91] 7.136460e-05 6.502477e-05 5.924815e-05 5.398470e-05 4.918885e-05
#>  [96] 4.481905e-05 4.083745e-05 3.720956e-05 3.390397e-05 3.089203e-05
#> 
#> $etas
#>              [,1]       [,2]         [,3]         [,4]       [,5]         [,6]
#>   [1,] 0.00000000 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [2,] 0.03624828 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [3,] 0.07013592 0.01117874  0.000000000  0.000000000 0.00000000  0.000000000
#>   [4,] 0.10127198 0.04873376  0.000000000  0.000000000 0.00000000  0.000000000
#>   [5,] 0.13074780 0.08455985  0.000000000  0.000000000 0.00000000  0.000000000
#>   [6,] 0.15850358 0.11853574  0.000000000  0.000000000 0.00000000  0.000000000
#>   [7,] 0.18453640 0.14820136  0.000000000  0.000000000 0.01263273  0.000000000
#>   [8,] 0.20902026 0.17534465  0.000000000  0.000000000 0.02975226  0.000000000
#>   [9,] 0.23195752 0.20099142  0.000000000  0.000000000 0.04573723  0.000000000
#>  [10,] 0.25334680 0.22508480  0.000000000  0.000000000 0.06059224  0.000000000
#>  [11,] 0.27320849 0.24760071  0.000000000  0.000000000 0.07433839  0.000000000
#>  [12,] 0.29104838 0.26730795  0.000000000  0.000000000 0.08698454  0.000000000
#>  [13,] 0.30742198 0.28544282  0.000000000  0.000000000 0.09880956  0.000000000
#>  [14,] 0.32344554 0.29904410  0.000000000  0.000000000 0.11148530  0.000000000
#>  [15,] 0.33897429 0.30865928  0.000000000  0.000000000 0.12460584  0.000000000
#>  [16,] 0.35315394 0.31751815  0.000000000  0.000000000 0.13657435  0.000000000
#>  [17,] 0.36608210 0.32566114  0.000000000  0.000000000 0.14747689  0.000000000
#>  [18,] 0.37785551 0.33313210  0.000000000  0.000000000 0.15739763  0.000000000
#>  [19,] 0.38696352 0.34031105 -0.004056181  0.000000000 0.16640843  0.000000000
#>  [20,] 0.39395082 0.34715203 -0.011050927  0.000000000 0.17460549  0.000000000
#>  [21,] 0.40034022 0.35339416 -0.017374371  0.000000000 0.18206842  0.000000000
#>  [22,] 0.40691441 0.35876341 -0.023421540  0.000000000 0.18890264  0.000000000
#>  [23,] 0.41293999 0.36366977 -0.028901157  0.000000000 0.19515246  0.000000000
#>  [24,] 0.41845722 0.36814793 -0.033867320  0.000000000 0.20086293  0.000000000
#>  [25,] 0.42388826 0.37225069 -0.038005906  0.000000000 0.20636630  0.000000000
#>  [26,] 0.42901076 0.37602179 -0.041570274  0.000000000 0.21152608  0.000000000
#>  [27,] 0.43367500 0.37948023 -0.044778194  0.000000000 0.21623991  0.000000000
#>  [28,] 0.43805289 0.38342116 -0.047780050 -0.001490303 0.22090435 -0.001646685
#>  [29,] 0.44258305 0.39031098 -0.050799830 -0.009684728 0.22685333 -0.008693578
#>  [30,] 0.44674019 0.39671685 -0.053544309 -0.017230002 0.23232628 -0.015166641
#>  [31,] 0.45055191 0.40265579 -0.056038182 -0.024167376 0.23735471 -0.021105166
#>  [32,] 0.45404428 0.40814908 -0.058304212 -0.030537978 0.24196936 -0.026547654
#>  [33,] 0.45724195 0.41321985 -0.060363170 -0.036381815 0.24620003 -0.031531067
#>  [34,] 0.46016807 0.41789270 -0.062234033 -0.041737593 0.25007532 -0.036090723
#>  [35,] 0.46284439 0.42219218 -0.063934036 -0.046642115 0.25362249 -0.040259906
#>  [36,] 0.46529109 0.42614233 -0.065478862 -0.051129900 0.25686721 -0.044069743
#>  [37,] 0.46752700 0.42976814 -0.066882853 -0.055234248 0.25983364 -0.047549775
#>  [38,] 0.46956948 0.43309365 -0.068159075 -0.058986203 0.26254428 -0.050727473
#>  [39,] 0.47143484 0.43614053 -0.069319165 -0.062414369 0.26502035 -0.053627959
#>  [40,] 0.47313793 0.43893037 -0.070373864 -0.065545553 0.26728131 -0.056274733
#>  [41,] 0.47469247 0.44148366 -0.071332924 -0.068404707 0.26934520 -0.058689534
#>  [42,] 0.47611120 0.44381863 -0.072205031 -0.071014533 0.27122880 -0.060892086
#>  [43,] 0.47740572 0.44595324 -0.072998209 -0.073396305 0.27294746 -0.062900788
#>  [44,] 0.47858672 0.44790391 -0.073719687 -0.075569511 0.27451533 -0.064732447
#>  [45,] 0.47966401 0.44968588 -0.074376025 -0.077552069 0.27594542 -0.066402473
#>  [46,] 0.48064657 0.45131325 -0.074973172 -0.079360432 0.27724966 -0.067924979
#>  [47,] 0.48154264 0.45279906 -0.075516523 -0.081009694 0.27843899 -0.069312883
#>  [48,] 0.48235976 0.45415531 -0.076010976 -0.082513681 0.27952341 -0.070577996
#>  [49,] 0.48310481 0.45539305 -0.076460975 -0.083885058 0.28051211 -0.071731114
#>  [50,] 0.48378411 0.45652244 -0.076870553 -0.085135409 0.28141346 -0.072782100
#>  [51,] 0.48440342 0.45755281 -0.077243373 -0.086275333 0.28223512 -0.073739959
#>  [52,] 0.48496800 0.45849271 -0.077582760 -0.087314514 0.28298410 -0.074612913
#>  [53,] 0.48548266 0.45934999 -0.077891736 -0.088261802 0.28366680 -0.075408462
#>  [54,] 0.48595179 0.46013182 -0.078173045 -0.089125281 0.28428905 -0.076133453
#>  [55,] 0.48637941 0.46084477 -0.078429180 -0.089912331 0.28485618 -0.076794129
#>  [56,] 0.48676917 0.46149487 -0.078662408 -0.090629689 0.28537307 -0.077396184
#>  [57,] 0.48712442 0.46208760 -0.078874789 -0.091283505 0.28584414 -0.077944811
#>  [58,] 0.48744819 0.46262800 -0.079068197 -0.091879390 0.28627345 -0.078444744
#>  [59,] 0.48774327 0.46312066 -0.079244334 -0.092422461 0.28666470 -0.078900300
#>  [60,] 0.48801220 0.46356976 -0.079404749 -0.092917389 0.28702124 -0.079315414
#>  [61,] 0.48825729 0.46397915 -0.079550852 -0.093368431 0.28734616 -0.079693673
#>  [62,] 0.48848064 0.46435232 -0.079683925 -0.093779472 0.28764225 -0.080038346
#>  [63,] 0.48868419 0.46469245 -0.079805132 -0.094154053 0.28791207 -0.080352415
#>  [64,] 0.48886969 0.46500247 -0.079915537 -0.094495403 0.28815794 -0.080638594
#>  [65,] 0.48903872 0.46528503 -0.080016103 -0.094806466 0.28838200 -0.080899358
#>  [66,] 0.48919276 0.46554256 -0.080107711 -0.095089926 0.28858616 -0.081136965
#>  [67,] 0.48933313 0.46577726 -0.080191161 -0.095348230 0.28877221 -0.081353470
#>  [68,] 0.48946105 0.46599116 -0.080267179 -0.095583608 0.28894173 -0.081550746
#>  [69,] 0.48957761 0.46618610 -0.080336431 -0.095798093 0.28909621 -0.081730501
#>  [70,] 0.48968382 0.46636375 -0.080399518 -0.095993538 0.28923697 -0.081894290
#>  [71,] 0.48978061 0.46652564 -0.080456991 -0.096171632 0.28936523 -0.082043531
#>  [72,] 0.48986880 0.46667318 -0.080509350 -0.096333915 0.28948211 -0.082179515
#>  [73,] 0.48994917 0.46680763 -0.080557051 -0.096481788 0.28958860 -0.082303422
#>  [74,] 0.49002240 0.46693014 -0.080600509 -0.096616532 0.28968564 -0.082416322
#>  [75,] 0.49008912 0.46704179 -0.080640102 -0.096739311 0.28977406 -0.082519193
#>  [76,] 0.49014993 0.46714352 -0.080676173 -0.096851187 0.28985463 -0.082612927
#>  [77,] 0.49020533 0.46723623 -0.080709036 -0.096953128 0.28992804 -0.082698334
#>  [78,] 0.49025581 0.46732071 -0.080738977 -0.097046016 0.28999493 -0.082776154
#>  [79,] 0.49030181 0.46739769 -0.080766257 -0.097130654 0.29005588 -0.082847062
#>  [80,] 0.49034373 0.46746784 -0.080791111 -0.097207776 0.29011142 -0.082911671
#>  [81,] 0.49038192 0.46753175 -0.080813755 -0.097278048 0.29016203 -0.082970541
#>  [82,] 0.49041672 0.46759000 -0.080834386 -0.097342079 0.29020814 -0.083024181
#>  [83,] 0.49044843 0.46764307 -0.080853184 -0.097400423 0.29025015 -0.083073056
#>  [84,] 0.49047733 0.46769143 -0.080870311 -0.097453584 0.29028843 -0.083117589
#>  [85,] 0.49050365 0.46773549 -0.080885915 -0.097502024 0.29032332 -0.083158167
#>  [86,] 0.49052764 0.46777564 -0.080900133 -0.097546161 0.29035510 -0.083195139
#>  [87,] 0.49054950 0.46781223 -0.080913087 -0.097586378 0.29038406 -0.083228827
#>  [88,] 0.49056942 0.46784557 -0.080924890 -0.097623022 0.29041045 -0.083259523
#>  [89,] 0.49058756 0.46787594 -0.080935644 -0.097656412 0.29043449 -0.083287492
#>  [90,] 0.49060410 0.46790362 -0.080945442 -0.097686835 0.29045640 -0.083312976
#>  [91,] 0.49061916 0.46792884 -0.080954370 -0.097714556 0.29047636 -0.083336196
#>  [92,] 0.49063289 0.46795182 -0.080962504 -0.097739815 0.29049455 -0.083357353
#>  [93,] 0.49064540 0.46797276 -0.080969916 -0.097762830 0.29051112 -0.083376631
#>  [94,] 0.49065680 0.46799184 -0.080976669 -0.097783800 0.29052622 -0.083394197
#>  [95,] 0.49066719 0.46800923 -0.080982822 -0.097802908 0.29053998 -0.083410202
#>  [96,] 0.49067665 0.46802507 -0.080988428 -0.097820318 0.29055252 -0.083424785
#>  [97,] 0.49068527 0.46803950 -0.080993536 -0.097836182 0.29056394 -0.083438072
#>  [98,] 0.49069313 0.46805265 -0.080998190 -0.097850636 0.29057435 -0.083450179
#>  [99,] 0.49070028 0.46806464 -0.081002431 -0.097863807 0.29058384 -0.083461211
#> [100,] 0.49070681 0.46807555 -0.081006295 -0.097875807 0.29059248 -0.083471263
#>               [,7]        [,8]         [,9]       [,10]
#>   [1,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [2,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [3,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [4,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [5,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [6,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [7,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [8,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [9,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [10,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [11,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [12,] -0.01129958 0.000000000 0.000000e+00 0.000000000
#>  [13,] -0.02388400 0.000000000 0.000000e+00 0.000000000
#>  [14,] -0.03676655 0.000000000 0.000000e+00 0.009760774
#>  [15,] -0.04977201 0.000000000 0.000000e+00 0.027188113
#>  [16,] -0.06170263 0.000000000 0.000000e+00 0.042706096
#>  [17,] -0.07261934 0.000000000 0.000000e+00 0.056535434
#>  [18,] -0.08258810 0.000000000 0.000000e+00 0.068872236
#>  [19,] -0.09134696 0.000000000 0.000000e+00 0.080328243
#>  [20,] -0.09903500 0.000000000 0.000000e+00 0.090927177
#>  [21,] -0.10601057 0.000000000 7.599179e-05 0.100413163
#>  [22,] -0.11220024 0.000000000 6.101133e-03 0.108903540
#>  [23,] -0.11781888 0.000000000 1.162767e-02 0.116523035
#>  [24,] -0.12291651 0.000000000 1.669035e-02 0.123368816
#>  [25,] -0.12860022 0.005384682 2.134492e-02 0.130745358
#>  [26,] -0.13427265 0.012695351 2.562583e-02 0.137925796
#>  [27,] -0.13945759 0.019334322 2.955225e-02 0.144389358
#>  [28,] -0.14412553 0.025279136 3.357610e-02 0.149913216
#>  [29,] -0.14828624 0.030423153 3.899612e-02 0.153529296
#>  [30,] -0.15213579 0.035123349 4.400876e-02 0.156741721
#>  [31,] -0.15568947 0.039415735 4.863620e-02 0.159600441
#>  [32,] -0.15896346 0.043333744 5.290126e-02 0.162148382
#>  [33,] -0.16197475 0.046908580 5.682686e-02 0.164422866
#>  [34,] -0.16474042 0.050169155 6.043564e-02 0.166456184
#>  [35,] -0.16727756 0.053142332 6.374963e-02 0.168276587
#>  [36,] -0.16960280 0.055852896 6.678998e-02 0.169908825
#>  [37,] -0.17173188 0.058323500 6.957712e-02 0.171374002
#>  [38,] -0.17367970 0.060574914 7.213036e-02 0.172690564
#>  [39,] -0.17546092 0.062626546 7.446785e-02 0.173875183
#>  [40,] -0.17708883 0.064495886 7.660666e-02 0.174942106
#>  [41,] -0.17857586 0.066198936 7.856280e-02 0.175903814
#>  [42,] -0.17993390 0.067750537 8.035106e-02 0.176771680
#>  [43,] -0.18117365 0.069164049 8.198528e-02 0.177555411
#>  [44,] -0.18230512 0.070451727 8.347825e-02 0.178263700
#>  [45,] -0.18333754 0.071624755 8.484179e-02 0.178904257
#>  [46,] -0.18427938 0.072693329 8.608682e-02 0.179483930
#>  [47,] -0.18513846 0.073666746 8.722338e-02 0.180008814
#>  [48,] -0.18592194 0.074553481 8.826072e-02 0.180484345
#>  [49,] -0.18663638 0.075361256 8.920734e-02 0.180915376
#>  [50,] -0.18728781 0.076097105 9.007103e-02 0.181306249
#>  [51,] -0.18788172 0.076767440 9.085897e-02 0.181660852
#>  [52,] -0.18842316 0.077378098 9.157769e-02 0.181982673
#>  [53,] -0.18891673 0.077934398 9.223322e-02 0.182274843
#>  [54,] -0.18936663 0.078441183 9.283105e-02 0.182540180
#>  [55,] -0.18977671 0.078902867 9.337620e-02 0.182781217
#>  [56,] -0.19015047 0.079323467 9.387329e-02 0.183000237
#>  [57,] -0.19049113 0.079706644 9.432652e-02 0.183199301
#>  [58,] -0.19080160 0.080055730 9.473973e-02 0.183380265
#>  [59,] -0.19108455 0.080373762 9.511644e-02 0.183544810
#>  [60,] -0.19134242 0.080663505 9.545985e-02 0.183694452
#>  [61,] -0.19157742 0.080927479 9.577289e-02 0.183830563
#>  [62,] -0.19179157 0.081167976 9.605824e-02 0.183954387
#>  [63,] -0.19198673 0.081387087 9.631833e-02 0.184067049
#>  [64,] -0.19216457 0.081586714 9.655539e-02 0.184169567
#>  [65,] -0.19232662 0.081768592 9.677146e-02 0.184262866
#>  [66,] -0.19247430 0.081934300 9.696838e-02 0.184347783
#>  [67,] -0.19260887 0.082085276 9.714786e-02 0.184425079
#>  [68,] -0.19273150 0.082222831 9.731143e-02 0.184495445
#>  [69,] -0.19284323 0.082348159 9.746050e-02 0.184559506
#>  [70,] -0.19294505 0.082462347 9.759635e-02 0.184617833
#>  [71,] -0.19303783 0.082566385 9.772015e-02 0.184670941
#>  [72,] -0.19312237 0.082661177 9.783298e-02 0.184719302
#>  [73,] -0.19319941 0.082747544 9.793579e-02 0.184763340
#>  [74,] -0.19326960 0.082826235 9.802948e-02 0.184803446
#>  [75,] -0.19333356 0.082897933 9.811486e-02 0.184839972
#>  [76,] -0.19339184 0.082963260 9.819266e-02 0.184873238
#>  [77,] -0.19344495 0.083022781 9.826356e-02 0.184903537
#>  [78,] -0.19349333 0.083077013 9.832816e-02 0.184931135
#>  [79,] -0.19353742 0.083126427 9.838703e-02 0.184956272
#>  [80,] -0.19357760 0.083171449 9.844068e-02 0.184979170
#>  [81,] -0.19361421 0.083212471 9.848956e-02 0.185000028
#>  [82,] -0.19364756 0.083249848 9.853410e-02 0.185019028
#>  [83,] -0.19367795 0.083283904 9.857469e-02 0.185036336
#>  [84,] -0.19370565 0.083314934 9.861167e-02 0.185052104
#>  [85,] -0.19373088 0.083343207 9.864537e-02 0.185066468
#>  [86,] -0.19375387 0.083368968 9.867607e-02 0.185079554
#>  [87,] -0.19377482 0.083392440 9.870405e-02 0.185091475
#>  [88,] -0.19379391 0.083413827 9.872955e-02 0.185102336
#>  [89,] -0.19381130 0.083433314 9.875278e-02 0.185112231
#>  [90,] -0.19382715 0.083451069 9.877395e-02 0.185121245
#>  [91,] -0.19384159 0.083467247 9.879324e-02 0.185129458
#>  [92,] -0.19385475 0.083481987 9.881081e-02 0.185136941
#>  [93,] -0.19386674 0.083495418 9.882682e-02 0.185143758
#>  [94,] -0.19387766 0.083507656 9.884141e-02 0.185149969
#>  [95,] -0.19388761 0.083518807 9.885471e-02 0.185155628
#>  [96,] -0.19389668 0.083528966 9.886682e-02 0.185160784
#>  [97,] -0.19390495 0.083538224 9.887786e-02 0.185165481
#>  [98,] -0.19391248 0.083546658 9.888792e-02 0.185169761
#>  [99,] -0.19391934 0.083554344 9.889708e-02 0.185173661
#> [100,] -0.19392559 0.083561347 9.890544e-02 0.185177214
#> 
#> $cvm
#>   [1] -4.152329 -4.106581 -4.040060 -3.936808 -3.839986 -3.709942 -3.596159
#>   [8] -3.502215 -3.420990 -3.344641 -3.290559 -3.251141 -3.219548 -3.196991
#>  [15] -3.173511 -3.161174 -3.147264 -3.136554 -3.134657 -3.146756 -3.165928
#>  [22] -3.186134 -3.211227 -3.236141 -3.263466 -3.291322 -3.318163 -3.342692
#>  [29] -3.365090 -3.386784 -3.407459 -3.427014 -3.443999 -3.456532 -3.468479
#>  [36] -3.479800 -3.490476 -3.499695 -3.506819 -3.514918 -3.523081 -3.530131
#>  [43] -3.536311 -3.542076 -3.547442 -3.552424 -3.557041 -3.561312 -3.565258
#>  [50] -3.568897 -3.572250 -3.575336 -3.578173 -3.580779 -3.583172 -3.585366
#>  [57] -3.587378 -3.589221 -3.590909 -3.592453 -3.593867 -3.595159 -3.596341
#>  [64] -3.597421 -3.598408 -3.599220 -3.599907 -3.600535 -3.601108 -3.601631
#>  [71] -3.602109 -3.602545 -3.602942 -3.603305 -3.603636 -3.603938 -3.604214
#>  [78] -3.604465 -3.604694 -3.604903 -3.605093 -3.605267 -3.605425 -3.605569
#>  [85] -3.605701 -3.605821 -3.605930 -3.606030 -3.606120 -3.606203 -3.606279
#>  [92] -3.606347 -3.606410 -3.606467 -3.606519 -3.606566 -3.606610 -3.606649
#>  [99] -3.606685 -3.606718
#> 
#> $cvsd
#>   [1] 2.885784 2.907532 2.897225 2.844494 2.778917 2.675856 2.589665 2.499533
#>   [9] 2.431459 2.377554 2.338423 2.307640 2.281700 2.258828 2.245847 2.244036
#>  [17] 2.241190 2.239259 2.240840 2.247792 2.256631 2.264452 2.275543 2.290818
#>  [25] 2.308547 2.326165 2.343056 2.359595 2.375712 2.391531 2.407295 2.422450
#>  [33] 2.436726 2.448541 2.459624 2.469989 2.479659 2.488945 2.498401 2.508929
#>  [41] 2.519489 2.530193 2.540737 2.550508 2.559548 2.567900 2.575606 2.582708
#>  [49] 2.589246 2.595259 2.600784 2.605857 2.610512 2.614780 2.618692 2.622274
#>  [57] 2.625554 2.628556 2.631301 2.633811 2.636106 2.638203 2.640119 2.641869
#>  [65] 2.643468 2.644955 2.646329 2.647584 2.648729 2.649774 2.650727 2.651597
#>  [73] 2.652390 2.653114 2.653774 2.654375 2.654924 2.655424 2.655880 2.656296
#>  [81] 2.656675 2.657021 2.657336 2.657623 2.657885 2.658123 2.658341 2.658539
#>  [89] 2.658719 2.658884 2.659034 2.659171 2.659295 2.659409 2.659512 2.659606
#>  [97] 2.659692 2.659771 2.659842 2.659907
#> 
#> attr(,"class")
#> [1] "cv_survtran"