Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

cbh_func

a function specifying the cumulative baseline hazard. The default is NULL, which means the cumulative baseline hazard is estimateds from the survode object fitted to the source data.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans.control containing control parameters for the fitting algorithm. Default is survtrans.control(...).

...

Other arguments passed to survtrans.control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6205607  1.3934978 -0.9954553  1.0047907 -0.8622008  1.0413994 -1.1190563 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0769175 
#> 
#> $coefficients$eta
#>  [1]  0.3830666143  0.3364964934 -0.0001210021  0.0000000000  0.1618263201
#>  [6]  0.0000000000 -0.0870384693  0.0000000000  0.0000000000  0.0743186973
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.06076032
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]          [,3]         [,4]        [,5]
#>   [1,] 0.08525555 0.02937811  0.0000000000  0.000000000 0.000000000
#>   [2,] 0.11560547 0.06612220  0.0000000000  0.000000000 0.000000000
#>   [3,] 0.14426287 0.10107474  0.0000000000  0.000000000 0.000000000
#>   [4,] 0.17117742 0.13349864  0.0000000000  0.000000000 0.003269573
#>   [5,] 0.19645868 0.16138758  0.0000000000  0.000000000 0.020975898
#>   [6,] 0.22020318 0.18782313  0.0000000000  0.000000000 0.037552408
#>   [7,] 0.24239759 0.21273070  0.0000000000  0.000000000 0.052994375
#>   [8,] 0.26305119 0.23606938  0.0000000000  0.000000000 0.067314567
#>   [9,] 0.28198608 0.25734278  0.0000000000  0.000000000 0.080509645
#>  [10,] 0.29905835 0.27615901  0.0000000000  0.000000000 0.092749554
#>  [11,] 0.31480520 0.29367295  0.0000000000  0.000000000 0.104192228
#>  [12,] 0.33105988 0.30374760  0.0000000000  0.000000000 0.117920573
#>  [13,] 0.34593018 0.31299541  0.0000000000  0.000000000 0.130478490
#>  [14,] 0.35949805 0.32150599  0.0000000000  0.000000000 0.141925628
#>  [15,] 0.37186103 0.32932147  0.0000000000  0.000000000 0.152347434
#>  [16,] 0.38306661 0.33649649 -0.0001210021  0.000000000 0.161826320
#>  [17,] 0.39038997 0.34366585 -0.0074961680  0.000000000 0.170431286
#>  [18,] 0.39708023 0.35021607 -0.0141585634  0.000000000 0.178268628
#>  [19,] 0.40356098 0.35602660 -0.0203460393  0.000000000 0.185418558
#>  [20,] 0.40986735 0.36116935 -0.0261142089  0.000000000 0.191967163
#>  [21,] 0.41564434 0.36586631 -0.0313414262  0.000000000 0.197953072
#>  [22,] 0.42102437 0.37015468 -0.0359936946  0.000000000 0.203489183
#>  [23,] 0.42640256 0.37409750 -0.0397606073  0.000000000 0.208896595
#>  [24,] 0.43129998 0.37771638 -0.0431492638  0.000000000 0.213837820
#>  [25,] 0.43575930 0.38103278 -0.0462001542  0.000000000 0.218351175
#>  [26,] 0.44027162 0.38678357 -0.0492628612 -0.005499953 0.223815910
#>  [27,] 0.44461957 0.39343945 -0.0521474962 -0.013378142 0.229532898
#>  [28,] 0.44860789 0.39961905 -0.0547689301 -0.020626963 0.234789004
#>  [29,] 0.45226343 0.40534176 -0.0571509396 -0.027287746 0.239615392
#>  [30,] 0.45561159 0.41063003 -0.0593153241 -0.033401222 0.244042388
#>  [31,] 0.45867637 0.41550694 -0.0612818946 -0.039006452 0.248099317
#>  [32,] 0.46148019 0.41999751 -0.0630688268 -0.044141382 0.251814088
#>  [33,] 0.46404403 0.42412659 -0.0646926324 -0.048841980 0.255213207
#>  [34,] 0.46638750 0.42791908 -0.0661683276 -0.053142434 0.258321664
#>  [35,] 0.46852866 0.43139814 -0.0675095476 -0.057074266 0.261162850
#>  [36,] 0.47048438 0.43458732 -0.0687287050 -0.060667623 0.263758613
#>  [37,] 0.47227021 0.43750851 -0.0698370369 -0.063950309 0.266129275
#>  [38,] 0.47390050 0.44018245 -0.0708447403 -0.066948138 0.268293669
#>  [39,] 0.47538848 0.44262868 -0.0717610670 -0.069685021 0.270269205
#>  [40,] 0.47674631 0.44486550 -0.0725944077 -0.072183039 0.272071942
#>  [41,] 0.47798517 0.44690998 -0.0733523712 -0.074462541 0.273716668
#>  [42,] 0.47911531 0.44877797 -0.0740418582 -0.076542251 0.275216976
#>  [43,] 0.48014614 0.45048415 -0.0746691276 -0.078439375 0.276585347
#>  [44,] 0.48108629 0.45204214 -0.0752398588 -0.080169709 0.277833224
#>  [45,] 0.48194363 0.45346444 -0.0757592004 -0.081747717 0.278971096
#>  [46,] 0.48272540 0.45476258 -0.0762318266 -0.083186658 0.280008564
#>  [47,] 0.48343820 0.45594720 -0.0766619797 -0.084498668 0.280954410
#>  [48,] 0.48408807 0.45702804 -0.0770535109 -0.085694851 0.281816666
#>  [49,] 0.48468053 0.45801404 -0.0774099173 -0.086785354 0.282602673
#>  [50,] 0.48522062 0.45891342 -0.0777343749 -0.087779455 0.283319134
#>  [51,] 0.48571295 0.45973369 -0.0780297685 -0.088685627 0.283972174
#>  [52,] 0.48616171 0.46048174 -0.0782987196 -0.089511610 0.284567383
#>  [53,] 0.48657076 0.46116386 -0.0785436101 -0.090264468 0.285109863
#>  [54,] 0.48694358 0.46178582 -0.0787666051 -0.090950651 0.285604271
#>  [55,] 0.48728339 0.46235288 -0.0789696729 -0.091576044 0.286054854
#>  [56,] 0.48759308 0.46286986 -0.0791546032 -0.092146016 0.286465488
#>  [57,] 0.48787533 0.46334115 -0.0793230235 -0.092665466 0.286839707
#>  [58,] 0.48813257 0.46377077 -0.0794764140 -0.093138861 0.287180733
#>  [59,] 0.48836699 0.46416239 -0.0796161212 -0.093570277 0.287491507
#>  [60,] 0.48858063 0.46451935 -0.0797433702 -0.093963428 0.287774707
#>  [61,] 0.48877532 0.46484471 -0.0798592755 -0.094321704 0.288032777
#>  [62,] 0.48895274 0.46514126 -0.0799648515 -0.094648193 0.288267944
#>  [63,] 0.48911441 0.46541153 -0.0800610213 -0.094945712 0.288482239
#>  [64,] 0.48926175 0.46565786 -0.0801486250 -0.095216828 0.288677512
#>  [65,] 0.48939601 0.46588236 -0.0802284275 -0.095463882 0.288855451
#>  [66,] 0.48951835 0.46608695 -0.0803011249 -0.095689008 0.289017592
#>  [67,] 0.48962983 0.46627341 -0.0803673510 -0.095894150 0.289165339
#>  [68,] 0.48973142 0.46644333 -0.0804276831 -0.096081081 0.289299967
#>  [69,] 0.48982399 0.46659818 -0.0804826464 -0.096251416 0.289422641
#>  [70,] 0.48990834 0.46673929 -0.0805327194 -0.096406628 0.289534422
#>  [71,] 0.48998521 0.46686788 -0.0805783379 -0.096548058 0.289636276
#>  [72,] 0.49005525 0.46698506 -0.0806198986 -0.096676930 0.289729086
#>  [73,] 0.49011907 0.46709185 -0.0806577628 -0.096794359 0.289813653
#>  [74,] 0.49017722 0.46718915 -0.0806922598 -0.096901360 0.289890710
#>  [75,] 0.49023021 0.46727782 -0.0807236892 -0.096998858 0.289960923
#>  [76,] 0.49027849 0.46735862 -0.0807523240 -0.097087698 0.290024900
#>  [77,] 0.49032249 0.46743224 -0.0807784129 -0.097168648 0.290083195
#>  [78,] 0.49036258 0.46749933 -0.0808021825 -0.097242409 0.290136312
#>  [79,] 0.49039910 0.46756047 -0.0808238391 -0.097309618 0.290184711
#>  [80,] 0.49043239 0.46761617 -0.0808435705 -0.097370858 0.290228812
#>  [81,] 0.49046272 0.46766693 -0.0808615481 -0.097426659 0.290268995
#>  [82,] 0.49049035 0.46771318 -0.0808779278 -0.097477503 0.290305609
#>  [83,] 0.49051553 0.46775533 -0.0808928518 -0.097523832 0.290338970
#>  [84,] 0.49053847 0.46779373 -0.0809064493 -0.097566045 0.290369369
#>  [85,] 0.49055938 0.46782872 -0.0809188384 -0.097604509 0.290397067
#>  [86,] 0.49057843 0.46786061 -0.0809301266 -0.097639556 0.290422304
#>  [87,] 0.49059578 0.46788966 -0.0809404116 -0.097671490 0.290445300
#>  [88,] 0.49061160 0.46791613 -0.0809497826 -0.097700587 0.290466253
#>  [89,] 0.49062601 0.46794025 -0.0809583209 -0.097727100 0.290485345
#>  [90,] 0.49063914 0.46796223 -0.0809661005 -0.097751257 0.290502741
#>  [91,] 0.49065110 0.46798226 -0.0809731889 -0.097773269 0.290518591
#>  [92,] 0.49066200 0.46800051 -0.0809796474 -0.097793325 0.290533034
#>  [93,] 0.49067194 0.46801713 -0.0809855321 -0.097811600 0.290546193
#>  [94,] 0.49068099 0.46803228 -0.0809908939 -0.097828251 0.290558184
#>  [95,] 0.49068923 0.46804609 -0.0809957792 -0.097843424 0.290569109
#>  [96,] 0.49069675 0.46805866 -0.0810002306 -0.097857248 0.290579064
#>  [97,] 0.49070359 0.46807013 -0.0810042864 -0.097869844 0.290588135
#>  [98,] 0.49070983 0.46808057 -0.0810079819 -0.097881321 0.290596399
#>  [99,] 0.49071551 0.46809008 -0.0810113490 -0.097891779 0.290603930
#> [100,] 0.49072069 0.46809875 -0.0810144170 -0.097901308 0.290610791
#>                [,6]         [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,]  0.000000000 -0.004351415 0.000000000 0.000000000 0.00000000
#>  [10,]  0.000000000 -0.017451542 0.000000000 0.000000000 0.00000000
#>  [11,]  0.000000000 -0.029568469 0.000000000 0.000000000 0.00000000
#>  [12,]  0.000000000 -0.043135140 0.000000000 0.000000000 0.01836346
#>  [13,]  0.000000000 -0.055618591 0.000000000 0.000000000 0.03484697
#>  [14,]  0.000000000 -0.067055449 0.000000000 0.000000000 0.04952993
#>  [15,]  0.000000000 -0.077509559 0.000000000 0.000000000 0.06262115
#>  [16,]  0.000000000 -0.087038469 0.000000000 0.000000000 0.07431870
#>  [17,]  0.000000000 -0.095123151 0.000000000 0.000000000 0.08555382
#>  [18,]  0.000000000 -0.102462565 0.000000000 0.000000000 0.09560260
#>  [19,]  0.000000000 -0.109050592 0.000000000 0.003027082 0.10459625
#>  [20,]  0.000000000 -0.114960338 0.000000000 0.008808991 0.11265658
#>  [21,]  0.000000000 -0.120323313 0.000000000 0.014108972 0.11989402
#>  [22,]  0.000000000 -0.125439453 0.001288177 0.018965854 0.12669267
#>  [23,]  0.000000000 -0.131381430 0.008975410 0.023441820 0.13428043
#>  [24,]  0.000000000 -0.136815228 0.015955966 0.027549623 0.14110700
#>  [25,]  0.000000000 -0.141780265 0.022295577 0.031315368 0.14725504
#>  [26,] -0.005097041 -0.146158599 0.027801362 0.036225059 0.15169944
#>  [27,] -0.011864000 -0.150168509 0.032728151 0.041447179 0.15511552
#>  [28,] -0.018076061 -0.153874377 0.037228662 0.046272436 0.15815276
#>  [29,] -0.023772237 -0.157291947 0.041337630 0.050723402 0.16085757
#>  [30,] -0.028990379 -0.160437824 0.045087394 0.054823018 0.16327001
#>  [31,] -0.033766444 -0.163329337 0.048508214 0.058594016 0.16542524
#>  [32,] -0.038134895 -0.165983474 0.051627917 0.062058815 0.16735338
#>  [33,] -0.042128132 -0.168417023 0.054472287 0.065239105 0.16908076
#>  [34,] -0.045776640 -0.170646171 0.057065053 0.068155764 0.17063020
#>  [35,] -0.049108545 -0.172686599 0.059428162 0.070828510 0.17202201
#>  [36,] -0.052150363 -0.174552931 0.061581608 0.073276180 0.17327349
#>  [37,] -0.054926498 -0.176259068 0.063543797 0.075516427 0.17440002
#>  [38,] -0.057459504 -0.177818020 0.065331578 0.077565783 0.17541511
#>  [39,] -0.059770179 -0.179241906 0.066960354 0.079439683 0.17633062
#>  [40,] -0.061877659 -0.180541982 0.068444200 0.081152480 0.17715703
#>  [41,] -0.063799522 -0.181728667 0.069795969 0.082717488 0.17790361
#>  [42,] -0.065551893 -0.182811583 0.071027390 0.084147028 0.17857855
#>  [43,] -0.067149549 -0.183799596 0.072149162 0.085452479 0.17918914
#>  [44,] -0.068606025 -0.184700856 0.073171039 0.086644334 0.17974185
#>  [45,] -0.069933690 -0.185522860 0.074101918 0.087732250 0.18024245
#>  [46,] -0.071143858 -0.186272480 0.074949903 0.088725109 0.18069609
#>  [47,] -0.072246868 -0.186956014 0.075722380 0.089631066 0.18110736
#>  [48,] -0.073252163 -0.187579228 0.076426078 0.090457610 0.18148040
#>  [49,] -0.074168364 -0.188147399 0.077067126 0.091211603 0.18181888
#>  [50,] -0.075003340 -0.188665351 0.077651107 0.091899333 0.18212613
#>  [51,] -0.075764269 -0.189137495 0.078183108 0.092526559 0.18240511
#>  [52,] -0.076457703 -0.189567860 0.078667760 0.093098552 0.18265850
#>  [53,] -0.077089615 -0.189960124 0.079109281 0.093620132 0.18288872
#>  [54,] -0.077665454 -0.190317646 0.079511516 0.094095706 0.18309794
#>  [55,] -0.078190187 -0.190643493 0.079877963 0.094529305 0.18328811
#>  [56,] -0.078668345 -0.190940461 0.080211809 0.094924608 0.18346101
#>  [57,] -0.079104057 -0.191211103 0.080515959 0.095284980 0.18361823
#>  [58,] -0.079501087 -0.191457748 0.080793056 0.095613490 0.18376123
#>  [59,] -0.079862866 -0.191682517 0.081045508 0.095912943 0.18389130
#>  [60,] -0.080192521 -0.191887349 0.081275510 0.096185898 0.18400964
#>  [61,] -0.080492904 -0.192074009 0.081485060 0.096434691 0.18411733
#>  [62,] -0.080766613 -0.192244105 0.081675977 0.096661453 0.18421532
#>  [63,] -0.081016014 -0.192399107 0.081849920 0.096868129 0.18430450
#>  [64,] -0.081243266 -0.192540353 0.082008399 0.097056494 0.18438568
#>  [65,] -0.081450335 -0.192669061 0.082152789 0.097228166 0.18445957
#>  [66,] -0.081639012 -0.192786344 0.082284344 0.097384620 0.18452684
#>  [67,] -0.081810932 -0.192893215 0.082404205 0.097527203 0.18458809
#>  [68,] -0.081967582 -0.192990598 0.082513412 0.097657143 0.18464385
#>  [69,] -0.082110318 -0.193079335 0.082612913 0.097775558 0.18469463
#>  [70,] -0.082240376 -0.193160192 0.082703570 0.097883469 0.18474087
#>  [71,] -0.082358881 -0.193233870 0.082786171 0.097981807 0.18478298
#>  [72,] -0.082466859 -0.193301005 0.082861431 0.098071420 0.18482133
#>  [73,] -0.082565247 -0.193362178 0.082930002 0.098153080 0.18485625
#>  [74,] -0.082654894 -0.193417919 0.082992480 0.098227494 0.18488806
#>  [75,] -0.082736579 -0.193468709 0.083049406 0.098295303 0.18491704
#>  [76,] -0.082811007 -0.193514988 0.083101274 0.098357093 0.18494343
#>  [77,] -0.082878824 -0.193557158 0.083148532 0.098413399 0.18496746
#>  [78,] -0.082940616 -0.193595581 0.083191592 0.098464706 0.18498936
#>  [79,] -0.082996920 -0.193630593 0.083230825 0.098511457 0.18500931
#>  [80,] -0.083048221 -0.193662494 0.083266572 0.098554058 0.18502748
#>  [81,] -0.083094966 -0.193691562 0.083299143 0.098592877 0.18504403
#>  [82,] -0.083137558 -0.193718048 0.083328821 0.098628248 0.18505911
#>  [83,] -0.083176366 -0.193742181 0.083355861 0.098660479 0.18507284
#>  [84,] -0.083211727 -0.193764171 0.083380499 0.098689847 0.18508536
#>  [85,] -0.083243947 -0.193784207 0.083402948 0.098716608 0.18509676
#>  [86,] -0.083273304 -0.193802464 0.083423402 0.098740991 0.18510715
#>  [87,] -0.083300054 -0.193819099 0.083442039 0.098763210 0.18511661
#>  [88,] -0.083324427 -0.193834256 0.083459020 0.098783455 0.18512523
#>  [89,] -0.083346635 -0.193848067 0.083474493 0.098801902 0.18513309
#>  [90,] -0.083366870 -0.193860651 0.083488591 0.098818711 0.18514024
#>  [91,] -0.083385307 -0.193872117 0.083501436 0.098834026 0.18514676
#>  [92,] -0.083402107 -0.193882564 0.083513140 0.098847982 0.18515270
#>  [93,] -0.083417414 -0.193892084 0.083523805 0.098860697 0.18515811
#>  [94,] -0.083431361 -0.193900758 0.083533522 0.098872284 0.18516304
#>  [95,] -0.083444070 -0.193908661 0.083542375 0.098882841 0.18516754
#>  [96,] -0.083455649 -0.193915862 0.083550442 0.098892460 0.18517163
#>  [97,] -0.083466200 -0.193922424 0.083557793 0.098901225 0.18517536
#>  [98,] -0.083475813 -0.193928402 0.083564490 0.098909212 0.18517876
#>  [99,] -0.083484573 -0.193933850 0.083570593 0.098916489 0.18518185
#> [100,] -0.083492554 -0.193938813 0.083576153 0.098923119 0.18518467
#> 
#> $cvm
#>   [1] -3.974617 -3.864589 -3.759502 -3.625773 -3.507278 -3.402227 -3.315299
#>   [8] -3.256535 -3.198631 -3.154065 -3.121081 -3.094853 -3.075702 -3.057503
#>  [15] -3.038036 -3.034053 -3.044859 -3.065720 -3.093459 -3.129216 -3.167766
#>  [22] -3.205931 -3.241665 -3.275536 -3.305596 -3.331527 -3.356021 -3.378784
#>  [29] -3.399301 -3.417748 -3.434294 -3.449525 -3.463833 -3.476845 -3.493176
#>  [36] -3.506298 -3.517897 -3.528824 -3.539078 -3.548672 -3.557623 -3.565953
#>  [43] -3.573659 -3.580121 -3.586111 -3.591655 -3.596674 -3.601011 -3.605014
#>  [50] -3.608705 -3.612103 -3.615228 -3.618100 -3.620738 -3.623158 -3.625378
#>  [57] -3.627412 -3.629275 -3.630980 -3.632541 -3.633969 -3.635274 -3.636468
#>  [64] -3.637558 -3.638555 -3.639465 -3.640296 -3.641055 -3.641747 -3.642380
#>  [71] -3.642957 -3.643483 -3.643963 -3.644401 -3.644801 -3.645165 -3.645497
#>  [78] -3.645800 -3.646077 -3.646328 -3.646558 -3.646767 -3.646958 -3.647132
#>  [85] -3.647291 -3.647436 -3.647567 -3.647687 -3.647797 -3.647897 -3.647988
#>  [92] -3.648070 -3.648146 -3.648215 -3.648278 -3.648335 -3.648387 -3.648434
#>  [99] -3.648477 -3.648517
#> 
#> $cvsd
#>   [1] 3.151781 3.144708 3.131093 3.097339 3.068904 3.038745 3.015194 3.001899
#>   [9] 2.974210 2.948069 2.930148 2.914096 2.900737 2.902591 2.920740 2.943773
#>  [17] 2.973835 3.007285 3.038408 3.069115 3.101141 3.132246 3.159010 3.184447
#>  [25] 3.208338 3.231157 3.254322 3.277156 3.299555 3.320161 3.338199 3.354566
#>  [33] 3.369892 3.383928 3.406650 3.430230 3.451842 3.472099 3.491031 3.508679
#>  [41] 3.525093 3.540328 3.554424 3.567026 3.578672 3.589421 3.599205 3.607867
#>  [49] 3.615844 3.623183 3.629929 3.636125 3.641811 3.647026 3.651806 3.656184
#>  [57] 3.660194 3.663863 3.667220 3.670289 3.673096 3.675661 3.678004 3.680145
#>  [65] 3.682099 3.683884 3.685514 3.687001 3.688358 3.689597 3.690727 3.691757
#>  [73] 3.692697 3.693555 3.694337 3.695050 3.695700 3.696293 3.696833 3.697326
#>  [81] 3.697775 3.698185 3.698558 3.698898 3.699208 3.699491 3.699748 3.699983
#>  [89] 3.700197 3.700392 3.700570 3.700732 3.700880 3.701014 3.701137 3.701248
#>  [97] 3.701350 3.701443 3.701527 3.701604
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x55e791b2ad50>
#> 
#> attr(,"class")
#> [1] "cv_survtran"