Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv_survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6000663  1.4130279 -1.0156804  1.0047907 -0.8386085  1.0413994 -1.1410685 
#>         X8         X9        X10 
#>  0.9215155 -0.9956178  1.1071950 
#> 
#> $coefficients$eta
#>  [1]  0.403560993  0.356026617 -0.020346033  0.000000000  0.185418550
#>  [6]  0.000000000 -0.109050585  0.000000000  0.003027084  0.104596247
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.04596296
#> 
#> 
#> $lambdas
#>   [1] 2.452905e-01 2.234996e-01 2.036445e-01 1.855533e-01 1.690692e-01
#>   [6] 1.540496e-01 1.403642e-01 1.278947e-01 1.165329e-01 1.061804e-01
#>  [11] 9.674762e-02 8.815283e-02 8.032157e-02 7.318603e-02 6.668438e-02
#>  [16] 6.076032e-02 5.536255e-02 5.044429e-02 4.596296e-02 4.187974e-02
#>  [21] 3.815925e-02 3.476929e-02 3.168048e-02 2.886608e-02 2.630169e-02
#>  [26] 2.396512e-02 2.183613e-02 1.989627e-02 1.812874e-02 1.651823e-02
#>  [31] 1.505080e-02 1.371373e-02 1.249544e-02 1.138538e-02 1.037393e-02
#>  [36] 9.452339e-03 8.612619e-03 7.847498e-03 7.150348e-03 6.515130e-03
#>  [41] 5.936344e-03 5.408976e-03 4.928457e-03 4.490627e-03 4.091692e-03
#>  [46] 3.728197e-03 3.396994e-03 3.095215e-03 2.820244e-03 2.569702e-03
#>  [51] 2.341416e-03 2.133412e-03 1.943885e-03 1.771196e-03 1.613848e-03
#>  [56] 1.470478e-03 1.339845e-03 1.220817e-03 1.112363e-03 1.013543e-03
#>  [61] 9.235029e-04 8.414615e-04 7.667083e-04 6.985961e-04 6.365347e-04
#>  [66] 5.799867e-04 5.284623e-04 4.815152e-04 4.387387e-04 3.997624e-04
#>  [71] 3.642486e-04 3.318897e-04 3.024056e-04 2.755407e-04 2.510624e-04
#>  [76] 2.287587e-04 2.084364e-04 1.899195e-04 1.730476e-04 1.576745e-04
#>  [81] 1.436671e-04 1.309042e-04 1.192750e-04 1.086789e-04 9.902419e-05
#>  [86] 9.022715e-05 8.221162e-05 7.490817e-05 6.825353e-05 6.219008e-05
#>  [91] 5.666528e-05 5.163129e-05 4.704451e-05 4.286521e-05 3.905718e-05
#>  [96] 3.558745e-05 3.242596e-05 2.954533e-05 2.692060e-05 2.452905e-05
#> 
#> $etas
#>              [,1]       [,2]          [,3]        [,4]        [,5]         [,6]
#>   [1,] 0.08525555 0.02937811  0.0000000000  0.00000000 0.000000000  0.000000000
#>   [2,] 0.11560547 0.06612220  0.0000000000  0.00000000 0.000000000  0.000000000
#>   [3,] 0.14426287 0.10107474  0.0000000000  0.00000000 0.000000000  0.000000000
#>   [4,] 0.17117741 0.13349865  0.0000000000  0.00000000 0.003269571  0.000000000
#>   [5,] 0.19645868 0.16138758  0.0000000000  0.00000000 0.020975898  0.000000000
#>   [6,] 0.22020319 0.18782314  0.0000000000  0.00000000 0.037552407  0.000000000
#>   [7,] 0.24239758 0.21273069  0.0000000000  0.00000000 0.052994376  0.000000000
#>   [8,] 0.26305120 0.23606940  0.0000000000  0.00000000 0.067314566  0.000000000
#>   [9,] 0.28198608 0.25734278  0.0000000000  0.00000000 0.080509645  0.000000000
#>  [10,] 0.29905835 0.27615901  0.0000000000  0.00000000 0.092749554  0.000000000
#>  [11,] 0.31480521 0.29367296  0.0000000000  0.00000000 0.104192228  0.000000000
#>  [12,] 0.33105989 0.30374752  0.0000000000  0.00000000 0.117920603  0.000000000
#>  [13,] 0.34593017 0.31299546  0.0000000000  0.00000000 0.130478471  0.000000000
#>  [14,] 0.35949803 0.32150615  0.0000000000  0.00000000 0.141925570  0.000000000
#>  [15,] 0.37186100 0.32932166  0.0000000000  0.00000000 0.152347365  0.000000000
#>  [16,] 0.38306672 0.33649636 -0.0001209746  0.00000000 0.161826367  0.000000000
#>  [17,] 0.39039005 0.34366578 -0.0074961453  0.00000000 0.170431310  0.000000000
#>  [18,] 0.39708029 0.35021603 -0.0141585445  0.00000000 0.178268639  0.000000000
#>  [19,] 0.40356099 0.35602662 -0.0203460334  0.00000000 0.185418550  0.000000000
#>  [20,] 0.40986722 0.36116960 -0.0261142211  0.00000000 0.191967078  0.000000000
#>  [21,] 0.41564434 0.36586655 -0.0313414000  0.00000000 0.197952984  0.000000000
#>  [22,] 0.42102407 0.37015500 -0.0359938464  0.00000000 0.203489039  0.000000000
#>  [23,] 0.42640232 0.37409786 -0.0397607012  0.00000000 0.208896432  0.000000000
#>  [24,] 0.43130009 0.37771632 -0.0431491911  0.00000000 0.213837849  0.000000000
#>  [25,] 0.43575910 0.38103299 -0.0462002412  0.00000000 0.218351078  0.000000000
#>  [26,] 0.44027159 0.38678373 -0.0492628864 -0.00550000 0.223815882 -0.005097081
#>  [27,] 0.44461955 0.39343957 -0.0521475143 -0.01337818 0.229532878 -0.011864030
#>  [28,] 0.44860787 0.39961912 -0.0547689410 -0.02062699 0.234788992 -0.018076079
#>  [29,] 0.45226343 0.40534179 -0.0571509425 -0.02728775 0.239615389 -0.023772242
#>  [30,] 0.45561160 0.41062997 -0.0593153147 -0.03340120 0.244042399 -0.028990358
#>  [31,] 0.45867638 0.41550681 -0.0612818750 -0.03900641 0.248099336 -0.033766407
#>  [32,] 0.46148024 0.41999771 -0.0630688096 -0.04414156 0.251814140 -0.038134964
#>  [33,] 0.46404411 0.42412659 -0.0646925830 -0.04884209 0.255213287 -0.042128145
#>  [34,] 0.46638756 0.42791857 -0.0661682502 -0.05314228 0.258321752 -0.045776498
#>  [35,] 0.46852874 0.43139756 -0.0675094563 -0.05707408 0.261162949 -0.049108380
#>  [36,] 0.47048435 0.43458697 -0.0687286874 -0.06066743 0.263758614 -0.052150249
#>  [37,] 0.47227018 0.43750813 -0.0698370159 -0.06395010 0.266129277 -0.054926377
#>  [38,] 0.47390048 0.44018204 -0.0708447162 -0.06694792 0.268293672 -0.057459375
#>  [39,] 0.47538846 0.44262825 -0.0717610401 -0.06968480 0.270269209 -0.059770044
#>  [40,] 0.47674629 0.44486505 -0.0725943780 -0.07218281 0.272071948 -0.061877518
#>  [41,] 0.47798508 0.44690987 -0.0733524044 -0.07446242 0.273716622 -0.063799476
#>  [42,] 0.47911522 0.44877788 -0.0740418946 -0.07654214 0.275216928 -0.065551855
#>  [43,] 0.48014605 0.45048410 -0.0746691670 -0.07843928 0.276585296 -0.067149520
#>  [44,] 0.48108620 0.45204209 -0.0752398993 -0.08016961 0.277833173 -0.068605998
#>  [45,] 0.48194354 0.45346440 -0.0757592418 -0.08174763 0.278971045 -0.069933665
#>  [46,] 0.48272531 0.45476255 -0.0762318689 -0.08318657 0.280008512 -0.071143835
#>  [47,] 0.48343811 0.45594718 -0.0766620229 -0.08449858 0.280954357 -0.072246847
#>  [48,] 0.48408798 0.45702802 -0.0770535550 -0.08569477 0.281816613 -0.073252143
#>  [49,] 0.48468044 0.45801402 -0.0774099622 -0.08678527 0.282602619 -0.074168345
#>  [50,] 0.48522053 0.45891341 -0.0777344204 -0.08777937 0.283319079 -0.075003322
#>  [51,] 0.48571285 0.45973368 -0.0780298147 -0.08868555 0.283972119 -0.075764252
#>  [52,] 0.48616162 0.46048173 -0.0782987664 -0.08951153 0.284567327 -0.076457687
#>  [53,] 0.48657066 0.46116386 -0.0785436574 -0.09026439 0.285109807 -0.077089600
#>  [54,] 0.48694349 0.46178582 -0.0787666529 -0.09095057 0.285604214 -0.077665439
#>  [55,] 0.48728329 0.46235289 -0.0789697211 -0.09157597 0.286054797 -0.078190174
#>  [56,] 0.48759298 0.46286987 -0.0791546519 -0.09214594 0.286465431 -0.078668332
#>  [57,] 0.48787524 0.46334116 -0.0793230725 -0.09266539 0.286839649 -0.079104045
#>  [58,] 0.48813247 0.46377078 -0.0794764633 -0.09313879 0.287180676 -0.079501075
#>  [59,] 0.48836689 0.46416240 -0.0796161709 -0.09357020 0.287491449 -0.079862854
#>  [60,] 0.48858053 0.46451936 -0.0797434201 -0.09396336 0.287774649 -0.080192510
#>  [61,] 0.48877522 0.46484472 -0.0798593257 -0.09432163 0.288032719 -0.080492894
#>  [62,] 0.48895263 0.46514127 -0.0799649019 -0.09464812 0.288267886 -0.080766603
#>  [63,] 0.48911431 0.46541155 -0.0800610719 -0.09494564 0.288482181 -0.081016004
#>  [64,] 0.48926165 0.46565788 -0.0801486758 -0.09521676 0.288677454 -0.081243256
#>  [65,] 0.48939590 0.46588238 -0.0802284784 -0.09546381 0.288855392 -0.081450325
#>  [66,] 0.48951825 0.46608697 -0.0803011760 -0.09568894 0.289017534 -0.081639004
#>  [67,] 0.48962973 0.46627343 -0.0803674023 -0.09589408 0.289165280 -0.081810924
#>  [68,] 0.48973132 0.46644335 -0.0804277344 -0.09608101 0.289299908 -0.081967574
#>  [69,] 0.48982389 0.46659820 -0.0804826978 -0.09625135 0.289422582 -0.082110310
#>  [70,] 0.48990824 0.46673931 -0.0805327710 -0.09640656 0.289534363 -0.082240367
#>  [71,] 0.48998511 0.46686790 -0.0805783896 -0.09654799 0.289636217 -0.082358873
#>  [72,] 0.49005514 0.46698509 -0.0806199503 -0.09667686 0.289729027 -0.082466852
#>  [73,] 0.49011896 0.46709187 -0.0806578147 -0.09679429 0.289813594 -0.082565239
#>  [74,] 0.49017712 0.46718918 -0.0806923117 -0.09690129 0.289890650 -0.082654887
#>  [75,] 0.49023011 0.46727785 -0.0807237412 -0.09699879 0.289960864 -0.082736571
#>  [76,] 0.49027839 0.46735864 -0.0807523761 -0.09708763 0.290024841 -0.082811000
#>  [77,] 0.49032239 0.46743227 -0.0807784651 -0.09716858 0.290083136 -0.082878817
#>  [78,] 0.49036247 0.46749936 -0.0808022347 -0.09724234 0.290136253 -0.082940609
#>  [79,] 0.49039900 0.46756049 -0.0808238913 -0.09730955 0.290184652 -0.082996913
#>  [80,] 0.49043229 0.46761620 -0.0808436228 -0.09737079 0.290228752 -0.083048214
#>  [81,] 0.49046261 0.46766696 -0.0808616004 -0.09742659 0.290268935 -0.083094959
#>  [82,] 0.49049025 0.46771321 -0.0808779802 -0.09747744 0.290305549 -0.083137551
#>  [83,] 0.49051543 0.46775536 -0.0808929041 -0.09752376 0.290338911 -0.083176359
#>  [84,] 0.49053837 0.46779376 -0.0809065017 -0.09756598 0.290369309 -0.083211720
#>  [85,] 0.49055928 0.46782875 -0.0809188908 -0.09760444 0.290397007 -0.083243940
#>  [86,] 0.49057832 0.46786063 -0.0809301790 -0.09763949 0.290422245 -0.083273297
#>  [87,] 0.49059568 0.46788969 -0.0809404640 -0.09767142 0.290445241 -0.083300047
#>  [88,] 0.49061150 0.46791616 -0.0809498351 -0.09770052 0.290466194 -0.083324420
#>  [89,] 0.49062591 0.46794028 -0.0809583734 -0.09772703 0.290485286 -0.083346628
#>  [90,] 0.49063904 0.46796226 -0.0809661531 -0.09775119 0.290502681 -0.083366863
#>  [91,] 0.49065100 0.46798229 -0.0809732414 -0.09777320 0.290518532 -0.083385301
#>  [92,] 0.49066190 0.46800053 -0.0809796999 -0.09779326 0.290532974 -0.083402100
#>  [93,] 0.49067183 0.46801716 -0.0809855846 -0.09781153 0.290546134 -0.083417408
#>  [94,] 0.49068088 0.46803231 -0.0809909464 -0.09782818 0.290558124 -0.083431355
#>  [95,] 0.49068913 0.46804612 -0.0809958318 -0.09784336 0.290569050 -0.083444063
#>  [96,] 0.49069664 0.46805869 -0.0810002831 -0.09785718 0.290579005 -0.083455643
#>  [97,] 0.49070349 0.46807015 -0.0810043390 -0.09786978 0.290588075 -0.083466193
#>  [98,] 0.49070973 0.46808060 -0.0810080345 -0.09788125 0.290596340 -0.083475807
#>  [99,] 0.49071541 0.46809011 -0.0810114016 -0.09789171 0.290603870 -0.083484566
#> [100,] 0.49072059 0.46809878 -0.0810144696 -0.09790124 0.290610732 -0.083492547
#>                [,7]        [,8]        [,9]      [,10]
#>   [1,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [2,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [3,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [4,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [5,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [6,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [7,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [8,]  0.000000000 0.000000000 0.000000000 0.00000000
#>   [9,] -0.004351415 0.000000000 0.000000000 0.00000000
#>  [10,] -0.017451542 0.000000000 0.000000000 0.00000000
#>  [11,] -0.029568468 0.000000000 0.000000000 0.00000000
#>  [12,] -0.043135163 0.000000000 0.000000000 0.01836349
#>  [13,] -0.055618577 0.000000000 0.000000000 0.03484696
#>  [14,] -0.067055403 0.000000000 0.000000000 0.04952987
#>  [15,] -0.077509503 0.000000000 0.000000000 0.06262108
#>  [16,] -0.087038504 0.000000000 0.000000000 0.07431875
#>  [17,] -0.095123169 0.000000000 0.000000000 0.08555385
#>  [18,] -0.102462573 0.000000000 0.000000000 0.09560261
#>  [19,] -0.109050585 0.000000000 0.003027084 0.10459625
#>  [20,] -0.114960271 0.000000000 0.008808981 0.11265648
#>  [21,] -0.120323238 0.000000000 0.014108975 0.11989393
#>  [22,] -0.125439279 0.001288048 0.018965837 0.12669252
#>  [23,] -0.131381209 0.008975241 0.023441808 0.13428025
#>  [24,] -0.136815249 0.015955976 0.027549630 0.14110702
#>  [25,] -0.141780143 0.022295486 0.031315354 0.14725494
#>  [26,] -0.146158544 0.027801323 0.036225070 0.15169938
#>  [27,] -0.150168469 0.032728123 0.041447187 0.15511548
#>  [28,] -0.153874352 0.037228645 0.046272442 0.15815274
#>  [29,] -0.157291939 0.041337625 0.050723404 0.16085756
#>  [30,] -0.160437850 0.045087411 0.054823010 0.16327004
#>  [31,] -0.163329383 0.048508244 0.058594003 0.16542529
#>  [32,] -0.165983485 0.051627924 0.062058855 0.16735331
#>  [33,] -0.168417102 0.054472337 0.065239126 0.16908077
#>  [34,] -0.170646367 0.057065178 0.068155715 0.17063040
#>  [35,] -0.172686821 0.059428302 0.070828452 0.17202223
#>  [36,] -0.174553011 0.061581658 0.073276126 0.17327362
#>  [37,] -0.176259155 0.063543851 0.075516369 0.17440016
#>  [38,] -0.177818113 0.065331636 0.077565722 0.17541525
#>  [39,] -0.179242006 0.066960415 0.079439619 0.17633078
#>  [40,] -0.180542087 0.068444265 0.081152413 0.17715720
#>  [41,] -0.181728652 0.069795959 0.082717454 0.17790365
#>  [42,] -0.182811560 0.071027375 0.084146997 0.17857858
#>  [43,] -0.183799565 0.072149143 0.085452452 0.17918916
#>  [44,] -0.184700822 0.073171018 0.086644308 0.17974186
#>  [45,] -0.185522824 0.074101895 0.087732225 0.18024246
#>  [46,] -0.186272441 0.074949879 0.088725084 0.18069609
#>  [47,] -0.186955973 0.075722355 0.089631043 0.18110737
#>  [48,] -0.187579186 0.076426052 0.090457587 0.18148040
#>  [49,] -0.188147355 0.077067099 0.091211580 0.18181889
#>  [50,] -0.188665307 0.077651080 0.091899310 0.18212613
#>  [51,] -0.189137450 0.078183080 0.092526537 0.18240511
#>  [52,] -0.189567813 0.078667731 0.093098530 0.18265850
#>  [53,] -0.189960076 0.079109252 0.093620110 0.18288872
#>  [54,] -0.190317598 0.079511486 0.094095685 0.18309793
#>  [55,] -0.190643443 0.079877933 0.094529284 0.18328810
#>  [56,] -0.190940411 0.080211779 0.094924588 0.18346100
#>  [57,] -0.191211052 0.080515928 0.095284959 0.18361822
#>  [58,] -0.191457696 0.080793025 0.095613470 0.18376122
#>  [59,] -0.191682465 0.081045477 0.095912923 0.18389130
#>  [60,] -0.191887297 0.081275479 0.096185878 0.18400964
#>  [61,] -0.192073955 0.081485028 0.096434671 0.18411732
#>  [62,] -0.192244052 0.081675945 0.096661433 0.18421531
#>  [63,] -0.192399053 0.081849888 0.096868110 0.18430449
#>  [64,] -0.192540298 0.082008366 0.097056475 0.18438567
#>  [65,] -0.192669006 0.082152756 0.097228147 0.18445956
#>  [66,] -0.192786289 0.082284311 0.097384601 0.18452683
#>  [67,] -0.192893160 0.082404172 0.097527184 0.18458808
#>  [68,] -0.192990543 0.082513379 0.097657124 0.18464384
#>  [69,] -0.193079279 0.082612879 0.097775539 0.18469462
#>  [70,] -0.193160136 0.082703537 0.097883450 0.18474086
#>  [71,] -0.193233814 0.082786137 0.097981788 0.18478297
#>  [72,] -0.193300949 0.082861397 0.098071401 0.18482131
#>  [73,] -0.193362122 0.082929968 0.098153062 0.18485624
#>  [74,] -0.193417862 0.082992446 0.098227475 0.18488805
#>  [75,] -0.193468652 0.083049372 0.098295285 0.18491702
#>  [76,] -0.193514932 0.083101239 0.098357075 0.18494341
#>  [77,] -0.193557101 0.083148498 0.098413380 0.18496745
#>  [78,] -0.193595525 0.083191557 0.098464687 0.18498935
#>  [79,] -0.193630536 0.083230791 0.098511439 0.18500929
#>  [80,] -0.193662437 0.083266538 0.098554040 0.18502746
#>  [81,] -0.193691505 0.083299109 0.098592858 0.18504402
#>  [82,] -0.193717990 0.083328786 0.098628230 0.18505910
#>  [83,] -0.193742124 0.083355827 0.098660460 0.18507283
#>  [84,] -0.193764113 0.083380464 0.098689829 0.18508535
#>  [85,] -0.193784150 0.083402913 0.098716589 0.18509675
#>  [86,] -0.193802406 0.083423368 0.098740973 0.18510713
#>  [87,] -0.193819041 0.083442005 0.098763192 0.18511660
#>  [88,] -0.193834198 0.083458986 0.098783437 0.18512522
#>  [89,] -0.193848009 0.083474458 0.098801884 0.18513307
#>  [90,] -0.193860593 0.083488556 0.098818692 0.18514023
#>  [91,] -0.193872059 0.083501402 0.098834008 0.18514675
#>  [92,] -0.193882507 0.083513106 0.098847963 0.18515269
#>  [93,] -0.193892026 0.083523770 0.098860679 0.18515810
#>  [94,] -0.193900700 0.083533487 0.098872265 0.18516303
#>  [95,] -0.193908603 0.083542341 0.098882823 0.18516752
#>  [96,] -0.193915805 0.083550408 0.098892442 0.18517162
#>  [97,] -0.193922366 0.083557758 0.098901207 0.18517535
#>  [98,] -0.193928345 0.083564456 0.098909194 0.18517874
#>  [99,] -0.193933792 0.083570558 0.098916471 0.18518184
#> [100,] -0.193938756 0.083576118 0.098923101 0.18518466
#> 
#> $cvm
#>   [1] -3.961421 -3.844333 -3.731037 -3.615849 -3.511357 -3.425466 -3.354931
#>   [8] -3.301088 -3.259981 -3.222278 -3.195211 -3.177179 -3.161097 -3.147823
#>  [15] -3.134548 -3.124416 -3.115770 -3.107145 -3.105928 -3.114443 -3.130871
#>  [22] -3.150520 -3.172940 -3.195182 -3.215751 -3.235125 -3.253742 -3.279649
#>  [29] -3.306881 -3.333146 -3.356614 -3.377313 -3.396696 -3.415294 -3.433022
#>  [36] -3.449828 -3.465684 -3.480584 -3.494538 -3.507565 -3.518897 -3.529019
#>  [43] -3.538088 -3.546010 -3.553335 -3.560098 -3.566334 -3.572076 -3.577250
#>  [50] -3.581774 -3.585932 -3.589751 -3.593254 -3.596467 -3.599411 -3.602108
#>  [57] -3.604576 -3.606836 -3.608902 -3.610792 -3.612519 -3.614097 -3.615539
#>  [64] -3.616857 -3.618059 -3.619157 -3.620159 -3.621074 -3.621909 -3.622671
#>  [71] -3.623365 -3.623999 -3.624577 -3.625105 -3.625585 -3.626024 -3.626424
#>  [78] -3.626788 -3.627120 -3.627423 -3.627700 -3.627951 -3.628181 -3.628390
#>  [85] -3.628581 -3.628754 -3.628913 -3.629057 -3.629189 -3.629308 -3.629418
#>  [92] -3.629517 -3.629608 -3.629691 -3.629766 -3.629835 -3.629897 -3.629954
#>  [99] -3.630006 -3.630054
#> 
#> $cvsd
#>   [1] 2.919570 2.936468 2.973703 3.006727 3.036029 3.063648 3.084119 3.100712
#>   [9] 3.125242 3.150440 3.176106 3.195443 3.212167 3.228390 3.237831 3.247250
#>  [17] 3.257141 3.264801 3.261911 3.256019 3.247108 3.238547 3.231619 3.223645
#>  [25] 3.213298 3.203902 3.195108 3.183941 3.173000 3.163445 3.156405 3.148120
#>  [33] 3.141221 3.135303 3.130265 3.126008 3.122437 3.119464 3.117008 3.114994
#>  [41] 3.113416 3.112147 3.111158 3.110433 3.109858 3.109406 3.109057 3.108792
#>  [49] 3.108756 3.109110 3.109464 3.109815 3.110157 3.110488 3.110807 3.111110
#>  [57] 3.111398 3.111670 3.111926 3.112165 3.112389 3.112598 3.112792 3.112972
#>  [65] 3.113139 3.113293 3.113436 3.113567 3.113688 3.113799 3.113902 3.113996
#>  [73] 3.114082 3.114161 3.114234 3.114300 3.114361 3.114417 3.114467 3.114514
#>  [81] 3.114557 3.114596 3.114631 3.114664 3.114693 3.114720 3.114745 3.114768
#>  [89] 3.114788 3.114807 3.114824 3.114840 3.114854 3.114867 3.114879 3.114890
#>  [97] 3.114900 3.114909 3.114917 3.114924
#> 
#> $formula
#> Surv(time, status) ~ .
#> <environment: 0x558b9a4a5190>
#> 
#> attr(,"class")
#> [1] "cv_survtran"