Cross-validation for survtran

Cross-validation for survtran

Usage

cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)

Arguments

data

a data frame containing the variables in the model.

fit_source

a survode object.

penalty

a character string specifying the type of penalty. Possible values are "lasso", "MCP" and "SCAD".

gamma

a numeric value specifying the tuning parameter for MCP or SCAD.

nfolds

an integer specifying the number of folds.

nlambdas

an integer specifying the number of lambda values.

lambda_min_ratio

a numeric value specifying the minimum lambda value as a fraction of lambda_max.

seed

an integer specifying the random seed.

control

Object of class survtrans_control containing control parameters for the fitting algorithm. Default is survtrans_control(...).

...

Other arguments passed to survtrans_control.

Value

a cv.survtran object.

Examples

library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
#> $coefficients
#> $coefficients$gamma
#> [1] 1
#> 
#> $coefficients$beta
#>         X1         X2         X3         X4         X5         X6         X7 
#> -0.6096766  1.4041534 -1.0063853  1.0047907 -0.8494216  1.0413994 -1.1310529 
#>         X8         X9        X10 
#>  0.9215155 -0.9986449  1.0935259 
#> 
#> $coefficients$eta
#>  [1]  0.39395076  0.34715213 -0.01105094  0.00000000  0.17460545  0.00000000
#>  [7] -0.09903498  0.00000000  0.00000000  0.09092714
#> 
#> $coefficients$penalty
#> [1] "lasso"
#> 
#> $coefficients$lambda
#> [1] 0.05274359
#> 
#> $coefficients$formula
#> Surv(time, status) ~ .
#> <environment: 0x55c19bcf9d00>
#> 
#> 
#> $lambdas
#>   [1] 3.089203e-01 2.814767e-01 2.564711e-01 2.336869e-01 2.129268e-01
#>   [6] 1.940110e-01 1.767756e-01 1.610713e-01 1.467622e-01 1.337242e-01
#>  [11] 1.218445e-01 1.110202e-01 1.011575e-01 9.217093e-02 8.398272e-02
#>  [16] 7.652192e-02 6.972393e-02 6.352984e-02 5.788603e-02 5.274359e-02
#>  [21] 4.805800e-02 4.378866e-02 3.989859e-02 3.635411e-02 3.312451e-02
#>  [26] 3.018182e-02 2.750055e-02 2.505748e-02 2.283144e-02 2.080316e-02
#>  [31] 1.895506e-02 1.727115e-02 1.573683e-02 1.433881e-02 1.306499e-02
#>  [36] 1.190433e-02 1.084678e-02 9.883186e-03 9.005191e-03 8.205195e-03
#>  [41] 7.476268e-03 6.812097e-03 6.206929e-03 5.655523e-03 5.153102e-03
#>  [46] 4.695314e-03 4.278195e-03 3.898132e-03 3.551833e-03 3.236298e-03
#>  [51] 2.948794e-03 2.686831e-03 2.448141e-03 2.230655e-03 2.032489e-03
#>  [56] 1.851929e-03 1.687408e-03 1.537504e-03 1.400916e-03 1.276463e-03
#>  [61] 1.163065e-03 1.059742e-03 9.655971e-04 8.798162e-04 8.016557e-04
#>  [66] 7.304388e-04 6.655487e-04 6.064231e-04 5.525502e-04 5.034632e-04
#>  [71] 4.587369e-04 4.179840e-04 3.808514e-04 3.470176e-04 3.161895e-04
#>  [76] 2.881001e-04 2.625061e-04 2.391858e-04 2.179372e-04 1.985762e-04
#>  [81] 1.809353e-04 1.648615e-04 1.502156e-04 1.368709e-04 1.247117e-04
#>  [86] 1.136326e-04 1.035378e-04 9.433980e-05 8.595891e-05 7.832256e-05
#>  [91] 7.136460e-05 6.502477e-05 5.924815e-05 5.398470e-05 4.918885e-05
#>  [96] 4.481905e-05 4.083745e-05 3.720956e-05 3.390397e-05 3.089203e-05
#> 
#> $etas
#>              [,1]       [,2]         [,3]         [,4]       [,5]         [,6]
#>   [1,] 0.00000000 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [2,] 0.03624828 0.00000000  0.000000000  0.000000000 0.00000000  0.000000000
#>   [3,] 0.07013592 0.01117874  0.000000000  0.000000000 0.00000000  0.000000000
#>   [4,] 0.10127198 0.04873376  0.000000000  0.000000000 0.00000000  0.000000000
#>   [5,] 0.13074780 0.08455985  0.000000000  0.000000000 0.00000000  0.000000000
#>   [6,] 0.15850358 0.11853574  0.000000000  0.000000000 0.00000000  0.000000000
#>   [7,] 0.18453639 0.14820139  0.000000000  0.000000000 0.01263272  0.000000000
#>   [8,] 0.20902026 0.17534464  0.000000000  0.000000000 0.02975226  0.000000000
#>   [9,] 0.23195752 0.20099142  0.000000000  0.000000000 0.04573723  0.000000000
#>  [10,] 0.25334680 0.22508480  0.000000000  0.000000000 0.06059224  0.000000000
#>  [11,] 0.27320850 0.24760072  0.000000000  0.000000000 0.07433839  0.000000000
#>  [12,] 0.29104838 0.26730795  0.000000000  0.000000000 0.08698454  0.000000000
#>  [13,] 0.30742198 0.28544282  0.000000000  0.000000000 0.09880956  0.000000000
#>  [14,] 0.32344556 0.29904392  0.000000000  0.000000000 0.11148537  0.000000000
#>  [15,] 0.33897430 0.30865923  0.000000000  0.000000000 0.12460586  0.000000000
#>  [16,] 0.35315396 0.31751801  0.000000000  0.000000000 0.13657440  0.000000000
#>  [17,] 0.36608209 0.32566108  0.000000000  0.000000000 0.14747692  0.000000000
#>  [18,] 0.37785552 0.33313202  0.000000000  0.000000000 0.15739766  0.000000000
#>  [19,] 0.38696345 0.34031120 -0.004056189  0.000000000 0.16640837  0.000000000
#>  [20,] 0.39395076 0.34715213 -0.011050938  0.000000000 0.17460545  0.000000000
#>  [21,] 0.40034019 0.35339418 -0.017374375  0.000000000 0.18206841  0.000000000
#>  [22,] 0.40691443 0.35876319 -0.023421552  0.000000000 0.18890272  0.000000000
#>  [23,] 0.41294000 0.36366947 -0.028901187  0.000000000 0.19515256  0.000000000
#>  [24,] 0.41845718 0.36814781 -0.033867341  0.000000000 0.20086298  0.000000000
#>  [25,] 0.42388846 0.37225043 -0.038005815  0.000000000 0.20636642  0.000000000
#>  [26,] 0.42901101 0.37602140 -0.041570169  0.000000000 0.21152626  0.000000000
#>  [27,] 0.43367492 0.37948024 -0.044778246  0.000000000 0.21623990  0.000000000
#>  [28,] 0.43805289 0.38342117 -0.047780049 -0.001490307 0.22090435 -0.001646687
#>  [29,] 0.44258305 0.39031099 -0.050799829 -0.009684732 0.22685333 -0.008693580
#>  [30,] 0.44674019 0.39671687 -0.053544306 -0.017230014 0.23232628 -0.015166646
#>  [31,] 0.45055192 0.40265581 -0.056038181 -0.024167388 0.23735472 -0.021105173
#>  [32,] 0.45404429 0.40814911 -0.058304212 -0.030537992 0.24196936 -0.026547662
#>  [33,] 0.45724195 0.41321988 -0.060363171 -0.036381827 0.24620003 -0.031531074
#>  [34,] 0.46016809 0.41789268 -0.062234024 -0.041737596 0.25007533 -0.036090720
#>  [35,] 0.46284446 0.42219188 -0.063933977 -0.046642022 0.25362253 -0.040259828
#>  [36,] 0.46529107 0.42614211 -0.065478858 -0.051129739 0.25686718 -0.044069671
#>  [37,] 0.46752696 0.42976816 -0.066882876 -0.055234188 0.25983359 -0.047549773
#>  [38,] 0.46956946 0.43309371 -0.068159094 -0.058986207 0.26254426 -0.050727489
#>  [39,] 0.47143481 0.43614064 -0.069319191 -0.062414393 0.26502033 -0.053627989
#>  [40,] 0.47313790 0.43893051 -0.070373896 -0.065545593 0.26728128 -0.056274775
#>  [41,] 0.47469248 0.44148347 -0.071332904 -0.068404626 0.26934521 -0.058689479
#>  [42,] 0.47611121 0.44381843 -0.072205010 -0.071014444 0.27122882 -0.060892027
#>  [43,] 0.47740573 0.44595302 -0.072998186 -0.073396208 0.27294747 -0.062900724
#>  [44,] 0.47858673 0.44790368 -0.073719664 -0.075569407 0.27451535 -0.064732379
#>  [45,] 0.47966402 0.44968564 -0.074376002 -0.077551960 0.27594544 -0.066402402
#>  [46,] 0.48064658 0.45131301 -0.074973148 -0.079360320 0.27724968 -0.067924906
#>  [47,] 0.48154262 0.45279916 -0.075516542 -0.081009729 0.27843897 -0.069312912
#>  [48,] 0.48235974 0.45415542 -0.076010996 -0.082513719 0.27952340 -0.070578026
#>  [49,] 0.48310479 0.45539317 -0.076460995 -0.083885097 0.28051209 -0.071731146
#>  [50,] 0.48378409 0.45652256 -0.076870574 -0.085135451 0.28141344 -0.072782134
#>  [51,] 0.48440340 0.45755294 -0.077243394 -0.086275377 0.28223510 -0.073739994
#>  [52,] 0.48496798 0.45849284 -0.077582782 -0.087314560 0.28298408 -0.074612949
#>  [53,] 0.48548264 0.45935012 -0.077891759 -0.088261850 0.28366678 -0.075408500
#>  [54,] 0.48595177 0.46013195 -0.078173068 -0.089125330 0.28428903 -0.076133492
#>  [55,] 0.48637939 0.46084491 -0.078429203 -0.089912382 0.28485616 -0.076794169
#>  [56,] 0.48676915 0.46149501 -0.078662432 -0.090629741 0.28537305 -0.077396224
#>  [57,] 0.48712439 0.46208775 -0.078874814 -0.091283559 0.28584412 -0.077944852
#>  [58,] 0.48744817 0.46262815 -0.079068221 -0.091879444 0.28627343 -0.078444786
#>  [59,] 0.48774325 0.46312081 -0.079244359 -0.092422517 0.28666468 -0.078900342
#>  [60,] 0.48801218 0.46356992 -0.079404775 -0.092917445 0.28702122 -0.079315457
#>  [61,] 0.48825726 0.46397930 -0.079550878 -0.093368488 0.28734614 -0.079693717
#>  [62,] 0.48848062 0.46435247 -0.079683950 -0.093779530 0.28764223 -0.080038391
#>  [63,] 0.48868417 0.46469261 -0.079805158 -0.094154112 0.28791205 -0.080352460
#>  [64,] 0.48886966 0.46500263 -0.079915563 -0.094495463 0.28815793 -0.080638639
#>  [65,] 0.48903870 0.46528519 -0.080016129 -0.094806526 0.28838198 -0.080899404
#>  [66,] 0.48919274 0.46554272 -0.080107738 -0.095089987 0.28858615 -0.081137012
#>  [67,] 0.48933311 0.46577742 -0.080191187 -0.095348291 0.28877219 -0.081353516
#>  [68,] 0.48946102 0.46599132 -0.080267206 -0.095583669 0.28894172 -0.081550792
#>  [69,] 0.48957758 0.46618626 -0.080336457 -0.095798154 0.28909619 -0.081730547
#>  [70,] 0.48968380 0.46636391 -0.080399545 -0.095993599 0.28923695 -0.081894336
#>  [71,] 0.48978058 0.46652580 -0.080457018 -0.096171693 0.28936522 -0.082043577
#>  [72,] 0.48986878 0.46667334 -0.080509377 -0.096333975 0.28948209 -0.082179561
#>  [73,] 0.48994914 0.46680778 -0.080557078 -0.096481848 0.28958858 -0.082303467
#>  [74,] 0.49002237 0.46693030 -0.080600536 -0.096616592 0.28968562 -0.082416367
#>  [75,] 0.49008910 0.46704195 -0.080640128 -0.096739371 0.28977404 -0.082519239
#>  [76,] 0.49014990 0.46714368 -0.080676199 -0.096851247 0.28985461 -0.082612972
#>  [77,] 0.49020531 0.46723639 -0.080709063 -0.096953188 0.28992802 -0.082698380
#>  [78,] 0.49025579 0.46732087 -0.080739004 -0.097046076 0.28999492 -0.082776200
#>  [79,] 0.49030179 0.46739785 -0.080766283 -0.097130714 0.29005587 -0.082847108
#>  [80,] 0.49034370 0.46746800 -0.080791137 -0.097207836 0.29011140 -0.082911717
#>  [81,] 0.49038190 0.46753191 -0.080813781 -0.097278108 0.29016201 -0.082970587
#>  [82,] 0.49041670 0.46759016 -0.080834413 -0.097342139 0.29020812 -0.083024227
#>  [83,] 0.49044841 0.46764323 -0.080853210 -0.097400483 0.29025013 -0.083073102
#>  [84,] 0.49047730 0.46769159 -0.080870337 -0.097453644 0.29028842 -0.083117635
#>  [85,] 0.49050363 0.46773565 -0.080885942 -0.097502084 0.29032330 -0.083158212
#>  [86,] 0.49052762 0.46777580 -0.080900159 -0.097546221 0.29035508 -0.083195185
#>  [87,] 0.49054947 0.46781239 -0.080913113 -0.097586438 0.29038404 -0.083228873
#>  [88,] 0.49056939 0.46784573 -0.080924916 -0.097623082 0.29041043 -0.083259569
#>  [89,] 0.49058754 0.46787610 -0.080935670 -0.097656471 0.29043447 -0.083287537
#>  [90,] 0.49060407 0.46790378 -0.080945469 -0.097686895 0.29045638 -0.083313022
#>  [91,] 0.49061914 0.46792900 -0.080954396 -0.097714616 0.29047634 -0.083336242
#>  [92,] 0.49063287 0.46795198 -0.080962531 -0.097739875 0.29049453 -0.083357399
#>  [93,] 0.49064538 0.46797292 -0.080969942 -0.097762890 0.29051111 -0.083376677
#>  [94,] 0.49065678 0.46799200 -0.080976695 -0.097783860 0.29052621 -0.083394242
#>  [95,] 0.49066716 0.46800939 -0.080982848 -0.097802968 0.29053997 -0.083410247
#>  [96,] 0.49067662 0.46802523 -0.080988454 -0.097820378 0.29055250 -0.083424830
#>  [97,] 0.49068525 0.46803966 -0.080993562 -0.097836242 0.29056393 -0.083438118
#>  [98,] 0.49069310 0.46805281 -0.080998217 -0.097850696 0.29057434 -0.083450225
#>  [99,] 0.49070026 0.46806480 -0.081002457 -0.097863867 0.29058382 -0.083461257
#> [100,] 0.49070678 0.46807571 -0.081006321 -0.097875867 0.29059246 -0.083471308
#>               [,7]        [,8]         [,9]       [,10]
#>   [1,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [2,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [3,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [4,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [5,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [6,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [7,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [8,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>   [9,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [10,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [11,]  0.00000000 0.000000000 0.000000e+00 0.000000000
#>  [12,] -0.01129958 0.000000000 0.000000e+00 0.000000000
#>  [13,] -0.02388400 0.000000000 0.000000e+00 0.000000000
#>  [14,] -0.03676660 0.000000000 0.000000e+00 0.009760838
#>  [15,] -0.04977202 0.000000000 0.000000e+00 0.027188130
#>  [16,] -0.06170267 0.000000000 0.000000e+00 0.042706143
#>  [17,] -0.07261935 0.000000000 0.000000e+00 0.056535454
#>  [18,] -0.08258812 0.000000000 0.000000e+00 0.068872262
#>  [19,] -0.09134692 0.000000000 0.000000e+00 0.080328188
#>  [20,] -0.09903498 0.000000000 0.000000e+00 0.090927140
#>  [21,] -0.10601057 0.000000000 7.598962e-05 0.100413154
#>  [22,] -0.11220030 0.000000000 6.101133e-03 0.108903620
#>  [23,] -0.11781897 0.000000000 1.162766e-02 0.116523145
#>  [24,] -0.12291655 0.000000000 1.669034e-02 0.123368858
#>  [25,] -0.12860037 0.005384795 2.134493e-02 0.130745485
#>  [26,] -0.13427290 0.012695535 2.562584e-02 0.137925991
#>  [27,] -0.13945760 0.019334335 2.955224e-02 0.144389360
#>  [28,] -0.14412553 0.025279136 3.357610e-02 0.149913215
#>  [29,] -0.14828624 0.030423153 3.899612e-02 0.153529295
#>  [30,] -0.15213579 0.035123348 4.400877e-02 0.156741715
#>  [31,] -0.15568947 0.039415733 4.863620e-02 0.159600434
#>  [32,] -0.15896346 0.043333740 5.290127e-02 0.162148372
#>  [33,] -0.16197475 0.046908577 5.682686e-02 0.164422858
#>  [34,] -0.16474043 0.050169162 6.043564e-02 0.166456191
#>  [35,] -0.16727766 0.053142396 6.374961e-02 0.168276698
#>  [36,] -0.16960281 0.055852902 6.678995e-02 0.169908900
#>  [37,] -0.17173182 0.058323461 6.957711e-02 0.171373987
#>  [38,] -0.17367967 0.060574894 7.213037e-02 0.172690542
#>  [39,] -0.17546087 0.062626517 7.446786e-02 0.173875142
#>  [40,] -0.17708877 0.064495849 7.660667e-02 0.174942050
#>  [41,] -0.17857591 0.066198969 7.856278e-02 0.175903883
#>  [42,] -0.17993395 0.067750572 8.035104e-02 0.176771755
#>  [43,] -0.18117371 0.069164085 8.198525e-02 0.177555491
#>  [44,] -0.18230518 0.070451765 8.347822e-02 0.178263785
#>  [45,] -0.18333760 0.071624793 8.484176e-02 0.178904344
#>  [46,] -0.18427945 0.072693368 8.608679e-02 0.179484020
#>  [47,] -0.18513843 0.073666725 8.722339e-02 0.180008776
#>  [48,] -0.18592190 0.074553459 8.826073e-02 0.180484305
#>  [49,] -0.18663635 0.075361233 8.920735e-02 0.180915334
#>  [50,] -0.18728777 0.076097082 9.007105e-02 0.181306205
#>  [51,] -0.18788168 0.076767415 9.085898e-02 0.181660806
#>  [52,] -0.18842312 0.077378073 9.157771e-02 0.181982626
#>  [53,] -0.18891668 0.077934372 9.223323e-02 0.182274795
#>  [54,] -0.18936658 0.078441157 9.283106e-02 0.182540130
#>  [55,] -0.18977666 0.078902840 9.337622e-02 0.182781166
#>  [56,] -0.19015043 0.079323440 9.387331e-02 0.183000186
#>  [57,] -0.19049109 0.079706616 9.432654e-02 0.183199248
#>  [58,] -0.19080156 0.080055702 9.473975e-02 0.183380212
#>  [59,] -0.19108451 0.080373734 9.511646e-02 0.183544755
#>  [60,] -0.19134237 0.080663477 9.545987e-02 0.183694396
#>  [61,] -0.19157737 0.080927450 9.577291e-02 0.183830507
#>  [62,] -0.19179152 0.081167947 9.605825e-02 0.183954331
#>  [63,] -0.19198668 0.081387057 9.631834e-02 0.184066992
#>  [64,] -0.19216452 0.081586684 9.655541e-02 0.184169509
#>  [65,] -0.19232657 0.081768562 9.677148e-02 0.184262807
#>  [66,] -0.19247425 0.081934270 9.696840e-02 0.184347724
#>  [67,] -0.19260882 0.082085246 9.714788e-02 0.184425021
#>  [68,] -0.19273144 0.082222801 9.731145e-02 0.184495386
#>  [69,] -0.19284318 0.082348129 9.746052e-02 0.184559448
#>  [70,] -0.19294500 0.082462317 9.759637e-02 0.184617775
#>  [71,] -0.19303778 0.082566355 9.772017e-02 0.184670883
#>  [72,] -0.19312232 0.082661147 9.783299e-02 0.184719243
#>  [73,] -0.19319936 0.082747514 9.793581e-02 0.184763282
#>  [74,] -0.19326955 0.082826205 9.802950e-02 0.184803388
#>  [75,] -0.19333351 0.082897903 9.811488e-02 0.184839914
#>  [76,] -0.19339179 0.082963230 9.819268e-02 0.184873180
#>  [77,] -0.19344489 0.083022751 9.826358e-02 0.184903479
#>  [78,] -0.19349328 0.083076984 9.832818e-02 0.184931077
#>  [79,] -0.19353737 0.083126397 9.838705e-02 0.184956214
#>  [80,] -0.19357755 0.083171419 9.844069e-02 0.184979112
#>  [81,] -0.19361415 0.083212441 9.848958e-02 0.184999970
#>  [82,] -0.19364751 0.083249818 9.853412e-02 0.185018970
#>  [83,] -0.19367790 0.083283874 9.857470e-02 0.185036278
#>  [84,] -0.19370560 0.083314904 9.861169e-02 0.185052046
#>  [85,] -0.19373083 0.083343177 9.864539e-02 0.185066410
#>  [86,] -0.19375382 0.083368938 9.867609e-02 0.185079496
#>  [87,] -0.19377477 0.083392411 9.870407e-02 0.185091417
#>  [88,] -0.19379386 0.083413797 9.872957e-02 0.185102278
#>  [89,] -0.19381125 0.083433284 9.875280e-02 0.185112173
#>  [90,] -0.19382710 0.083451039 9.877397e-02 0.185121187
#>  [91,] -0.19384154 0.083467217 9.879325e-02 0.185129400
#>  [92,] -0.19385470 0.083481958 9.881083e-02 0.185136883
#>  [93,] -0.19386669 0.083495389 9.882684e-02 0.185143700
#>  [94,] -0.19387761 0.083507626 9.884143e-02 0.185149911
#>  [95,] -0.19388756 0.083518777 9.885473e-02 0.185155570
#>  [96,] -0.19389663 0.083528937 9.886684e-02 0.185160726
#>  [97,] -0.19390490 0.083538194 9.887788e-02 0.185165423
#>  [98,] -0.19391242 0.083546629 9.888794e-02 0.185169703
#>  [99,] -0.19391929 0.083554314 9.889710e-02 0.185173603
#> [100,] -0.19392554 0.083561317 9.890545e-02 0.185177156
#> 
#> $cvm
#>   [1] -4.157059 -4.126366 -4.045741 -3.955655 -3.877995 -3.751915 -3.633486
#>   [8] -3.535310 -3.457671 -3.393615 -3.331382 -3.275356 -3.228379 -3.187720
#>  [15] -3.148372 -3.112677 -3.081873 -3.053910 -3.030313 -3.017084 -3.018981
#>  [22] -3.037236 -3.060001 -3.085858 -3.114539 -3.143089 -3.170262 -3.197163
#>  [29] -3.220944 -3.242997 -3.263910 -3.283244 -3.300412 -3.315790 -3.330290
#>  [36] -3.343971 -3.356826 -3.368859 -3.378459 -3.387352 -3.395616 -3.401606
#>  [43] -3.406622 -3.410935 -3.414670 -3.418136 -3.421346 -3.424314 -3.427055
#>  [50] -3.429583 -3.431911 -3.434053 -3.436021 -3.437829 -3.439489 -3.441010
#>  [57] -3.442405 -3.443683 -3.444853 -3.445923 -3.446903 -3.447798 -3.448617
#>  [64] -3.449365 -3.450049 -3.450673 -3.451243 -3.451764 -3.452239 -3.452673
#>  [71] -3.453069 -3.453430 -3.453760 -3.454060 -3.454334 -3.454584 -3.454812
#>  [78] -3.455020 -3.455210 -3.455383 -3.455541 -3.455684 -3.455815 -3.455935
#>  [85] -3.456043 -3.456143 -3.456233 -3.456316 -3.456391 -3.456459 -3.456521
#>  [92] -3.456578 -3.456630 -3.456677 -3.456720 -3.456760 -3.456795 -3.456828
#>  [99] -3.456858 -3.456885
#> 
#> $cvsd
#>   [1] 3.272380 3.230694 3.185130 3.147138 3.123246 3.024478 2.950203 2.898017
#>   [9] 2.861392 2.835061 2.800300 2.750202 2.708828 2.675879 2.638738 2.617834
#>  [17] 2.608775 2.603662 2.608396 2.616135 2.623703 2.644889 2.666473 2.687242
#>  [25] 2.710399 2.733688 2.755736 2.776434 2.797015 2.816329 2.834468 2.850824
#>  [33] 2.863272 2.874293 2.884527 2.894049 2.902883 2.911061 2.921082 2.930391
#>  [41] 2.938960 2.948336 2.957082 2.964987 2.972141 2.978693 2.984691 2.990179
#>  [49] 2.995198 2.999786 3.003979 3.007810 3.011308 3.014503 3.017419 3.020081
#>  [57] 3.022510 3.024727 3.026749 3.028594 3.030276 3.031810 3.033210 3.034486
#>  [65] 3.035649 3.036710 3.037677 3.038559 3.039362 3.040095 3.040763 3.041371
#>  [73] 3.041926 3.042432 3.042892 3.043312 3.043695 3.044044 3.044362 3.044652
#>  [81] 3.044916 3.045156 3.045375 3.045575 3.045757 3.045923 3.046074 3.046212
#>  [89] 3.046338 3.046452 3.046556 3.046651 3.046738 3.046817 3.046888 3.046954
#>  [97] 3.047014 3.047068 3.047117 3.047163
#> 
#> attr(,"class")
#> [1] "cv.survtran"