Simulate survival data for a multi-source Cox model

Simulate survival data for a multi-source Cox model

Usage

simsurv_tl(beta, eta, lambda, gamma, dist, maxt, n_samples, seed = 0)

Arguments

beta

A vector of length p representing the common coefficients, where p is the number of features.

eta

A matrix of size p x K representing the group-specific coefficients, where K is the number of groups.

lambda

A vector of length K representing the baseline hazard's scale parameters.

gamma

A vector of length K representing the baseline hazard's shape parameters.

dist

A string specifying the distribution of the baseline hazard, either "exponential", "weibull", or "gompertz".

maxt

A positive number specifying the maximum time to simulate.

n_samples

A vector of length K specifying the number of samples per group, or a single number specifying the total number of samples.

seed

An integer specifying the random seed, with a default value of 0.

Value

A data frame with columns "id", "group", "X1", "X2", ..., "Xp", "time", and "status".

Examples

beta <- c(1, 1)
eta <- matrix(c(0, 0, 1, 1), nrow = 2, ncol = 2)
lambda <- c(1, 2)
gamma <- c(2, 1)
dist <- c("gompertz", "weibull")
maxt <- 3
n_samples <- 100
df <- simsurv_tl(beta, eta, lambda, gamma, dist, maxt, n_samples)
head(df)
#>   id group         X1         X2        time status
#> 1  1     1 -0.7818592  1.2629543 0.538752664      1
#> 2  2     2 -1.3143791  2.0573531 0.368023087      1
#> 3  3     1 -1.6782972  2.4046534 0.159082214      1
#> 4  4     2  0.7567252  1.0222779 0.008187544      1
#> 5  5     1  0.4908328  0.9965439 0.262357935      1
#> 6  6     2 -0.5176900 -0.5946549 0.525280653      1